Compute matrix of conditional likelihoods.

computes matrix of condition likelihoods for each of J rows of Bhat for each of P prior covariances.

This is an internal (non-exported) function. This help page provides additional documentation mainly intended for developers and expert users.

calc_lik_matrix(
  data,
  Ulist,
  log = FALSE,
  mc.cores = 1,
  algorithm.version = c("Rcpp", "R")
)

Arguments

data
A mash data object; e.g., created by mash_set_data.

Ulist
List containing the prior covariance matrices.

log
If TRUE, the return value is a matrix of log- likelihoods.

mc.cores
The argument supplied to openmp specifying the number of cores to use. Note that this is only has an effect for the Rcpp version.

algorithm.version
Indicate R or Rcpp version

Value

J x P matrix of multivariate normal likelihoods, p(bhat | Ulist[p], V).