Perform PCA on data and return list of candidate covariance matrices

Perform PCA on data and return list of candidate covariance matrices

cov_pca(data, npc, subset = NULL)

Arguments

data
a mash data object

npc
the number of PCs to use

subset
indices of the subset of data to use (set to NULL for all data)

Value

Returns a list of covariance matrices: the npc rank-one covariance matrices based on the first npc PCs, and the rank npc covariance matrix