Compute the proportion of (significant) signals shared by magnitude in each pair of conditions

Compute the proportion of (significant) signals shared by magnitude in each pair of conditions

get_pairwise_sharing_from_samples(
  m,
  factor = 0.5,
  lfsr_thresh = 0.05,
  FUN = identity
)

Arguments

m
the mash fit with samples from posteriors

factor
a number in [0,1] the factor within which effects are considered to be shared

lfsr_thresh
the lfsr threshold for including an effect in the assessment

FUN
a function to be applied to the estimated effect sizes before assessing sharing. The most obvious choice beside the default 'FUN=identity' would be 'FUN=abs' if you want to ignore the sign of the effects when assesing sharing.

Details

For each pair of conditions, compute the fraction of effects that are within a factor `factor` of one another. The results are returned as an R by R matrix.

Examples

simdata = simple_sims(50,5,1)
data = mash_set_data(simdata$Bhat, simdata$Shat)
m = mash(data, cov_canonical(data), posterior_samples=5, algorithm='R')

#>  - Computing 200 x 151 likelihood matrix.
#>  - Likelihood calculations took 0.04 seconds.
#>  - Fitting model with 151 mixture components.
#>  - Model fitting took 0.18 seconds.
#>  - Computing posterior matrices.
#>  - Computation allocated took 0.12 seconds.

get_pairwise_sharing_from_samples(m) # sharing by magnitude (same sign)

#>             condition_1 condition_2 condition_3 condition_4 condition_5
#> condition_1   1.0000000   0.6444444   0.4727273   0.5200000   0.6222222
#> condition_2   0.6444444   1.0000000   0.6000000   0.6400000   0.6000000
#> condition_3   0.4727273   0.6000000   1.0000000   0.6444444   0.6000000
#> condition_4   0.5200000   0.6400000   0.6444444   1.0000000   0.6666667
#> condition_5   0.6222222   0.6000000   0.6000000   0.6666667   1.0000000

get_pairwise_sharing_from_samples(m, factor=0) # sharing by sign

#>             condition_1 condition_2 condition_3 condition_4 condition_5
#> condition_1   1.0000000   0.7777778   0.7272727   0.7200000   0.8888889
#> condition_2   0.7777778   1.0000000   0.8545455   0.8600000   0.8444444
#> condition_3   0.7272727   0.8545455   1.0000000   0.8666667   0.7777778
#> condition_4   0.7200000   0.8600000   0.8666667   1.0000000   0.8666667
#> condition_5   0.8888889   0.8444444   0.7777778   0.8666667   1.0000000

get_pairwise_sharing_from_samples(m, FUN=abs) # sharing by magnitude when sign is ignored

#>             condition_1 condition_2 condition_3 condition_4 condition_5
#> condition_1   1.0000000   0.7333333   0.5090909   0.5800000   0.6222222
#> condition_2   0.7333333   1.0000000   0.6727273   0.7000000   0.6444444
#> condition_3   0.5090909   0.6727273   1.0000000   0.6888889   0.7333333
#> condition_4   0.5800000   0.7000000   0.6888889   1.0000000   0.7333333
#> condition_5   0.6222222   0.6444444   0.7333333   0.7333333   1.0000000