Perform condition-by-condition analyses

Performs simple "condition-by-condition" analysis by running ash from package ashr on data from each condition, one at a time. May be a useful first step to identify top hits in each condition before a mash analysis.

mash_1by1(data, alpha = 0, ...)

Arguments

data

A list with the following two elements: Bhat an n by R matrix of observations (n units in R conditions); and Shat, an n by R matrix of standard errors (n units in R conditions),

alpha

Numeric value of alpha parameter in the model. alpha = 0 for Exchangeable Effects (EE), alpha = 1 for Exchangeable Z-scores (EZ).

...

optionally, other parameters to be passed to ash

Value

A list similar to the output of mash, particularly including posterior matrices.

Examples

simdata = simple_sims(50,5,1)
mash_1by1(simdata)
#> $result
#> $result$PosteriorMean
#>             condition_1   condition_2   condition_3   condition_4   condition_5
#> effect_1   -0.018250436 -0.0504477040 -0.3249518798 -0.1156385567  2.011049e-02
#> effect_2    0.481924866 -0.0317211818 -0.0832436992  0.0106861128 -6.971801e-02
#> effect_3   -0.375495751 -0.0107944301  0.4299473767 -0.1153484998 -1.738044e-01
#> effect_4   -0.050205107  0.0100401003  0.2358438794 -0.0074717319  1.652739e-02
#> effect_5   -0.294814811 -0.0598464364 -0.0453165386 -0.0976037799  1.763458e-02
#> effect_6    0.295243292 -0.0774858263  0.0069810608  0.0192120025 -7.897321e-02
#> effect_7    0.140096423 -0.0048268033 -0.2033467641 -0.3191984079  7.192998e-02
#> effect_8    0.721958662 -0.3365947354  0.3498046211  0.0412758827 -6.359684e-02
#> effect_9   -0.613852256 -0.0125200008  0.0303316697  0.1133927714  3.435620e-02
#> effect_10  -0.104472938 -0.0179722646 -0.0556289821 -0.0083264243 -1.811672e-02
#> effect_11   0.135975990 -0.0445724266 -0.0329792289  0.0742991516 -5.753340e-02
#> effect_12   0.373067787  0.4155320229 -0.0827415631 -0.0504606652  4.836721e-02
#> effect_13  -0.029741503 -0.2927599283  0.0482844060 -0.0685395591 -2.739608e-02
#> effect_14  -0.271447562  0.0464613462  0.1579729131 -0.1370962754 -4.392470e-02
#> effect_15  -0.603061708  0.1804918062  0.1186979833  0.0488409531 -2.911106e-02
#> effect_16   0.175902479 -0.0652413974  0.0587531708  0.0885687068 -5.411655e-02
#> effect_17   1.356208109 -0.0882438047 -0.0011197767  0.0780089821  7.509434e-03
#> effect_18   0.232799874  0.0462888240 -0.0689814317  0.3250679416  2.043724e-02
#> effect_19   0.155723834  0.0061540927 -0.2448958909  0.0119495934  2.258648e-02
#> effect_20   0.182011440 -0.0346648578 -0.1282098665  0.0183486749  4.831392e-01
#> effect_21  -0.031366776 -0.0892985865  0.0274820626  0.0145854953 -5.434030e-02
#> effect_22  -0.038332450 -0.0360834524 -0.0936970571 -0.0333583663 -1.495784e+00
#> effect_23  -0.022041201  0.0079440027 -0.0271705502 -0.0274554990 -3.177983e-02
#> effect_24  -0.154947221 -0.2293217750  0.2904689705  0.0316549932  9.133586e-02
#> effect_25  -0.104070005 -0.0992878779  0.1589779674 -0.0036809144 -1.396983e-01
#> effect_26   0.067224561 -0.1093272673 -0.0106811247 -0.1662997091  2.942194e-03
#> effect_27  -0.270473604 -0.1153465530 -0.0042829274 -0.0125360240  1.346927e-02
#> effect_28  -0.062452673  0.0348547120 -0.0145758785 -0.0470567513  4.014248e-03
#> effect_29   0.075682488  0.0012499092  0.0458513714 -0.0549800937 -1.513451e+00
#> effect_30  -0.272941426  0.0378978688 -0.1630443683  0.0090605091 -5.485755e-02
#> effect_31   0.016918626  0.1663259929  0.1642822832  0.0138455987 -2.790162e-02
#> effect_32   0.190306724 -0.0805527398  0.9367956811  0.0026397377 -3.519879e-02
#> effect_33  -0.171297678  0.0431987880 -0.2680516857  0.0287514532 -7.926614e-03
#> effect_34   0.940158479 -0.0233787283 -0.3648962142  0.0189297487 -3.101338e-02
#> effect_35  -0.021662452 -0.0063606015 -0.0694329545  0.0313104116 -1.472215e-01
#> effect_36   0.142363012  0.0425184538  0.0062247941  0.0494823944  9.411676e-02
#> effect_37  -0.007647168  0.8520206466  0.0071224115  0.0360242135  3.897161e-01
#> effect_38   0.108217960 -0.0022673849 -0.1297601652 -0.0330508707  8.203312e-03
#> effect_39  -1.207430059 -0.2266271272 -0.0192422763  0.1630595847 -5.074887e-02
#> effect_40   0.016264682  0.0157698006 -0.0155745950  0.0175494034 -5.100648e-01
#> effect_41   0.070439289  0.0012909853 -0.8202951081  0.0374721873 -5.882512e-01
#> effect_42  -0.027853245  0.0040446241  0.0050270583 -0.0204676458  8.173598e-02
#> effect_43   0.018185139 -0.0594485687  0.0098602256 -0.0032948562  6.458203e-02
#> effect_44   0.038607100 -1.5869238073 -0.0127099853  0.0084026363  3.128433e-02
#> effect_45   0.426596981  0.0453382139 -0.0294262845  0.0173016351 -1.525175e-01
#> effect_46  -0.108642451  0.0291485614 -0.0041851129  0.0177537405 -8.790888e-02
#> effect_47   0.025410366 -0.0289936806  0.0685068865  0.0034411893 -5.299121e-01
#> effect_48  -0.026168402  0.0009545635 -0.0442278729 -0.0850720336  6.488606e-02
#> effect_49  -0.014538628  0.0075831332 -0.0182923642 -0.0142041366 -1.511770e-02
#> effect_50  -0.066440133  0.0003405478  0.0414787000  0.0229946409  3.083817e-02
#> effect_51  -0.624166576  0.1207036531  0.0390430093 -0.1216180205  8.422981e-03
#> effect_52   0.422145092  0.0081126142 -0.1522928442  0.0242964102 -2.540617e-02
#> effect_53  -0.321583220 -0.0744069478 -0.0754536348  0.0969724410 -3.573087e-02
#> effect_54  -0.137957013 -0.2097118624  0.0434431902 -0.0398612436 -8.612164e-03
#> effect_55  -0.199389709  0.0179258197  0.1052571661 -0.0455384410 -1.196323e-01
#> effect_56   0.195565329  0.1095482577  0.0099136812 -0.0087550153 -6.910769e-01
#> effect_57  -0.370299855  0.0560440667  0.6757696138  0.1682263235  3.097923e-02
#> effect_58  -0.050029758  0.5650748013 -0.1110512865  0.0695344122 -1.117070e+00
#> effect_59   0.533800198 -0.0048021328  0.0291415633 -0.0193896122  1.313533e-01
#> effect_60   0.096031373  2.6477482872 -1.9790811159 -0.0477110190  1.529633e+00
#> effect_61   1.194650012  0.3994363799  0.0723432418  0.0095761114  4.344033e-01
#> effect_62   0.164164121 -0.6956223820 -1.4445340500  0.7903017438  4.961722e-02
#> effect_63  -0.002113468  0.0300279286  0.1042114211  0.0954644043 -7.611963e-03
#> effect_64   0.478328707 -0.0558163035 -0.1004129395 -0.0093246454  2.002603e-02
#> effect_65  -0.314655001  0.0516345427 -0.0543928146  0.0429861189 -1.866936e-01
#> effect_66   0.196317308 -0.0070333166 -0.0444105746 -0.0110016477 -4.327072e-02
#> effect_67   0.925955740  0.0020359088 -0.0611972505  0.3479999058 -7.905442e-01
#> effect_68   0.044811397 -0.2004444554 -1.7323485659 -0.0281584210  9.128184e-02
#> effect_69  -0.499404226  0.0041617743 -1.5603308433 -0.0205154399 -3.402488e-02
#> effect_70  -0.006704740 -0.0312519699  0.3737131409  0.4477985792  5.084578e-02
#> effect_71  -0.093368370 -0.0137086465 -0.1044674923 -0.0377297899 -6.213273e-02
#> effect_72  -0.704411567 -0.0574310027  2.8389968648 -0.0656457951  9.640963e-02
#> effect_73  -1.341534346  0.0685777373  0.0076453502  0.0145448929 -2.292437e-03
#> effect_74   0.174684860  0.0008543948  0.2909412221 -0.0214656071  7.451931e-02
#> effect_75  -0.168527999  1.0352118605 -0.1308988478  0.1075662599  6.324312e-03
#> effect_76  -0.221498920  0.0115716355  2.5073320980 -0.0216788653 -1.009098e-01
#> effect_77  -0.039216832 -0.0467603922  0.6237720777  0.0971471106 -1.712017e-01
#> effect_78  -1.099403553 -0.0143943335  0.1033689728 -1.1481103561  4.306797e-02
#> effect_79   0.041924253 -0.0165693946  0.0162981943  0.1056758226 -1.309901e-01
#> effect_80   0.078534758 -0.0249626847 -0.4451572687 -0.0951402263 -3.268417e-02
#> effect_81   0.183705583 -0.2509681754 -0.0964761986 -0.0015563300 -5.076499e-03
#> effect_82  -0.531598202  0.7993305542 -0.1863457334  0.1680852185 -2.605520e-03
#> effect_83  -0.041330157  0.1547307778 -0.0780688767  0.0073668952  3.873336e-02
#> effect_84  -0.289531380 -0.0490362430 -0.1388577393  1.4501028057  1.636611e-01
#> effect_85  -0.046562156  0.0414550205 -0.0899784988  1.9353858508  2.589211e-03
#> effect_86  -0.120302363  0.2039163129 -1.2547617702  0.0008991903  2.203608e-02
#> effect_87   0.310845132 -0.0109288972 -0.1317586095 -0.0143339858  6.840698e-04
#> effect_88   0.141942930 -0.2743225840  0.0434408119  0.1000914625 -7.516596e-02
#> effect_89   0.480773076 -0.6405935950  0.2196618382 -0.0514501962 -7.517977e-01
#> effect_90   0.100639817  0.0047130833  2.3369804656  0.0025618366  1.129856e-02
#> effect_91   0.582653419 -0.2605355574 -0.0016085373  0.0746810237  1.069589e-02
#> effect_92  -0.131166477 -0.0041900449 -0.5594092605  0.0284290993  4.422218e-02
#> effect_93   0.039869017 -0.0032335295 -0.0997871640  0.0088707736  8.605883e-03
#> effect_94   0.137419891 -0.0313814415  0.0231665117 -0.0648030734 -5.860051e-02
#> effect_95   0.350043718  0.1637079639  0.0582291155 -0.0872281290 -5.855104e-02
#> effect_96   1.187275072  0.1316150392 -1.2589486509  0.0092738342  1.838902e-01
#> effect_97   0.282766684  0.0511068865 -0.6701648788  0.0227279696 -3.214778e-02
#> effect_98  -0.201318408  0.3151442002 -0.3124815046 -0.0085870963 -1.366764e-01
#> effect_99   0.122707438 -0.1301851623 -0.3440117042  0.3264499157  1.517850e-01
#> effect_100  0.195224636  0.2185967983  0.0719369347 -0.0415603933 -9.332528e-02
#> effect_101  0.469881795  0.0169863606  0.0929814447 -0.0356388793 -1.116500e-01
#> effect_102  1.436754996 -0.0113692685 -0.0297156274 -0.0185300591  1.812433e-02
#> effect_103 -0.044618858 -0.0389277588  0.0491373592  0.0053522443 -6.659394e-02
#> effect_104  0.041384348  0.0081730133  0.0095301181 -0.0087360112 -1.306666e-01
#> effect_105 -0.244478399 -0.0007443064  0.0974266133 -0.1100282892  7.021003e-02
#> effect_106  2.155229166  0.0357078615 -0.3850194474 -0.0267825712  2.551182e-02
#> effect_107 -0.107094074  0.0542980021 -0.0478368814  0.0125836897  8.469233e-03
#> effect_108  0.323255300 -0.3085183069 -0.0460972232 -0.0133925686 -4.481384e-02
#> effect_109 -0.118989808 -0.0229875474  0.3055450192 -0.0074256123  2.257846e-01
#> effect_110  1.714359360 -0.0426346053  0.0922774629 -0.2374604272 -1.788374e-02
#> effect_111 -0.287988386  0.0368962368  0.1881970328 -0.0265928549  4.622268e-02
#> effect_112 -0.644215559 -0.3594819062  0.0120782773  0.0211192006 -1.126607e-01
#> effect_113 -1.013937969 -0.0349738612  0.0490466522 -0.0324960004  6.728514e-02
#> effect_114  0.976571139  0.0803697543 -0.1145701730  0.1313134114  6.816256e-02
#> effect_115 -2.497572819  0.0110893849  0.1220757879 -0.0515145072  2.421359e-01
#> effect_116 -0.078224472  0.0150313808  0.1067000069 -0.0295939309  2.134350e-04
#> effect_117  0.126643697 -0.0408812365 -0.0172034377  0.0428002779 -5.895358e-02
#> effect_118 -0.021547334 -0.1581041965 -0.0045516036 -0.0264753214 -2.111648e-02
#> effect_119  0.253813639  0.0964128815  0.5850086049  0.0333932815  2.849853e-03
#> effect_120  0.193103028  0.1518577394  0.0392147783 -0.1933944507 -3.744833e-01
#> effect_121 -0.310318656  0.0173815563  0.0465066593 -0.0803701050  6.469643e-03
#> effect_122 -0.342534051 -0.0659312026  0.1427671508 -0.0328541036  1.143773e-01
#> effect_123 -0.235552485 -0.0455720198  0.0202779884 -0.0006466889 -4.675273e-02
#> effect_124 -0.094511181  0.0105035343 -0.0201873179  0.1132768313  6.928801e-02
#> effect_125  0.130165179 -0.0034900329  0.1713230525  0.0526581569  3.181346e-02
#> effect_126 -1.554151795 -0.0187160878 -0.0776652571  0.0232344072 -6.395080e-02
#> effect_127 -1.017999412 -0.0026974646 -0.0202655479 -0.0433718760  7.473889e-03
#> effect_128 -0.033507390  0.0454683425 -0.0097333771 -0.0224338487  1.470489e-01
#> effect_129  2.038349345  0.0118689486  0.0293630024  0.0220582239 -2.034564e-01
#> effect_130 -0.077709183 -1.2354740011  0.0243591747  0.0428723962  8.553722e-02
#> effect_131 -0.834057904 -0.0261338163 -0.1019552746 -0.0179459951  7.418920e-03
#> effect_132 -0.009191390  0.0810844808 -0.2363438861  0.0196738219  6.450582e-01
#> effect_133 -0.047014258  0.3173959306  0.0847173007  0.0977567118 -2.699875e-02
#> effect_134  0.222629799 -0.2597246145 -0.0056697473 -0.0965918722  4.582457e-02
#> effect_135  0.016233856 -0.0457524075  0.0818929377  0.0826094202 -7.570471e-02
#> effect_136 -0.761950718  0.0045489152  0.0661506581  0.0167472203  6.129213e-02
#> effect_137  1.273157872  0.1044758322  0.0617026171  0.0509703367 -7.451165e-02
#> effect_138 -1.422345265  0.0265937514  0.0466391119  0.0233181612 -6.626584e-02
#> effect_139 -0.674583579  0.0628197061 -0.0527312151 -0.0458452427 -7.801485e-02
#> effect_140  0.643640129 -0.2217202755  1.6705126928 -0.0056854716 -2.108319e-01
#> effect_141 -0.020489913 -0.0135764249 -0.0332352912 -0.0971478437 -4.329317e-02
#> effect_142  0.184828344 -0.0845955652  0.0621452752  0.7682686130 -1.690194e-02
#> effect_143  1.188960251 -0.1011548142  0.0164296745  0.0282332199  1.001428e-01
#> effect_144 -0.149991114 -0.1875170118 -0.0367590074 -0.0219759297  2.328050e-02
#> effect_145  0.732950932 -0.0522541333  0.2117846582 -0.2174217052  8.045860e-03
#> effect_146  0.054187464 -0.0286224448  0.0673434616 -0.2371930007  5.608292e-02
#> effect_147 -0.132728416  0.0033632686  0.5655594275 -0.0524847795 -3.851527e-02
#> effect_148 -0.349197053  0.1573015796 -0.0004648375  0.0372403589  4.818358e-02
#> effect_149  0.031849122  0.0465952813 -0.3960573633  0.0713479090  7.217597e-03
#> effect_150 -0.983806496  0.0838131939 -0.0206747374  0.0197175805  5.707766e-03
#> effect_151  0.723336383 -0.0412791193 -0.0394363736 -0.0024265728 -1.557618e-01
#> effect_152  0.350641909 -0.0648533594  0.1610832585  0.1691182680  1.105816e-02
#> effect_153 -0.367920664 -0.0252663751 -0.0288556364  0.3229808057 -9.771924e-03
#> effect_154  0.044945066  0.0289963499 -0.0087031125  0.0614561887 -1.037870e-01
#> effect_155  0.180561539  0.1029887118  0.6234339603  0.1071715819  1.695710e+00
#> effect_156 -1.084025432 -0.2611023124 -0.8978927922 -0.4461126282 -6.543818e-02
#> effect_157  0.113358290 -0.0319117697  0.0319148803  0.0614667347  2.061702e-01
#> effect_158 -0.560744871 -1.2822906641 -2.8745730104 -0.4784534192 -2.256218e+00
#> effect_159 -0.056206739  0.1313105663 -0.0243864685  0.1357834373 -1.573448e-02
#> effect_160  0.328790778  0.0159703895 -0.0120587184  0.0407536374  2.157774e+00
#> effect_161 -1.547453857 -1.5466731033 -0.5101745189 -0.2155844304 -2.097752e+00
#> effect_162 -0.138665525  0.1218079394  0.0656016834  0.0071186822  4.837560e-02
#> effect_163  0.152672977 -0.0551433742 -0.3811842516  1.0792104113 -3.626713e-01
#> effect_164 -0.393071608 -0.0104288371 -0.0650077966 -0.0431296419 -2.293072e-02
#> effect_165  0.338935294  0.0209146107  0.5289845584 -0.0321299962 -3.058961e-02
#> effect_166  0.815182661  0.0144745996  0.1384745416  0.0281801146  1.440996e-01
#> effect_167  0.637785748  0.0625972797  0.0090587820  0.2533604975  1.173356e-02
#> effect_168 -0.102142663  0.2843056283 -0.0567212937  0.0043456078  7.645212e-02
#> effect_169  0.594452863  0.1572744307  0.1024055644  0.5322323638  7.540120e-01
#> effect_170 -0.044037731 -0.0133565419 -0.0537250498 -0.0126574510 -4.161895e-01
#> effect_171  0.104999587 -0.5548288543 -0.3133889485 -0.0531583710  1.164844e-01
#> effect_172  0.016628871 -0.0320670674 -0.1011630550  0.0427932895 -5.740449e-02
#> effect_173  0.054520188  2.1135773789 -0.0598852203 -0.0334014503 -3.283992e-01
#> effect_174 -0.037951129 -0.0768203894 -0.0917611801 -0.0331781153  2.308910e-02
#> effect_175  0.147583828  0.0095021323 -0.0805434684 -0.0007365210 -5.668714e-02
#> effect_176  0.089527633  0.9962426589  0.2890449897  0.1357385318  2.425212e-01
#> effect_177 -1.912022992 -2.2225689501 -0.1023497875 -0.1039545332 -3.491319e-01
#> effect_178 -0.003783425 -0.0344258869 -0.0530656041 -0.1400541128 -9.579405e-02
#> effect_179  0.260293912 -0.0095190060  0.0014643004 -0.0552652465  5.648345e-02
#> effect_180  2.295404163  0.4457647660  2.2117831364  0.2277813833  2.315984e+00
#> effect_181  0.039439854  0.0189779608 -0.0833567298 -0.0425914799 -2.081187e-02
#> effect_182 -0.431164900 -0.0087384191 -0.0778411921 -0.0214164352  5.946835e-02
#> effect_183  0.006476798  0.0142182187  0.4190887835  0.0347619258 -3.517748e-02
#> effect_184  0.260848217 -0.1280260048  0.0156397948 -0.5122048872 -5.719878e-02
#> effect_185  0.248002321  0.1416761073  0.0404813925  0.0165596748 -5.034588e-02
#> effect_186 -0.341432233  0.0074878909 -0.1214974793 -0.0024871603  1.480031e-02
#> effect_187  0.133850290  0.0103099973 -0.1803284228 -0.0078721356 -1.234826e-01
#> effect_188  1.254452104  0.0355647466 -0.3400318952  0.0383006477  6.598225e-03
#> effect_189 -0.115741262  0.1958026050  0.4550142638  0.0160703550  4.986707e-02
#> effect_190  0.173889795 -0.0249007918  0.4339799724  0.2207213408  3.739430e-02
#> effect_191  0.270082952  0.3712810424 -0.0094030057  0.0809857804  3.340764e-02
#> effect_192  0.089412803 -0.0704988393  0.0012960054 -0.0496223540 -7.914565e-03
#> effect_193  0.241968015  0.0234249308  0.1473844058  0.0095343872  3.640824e-01
#> effect_194  0.737709523  0.0541477846  0.0690741030  0.5410386294  3.585840e-01
#> effect_195  0.086403956 -0.1579959296 -0.0072422082 -0.2828977827 -3.806299e-05
#> effect_196 -0.292067932  0.7199306786  0.0992172705  0.0870692352  1.059972e-01
#> effect_197 -0.561787965  2.0167720702  0.0860261671  0.0868512994 -1.215833e-01
#> effect_198 -0.229335335  0.0541728030 -0.1652909213  0.1409549932 -1.228745e-01
#> effect_199 -0.022524730  0.0309517830  0.0152776298  0.1516828458  1.019028e-01
#> effect_200 -0.172896798  0.0830403891  0.8535328123 -0.0321379746  6.090676e-02
#> 
#> $result$PosteriorSD
#>            condition_1 condition_2 condition_3 condition_4 condition_5
#> effect_1     0.4232130   0.3065798   0.7059320   0.4136853   0.2564153
#> effect_2     0.7527942   0.2709264   0.3816538   0.2237399   0.3449595
#> effect_3     0.6827181   0.2428228   0.8056787   0.4131779   0.5258522
#> effect_4     0.4329818   0.2422438   0.6044621   0.2214306   0.2522815
#> effect_5     0.6219468   0.3255233   0.3222502   0.3812726   0.2534909
#> effect_6     0.6222864   0.3609955   0.2859206   0.2331550   0.3629203
#> effect_7     0.4929866   0.2393332   0.5626322   0.6876552   0.3492708
#> effect_8     0.8724995   0.7377295   0.7312684   0.2710396   0.3329951
#> effect_9     0.8246515   0.2442887   0.3031565   0.4097448   0.2778457
#> effect_10    0.4653780   0.2501126   0.3374425   0.2219705   0.2540371
#> effect_11    0.4896476   0.2949378   0.3061864   0.3368174   0.3211448
#> effect_12    0.6809854   0.8160937   0.3808320   0.2892102   0.3034309
#> effect_13    0.4257394   0.6888517   0.3264930   0.3254358   0.2665417
#> effect_14    0.6031768   0.2986530   0.4991333   0.4500150   0.2950443
#> effect_15    0.8193533   0.5403663   0.4390921   0.2859722   0.2692017
#> effect_16    0.5228728   0.3364352   0.3422556   0.3643705   0.3144982
#> effect_17    0.9997817   0.3821307   0.2848979   0.3440751   0.2450345
#> effect_18    0.5711798   0.2983124   0.3584558   0.6937932   0.2568228
#> effect_19    0.5058759   0.2398877   0.6156152   0.2248489   0.2596203
#> effect_20    0.5280625   0.2761653   0.4540518   0.2320122   0.8652683
#> effect_21    0.4261897   0.3841752   0.3000970   0.2275020   0.3149323
#> effect_22    0.4283705   0.2787568   0.3987614   0.2560675   1.2383860
#> effect_23    0.4239172   0.2408398   0.2997761   0.2457085   0.2735119
#> effect_24    0.5052284   0.6096977   0.6687580   0.2529900   0.3864719
#> effect_25    0.4650861   0.4032699   0.5006080   0.2197174   0.4719545
#> effect_26    0.4413636   0.4219537   0.2873125   0.4959105   0.2433599
#> effect_27    0.6023842   0.4329081   0.2852671   0.2254003   0.2492826
#> effect_28    0.4388196   0.2765098   0.2893740   0.2824178   0.2436235
#> effect_29    0.4462091   0.2385000   0.3230058   0.2982790   1.2392987
#> effect_30    0.6043909   0.2821263   0.5065422   0.2224777   0.3159364
#> effect_31    0.4229964   0.5186001   0.5083385   0.2267126   0.2673162
#> effect_32    0.5351245   0.3670705   1.1035962   0.2194475   0.2792957
#> effect_33    0.5189709   0.2922571   0.6432229   0.2479040   0.2452581
#> effect_34    0.9431103   0.2574198   0.7461004   0.2327773   0.2722541
#> effect_35    0.4238411   0.2399857   0.3591825   0.2523753   0.4842721
#> effect_36    0.4948352   0.2909368   0.2857063   0.2872534   0.3916837
#> effect_37    0.4219381   1.1032895   0.2859634   0.2610002   0.7833387
#> effect_38    0.4681175   0.2386377   0.4564660   0.2555073   0.2454128
#> effect_39    0.9898241   0.6060810   0.2926011   0.4910047   0.3079926
#> effect_40    0.4228959   0.2475539   0.2899972   0.2309886   0.8865387
#> effect_41    0.4431563   0.2385040   1.0552943   0.2637305   0.9431760
#> effect_42    0.4252447   0.2390680   0.2854165   0.2348826   0.3682334
#> effect_43    0.4232020   0.3247183   0.2869551   0.2196071   0.3349225
#> effect_44    0.4284647   1.2661443   0.2883120   0.2220213   0.2726970
#> effect_45    0.7177862   0.2964402   0.3021609   0.2306782   0.4927933
#> effect_46    0.4684310   0.2665264   0.2852493   0.2312471   0.3800036
#> effect_47    0.4246506   0.2662674   0.3576928   0.2196475   0.9016154
#> effect_48    0.4248294   0.2384749   0.3207250   0.3577125   0.3355173
#> effect_49    0.4226494   0.2406292   0.2918804   0.2270909   0.2508350
#> effect_50    0.4409357   0.2384442   0.3169558   0.2385769   0.2719688
#> effect_51    0.8296243   0.4425019   0.3137238   0.4240466   0.2455391
#> effect_52    0.7148368   0.2409413   0.4907403   0.2405834   0.2635756
#> effect_53    0.6428280   0.3548614   0.3689330   0.3801059   0.2802189
#> effect_54    0.4912493   0.5828443   0.3196369   0.2683006   0.2456505
#> effect_55    0.5428637   0.2500559   0.4175611   0.2794061   0.4378307
#> effect_56    0.5396052   0.4223592   0.2869775   0.2222618   1.0074536
#> effect_57    0.6790027   0.3178354   0.9805948   0.4988066   0.2721985
#> effect_58    0.4329059   0.9380383   0.4268939   0.3274107   1.1800967
#> effect_59    0.7829339   0.2393242   0.3018523   0.2333947   0.4579915
#> effect_60    0.4593874   1.0857310   1.2007665   0.2837195   1.2400035
#> effect_61    0.9884678   0.8010159   0.3638831   0.2228575   0.8242002
#> effect_62    0.5129499   1.0227164   1.2160868   1.0213216   0.3058204
#> effect_63    0.4216860   0.2680098   0.4158694   0.3773103   0.2450884
#> effect_64    0.7506101   0.3173756   0.4097079   0.2226698   0.2563108
#> effect_65    0.6374899   0.3089572   0.3355609   0.2743781   0.5449946
#> effect_66    0.5402459   0.2403264   0.3209797   0.2240066   0.2938264
#> effect_67    0.9394720   0.2385994   0.3460720   0.7170950   1.0601397
#> effect_68    0.4307521   0.5697060   1.2223436   0.2468931   0.3863703
#> effect_69    0.7632334   0.2391048   1.2230152   0.2349498   0.2772797
#> effect_70    0.4218750   0.2701106   0.7545808   0.8074508   0.3081790
#> effect_71    0.4575554   0.2454088   0.4162839   0.2642196   0.3301308
#> effect_72    0.8653577   0.3206377   1.0103058   0.3196729   0.3959551
#> effect_73    0.9992592   0.3431658   0.2861287   0.2274578   0.2432394
#> effect_74    0.5218401   0.2384679   0.6692839   0.2363082   0.3543042
#> effect_75    0.5166294   1.1739254   0.4582347   0.3993955   0.2444628
#> effect_76    0.5616477   0.2434590   1.0933775   0.2366186   0.4042688
#> effect_77    0.4286759   0.2992439   0.9491526   0.3804289   0.5219090
#> effect_78    0.9756961   0.2460945   0.4145051   1.1401418   0.2934499
#> effect_79    0.4296499   0.2484521   0.2904722   0.3959981   0.4573764
#> effect_80    0.4479337   0.2598009   0.8186691   0.3767077   0.2750151
#> effect_81    0.5295037   0.6379547   0.4032982   0.2192606   0.2439638
#> effect_82    0.7817056   1.0784189   0.5395650   0.4985950   0.2432937
#> effect_83    0.4294312   0.5001612   0.3731947   0.2213682   0.2855296
#> effect_84    0.6177454   0.3037617   0.4704901   1.1814707   0.5103318
#> effect_85    0.4314518   0.2888837   0.3926802   1.1609527   0.2432907
#> effect_86    0.4772529   0.5746630   1.1907007   0.2191942   0.2588851
#> effect_87    0.6345344   0.2429300   0.4595676   0.2272298   0.2430695
#> effect_88    0.4944920   0.6669557   0.3196336   0.3858487   0.3555587
#> effect_89    0.7520960   0.9891644   0.5839921   0.2911923   1.0406416
#> effect_90    0.4626247   0.2392918   1.1345565   0.2194309   0.2474796
#> effect_91    0.8090605   0.6500137   0.2849267   0.3375673   0.2470296
#> effect_92    0.4857890   0.2391139   0.9064262   0.2473532   0.2955999
#> effect_93    0.4289052   0.2388418   0.4086906   0.2223428   0.2456468
#> effect_94    0.4908143   0.2703351   0.2959040   0.3179899   0.3232268
#> effect_95    0.6642542   0.5144872   0.3414428   0.3618255   0.3231303
#> effect_96    0.9876475   0.4615968   1.1914600   0.2226325   0.5408847
#> effect_97    0.6123292   0.3078994   0.9773303   0.2381742   0.2741210
#> effect_98    0.5445065   0.7143434   0.6927695   0.2221461   0.4669351
#> effect_99    0.4791188   0.4591280   0.7254661   0.6952279   0.4916219
#> effect_100   0.5393149   0.5951662   0.3632252   0.2715930   0.3902038
#> effect_101   0.7454305   0.2489346   0.3975919   0.2602793   0.4237263
#> effect_102   1.0009653   0.2432896   0.3024767   0.2322492   0.2540458
#> effect_103   0.4306766   0.2840635   0.3277337   0.2203332   0.3388574
#> effect_104   0.4294510   0.2409782   0.2868191   0.2222486   0.4568281
#> effect_105   0.5809572   0.2384611   0.4048476   0.4037910   0.3459193
#> effect_106   0.9193275   0.2780668   0.7652623   0.2445887   0.2637297
#> effect_107   0.4672904   0.3143130   0.3258457   0.2254461   0.2455662
#> effect_108   0.6441092   0.7069173   0.3233545   0.2262462   0.2967077
#> effect_109   0.4762408   0.2568472   0.6853124   0.2214031   0.5994396
#> effect_110   0.9855991   0.2911618   0.3964410   0.5937427   0.2537717
#> effect_111   0.6165136   0.2802589   0.5421181   0.2442756   0.2993597
#> effect_112   0.8390376   0.7616390   0.2879832   0.2358082   0.4255291
#> effect_113   0.9600376   0.2767263   0.3276013   0.2545016   0.3402085
#> effect_114   0.9518723   0.3667091   0.4325251   0.4404537   0.3419229
#> effect_115   0.8621832   0.2430596   0.4444329   0.2913213   0.6207590
#> effect_116   0.4477439   0.2467567   0.4198917   0.2493570   0.2430545
#> effect_117   0.4822028   0.2877817   0.2910935   0.2740140   0.3239162
#> effect_118   0.4238182   0.5055856   0.2853183   0.2440823   0.2576853
#> effect_119   0.5887094   0.3978249   0.9239500   0.2561312   0.2433410
#> effect_120   0.5375070   0.4955018   0.3139482   0.5352820   0.7686399
#> effect_121   0.6341249   0.2494005   0.3239371   0.3486612   0.2445278
#> effect_122   0.6586800   0.3378284   0.4764399   0.2551499   0.4285797
#> effect_123   0.5734915   0.2969000   0.2934228   0.2191781   0.3003624
#> effect_124   0.4583380   0.2425950   0.2933494   0.4095407   0.3441204
#> effect_125   0.4849913   0.2389080   0.5184652   0.2936152   0.2735674
#> effect_126   0.9978971   0.2510350   0.3725362   0.2389414   0.3336876
#> effect_127   0.9608757   0.2387198   0.2934127   0.2751348   0.2450160
#> effect_128   0.4268169   0.2966961   0.2869023   0.2377334   0.4839924
#> effect_129   0.9397092   0.2437129   0.3020921   0.2371757   0.5689769
#> effect_130   0.4474300   1.2264769   0.2970067   0.2741553   0.3754989
#> effect_131   0.9128152   0.2616224   0.4122127   0.2314923   0.2449875
#> effect_132   0.4220593   0.3681199   0.6050836   0.2337816   0.9800114
#> effect_133   0.4316363   0.7168449   0.3840661   0.3815549   0.2659387
#> effect_134   0.5626042   0.6490014   0.2855644   0.3794016   0.2986084
#> effect_135   0.4228912   0.2972550   0.3794435   0.3529858   0.3566030
#> effect_136   0.8879163   0.2392337   0.3539182   0.2299957   0.3284867
#> effect_137   0.9955212   0.4129894   0.3468660   0.2902308   0.3542894
#> effect_138   1.0009581   0.2623512   0.3241261   0.2390693   0.3382159
#> effect_139   0.8526737   0.3315394   0.3330547   0.2800135   0.3610714
#> effect_140   0.8387720   0.5994362   1.2240314   0.2204826   0.5792199
#> effect_141   0.4236134   0.2452799   0.3064885   0.3804303   0.2938682
#> effect_142   0.5304592   0.3750195   0.3475629   1.0109946   0.2526835
#> effect_143   0.9878373   0.4067833   0.2905606   0.2470201   0.4028585
#> effect_144   0.5011114   0.5508665   0.3107944   0.2370543   0.2605651
#> effect_145   0.8768550   0.3102008   0.5737719   0.5679437   0.2453242
#> effect_146   0.4347656   0.2656498   0.3558261   0.5934064   0.3183189
#> effect_147   0.4870372   0.2388747   0.9107023   0.2932673   0.2851385
#> effect_148   0.6636286   0.5042995   0.2848754   0.2632912   0.3030809
#> effect_149   0.4263276   0.2989175   0.7754865   0.3310015   0.2448850
#> effect_150   0.9535179   0.3734867   0.2937472   0.2338415   0.2442029
#> effect_151   0.8730504   0.2885454   0.3142385   0.2194032   0.4979534
#> effect_152   0.6646958   0.3356512   0.5036867   0.5001422   0.2472974
#> effect_153   0.6772922   0.2602684   0.3015445   0.6916190   0.2463839
#> effect_154   0.4308047   0.2662719   0.2864981   0.3112890   0.4095350
#> effect_155   0.5268296   0.4102173   0.9489394   0.3986878   1.2402085
#> effect_156   0.9731799   0.6507202   1.0885527   0.8060572   0.3365973
#> effect_157   0.4719513   0.2712594   0.3049477   0.3113101   0.5727671
#> effect_158   0.7976073   1.2354353   1.0020494   0.8320785   1.1581370
#> effect_159   0.4357131   0.4610719   0.2970324   0.4478588   0.2514547
#> effect_160   0.6483310   0.2477759   0.2879733   0.2700259   1.1805020
#> effect_161   0.9982160   1.2646827   0.8706161   0.5655150   1.1926871
#> effect_162   0.4918238   0.4444614   0.3530422   0.2212238   0.3034469
#> effect_163   0.5033357   0.3160175   0.7616630   1.1237288   0.7569538
#> effect_164   0.6950805   0.2425374   0.3520963   0.2746595   0.2600865
#> effect_165   0.6559884   0.2539217   0.8846361   0.2538418   0.2715654
#> effect_166   0.9066473   0.2461766   0.4699044   0.2469299   0.4791913
#> effect_167   0.8360547   0.3310894   0.2866327   0.6133634   0.2478182
#> effect_168   0.4636979   0.6789162   0.3391162   0.2199356   0.3580503
#> effect_169   0.8150552   0.5042560   0.4129433   0.8722367   1.0417902
#> effect_170   0.4304504   0.2450679   0.3345505   0.2255173   0.8079372
#> effect_171   0.4657604   0.9306111   0.6937378   0.2946193   0.4323050
#> effect_172   0.4229513   0.2715314   0.4109266   0.2740003   0.3208934
#> effect_173   0.4349198   1.2216565   0.3440181   0.2561462   0.7215307
#> effect_174   0.4282408   0.3596726   0.3955968   0.2557389   0.2603026
#> effect_175   0.4991223   0.2418544   0.3772366   0.2191832   0.3194954
#> effect_176   0.4549660   1.1608274   0.6671692   0.4477849   0.6212518
#> effect_177   0.9601197   1.1995436   0.4128528   0.3928879   0.7432364
#> effect_178   0.4217320   0.2757327   0.3335569   0.4548424   0.3948106
#> effect_179   0.5940541   0.2418663   0.2849172   0.2988521   0.3190987
#> effect_180   0.8946385   0.8432886   1.1615661   0.5814363   1.1432692
#> effect_181   0.4287540   0.2513665   0.3818388   0.2736053   0.2572960
#> effect_182   0.7207917   0.2413358   0.3728232   0.2362369   0.3249217
#> effect_183   0.4218610   0.2459157   0.7961963   0.2586483   0.2792588
#> effect_184   0.5945098   0.4553812   0.2900392   0.8577140   0.3204924
#> effect_185   0.5838906   0.4786884   0.3156196   0.2297687   0.3072181
#> effect_186   0.6578573   0.2405752   0.4435207   0.2194154   0.2505242
#> effect_187   0.4879367   0.2424465   0.5311956   0.2216767   0.4445251
#> effect_188   0.9941145   0.2778046   0.7214440   0.2653067   0.2445865
#> effect_189   0.4737556   0.5630115   0.8269110   0.2291857   0.3062993
#> effect_190   0.5211661   0.2597060   0.8091556   0.5722781   0.2831410
#> effect_191   0.6020660   0.7735692   0.2867679   0.3498527   0.2762319
#> effect_192   0.4548896   0.3470306   0.2849071   0.2875332   0.2452515
#> effect_193   0.5788625   0.2574878   0.4834066   0.2228260   0.7583635
#> effect_194   0.8787126   0.3140103   0.3586049   0.8784672   0.7528496
#> effect_195   0.4529086   0.5054123   0.2860002   0.6479956   0.2430530
#> effect_196   0.6197657   1.0366224   0.4077635   0.3615232   0.4135538
#> effect_197   0.7981623   1.2379685   0.3862093   0.3611084   0.4412316
#> effect_198   0.5682642   0.3140607   0.5097987   0.4563043   0.4434725
#> effect_199   0.4240163   0.2695916   0.2898080   0.4734206   0.4060908
#> effect_200   0.5203248   0.3719700   1.0700860   0.2538561   0.3277331
#> 
#> $result$lfsr
#>            condition_1 condition_2 condition_3 condition_4 condition_5
#> effect_1    0.85275669   0.9326932  0.76016863   0.8851794   0.9493388
#> effect_2    0.58013647   0.9444819  0.89652272   0.9554557   0.9181717
#> effect_3    0.63785119   0.9567624  0.70831788   0.8853651   0.8560052
#> effect_4    0.83492431   0.9571669  0.80715620   0.9573344   0.9514143
#> effect_5    0.68431388   0.9267428  0.92023029   0.8968706   0.9507792
#> effect_6    0.68406073   0.9156830  0.94207212   0.9501429   0.9123351
#> effect_7    0.77958843   0.9598583  0.82513930   0.7680981   0.9167720
#> effect_8    0.46306108   0.7750177  0.74758640   0.9351912   0.9220580
#> effect_9    0.51371268   0.9558240  0.92926894   0.8866193   0.9406557
#> effect_10   0.80199530   0.9527509  0.91383402   0.9568426   0.9505008
#> effect_11   0.78218975   0.9364151  0.92770412   0.9124245   0.9259227
#> effect_12   0.63921436   0.7374202  0.89683693   0.9288116   0.9317760
#> effect_13   0.84652848   0.7966885  0.91839937   0.9163452   0.9449670
#> effect_14   0.69822234   0.9352193  0.85115208   0.8716357   0.9346076
#> effect_15   0.51894844   0.8555394  0.87461143   0.9299364   0.9439146
#> effect_16   0.75701275   0.9233411  0.91188147   0.9028403   0.9281044
#> effect_17   0.22330299   0.9090410  0.94490833   0.9099146   0.9563413
#> effect_18   0.72165754   0.9353285  0.90546732   0.7650107   0.9491468
#> effect_19   0.76972217   0.9591912  0.80223458   0.9546966   0.9478734
#> effect_20   0.75317777   0.9426508  0.86884685   0.9506991   0.7008322
#> effect_21   0.84562971   0.9083945  0.93093324   0.9530789   0.9279615
#> effect_22   0.84173094   0.9417638  0.89000131   0.9406619   0.3184282
#> effect_23   0.85072728   0.9582716  0.93111383   0.9446847   0.9422632
#> effect_24   0.77021224   0.8292675  0.77798386   0.9418295   0.9046284
#> effect_25   0.80224688   0.9023143  0.85056357   0.9594428   0.8755548
#> effect_26   0.82486523   0.8962813  0.94019229   0.8537742   0.9586383
#> effect_27   0.69880641   0.8927009  0.94340001   0.9543405   0.9531370
#> effect_28   0.82771720   0.9425323  0.93814473   0.9311759   0.9581126
#> effect_29   0.81975805   0.9615899  0.91990107   0.9256775   0.3129489
#> effect_30   0.69732717   0.9406258  0.84818837   0.9564156   0.9276311
#> effect_31   0.85346357   0.8633939  0.84746717   0.9535374   0.9446575
#> effect_32   0.74798272   0.9137807  0.49274186   0.9599999   0.9401275
#> effect_33   0.75990798   0.9372840  0.78980644   0.9438076   0.9561243
#> effect_34   0.37004762   0.9495705  0.74004447   0.9503251   0.9427390
#> effect_35   0.85093119   0.9590862  0.90518383   0.9420651   0.8711797
#> effect_36   0.77815724   0.9377139  0.94244806   0.9294909   0.9029100
#> effect_37   0.85829297   0.5525608  0.94200154   0.9388265   0.7445658
#> effect_38   0.79965410   0.9611074  0.86791231   0.9408730   0.9559797
#> effect_39   0.27135657   0.8306877  0.93560691   0.8557261   0.9302554
#> effect_40   0.85380947   0.9540106  0.93760901   0.9512108   0.6886010
#> effect_41   0.82293154   0.9615705  0.53822904   0.9378263   0.6539223
#> effect_42   0.84756730   0.9602453  0.94303751   0.9493279   0.9106035
#> effect_43   0.85279142   0.9269942  0.94061503   0.9596505   0.9214313
#> effect_44   0.84157571   0.3008316  0.93913413   0.9567984   0.9425710
#> effect_45   0.60965036   0.9359305  0.92980007   0.9513688   0.8681218
#> effect_46   0.79938841   0.9460692  0.94344745   0.9510803   0.9067538
#> effect_47   0.84890247   0.9461643  0.90576525   0.9595719   0.6796828
#> effect_48   0.84848923   0.9617283  0.92089934   0.9051713   0.9212380
#> effect_49   0.85471870   0.9584588  0.93613051   0.9533156   0.9522144
#> effect_50   0.82533559   0.9620140  0.92258188   0.9476681   0.9428476
#> effect_51   0.50873938   0.8895371  0.92406315   0.8813667   0.9558646
#> effect_52   0.61207054   0.9581838  0.85448901   0.9468039   0.9461781
#> effect_53   0.66863029   0.9175994  0.90140447   0.8972852   0.9397934
#> effect_54   0.78093919   0.8396784  0.92138064   0.9361722   0.9557651
#> effect_55   0.74231296   0.9527777  0.88284465   0.9322309   0.8874121
#> effect_56   0.74469769   0.8961494  0.94058760   0.9565938   0.6100297
#> effect_57   0.64077102   0.9291476  0.59774176   0.8526169   0.9427602
#> effect_58   0.83502620   0.6702468  0.87928323   0.9156660   0.4451968
#> effect_59   0.55337746   0.9598706  0.92996666   0.9500281   0.8804520
#> effect_60   0.80725149   0.0620715  0.17087766   0.9307214   0.3079643
#> effect_61   0.27571147   0.7449514  0.90335665   0.9561132   0.7233828
#> effect_62   0.76439967   0.6150931  0.31725045   0.5487717   0.9309780
#> effect_63   0.86109620   0.9455281  0.88348933   0.8982772   0.9562881
#> effect_64   0.58202367   0.9292917  0.88583570   0.9562610   0.9493884
#> effect_65   0.67266388   0.9319409  0.91460515   0.9340040   0.8487962
#> effect_66   0.74422849   0.9587423  0.92078716   0.9552671   0.9350236
#> effect_67   0.37574932   0.9612179  0.91035107   0.7530635   0.5692255
#> effect_68   0.83804035   0.8446618  0.23340686   0.9442095   0.9046618
#> effect_69   0.57102376   0.9601877  0.28214992   0.9492968   0.9408630
#> effect_70   0.85877466   0.9447724  0.73567131   0.7029766   0.9301935
#> effect_71   0.80890260   0.9551674  0.88333141   0.9376481   0.9229901
#> effect_72   0.47106561   0.9282699  0.03106897   0.9183257   0.9014975
#> effect_73   0.22782645   0.9212451  0.94173954   0.9531041   0.9589528
#> effect_74   0.75777796   0.9617752  0.77773691   0.9486754   0.9151373
#> effect_75   0.76165105   0.4836117  0.86722679   0.8903760   0.9569514
#> effect_76   0.72860615   0.9563419  0.06953417   0.9485355   0.8987367
#> effect_77   0.84123071   0.9350298  0.62006908   0.8971705   0.8574723
#> effect_78   0.30930789   0.9547851  0.88400909   0.4094113   0.9351526
#> effect_79   0.83969227   0.9535560  0.93721828   0.8916014   0.8806662
#> effect_80   0.81802218   0.9486188  0.70107020   0.8984907   0.9417003
#> effect_81   0.75211557   0.8179722  0.88827498   0.9605690   0.9575834
#> effect_82   0.55449599   0.5732289  0.83475611   0.8527016   0.9588017
#> effect_83   0.84003075   0.8699107  0.89976412   0.9573943   0.9379011
#> effect_84   0.68744099   0.9335878  0.86245670   0.3064845   0.8617449
#> effect_85   0.83703309   0.9383855  0.89231640   0.1694890   0.9588096
#> effect_86   0.79206979   0.8427899  0.37869354   0.9609089   0.9482012
#> effect_87   0.67488961   0.9566899  0.86670970   0.9532351   0.9597176
#> effect_88   0.77842250   0.8059979  0.92138210   0.8952403   0.9147296
#> effect_89   0.58074045   0.6379878  0.81604778   0.9281247   0.5849375
#> effect_90   0.80438592   0.9599149  0.09701706   0.9600412   0.9543290
#> effect_91   0.52894386   0.8130412  0.94467876   0.9121656   0.9546551
#> effect_92   0.78522469   0.9601737  0.64844575   0.9440262   0.9344183
#> effect_93   0.84086108   0.9606418  0.88622295   0.9565263   0.9557685
#> effect_94   0.78127828   0.9446922  0.93340900   0.9189037   0.9252418
#> effect_95   0.65224824   0.8648582  0.91220932   0.9037326   0.9252733
#> effect_96   0.27824095   0.8831574  0.37728543   0.9562907   0.8503563
#> effect_97   0.69145992   0.9322753  0.60012385   0.9478444   0.9420344
#> effect_98   0.74111174   0.7855445  0.76656168   0.9566914   0.8773227
#> effect_99   0.79055611   0.8839886  0.75050081   0.7642856   0.8685437
#> effect_100  0.74491031   0.8349394  0.90361172   0.9349937   0.9033985
#> effect_101  0.58647321   0.9533177  0.89044638   0.9390923   0.8922087
#> effect_102  0.19931494   0.9564518  0.92963056   0.9505825   0.9504964
#> effect_103  0.83815087   0.9399783  0.91787147   0.9585283   0.9201529
#> effect_104  0.83999990   0.9581523  0.94078413   0.9566048   0.8808571
#> effect_105  0.71452103   0.9618265  0.88768541   0.8887849   0.9178601
#> effect_106  0.04963919   0.9419989  0.73009809   0.9451385   0.9461141
#> effect_107  0.80035723   0.9302532  0.91867606   0.9543115   0.9558403
#> effect_108  0.66765950   0.7888258  0.91974963   0.9538164   0.9340416
#> effect_109  0.79289537   0.9498043  0.77014172   0.9573608   0.8274589
#> effect_110  0.12764649   0.9376405  0.89088444   0.8124981   0.9506354
#> effect_111  0.68835620   0.9412545  0.83370158   0.9452662   0.9331439
#> effect_112  0.49915902   0.7639394  0.93946561   0.9489024   0.8915988
#> effect_113  0.34118515   0.9424579  0.91792764   0.9412536   0.9197141
#> effect_114  0.35564544   0.9138940  0.87712926   0.8752486   0.9191575
#> effect_115  0.01802668   0.9566033  0.87255833   0.9280801   0.8187468
#> effect_116  0.81821133   0.9544276  0.88195611   0.9432354   0.9599376
#> effect_117  0.78807651   0.9387475  0.93672645   0.9341330   0.9250165
#> effect_118  0.85099311   0.8680063  0.94326946   0.9453452   0.9487463
#> effect_119  0.70885012   0.9040563  0.63705744   0.9406379   0.9586832
#> effect_120  0.74623504   0.8715382  0.92395900   0.8377165   0.7519064
#> effect_121  0.67519760   0.9530910  0.91949726   0.9083238   0.9568771
#> effect_122  0.65654164   0.9229073  0.86012730   0.9410080   0.8905646
#> effect_123  0.71997134   0.9357824  0.93503216   0.9610384   0.9328060
#> effect_124  0.80819449   0.9569188  0.93508263   0.8866938   0.9184441
#> effect_125  0.78585627   0.9605170  0.84338169   0.9272867   0.9422423
#> effect_126  0.16691294   0.9523204  0.90001720   0.9475094   0.9218328
#> effect_127  0.33963276   0.9609010  0.93503909   0.9337361   0.9563598
#> effect_128  0.84443955   0.9358481  0.94068007   0.9480386   0.8712796
#> effect_129  0.06602413   0.9561801  0.92983712   0.9482861   0.8395536
#> effect_130  0.81852528   0.4130628  0.93273058   0.9340830   0.9082297
#> effect_131  0.41380575   0.9479102  0.88488210   0.9509573   0.9563882
#> effect_132  0.85749996   0.9134516  0.80688339   0.9498440   0.6294468
#> effect_133  0.83677232   0.7844327  0.89560104   0.8967702   0.9452097
#> effect_134  0.72790897   0.8134578  0.94272213   0.8975354   0.9333977
#> effect_135  0.85382576   0.9356682  0.89736814   0.9068198   0.9143902
#> effect_136  0.44512036   0.9599963  0.90724436   0.9517211   0.9235256
#> effect_137  0.24952772   0.8991871  0.91003438   0.9284578   0.9151421
#> effect_138  0.20350165   0.9476308  0.91941557   0.9474539   0.9203613
#> effect_139  0.48486217   0.9248664  0.91564013   0.9320178   0.9129370
#> effect_140  0.49943241   0.8332822  0.25045465   0.9583431   0.8355321
#> effect_141  0.85156085   0.9552408  0.92755190   0.8971700   0.9350094
#> effect_142  0.75141195   0.9112837  0.90975692   0.5580169   0.9512001
#> effect_143  0.27766190   0.9011866  0.93714705   0.9441588   0.8992061
#> effect_144  0.77334061   0.8516855  0.92544300   0.9483402   0.9474586
#> effect_145  0.45808822   0.9315482  0.82042098   0.8238389   0.9560620
#> effect_146  0.83259998   0.9463921  0.90649564   0.8126481   0.9268486
#> effect_147  0.78423928   0.9605787  0.64569712   0.9274069   0.9380389
#> effect_148  0.65273125   0.8684588  0.94521391   0.9379865   0.9318932
#> effect_149  0.84536216   0.9351344  0.72469383   0.9144299   0.9564924
#> effect_150  0.35282036   0.9117660  0.93481095   0.9498156   0.9572650
#> effect_151  0.46243606   0.9384965  0.92382457   0.9601127   0.8662574
#> effect_152  0.65190714   0.9235853  0.84933268   0.8520820   0.9544594
#> effect_153  0.64211129   0.9484354  0.93013374   0.7661071   0.9551507
#> effect_154  0.83796359   0.9461627  0.94120547   0.9212044   0.8969794
#> effect_155  0.75408730   0.9000814  0.62021597   0.8906316   0.2587292
#> effect_156  0.31492199   0.8127502  0.50769781   0.7037994   0.9208871
#> effect_157  0.79643291   0.9443638  0.92833521   0.9211972   0.8380708
#> effect_158  0.53981169   0.3972664  0.02809822   0.6881433   0.1207014
#> effect_159  0.83141415   0.8833343  0.93271500   0.8724536   0.9518656
#> effect_160  0.66445309   0.9538969  0.93947584   0.9355535   0.1415117
#> effect_161  0.16867951   0.3129388  0.67074846   0.8248889   0.1549640
#> effect_162  0.78049188   0.8888876  0.90758890   0.9575358   0.9317706
#> effect_163  0.77164760   0.9297178  0.73198427   0.4347411   0.7576430
#> effect_164  0.62804683   0.9569589  0.90796157   0.9339043   0.9476679
#> effect_165  0.65860652   0.9510342  0.66216387   0.9415043   0.9430016
#> effect_166  0.42187688   0.9547402  0.86268551   0.9441948   0.8729907
#> effect_167  0.50221910   0.9250066  0.94102467   0.8036380   0.9540923
#> effect_168  0.80344936   0.8009422  0.91315190   0.9590820   0.9139197
#> effect_169  0.52314989   0.8684741  0.88460392   0.6626738   0.5840334
#> effect_170  0.83848414   0.9553627  0.91502131   0.9542665   0.7319554
#> effect_171  0.80166638   0.6747031  0.76609461   0.9269398   0.8892981
#> effect_172  0.85361693   0.9442675  0.88537177   0.9341379   0.9260050
#> effect_173  0.83240491   0.1614012  0.91117289   0.9406323   0.7745198
#> effect_174  0.84194628   0.9160966  0.89120583   0.9407857   0.9475732
#> effect_175  0.77486053   0.9574532  0.89821324   0.9609924   0.9264629
#> effect_176  0.81127715   0.4979144  0.77872900   0.8724816   0.8185429
#> effect_177  0.08720170   0.1373473  0.88463837   0.8927200   0.7642680
#> effect_178  0.86025666   0.9428000  0.91543199   0.8697980   0.9018764
#> effect_179  0.70493156   0.9574442  0.94474665   0.9254800   0.9265929
#> effect_180  0.03390431   0.7234454  0.12045227   0.8179507   0.1088555
#> effect_181  0.84110438   0.9521682  0.89645200   0.9342780   0.9489261
#> effect_182  0.60717425   0.9578563  0.89990688   0.9487077   0.9246882
#> effect_183  0.85889090   0.9548835  0.71353010   0.9396966   0.9401409
#> effect_184  0.70459708   0.8852465  0.93757390   0.6720795   0.9261363
#> effect_185  0.71237678   0.8773484  0.92318954   0.9518399   0.9305128
#> effect_186  0.65717333   0.9585080  0.87290937   0.9600807   0.9523932
#> effect_187  0.78353136   0.9570226  0.83819634   0.9571047   0.8851150
#> effect_188  0.25564334   0.9420885  0.75250944   0.9372532   0.9568113
#> effect_189  0.79493698   0.8471739  0.69640558   0.9521490   0.9308185
#> effect_190  0.75827776   0.9486561  0.70639032   0.8219575   0.9387464
#> effect_191  0.69904078   0.7582868  0.94084910   0.9079098   0.9412489
#> effect_192  0.81134801   0.9200412  0.94482571   0.9293937   0.9561306
#> effect_193  0.71605118   0.9495429  0.85738770   0.9561378   0.7569556
#> effect_194  0.45594530   0.9303484  0.90540913   0.6585668   0.7596374
#> effect_195  0.81320195   0.8680674  0.94194164   0.7874772   0.9600192
#> effect_196  0.68593839   0.6051341  0.88657578   0.9038385   0.8956338
#> effect_197  0.53929099   0.1842337  0.89478298   0.9039837   0.8862468
#> effect_198  0.72378340   0.9303325  0.84688018   0.8692396   0.8854771
#> effect_199  0.85046659   0.9449580  0.93776875   0.8626371   0.8981295
#> effect_200  0.75890214   0.9122428  0.52503537   0.9414989   0.9237712
#> 
#> 
#> $loglik
#> [1] -1636.15
#> 
#> attr(,"class")
#> [1] "mash_1by1"