Update the data object for mash analysis.

This function can update two parts of the mash data. The first one is setting the reference group, so the mash data can be used for commonbaseline analysis. The other one is updating the null correlation matrix.

mash_update_data(mashdata, ref = NULL, V = NULL)

Arguments

mashdata

mash data object ontaining the Bhat matrix, standard errors, V; created using mash_set_data

ref

the reference group. It could be a number between 1,..., R, R is number of conditions, or the name of reference group. If there is no reference group, it can be the string 'mean'.

V

an R by R matrix / [R x R x N] array of correlation matrix of error correlations

Value

a updated mash data object

Examples

simdata = simple_sims(50,5,1)
data = mash_set_data(simdata$Bhat, simdata$Shat)
mash_update_data(data, 'mean')
#> $Bhat
#>            condition_1-mean condition_2-mean condition_3-mean condition_4-mean
#> effect_1         1.16629112     -0.502379393      -1.11867970     -0.246822147
#> effect_2         0.37890373      0.242329714       0.21855071     -0.014682579
#> effect_3        -0.32162549     -1.758069262       1.08564385     -0.495873238
#> effect_4         0.12253457     -0.312513634       1.13753921     -0.187651350
#> effect_5         0.15661092     -0.105296448      -0.45995382     -0.874304974
#> effect_6        -0.32869100      0.563302509       0.97937479      0.654231634
#> effect_7        -1.43742195      0.831700423       0.22234093     -0.663371912
#> effect_8         0.21238480     -0.309840349       0.87817178     -0.696442219
#> effect_9         0.79187957      0.329132135       0.87733519     -1.079033769
#> effect_10        0.42665355      0.012676442      -0.19813016      0.154279835
#> effect_11       -0.63816945     -0.661960674       1.29036130      0.008159877
#> effect_12        0.36886058      0.722099676      -0.07669161     -0.220808187
#> effect_13       -0.47400112      0.627029094       0.22194956     -0.094320495
#> effect_14       -1.47505789      2.364686492      -0.97110415     -0.459488695
#> effect_15        0.94838706     -0.097187232       0.42465613     -0.540771945
#> effect_16        1.09549296     -0.934828269      -1.02412610      2.890417998
#> effect_17       -0.71647250      0.857099224      -0.47345624     -0.360446532
#> effect_18        0.91177398     -1.340231387      -1.02523396      1.398284708
#> effect_19       -1.64308219      1.621683927       0.54670900     -0.011065379
#> effect_20        1.54166648     -0.013961481      -0.12409735     -0.071142112
#> effect_21        0.18681211     -0.777490070      -0.04772111      0.368961575
#> effect_22       -1.49822266      0.106856097       0.57188254     -0.142065650
#> effect_23        1.24204204      0.703433187      -0.73945892      0.055862248
#> effect_24       -0.64998962     -0.029456694       0.58111957      0.140841801
#> effect_25       -1.38964770     -0.430175897       0.14067692      1.352452194
#> effect_26        0.24296652     -0.810498399       1.31269291      0.113012425
#> effect_27       -1.23530402      0.479749638       0.81926654      0.768980154
#> effect_28        0.03563576     -0.828209010      -0.31961487      0.586306400
#> effect_29       -0.57849438      0.212199885       0.41918291      0.244078559
#> effect_30        0.91326621      0.398016727      -2.65995005      1.067522310
#> effect_31        0.07267138     -0.500858335       1.61079401      0.368088651
#> effect_32       -1.14769485     -1.244831947      -0.53237279      0.785009270
#> effect_33       -0.51470594      1.246725072       1.42273716     -1.875913044
#> effect_34       -1.73806427      0.618811501      -1.24027599      1.401602811
#> effect_35       -1.19929830      1.217839815       1.26002805     -2.060666208
#> effect_36        1.00012011     -0.795030012       0.13490456      0.748962850
#> effect_37       -0.25132360      1.074899535      -0.27868534     -0.796710884
#> effect_38        0.77428476     -0.466388234      -0.89384168      0.442585094
#> effect_39        0.88384643     -0.724221009       0.60427212     -1.088533588
#> effect_40       -0.11716359     -0.077190583       0.29455869      1.066387774
#> effect_41       -0.12827447      0.752434446       0.22051659      0.602097747
#> effect_42       -0.25795329     -0.388875391       0.78821629     -0.126641727
#> effect_43       -0.49502958      0.282183198       0.83636474     -0.656370984
#> effect_44       -0.50856778      1.106115923       1.12091247     -0.703480285
#> effect_45        0.09631091      0.136838482      -1.00411187     -0.773071518
#> effect_46        0.16484785      1.111028895      -0.40078177      0.346319462
#> effect_47        0.15521587     -1.269834645      -0.30609360     -0.148599730
#> effect_48       -0.21060018     -0.173787097       0.78298988     -0.170548554
#> effect_49       -0.74485090      2.740885834      -1.33603543     -0.457732386
#> effect_50        0.15581482     -0.303339218       0.40581243     -0.025811641
#> effect_51        1.60125259     -1.076147355       0.62523909     -0.037188178
#> effect_52        2.20111949     -0.896546770      -1.73678055     -0.437546418
#> effect_53        1.17452134     -0.027315708       0.70103363     -2.118255020
#> effect_54       -2.14026306      0.117648059       0.95378209      0.871263552
#> effect_55        1.26722865     -1.087789707      -2.31841156      1.831928719
#> effect_56       -2.15659438     -2.816898268       1.17617573      1.500285611
#> effect_57       -1.88806718      1.446246310      -0.52811551      1.211428152
#> effect_58        0.96995400      0.353974199       1.36771198     -2.585154941
#> effect_59       -0.64870890      1.097375856      -1.05303233      1.192250597
#> effect_60       -0.42882732     -0.612147409       0.64002379      0.011296828
#> effect_61       -2.32092129      1.110242434       2.14905628      0.687344139
#> effect_62       -0.80257710      0.115094906      -0.83070324      0.601858138
#> effect_63       -1.04094529      1.064256421      -0.63405261      0.405218820
#> effect_64        0.48243570      0.654707508       1.90322441     -0.854438144
#> effect_65        1.89346983     -1.248058073      -0.31613438     -2.182405628
#> effect_66       -1.68695626     -0.971812431       2.29437932      0.913138765
#> effect_67       -1.06168102     -0.008127938      -0.35780071      0.501475866
#> effect_68        0.30190069     -0.493923928       0.99604266     -0.995873582
#> effect_69        0.60597284      0.705708902      -0.94570370     -0.344653056
#> effect_70        0.13850178     -0.307606807       0.20276913     -0.107239882
#> effect_71       -1.82001659      0.985139953       0.74684306     -1.062341672
#> effect_72       -1.43127702      1.997843210      -0.53776912     -0.211221599
#> effect_73        1.11237278     -1.516223033      -0.14171665      1.137310748
#> effect_74        1.04874979      0.569161919       1.44058189     -2.001449623
#> effect_75       -0.74074388      0.814951557      -0.62295312      0.275362091
#> effect_76       -1.75489291     -0.972178452      -0.50273193      1.905480242
#> effect_77       -0.36003722     -0.421646942       0.39661826      0.022756485
#> effect_78       -0.20134167      1.766823996      -0.39048725     -0.869696478
#> effect_79        0.38692025      0.431421608      -2.38274438      1.965368360
#> effect_80       -1.04677187     -0.371681003      -0.42414378      1.514880412
#> effect_81        0.21793638     -0.280958414      -0.42016793      1.201710448
#> effect_82        0.98671461      0.920475748       0.55561868     -1.073603647
#> effect_83       -0.66240055      1.573858214      -0.41548108      0.688583838
#> effect_84       -3.36834059      2.051012475      -1.29267981      1.983666705
#> effect_85        2.82343785     -2.447642996      -1.74589937      0.607754576
#> effect_86        0.63055163     -0.456127426       0.10485105     -2.847573540
#> effect_87        0.24482791     -1.261343677       1.04809276      1.044536446
#> effect_88        0.17359977      0.124675554       1.08044437     -1.592406864
#> effect_89        0.05962431     -1.603512322       1.07073723     -0.155058915
#> effect_90        0.17125524     -0.231264931       1.31368649      0.607086145
#> effect_91        1.56986544     -0.565445553       1.39327610      0.426915674
#> effect_92        0.94452792     -0.793921018       0.28852327     -1.724706249
#> effect_93       -0.02515383      1.053702785      -0.45792498      0.255421956
#> effect_94        1.14814874     -0.944679082      -0.20445397     -0.968431796
#> effect_95       -0.48608395     -0.544993909      -0.29710895     -0.272010226
#> effect_96        0.54470501      0.591977359       0.90410725     -2.134588262
#> effect_97       -0.59743985      0.592766609       0.57246122      0.168639528
#> effect_98       -3.13227667     -0.910992764       0.22581345      1.535625491
#> effect_99       -2.54736018      1.109108155      -0.81308479      1.817991911
#> effect_100       0.43183664     -0.240423765       0.34705159     -0.366161685
#> effect_101      -0.52413718      0.724108589      -0.69116554      0.370815572
#> effect_102      -1.75561612      1.599710302       0.78702284     -0.142061305
#> effect_103      -2.62277116     -0.099884098      -0.08789254      0.766694984
#> effect_104       0.60768361      0.417764322       1.58334639     -1.388645645
#> effect_105      -0.63352975     -0.350031127      -0.84339085      1.357144143
#> effect_106      -0.70745581      0.616065937      -0.01389634      0.700362277
#> effect_107       0.64758188      1.054706463      -1.67415816      0.199168872
#> effect_108       0.60241477      0.040833320       0.17147850     -0.394514820
#> effect_109      -0.10198296     -1.769678561       0.34039898      0.910979804
#> effect_110      -0.54588572      0.556956940       0.34449146     -0.453634498
#> effect_111      -0.14426336     -0.090661682       1.03550216     -0.976245615
#> effect_112       1.32633903     -0.711643632       0.29558351     -0.654858535
#> effect_113       0.80179170     -0.258353145       0.15163993      0.832111703
#> effect_114      -0.40209236     -0.311038460       1.02555173     -0.364809193
#> effect_115      -1.16261974     -1.109047519       0.66024386      1.638782548
#> effect_116       1.33321318     -1.568244390       0.01376441     -0.463689826
#> effect_117      -0.58859824      0.905303626       1.54856608     -1.565939139
#> effect_118       0.63191951     -0.503645609       1.42349545     -1.604193576
#> effect_119       0.77973404     -0.729950884       0.60431483      0.207782084
#> effect_120      -0.00645734      0.147102976       0.34743982     -1.162786164
#> effect_121      -0.85482241     -1.288052643       0.17683447      0.650550354
#> effect_122      -2.25007692      0.802267877       1.11217133     -0.403059720
#> effect_123       0.68577099     -1.041878714      -0.29000994     -1.321690071
#> effect_124       0.10438776     -1.171602171       0.72305054     -0.686287921
#> effect_125      -1.13493050      0.869951644       1.36525284     -1.258170369
#> effect_126       0.36448209     -1.157727705       0.43153603     -0.359886828
#> effect_127       1.33977694      0.740503940      -0.73721583     -0.357066705
#> effect_128      -0.19818625      0.991328547       0.37153001     -0.345102077
#> effect_129      -1.94771136      0.834066221       1.46703401      0.640121654
#> effect_130      -0.74823506     -0.516784154      -0.34980408     -0.228038326
#> effect_131      -0.42815140      0.600723311       1.60991053      1.013720664
#> effect_132       0.55288539      0.085291782      -1.47271357     -0.016637082
#> effect_133      -0.58109468     -0.209533523      -0.41442307      0.909089634
#> effect_134      -0.43251892      0.148991186       0.27556930      0.085622508
#> effect_135      -0.99885225      0.402219812      -0.92502342     -0.234674499
#> effect_136      -2.49029683      1.553543993      -0.02251533      0.042376203
#> effect_137       0.57037309     -0.729565186      -1.43600081      1.387772331
#> effect_138      -0.16420132      0.558207149      -0.16850738     -0.731494294
#> effect_139      -0.08266775      0.012668459      -0.17459515      0.951331972
#> effect_140      -0.99966188     -0.023151268      -0.35585508      0.551850458
#> effect_141       0.53978265     -1.179063376       0.39763543      0.392830141
#> effect_142      -0.29649711      0.576602367      -1.19697580     -0.277359627
#> effect_143       1.12174166     -0.584043457      -0.47883460     -0.472324514
#> effect_144       1.50219669     -0.716598144      -1.20146890      0.033764430
#> effect_145       2.02962884     -0.124105177      -1.11853043     -1.870107369
#> effect_146      -0.67450813      0.314333715       0.36025630      0.488683605
#> effect_147       2.37492849     -0.947036215      -0.73845301     -0.554762800
#> effect_148      -0.30982311     -0.624574930       0.23911141      0.534846981
#> effect_149       0.34406101      0.959305270       0.03666553      0.123300976
#> effect_150       0.79507090     -0.112874955       0.34474075     -0.002675595
#> effect_151      -1.70654316      0.342582346       0.92696411     -0.211628704
#> effect_152       0.43267158      0.569708861      -0.37745559      0.558552681
#> effect_153      -0.68244064     -0.024670368      -0.72367396      1.623370664
#> effect_154      -0.03247963     -0.177427671       0.15618377      0.109861606
#> effect_155      -0.14217620      0.391294775       0.15375376      0.180177591
#> effect_156       0.67329171     -0.199206945       0.10280105      0.388218013
#> effect_157      -0.71391363      0.232474480      -1.22484651      0.175548996
#> effect_158       0.82665439     -0.101352762      -1.56433649      1.630942462
#> effect_159       0.67472142     -0.906390838       1.22871727     -0.812332773
#> effect_160       0.29078976      0.653331002      -0.91300363     -0.309335791
#> effect_161      -0.17532227      0.314139324      -0.86800170     -0.412537342
#> effect_162       0.86083413     -1.875658635      -1.22218873      0.852991220
#> effect_163       0.60646147     -1.528002196      -0.45216974      0.854962803
#> effect_164       0.79460890      0.353027861      -0.02430671     -1.048437576
#> effect_165      -0.72312746     -0.883983069       1.02178925     -0.179417967
#> effect_166      -0.81772153      0.998698097       0.96420729     -1.821413609
#> effect_167      -0.18004138      1.241786369      -0.67053293     -0.913876274
#> effect_168      -0.27291619      1.308102493      -0.37060825     -1.445350209
#> effect_169      -1.22667482     -0.295228817      -0.41346754      0.804572451
#> effect_170       0.68601986     -0.991282443       1.91755060     -1.440836309
#> effect_171       1.94549889     -1.207032403      -1.01903637     -0.575885565
#> effect_172       1.28726702     -0.649348920      -0.30885705      0.265452741
#> effect_173      -0.66346625     -0.808958002       0.33922405      1.002692248
#> effect_174      -0.15661311     -0.398475559       0.43097384     -0.231471653
#> effect_175       0.51967944      0.447851051       0.30638775     -0.653173050
#> effect_176      -1.20062308     -0.504676546       1.32383714     -1.058390762
#> effect_177      -1.30831178     -0.497053902       0.91473416      0.807343382
#> effect_178       0.90872115      0.036275315      -0.74024602      0.612885541
#> effect_179       0.34567893      0.754305130      -0.34597813      0.412963386
#> effect_180       0.40488044      1.017847389       0.08088572     -0.935819741
#> effect_181      -1.06014663      0.213322923      -0.45386799      1.191330778
#> effect_182       0.09181141      0.478840408      -0.13161746     -0.784222750
#> effect_183       0.99156611     -1.447637787       0.75229675     -0.270804675
#> effect_184       0.50608788      1.223211430       0.77531437     -1.479265285
#> effect_185      -0.08278051      0.477082563       0.26571376     -0.840006038
#> effect_186       1.14859495     -0.270126864      -0.77717121      0.652940602
#> effect_187       0.96189342     -0.245912663      -0.37625654     -0.068217569
#> effect_188       0.41482942      1.005868509      -0.64437789      0.004529797
#> effect_189       0.47144546     -0.623349540       0.15227382     -0.953264532
#> effect_190       0.54507398      0.568396772      -0.49692349     -0.886721384
#> effect_191      -1.21763277     -0.628272015       1.21944573      0.569622857
#> effect_192       0.67697271     -0.112338278      -0.19908558     -0.942814885
#> effect_193      -0.47429922     -0.379158629       0.81254050     -0.568324235
#> effect_194      -0.65210776      0.770624505      -0.12232956     -1.367132296
#> effect_195      -1.22767252      0.431587126      -0.05142670      1.063192912
#> effect_196       1.31469535     -0.755004099       1.07944236     -0.056824257
#> effect_197       0.34412139      0.392490762      -2.20163909      1.277117181
#> effect_198       0.82545084     -0.100808411      -0.92370179      0.130622472
#> effect_199      -0.52216417      0.711849031      -1.52688218      0.767921221
#> effect_200      -0.62390911      0.932090914       1.20662797     -1.246293370
#> 
#> $Shat
#>             [,1]      [,2]      [,3]      [,4]
#>   [1,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [2,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [3,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [4,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [5,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [6,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [7,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [8,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [9,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [10,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [11,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [12,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [13,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [14,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [15,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [16,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [17,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [18,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [19,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [20,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [21,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [22,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [23,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [24,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [25,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [26,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [27,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [28,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [29,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [30,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [31,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [32,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [33,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [34,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [35,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [36,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [37,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [38,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [39,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [40,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [41,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [42,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [43,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [44,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [45,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [46,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [47,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [48,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [49,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [50,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [51,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [52,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [53,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [54,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [55,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [56,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [57,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [58,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [59,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [60,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [61,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [62,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [63,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [64,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [65,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [66,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [67,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [68,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [69,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [70,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [71,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [72,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [73,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [74,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [75,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [76,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [77,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [78,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [79,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [80,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [81,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [82,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [83,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [84,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [85,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [86,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [87,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [88,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [89,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [90,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [91,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [92,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [93,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [94,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [95,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [96,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [97,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [98,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [99,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [100,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [101,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [102,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [103,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [104,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [105,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [106,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [107,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [108,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [109,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [110,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [111,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [112,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [113,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [114,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [115,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [116,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [117,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [118,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [119,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [120,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [121,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [122,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [123,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [124,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [125,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [126,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [127,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [128,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [129,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [130,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [131,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [132,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [133,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [134,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [135,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [136,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [137,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [138,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [139,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [140,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [141,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [142,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [143,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [144,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [145,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [146,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [147,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [148,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [149,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [150,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [151,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [152,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [153,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [154,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [155,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [156,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [157,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [158,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [159,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [160,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [161,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [162,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [163,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [164,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [165,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [166,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [167,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [168,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [169,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [170,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [171,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [172,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [173,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [174,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [175,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [176,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [177,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [178,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [179,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [180,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [181,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [182,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [183,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [184,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [185,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [186,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [187,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [188,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [189,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [190,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [191,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [192,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [193,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [194,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [195,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [196,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [197,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [198,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [199,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [200,] 0.8944272 0.8944272 0.8944272 0.8944272
#> 
#> $Shat_orig
#>            condition_1 condition_2 condition_3 condition_4 condition_5
#> effect_1             1           1           1           1           1
#> effect_2             1           1           1           1           1
#> effect_3             1           1           1           1           1
#> effect_4             1           1           1           1           1
#> effect_5             1           1           1           1           1
#> effect_6             1           1           1           1           1
#> effect_7             1           1           1           1           1
#> effect_8             1           1           1           1           1
#> effect_9             1           1           1           1           1
#> effect_10            1           1           1           1           1
#> effect_11            1           1           1           1           1
#> effect_12            1           1           1           1           1
#> effect_13            1           1           1           1           1
#> effect_14            1           1           1           1           1
#> effect_15            1           1           1           1           1
#> effect_16            1           1           1           1           1
#> effect_17            1           1           1           1           1
#> effect_18            1           1           1           1           1
#> effect_19            1           1           1           1           1
#> effect_20            1           1           1           1           1
#> effect_21            1           1           1           1           1
#> effect_22            1           1           1           1           1
#> effect_23            1           1           1           1           1
#> effect_24            1           1           1           1           1
#> effect_25            1           1           1           1           1
#> effect_26            1           1           1           1           1
#> effect_27            1           1           1           1           1
#> effect_28            1           1           1           1           1
#> effect_29            1           1           1           1           1
#> effect_30            1           1           1           1           1
#> effect_31            1           1           1           1           1
#> effect_32            1           1           1           1           1
#> effect_33            1           1           1           1           1
#> effect_34            1           1           1           1           1
#> effect_35            1           1           1           1           1
#> effect_36            1           1           1           1           1
#> effect_37            1           1           1           1           1
#> effect_38            1           1           1           1           1
#> effect_39            1           1           1           1           1
#> effect_40            1           1           1           1           1
#> effect_41            1           1           1           1           1
#> effect_42            1           1           1           1           1
#> effect_43            1           1           1           1           1
#> effect_44            1           1           1           1           1
#> effect_45            1           1           1           1           1
#> effect_46            1           1           1           1           1
#> effect_47            1           1           1           1           1
#> effect_48            1           1           1           1           1
#> effect_49            1           1           1           1           1
#> effect_50            1           1           1           1           1
#> effect_51            1           1           1           1           1
#> effect_52            1           1           1           1           1
#> effect_53            1           1           1           1           1
#> effect_54            1           1           1           1           1
#> effect_55            1           1           1           1           1
#> effect_56            1           1           1           1           1
#> effect_57            1           1           1           1           1
#> effect_58            1           1           1           1           1
#> effect_59            1           1           1           1           1
#> effect_60            1           1           1           1           1
#> effect_61            1           1           1           1           1
#> effect_62            1           1           1           1           1
#> effect_63            1           1           1           1           1
#> effect_64            1           1           1           1           1
#> effect_65            1           1           1           1           1
#> effect_66            1           1           1           1           1
#> effect_67            1           1           1           1           1
#> effect_68            1           1           1           1           1
#> effect_69            1           1           1           1           1
#> effect_70            1           1           1           1           1
#> effect_71            1           1           1           1           1
#> effect_72            1           1           1           1           1
#> effect_73            1           1           1           1           1
#> effect_74            1           1           1           1           1
#> effect_75            1           1           1           1           1
#> effect_76            1           1           1           1           1
#> effect_77            1           1           1           1           1
#> effect_78            1           1           1           1           1
#> effect_79            1           1           1           1           1
#> effect_80            1           1           1           1           1
#> effect_81            1           1           1           1           1
#> effect_82            1           1           1           1           1
#> effect_83            1           1           1           1           1
#> effect_84            1           1           1           1           1
#> effect_85            1           1           1           1           1
#> effect_86            1           1           1           1           1
#> effect_87            1           1           1           1           1
#> effect_88            1           1           1           1           1
#> effect_89            1           1           1           1           1
#> effect_90            1           1           1           1           1
#> effect_91            1           1           1           1           1
#> effect_92            1           1           1           1           1
#> effect_93            1           1           1           1           1
#> effect_94            1           1           1           1           1
#> effect_95            1           1           1           1           1
#> effect_96            1           1           1           1           1
#> effect_97            1           1           1           1           1
#> effect_98            1           1           1           1           1
#> effect_99            1           1           1           1           1
#> effect_100           1           1           1           1           1
#> effect_101           1           1           1           1           1
#> effect_102           1           1           1           1           1
#> effect_103           1           1           1           1           1
#> effect_104           1           1           1           1           1
#> effect_105           1           1           1           1           1
#> effect_106           1           1           1           1           1
#> effect_107           1           1           1           1           1
#> effect_108           1           1           1           1           1
#> effect_109           1           1           1           1           1
#> effect_110           1           1           1           1           1
#> effect_111           1           1           1           1           1
#> effect_112           1           1           1           1           1
#> effect_113           1           1           1           1           1
#> effect_114           1           1           1           1           1
#> effect_115           1           1           1           1           1
#> effect_116           1           1           1           1           1
#> effect_117           1           1           1           1           1
#> effect_118           1           1           1           1           1
#> effect_119           1           1           1           1           1
#> effect_120           1           1           1           1           1
#> effect_121           1           1           1           1           1
#> effect_122           1           1           1           1           1
#> effect_123           1           1           1           1           1
#> effect_124           1           1           1           1           1
#> effect_125           1           1           1           1           1
#> effect_126           1           1           1           1           1
#> effect_127           1           1           1           1           1
#> effect_128           1           1           1           1           1
#> effect_129           1           1           1           1           1
#> effect_130           1           1           1           1           1
#> effect_131           1           1           1           1           1
#> effect_132           1           1           1           1           1
#> effect_133           1           1           1           1           1
#> effect_134           1           1           1           1           1
#> effect_135           1           1           1           1           1
#> effect_136           1           1           1           1           1
#> effect_137           1           1           1           1           1
#> effect_138           1           1           1           1           1
#> effect_139           1           1           1           1           1
#> effect_140           1           1           1           1           1
#> effect_141           1           1           1           1           1
#> effect_142           1           1           1           1           1
#> effect_143           1           1           1           1           1
#> effect_144           1           1           1           1           1
#> effect_145           1           1           1           1           1
#> effect_146           1           1           1           1           1
#> effect_147           1           1           1           1           1
#> effect_148           1           1           1           1           1
#> effect_149           1           1           1           1           1
#> effect_150           1           1           1           1           1
#> effect_151           1           1           1           1           1
#> effect_152           1           1           1           1           1
#> effect_153           1           1           1           1           1
#> effect_154           1           1           1           1           1
#> effect_155           1           1           1           1           1
#> effect_156           1           1           1           1           1
#> effect_157           1           1           1           1           1
#> effect_158           1           1           1           1           1
#> effect_159           1           1           1           1           1
#> effect_160           1           1           1           1           1
#> effect_161           1           1           1           1           1
#> effect_162           1           1           1           1           1
#> effect_163           1           1           1           1           1
#> effect_164           1           1           1           1           1
#> effect_165           1           1           1           1           1
#> effect_166           1           1           1           1           1
#> effect_167           1           1           1           1           1
#> effect_168           1           1           1           1           1
#> effect_169           1           1           1           1           1
#> effect_170           1           1           1           1           1
#> effect_171           1           1           1           1           1
#> effect_172           1           1           1           1           1
#> effect_173           1           1           1           1           1
#> effect_174           1           1           1           1           1
#> effect_175           1           1           1           1           1
#> effect_176           1           1           1           1           1
#> effect_177           1           1           1           1           1
#> effect_178           1           1           1           1           1
#> effect_179           1           1           1           1           1
#> effect_180           1           1           1           1           1
#> effect_181           1           1           1           1           1
#> effect_182           1           1           1           1           1
#> effect_183           1           1           1           1           1
#> effect_184           1           1           1           1           1
#> effect_185           1           1           1           1           1
#> effect_186           1           1           1           1           1
#> effect_187           1           1           1           1           1
#> effect_188           1           1           1           1           1
#> effect_189           1           1           1           1           1
#> effect_190           1           1           1           1           1
#> effect_191           1           1           1           1           1
#> effect_192           1           1           1           1           1
#> effect_193           1           1           1           1           1
#> effect_194           1           1           1           1           1
#> effect_195           1           1           1           1           1
#> effect_196           1           1           1           1           1
#> effect_197           1           1           1           1           1
#> effect_198           1           1           1           1           1
#> effect_199           1           1           1           1           1
#> effect_200           1           1           1           1           1
#> 
#> $Shat_alpha
#>        [,1] [,2] [,3] [,4]
#>   [1,]    1    1    1    1
#>   [2,]    1    1    1    1
#>   [3,]    1    1    1    1
#>   [4,]    1    1    1    1
#>   [5,]    1    1    1    1
#>   [6,]    1    1    1    1
#>   [7,]    1    1    1    1
#>   [8,]    1    1    1    1
#>   [9,]    1    1    1    1
#>  [10,]    1    1    1    1
#>  [11,]    1    1    1    1
#>  [12,]    1    1    1    1
#>  [13,]    1    1    1    1
#>  [14,]    1    1    1    1
#>  [15,]    1    1    1    1
#>  [16,]    1    1    1    1
#>  [17,]    1    1    1    1
#>  [18,]    1    1    1    1
#>  [19,]    1    1    1    1
#>  [20,]    1    1    1    1
#>  [21,]    1    1    1    1
#>  [22,]    1    1    1    1
#>  [23,]    1    1    1    1
#>  [24,]    1    1    1    1
#>  [25,]    1    1    1    1
#>  [26,]    1    1    1    1
#>  [27,]    1    1    1    1
#>  [28,]    1    1    1    1
#>  [29,]    1    1    1    1
#>  [30,]    1    1    1    1
#>  [31,]    1    1    1    1
#>  [32,]    1    1    1    1
#>  [33,]    1    1    1    1
#>  [34,]    1    1    1    1
#>  [35,]    1    1    1    1
#>  [36,]    1    1    1    1
#>  [37,]    1    1    1    1
#>  [38,]    1    1    1    1
#>  [39,]    1    1    1    1
#>  [40,]    1    1    1    1
#>  [41,]    1    1    1    1
#>  [42,]    1    1    1    1
#>  [43,]    1    1    1    1
#>  [44,]    1    1    1    1
#>  [45,]    1    1    1    1
#>  [46,]    1    1    1    1
#>  [47,]    1    1    1    1
#>  [48,]    1    1    1    1
#>  [49,]    1    1    1    1
#>  [50,]    1    1    1    1
#>  [51,]    1    1    1    1
#>  [52,]    1    1    1    1
#>  [53,]    1    1    1    1
#>  [54,]    1    1    1    1
#>  [55,]    1    1    1    1
#>  [56,]    1    1    1    1
#>  [57,]    1    1    1    1
#>  [58,]    1    1    1    1
#>  [59,]    1    1    1    1
#>  [60,]    1    1    1    1
#>  [61,]    1    1    1    1
#>  [62,]    1    1    1    1
#>  [63,]    1    1    1    1
#>  [64,]    1    1    1    1
#>  [65,]    1    1    1    1
#>  [66,]    1    1    1    1
#>  [67,]    1    1    1    1
#>  [68,]    1    1    1    1
#>  [69,]    1    1    1    1
#>  [70,]    1    1    1    1
#>  [71,]    1    1    1    1
#>  [72,]    1    1    1    1
#>  [73,]    1    1    1    1
#>  [74,]    1    1    1    1
#>  [75,]    1    1    1    1
#>  [76,]    1    1    1    1
#>  [77,]    1    1    1    1
#>  [78,]    1    1    1    1
#>  [79,]    1    1    1    1
#>  [80,]    1    1    1    1
#>  [81,]    1    1    1    1
#>  [82,]    1    1    1    1
#>  [83,]    1    1    1    1
#>  [84,]    1    1    1    1
#>  [85,]    1    1    1    1
#>  [86,]    1    1    1    1
#>  [87,]    1    1    1    1
#>  [88,]    1    1    1    1
#>  [89,]    1    1    1    1
#>  [90,]    1    1    1    1
#>  [91,]    1    1    1    1
#>  [92,]    1    1    1    1
#>  [93,]    1    1    1    1
#>  [94,]    1    1    1    1
#>  [95,]    1    1    1    1
#>  [96,]    1    1    1    1
#>  [97,]    1    1    1    1
#>  [98,]    1    1    1    1
#>  [99,]    1    1    1    1
#> [100,]    1    1    1    1
#> [101,]    1    1    1    1
#> [102,]    1    1    1    1
#> [103,]    1    1    1    1
#> [104,]    1    1    1    1
#> [105,]    1    1    1    1
#> [106,]    1    1    1    1
#> [107,]    1    1    1    1
#> [108,]    1    1    1    1
#> [109,]    1    1    1    1
#> [110,]    1    1    1    1
#> [111,]    1    1    1    1
#> [112,]    1    1    1    1
#> [113,]    1    1    1    1
#> [114,]    1    1    1    1
#> [115,]    1    1    1    1
#> [116,]    1    1    1    1
#> [117,]    1    1    1    1
#> [118,]    1    1    1    1
#> [119,]    1    1    1    1
#> [120,]    1    1    1    1
#> [121,]    1    1    1    1
#> [122,]    1    1    1    1
#> [123,]    1    1    1    1
#> [124,]    1    1    1    1
#> [125,]    1    1    1    1
#> [126,]    1    1    1    1
#> [127,]    1    1    1    1
#> [128,]    1    1    1    1
#> [129,]    1    1    1    1
#> [130,]    1    1    1    1
#> [131,]    1    1    1    1
#> [132,]    1    1    1    1
#> [133,]    1    1    1    1
#> [134,]    1    1    1    1
#> [135,]    1    1    1    1
#> [136,]    1    1    1    1
#> [137,]    1    1    1    1
#> [138,]    1    1    1    1
#> [139,]    1    1    1    1
#> [140,]    1    1    1    1
#> [141,]    1    1    1    1
#> [142,]    1    1    1    1
#> [143,]    1    1    1    1
#> [144,]    1    1    1    1
#> [145,]    1    1    1    1
#> [146,]    1    1    1    1
#> [147,]    1    1    1    1
#> [148,]    1    1    1    1
#> [149,]    1    1    1    1
#> [150,]    1    1    1    1
#> [151,]    1    1    1    1
#> [152,]    1    1    1    1
#> [153,]    1    1    1    1
#> [154,]    1    1    1    1
#> [155,]    1    1    1    1
#> [156,]    1    1    1    1
#> [157,]    1    1    1    1
#> [158,]    1    1    1    1
#> [159,]    1    1    1    1
#> [160,]    1    1    1    1
#> [161,]    1    1    1    1
#> [162,]    1    1    1    1
#> [163,]    1    1    1    1
#> [164,]    1    1    1    1
#> [165,]    1    1    1    1
#> [166,]    1    1    1    1
#> [167,]    1    1    1    1
#> [168,]    1    1    1    1
#> [169,]    1    1    1    1
#> [170,]    1    1    1    1
#> [171,]    1    1    1    1
#> [172,]    1    1    1    1
#> [173,]    1    1    1    1
#> [174,]    1    1    1    1
#> [175,]    1    1    1    1
#> [176,]    1    1    1    1
#> [177,]    1    1    1    1
#> [178,]    1    1    1    1
#> [179,]    1    1    1    1
#> [180,]    1    1    1    1
#> [181,]    1    1    1    1
#> [182,]    1    1    1    1
#> [183,]    1    1    1    1
#> [184,]    1    1    1    1
#> [185,]    1    1    1    1
#> [186,]    1    1    1    1
#> [187,]    1    1    1    1
#> [188,]    1    1    1    1
#> [189,]    1    1    1    1
#> [190,]    1    1    1    1
#> [191,]    1    1    1    1
#> [192,]    1    1    1    1
#> [193,]    1    1    1    1
#> [194,]    1    1    1    1
#> [195,]    1    1    1    1
#> [196,]    1    1    1    1
#> [197,]    1    1    1    1
#> [198,]    1    1    1    1
#> [199,]    1    1    1    1
#> [200,]    1    1    1    1
#> 
#> $V
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    1    0    0    0    0
#> [2,]    0    1    0    0    0
#> [3,]    0    0    1    0    0
#> [4,]    0    0    0    1    0
#> [5,]    0    0    0    0    1
#> 
#> $commonV
#> [1] TRUE
#> 
#> $alpha
#> [1] 0
#> 
#> $L
#>                  condition_1 condition_2 condition_3 condition_4 condition_5
#> condition_1-mean         0.8        -0.2        -0.2        -0.2        -0.2
#> condition_2-mean        -0.2         0.8        -0.2        -0.2        -0.2
#> condition_3-mean        -0.2        -0.2         0.8        -0.2        -0.2
#> condition_4-mean        -0.2        -0.2        -0.2         0.8        -0.2
#> attr(,"reference")
#> [1] "mean"
#> 
#> $LSVSLt
#>                  condition_1-mean condition_2-mean condition_3-mean
#> condition_1-mean             1.00            -0.25            -0.25
#> condition_2-mean            -0.25             1.00            -0.25
#> condition_3-mean            -0.25            -0.25             1.00
#> condition_4-mean            -0.25            -0.25            -0.25
#>                  condition_4-mean
#> condition_1-mean            -0.25
#> condition_2-mean            -0.25
#> condition_3-mean            -0.25
#> condition_4-mean             1.00
#> 
#> attr(,"class")
#> [1] "mash"