Create simulation with signal data used for contrast analysis.

Create simulation with signal data used for contrast analysis.

sim_contrast2(nsamp = 1000, ncond = 5, err_sd = sqrt(0.5))

Arguments

nsamp

Number of samples of each type.

ncond

Number of conditions.

err_sd

The standard deviation of the errors.

Details

The first condition is the reference group. The deviations are the difference between the subsequent conditions with the reference group. The simulation consists of 90 10 different types of deviations: equal among conditions, present only in the first subsequent condition, independent across conditions.

Examples

sim_contrast2(100,5)
#> $C
#>            condition_1 condition_2 condition_3 condition_4 condition_5
#> sample_1     11.112876   11.112876   11.112876   11.112876   11.112876
#> sample_2      9.476985    9.476985    9.476985    9.476985    9.476985
#> sample_3     10.332203   10.332203   10.332203   10.332203   10.332203
#> sample_4      9.321691    9.321691    9.321691    9.321691    9.321691
#> sample_5      8.779544    8.779544    8.779544    8.779544    8.779544
#> sample_6     11.038828   11.038828   11.038828   11.038828   11.038828
#> sample_7     10.285455   10.285455   10.285455   10.285455   10.285455
#> sample_8     10.703378   10.703378   10.703378   10.703378   10.703378
#> sample_9      9.811245    9.811245    9.811245    9.811245    9.811245
#> sample_10     9.928538    9.928538    9.928538    9.928538    9.928538
#> sample_11    10.009402   10.009402   10.009402   10.009402   10.009402
#> sample_12     7.823162    7.823162    7.823162    7.823162    7.823162
#> sample_13    10.558526   10.558526   10.558526   10.558526   10.558526
#> sample_14    10.530716   10.530716   10.530716   10.530716   10.530716
#> sample_15     9.805381    9.805381    9.805381    9.805381    9.805381
#> sample_16     8.758914    8.758914    8.758914    8.758914    8.758914
#> sample_17     8.557645    8.557645    8.557645    8.557645    8.557645
#> sample_18     8.361914    8.361914    8.361914    8.361914    8.361914
#> sample_19    10.795674   10.795674   10.795674   10.795674   10.795674
#> sample_20    10.172718   10.172718   10.172718   10.172718   10.172718
#> sample_21     9.915979    9.915979    9.915979    9.915979    9.915979
#> sample_22     9.892695    9.892695    9.892695    9.892695    9.892695
#> sample_23    10.754522   10.754522   10.754522   10.754522   10.754522
#> sample_24     9.983773    9.983773    9.983773    9.983773    9.983773
#> sample_25     8.355015    8.355015    8.355015    8.355015    8.355015
#> sample_26    10.326767   10.326767   10.326767   10.326767   10.326767
#> sample_27     9.373130    9.373130    9.373130    9.373130    9.373130
#> sample_28     8.982265    8.982265    8.982265    8.982265    8.982265
#> sample_29    11.798468   11.798468   11.798468   11.798468   11.798468
#> sample_30     9.691354    9.691354    9.691354    9.691354    9.691354
#> sample_31    11.483627   11.483627   11.483627   11.483627   11.483627
#> sample_32     9.929004    9.929004    9.929004    9.929004    9.929004
#> sample_33     9.568834    9.568834    9.568834    9.568834    9.568834
#> sample_34     8.301811    8.301811    8.301811    8.301811    8.301811
#> sample_35    11.770027   11.770027   11.770027   11.770027   11.770027
#> sample_36    10.866939   10.866939   10.866939   10.866939   10.866939
#> sample_37    12.027790   12.027790   12.027790   12.027790   12.027790
#> sample_38     9.374998    9.374998    9.374998    9.374998    9.374998
#> sample_39     9.647529    9.647529    9.647529    9.647529    9.647529
#> sample_40     9.551753    9.551753    9.551753    9.551753    9.551753
#> sample_41     9.377970    9.377970    9.377970    9.377970    9.377970
#> sample_42    11.317134   11.317134   11.317134   11.317134   11.317134
#> sample_43    10.461377   10.461377   10.461377   10.461377   10.461377
#> sample_44     8.114421    8.114421    8.114421    8.114421    8.114421
#> sample_45    11.315074   11.315074   11.315074   11.315074   11.315074
#> sample_46    10.998298   10.998298   10.998298   10.998298   10.998298
#> sample_47    11.056026   11.056026   11.056026   11.056026   11.056026
#> sample_48     9.529399    9.529399    9.529399    9.529399    9.529399
#> sample_49    11.674457   11.674457   11.674457   11.674457   11.674457
#> sample_50    10.510563   10.510563   10.510563   10.510563   10.510563
#> sample_51     9.565251    9.565251    9.565251    9.565251    9.565251
#> sample_52    10.752365   10.752365   10.752365   10.752365   10.752365
#> sample_53     8.829054    8.829054    8.829054    8.829054    8.829054
#> sample_54     9.863128    9.863128    9.863128    9.863128    9.863128
#> sample_55     8.978646    8.978646    8.978646    8.978646    8.978646
#> sample_56     9.721070    9.721070    9.721070    9.721070    9.721070
#> sample_57    10.789574   10.789574   10.789574   10.789574   10.789574
#> sample_58    11.659852   11.659852   11.659852   11.659852   11.659852
#> sample_59    11.516638   11.516638   11.516638   11.516638   11.516638
#> sample_60     9.542425    9.542425    9.542425    9.542425    9.542425
#> sample_61    10.992088   10.992088   10.992088   10.992088   10.992088
#> sample_62    10.291332   10.291332   10.291332   10.291332   10.291332
#> sample_63    12.097965   12.097965   12.097965   12.097965   12.097965
#> sample_64     9.080748    9.080748    9.080748    9.080748    9.080748
#> sample_65    11.224160   11.224160   11.224160   11.224160   11.224160
#> sample_66    11.623733   11.623733   11.623733   11.623733   11.623733
#> sample_67     9.373639    9.373639    9.373639    9.373639    9.373639
#> sample_68     7.880975    7.880975    7.880975    7.880975    7.880975
#> sample_69     9.577724    9.577724    9.577724    9.577724    9.577724
#> sample_70    10.298510   10.298510   10.298510   10.298510   10.298510
#> sample_71     8.636886    8.636886    8.636886    8.636886    8.636886
#> sample_72     9.379306    9.379306    9.379306    9.379306    9.379306
#> sample_73     8.520911    8.520911    8.520911    8.520911    8.520911
#> sample_74     8.488139    8.488139    8.488139    8.488139    8.488139
#> sample_75     7.860777    7.860777    7.860777    7.860777    7.860777
#> sample_76     9.224123    9.224123    9.224123    9.224123    9.224123
#> sample_77    10.135362   10.135362   10.135362   10.135362   10.135362
#> sample_78    10.417767   10.417767   10.417767   10.417767   10.417767
#> sample_79     9.960245    9.960245    9.960245    9.960245    9.960245
#> sample_80    10.024058   10.024058   10.024058   10.024058   10.024058
#> sample_81     9.639982    9.639982    9.639982    9.639982    9.639982
#> sample_82    10.330851   10.330851   10.330851   10.330851   10.330851
#> sample_83     7.914936    7.914936    7.914936    7.914936    7.914936
#> sample_84     8.776339    8.776339    8.776339    8.776339    8.776339
#> sample_85     9.352779    9.352779    9.352779    9.352779    9.352779
#> sample_86     9.410347    9.410347    9.410347    9.410347    9.410347
#> sample_87     9.821179    9.821179    9.821179    9.821179    9.821179
#> sample_88    10.131172   10.131172   10.131172   10.131172   10.131172
#> sample_89    12.796328   12.796328   12.796328   12.796328   12.796328
#> sample_90     9.547580    9.547580    9.547580    9.547580    9.547580
#> sample_91     8.876766    7.123230    7.495160   10.130116    9.203251
#> sample_92     7.538837    9.187422    8.173299    8.910710    9.302405
#> sample_93     9.321287    7.996022   11.989409   10.794938   10.552587
#> sample_94     7.827639    7.827639    7.827639    7.827639    7.682084
#> sample_95     7.653795    7.653795    7.653795    7.653795    9.693960
#> sample_96    11.952737   11.952737   11.952737   11.952737   11.403299
#> sample_97     9.009891    9.952479    9.952479    9.952479    9.952479
#> sample_98    10.640522   10.709927   10.709927   10.709927   10.709927
#> sample_99     9.593289   10.473760   10.473760   10.473760   10.473760
#> sample_100    8.645291    9.210152    9.210152    9.210152    9.210152
#> 
#> $Chat
#>            condition_1 condition_2 condition_3 condition_4 condition_5
#> sample_1     11.612939   11.920774    9.779221   10.272481   11.043696
#> sample_2     10.073335    8.701137    9.235700   10.249176    7.895511
#> sample_3     10.048753   10.964240   10.000185   10.958880   10.222161
#> sample_4      9.470914    8.988463    9.459212    8.561571    9.852436
#> sample_5      8.761785    7.986512    9.035379    9.502620    9.196170
#> sample_6      9.828917   11.461365   10.911319   10.864461    9.366800
#> sample_7     10.467011    9.185101   10.885105    9.836431    9.210986
#> sample_8     10.371701   10.683230   10.188096   11.209452   11.255216
#> sample_9     10.867837    9.549636    9.502382   10.179315    9.147487
#> sample_10     9.733837    8.993628   10.118728   10.021149    8.354110
#> sample_11    10.685891    9.863536    9.712434   10.183740   10.401833
#> sample_12     8.014373    8.627455    7.221493    8.197938    8.068810
#> sample_13     9.354545   10.421923   10.125670   10.119508   10.264824
#> sample_14     9.478231   10.560416   11.136864   11.442282    9.726815
#> sample_15     8.951141    9.459801   10.633297   10.057807   10.268088
#> sample_16     9.061714    9.192178    9.169471    8.048107    8.114971
#> sample_17     9.666290    8.444018    9.329960    8.569957    7.738108
#> sample_18     9.014909    7.712705    9.337310    9.424684    8.434758
#> sample_19    10.871268   10.631069   11.246468   10.572334   11.628324
#> sample_20    11.755718   10.956356   10.389807   10.114504   10.578387
#> sample_21     9.327095    9.108092   10.363594   10.142456    9.213503
#> sample_22     8.989522    9.524933    9.551194    9.682637    9.552772
#> sample_23    10.162268   10.542809   10.429295   12.185004   10.481481
#> sample_24    10.801179    9.948522    9.084018   11.107587    9.347739
#> sample_25     9.099815    8.096744    9.113172    8.869091    7.069250
#> sample_26     9.769580   11.792588   10.461777   10.708981    9.490981
#> sample_27    10.075448   10.474575    9.354416    9.318598    8.238943
#> sample_28     8.920015    9.296351   10.357068    6.821775    9.127073
#> sample_29    12.360760   12.963458   11.985304   10.836331   11.920721
#> sample_30     8.256713    9.950940   10.062025   10.002933    9.871062
#> sample_31    10.986812   12.803184   11.248194   12.472922   12.123940
#> sample_32     9.759519   10.209265   11.407914    9.280117   10.737185
#> sample_33     8.648682    9.123768    9.934220    8.856445   10.667477
#> sample_34     8.183303    7.730047    8.175802    7.203459    8.148803
#> sample_35    10.993553   12.460221   12.130559   10.330741   10.625608
#> sample_36    11.320347   11.382371   10.802519   11.083037   11.548823
#> sample_37    12.131099   12.649183   11.906146   11.289726   13.186872
#> sample_38     9.719832   10.041557   10.256391    8.726802    9.644104
#> sample_39     9.071742    9.641730   10.050009    9.459925    8.953806
#> sample_40     9.492893    8.815140   10.573806    9.804967   10.054464
#> sample_41     9.806247    8.611052    9.358086    9.949587    9.500092
#> sample_42    11.739986   10.439474   10.625322   11.471500    9.839465
#> sample_43    10.160425   11.279384    9.855807   10.001094   10.531547
#> sample_44     8.436280    8.470219    8.309925    8.012600    8.312650
#> sample_45    10.812433   11.576739   10.325465   11.809582   11.474549
#> sample_46    10.533977   10.692670   10.110578   11.749925   11.425145
#> sample_47    10.730390   10.149973   12.773542   10.121748   11.271366
#> sample_48     9.713253    8.528516    8.722669    9.952589    9.316014
#> sample_49    11.755312   11.787535   10.774421   10.899761   12.320109
#> sample_50     9.632640    9.448960   10.987263   11.162096   10.914650
#> sample_51     9.895877    9.606795   10.004906    9.953448   10.473073
#> sample_52     9.296556   10.409852    9.603942    9.361771   10.774521
#> sample_53     9.469768    8.849592    9.899888    8.104049    8.897747
#> sample_54    10.142261   10.978081    9.171521   10.591739    9.877185
#> sample_55     9.795235    8.961324    8.080844    9.407419   10.202260
#> sample_56    11.050922    8.512319    8.784869   10.262031    9.369246
#> sample_57    10.267875   10.977363   11.621719   10.042298   10.460889
#> sample_58    13.016399   10.436669   11.641207   11.748672   10.614006
#> sample_59    10.878206   11.562510   10.615273   12.160128   10.846232
#> sample_60     9.585527   10.006147    9.657519    8.692155    9.364524
#> sample_61    10.543851   11.084211   10.773174   11.173954   11.617265
#> sample_62     9.532162    9.277329   10.585792    9.830045    9.884725
#> sample_63    11.924266   12.321342   13.297964   12.053497   12.289358
#> sample_64     8.158957    9.844060    8.536411    9.319391    9.722231
#> sample_65    11.409154   10.519990   11.258750   10.989933   11.225182
#> sample_66    11.843503   12.139452   11.172351   12.490114   11.969018
#> sample_67     8.413402    9.301622   10.066273    8.560966    9.785060
#> sample_68     7.584451    6.970632    8.472406    7.227208    7.880818
#> sample_69    10.044277    8.646801    9.687549   11.420590    8.716160
#> sample_70    10.303451    8.754966   10.017321   10.567668   10.584749
#> sample_71     8.697739    6.354508    9.123155    8.387458    8.216173
#> sample_72    10.005642    9.645300   10.447807   10.515162    9.331613
#> sample_73     8.225850    8.662388    8.881044    7.274345    8.457132
#> sample_74     8.502206   10.289030    7.850329    7.477841    8.734841
#> sample_75     9.528359    8.042644    7.475723    8.262549    8.156939
#> sample_76     8.863283    8.906678    9.330385   10.412297    9.102246
#> sample_77     8.435700   10.196003    9.897381   10.629885    9.679073
#> sample_78    10.868148   10.435757    9.138783   11.030527    9.780341
#> sample_79    10.901947   10.594643    8.941887   10.479099   10.094622
#> sample_80    10.254727    8.629706    9.878878   10.289294    9.929535
#> sample_81    10.705339   10.516198   10.578385    9.199003    9.376651
#> sample_82    10.316872    9.686141   11.484874    9.232379   10.365578
#> sample_83     6.642212    9.791725    7.892689    6.664461    7.185101
#> sample_84     9.582537    8.906365    9.492733   10.351326    9.651220
#> sample_85     8.523558    8.837608   10.331636    9.139476   10.310514
#> sample_86    10.863320    9.754889    9.348182    8.764619    9.887433
#> sample_87     8.633519    9.808937   10.080948   10.128097    9.158938
#> sample_88    10.334467    8.885906   10.681905    9.949083   10.041725
#> sample_89    13.373149   12.420792   13.543795   14.647118   13.012258
#> sample_90    10.349712    9.016755    8.790160    9.743563    9.331780
#> sample_91    10.144071    8.719036    6.463988   10.305880    8.950317
#> sample_92     6.995637    9.849778    8.082554    9.772127    8.817013
#> sample_93     9.227238    7.684959   11.850470   12.040445   11.313421
#> sample_94     7.630067    6.895703    7.747427    8.802946    8.945179
#> sample_95     8.238259    7.939007    8.418984    6.806972    8.943941
#> sample_96    11.723776   12.716442   12.463943   12.080509   11.818592
#> sample_97    10.010867    9.991878   10.077726   10.877267    9.513966
#> sample_98     9.768997   11.061356   11.935893   10.697473   11.265070
#> sample_99     9.088488   10.815952   10.757100   11.180100    9.819506
#> sample_100    7.104634    9.216655    9.417679    8.047312    8.589307
#> 
#> $Shat
#>            condition_1 condition_2 condition_3 condition_4 condition_5
#> sample_1     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_2     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_3     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_4     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_5     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_6     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_7     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_8     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_9     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_10    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_11    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_12    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_13    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_14    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_15    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_16    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_17    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_18    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_19    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_20    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_21    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_22    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_23    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_24    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_25    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_26    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_27    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_28    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_29    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_30    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_31    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_32    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_33    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_34    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_35    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_36    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_37    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_38    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_39    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_40    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_41    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_42    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_43    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_44    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_45    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_46    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_47    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_48    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_49    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_50    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_51    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_52    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_53    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_54    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_55    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_56    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_57    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_58    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_59    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_60    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_61    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_62    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_63    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_64    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_65    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_66    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_67    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_68    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_69    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_70    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_71    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_72    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_73    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_74    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_75    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_76    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_77    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_78    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_79    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_80    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_81    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_82    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_83    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_84    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_85    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_86    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_87    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_88    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_89    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_90    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_91    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_92    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_93    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_94    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_95    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_96    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_97    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_98    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_99    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_100   0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#>