Create some more simple simulated data for testing purposes

Create some more simple simulated data for testing purposes

simple_sims2(nsamp = 100, err_sd = 0.01)

Arguments

nsamp

number of samples of each type

err_sd

the standard deviation of the errors

Details

The simulation consists of five conditions with two types of effecc those present (and identical) in first two conditions and those present (and identical) in last three conditions

Examples

simple_sims2(100, 5)
#> $B
#>            condition_1 condition_2 condition_3 condition_4 condition_5
#> effect_1   -0.47017291 -0.47017291  0.00000000  0.00000000  0.00000000
#> effect_2    0.93933364  0.93933364  0.00000000  0.00000000  0.00000000
#> effect_3   -0.04115531 -0.04115531  0.00000000  0.00000000  0.00000000
#> effect_4    0.92115623  0.92115623  0.00000000  0.00000000  0.00000000
#> effect_5    1.66948220  1.66948220  0.00000000  0.00000000  0.00000000
#> effect_6    0.67326440  0.67326440  0.00000000  0.00000000  0.00000000
#> effect_7    0.11059285  0.11059285  0.00000000  0.00000000  0.00000000
#> effect_8   -0.60998491 -0.60998491  0.00000000  0.00000000  0.00000000
#> effect_9   -1.44015854 -1.44015854  0.00000000  0.00000000  0.00000000
#> effect_10   0.35033993  0.35033993  0.00000000  0.00000000  0.00000000
#> effect_11  -1.15950513 -1.15950513  0.00000000  0.00000000  0.00000000
#> effect_12  -0.48934179 -0.48934179  0.00000000  0.00000000  0.00000000
#> effect_13  -0.41292740 -0.41292740  0.00000000  0.00000000  0.00000000
#> effect_14   0.97282873  0.97282873  0.00000000  0.00000000  0.00000000
#> effect_15  -0.83657624 -0.83657624  0.00000000  0.00000000  0.00000000
#> effect_16  -2.18732229 -2.18732229  0.00000000  0.00000000  0.00000000
#> effect_17   0.69174077  0.69174077  0.00000000  0.00000000  0.00000000
#> effect_18   2.19096665  2.19096665  0.00000000  0.00000000  0.00000000
#> effect_19  -0.99151866 -0.99151866  0.00000000  0.00000000  0.00000000
#> effect_20  -2.11543524 -2.11543524  0.00000000  0.00000000  0.00000000
#> effect_21   0.22653181  0.22653181  0.00000000  0.00000000  0.00000000
#> effect_22  -0.03382604 -0.03382604  0.00000000  0.00000000  0.00000000
#> effect_23   0.32923555  0.32923555  0.00000000  0.00000000  0.00000000
#> effect_24   0.30320277  0.30320277  0.00000000  0.00000000  0.00000000
#> effect_25   0.41958422  0.41958422  0.00000000  0.00000000  0.00000000
#> effect_26   0.39062682  0.39062682  0.00000000  0.00000000  0.00000000
#> effect_27  -1.25348750 -1.25348750  0.00000000  0.00000000  0.00000000
#> effect_28   0.43427311  0.43427311  0.00000000  0.00000000  0.00000000
#> effect_29  -0.21197051 -0.21197051  0.00000000  0.00000000  0.00000000
#> effect_30   0.43898526  0.43898526  0.00000000  0.00000000  0.00000000
#> effect_31   0.14660069  0.14660069  0.00000000  0.00000000  0.00000000
#> effect_32  -1.90113668 -1.90113668  0.00000000  0.00000000  0.00000000
#> effect_33  -1.55376622 -1.55376622  0.00000000  0.00000000  0.00000000
#> effect_34  -0.13215282 -0.13215282  0.00000000  0.00000000  0.00000000
#> effect_35   1.17289689  1.17289689  0.00000000  0.00000000  0.00000000
#> effect_36  -0.19139978 -0.19139978  0.00000000  0.00000000  0.00000000
#> effect_37   1.92173782  1.92173782  0.00000000  0.00000000  0.00000000
#> effect_38  -0.84093952 -0.84093952  0.00000000  0.00000000  0.00000000
#> effect_39   2.04911149  2.04911149  0.00000000  0.00000000  0.00000000
#> effect_40   1.10859104  1.10859104  0.00000000  0.00000000  0.00000000
#> effect_41   0.52205248  0.52205248  0.00000000  0.00000000  0.00000000
#> effect_42  -0.88245979 -0.88245979  0.00000000  0.00000000  0.00000000
#> effect_43   0.32009440  0.32009440  0.00000000  0.00000000  0.00000000
#> effect_44   0.03697823  0.03697823  0.00000000  0.00000000  0.00000000
#> effect_45  -0.56786346 -0.56786346  0.00000000  0.00000000  0.00000000
#> effect_46  -0.03933711 -0.03933711  0.00000000  0.00000000  0.00000000
#> effect_47   0.86085538  0.86085538  0.00000000  0.00000000  0.00000000
#> effect_48   1.10110417  1.10110417  0.00000000  0.00000000  0.00000000
#> effect_49  -0.14335759 -0.14335759  0.00000000  0.00000000  0.00000000
#> effect_50  -2.90490788 -2.90490788  0.00000000  0.00000000  0.00000000
#> effect_51  -0.90238904 -0.90238904  0.00000000  0.00000000  0.00000000
#> effect_52   1.33262814  1.33262814  0.00000000  0.00000000  0.00000000
#> effect_53   1.18089927  1.18089927  0.00000000  0.00000000  0.00000000
#> effect_54  -0.56799973 -0.56799973  0.00000000  0.00000000  0.00000000
#> effect_55  -1.72817818 -1.72817818  0.00000000  0.00000000  0.00000000
#> effect_56  -0.13570099 -0.13570099  0.00000000  0.00000000  0.00000000
#> effect_57  -0.08982261 -0.08982261  0.00000000  0.00000000  0.00000000
#> effect_58   0.53022623  0.53022623  0.00000000  0.00000000  0.00000000
#> effect_59   1.14510389  1.14510389  0.00000000  0.00000000  0.00000000
#> effect_60  -0.09785982 -0.09785982  0.00000000  0.00000000  0.00000000
#> effect_61   0.27257229  0.27257229  0.00000000  0.00000000  0.00000000
#> effect_62   0.24659655  0.24659655  0.00000000  0.00000000  0.00000000
#> effect_63   1.04838823  1.04838823  0.00000000  0.00000000  0.00000000
#> effect_64   0.76543084  0.76543084  0.00000000  0.00000000  0.00000000
#> effect_65  -0.35922833 -0.35922833  0.00000000  0.00000000  0.00000000
#> effect_66   1.89453579  1.89453579  0.00000000  0.00000000  0.00000000
#> effect_67   0.91746653  0.91746653  0.00000000  0.00000000  0.00000000
#> effect_68   0.15221023  0.15221023  0.00000000  0.00000000  0.00000000
#> effect_69   1.00698428  1.00698428  0.00000000  0.00000000  0.00000000
#> effect_70  -1.69693175 -1.69693175  0.00000000  0.00000000  0.00000000
#> effect_71   0.17538289  0.17538289  0.00000000  0.00000000  0.00000000
#> effect_72   1.52558977  1.52558977  0.00000000  0.00000000  0.00000000
#> effect_73   0.20491041  0.20491041  0.00000000  0.00000000  0.00000000
#> effect_74  -0.77088801 -0.77088801  0.00000000  0.00000000  0.00000000
#> effect_75  -1.12049302 -1.12049302  0.00000000  0.00000000  0.00000000
#> effect_76  -1.58471861 -1.58471861  0.00000000  0.00000000  0.00000000
#> effect_77   0.16334464  0.16334464  0.00000000  0.00000000  0.00000000
#> effect_78   0.86638712  0.86638712  0.00000000  0.00000000  0.00000000
#> effect_79  -0.72015580 -0.72015580  0.00000000  0.00000000  0.00000000
#> effect_80   0.17895399  0.17895399  0.00000000  0.00000000  0.00000000
#> effect_81  -1.34236577 -1.34236577  0.00000000  0.00000000  0.00000000
#> effect_82   1.69349322  1.69349322  0.00000000  0.00000000  0.00000000
#> effect_83  -0.51056868 -0.51056868  0.00000000  0.00000000  0.00000000
#> effect_84  -1.48074672 -1.48074672  0.00000000  0.00000000  0.00000000
#> effect_85   1.19886375  1.19886375  0.00000000  0.00000000  0.00000000
#> effect_86  -0.74858528 -0.74858528  0.00000000  0.00000000  0.00000000
#> effect_87   0.93747486  0.93747486  0.00000000  0.00000000  0.00000000
#> effect_88  -1.08985445 -1.08985445  0.00000000  0.00000000  0.00000000
#> effect_89  -1.37049341 -1.37049341  0.00000000  0.00000000  0.00000000
#> effect_90   0.34887848  0.34887848  0.00000000  0.00000000  0.00000000
#> effect_91   2.17079564  2.17079564  0.00000000  0.00000000  0.00000000
#> effect_92   0.32915418  0.32915418  0.00000000  0.00000000  0.00000000
#> effect_93   1.86460225  1.86460225  0.00000000  0.00000000  0.00000000
#> effect_94  -1.96433227 -1.96433227  0.00000000  0.00000000  0.00000000
#> effect_95   1.10312884  1.10312884  0.00000000  0.00000000  0.00000000
#> effect_96   0.07299956  0.07299956  0.00000000  0.00000000  0.00000000
#> effect_97   0.16232479  0.16232479  0.00000000  0.00000000  0.00000000
#> effect_98   0.35866558  0.35866558  0.00000000  0.00000000  0.00000000
#> effect_99  -1.46543700 -1.46543700  0.00000000  0.00000000  0.00000000
#> effect_100 -0.15943127 -0.15943127  0.00000000  0.00000000  0.00000000
#> effect_101  0.00000000  0.00000000  1.59727408  1.59727408  1.59727408
#> effect_102  0.00000000  0.00000000  0.92181695  0.92181695  0.92181695
#> effect_103  0.00000000  0.00000000  0.91867703  0.91867703  0.91867703
#> effect_104  0.00000000  0.00000000 -0.71948898 -0.71948898 -0.71948898
#> effect_105  0.00000000  0.00000000  0.32616898  0.32616898  0.32616898
#> effect_106  0.00000000  0.00000000 -0.63814625 -0.63814625 -0.63814625
#> effect_107  0.00000000  0.00000000 -0.68451793 -0.68451793 -0.68451793
#> effect_108  0.00000000  0.00000000 -0.58777402 -0.58777402 -0.58777402
#> effect_109  0.00000000  0.00000000  0.46213510  0.46213510  0.46213510
#> effect_110  0.00000000  0.00000000  0.07112818  0.07112818  0.07112818
#> effect_111  0.00000000  0.00000000 -1.08991613 -1.08991613 -1.08991613
#> effect_112  0.00000000  0.00000000  1.06792110  1.06792110  1.06792110
#> effect_113  0.00000000  0.00000000  0.90481074  0.90481074  0.90481074
#> effect_114  0.00000000  0.00000000 -0.97765781 -0.97765781 -0.97765781
#> effect_115  0.00000000  0.00000000  0.13411009  0.13411009  0.13411009
#> effect_116  0.00000000  0.00000000 -0.19389609 -0.19389609 -0.19389609
#> effect_117  0.00000000  0.00000000  1.05867804  1.05867804  1.05867804
#> effect_118  0.00000000  0.00000000  0.29453073  0.29453073  0.29453073
#> effect_119  0.00000000  0.00000000 -0.08642489 -0.08642489 -0.08642489
#> effect_120  0.00000000  0.00000000 -1.14528786 -1.14528786 -1.14528786
#> effect_121  0.00000000  0.00000000 -0.56856509 -0.56856509 -0.56856509
#> effect_122  0.00000000  0.00000000 -0.09353518 -0.09353518 -0.09353518
#> effect_123  0.00000000  0.00000000  0.10035557  0.10035557  0.10035557
#> effect_124  0.00000000  0.00000000  1.56534500  1.56534500  1.56534500
#> effect_125  0.00000000  0.00000000  0.85442343  0.85442343  0.85442343
#> effect_126  0.00000000  0.00000000 -1.56697090 -1.56697090 -1.56697090
#> effect_127  0.00000000  0.00000000 -0.34462894 -0.34462894 -0.34462894
#> effect_128  0.00000000  0.00000000 -1.27237680 -1.27237680 -1.27237680
#> effect_129  0.00000000  0.00000000  1.41133504  1.41133504  1.41133504
#> effect_130  0.00000000  0.00000000  1.83210971  1.83210971  1.83210971
#> effect_131  0.00000000  0.00000000 -1.00385401 -1.00385401 -1.00385401
#> effect_132  0.00000000  0.00000000  0.29265498  0.29265498  0.29265498
#> effect_133  0.00000000  0.00000000 -2.26688542 -2.26688542 -2.26688542
#> effect_134  0.00000000  0.00000000  1.23323069  1.23323069  1.23323069
#> effect_135  0.00000000  0.00000000 -1.71479088 -1.71479088 -1.71479088
#> effect_136  0.00000000  0.00000000 -1.26538790 -1.26538790 -1.26538790
#> effect_137  0.00000000  0.00000000 -1.36633626 -1.36633626 -1.36633626
#> effect_138  0.00000000  0.00000000  0.64406959  0.64406959  0.64406959
#> effect_139  0.00000000  0.00000000 -0.41141230 -0.41141230 -0.41141230
#> effect_140  0.00000000  0.00000000 -1.52521614 -1.52521614 -1.52521614
#> effect_141  0.00000000  0.00000000  0.90909037  0.90909037  0.90909037
#> effect_142  0.00000000  0.00000000  0.68653721  0.68653721  0.68653721
#> effect_143  0.00000000  0.00000000  0.05787257  0.05787257  0.05787257
#> effect_144  0.00000000  0.00000000  0.21048445  0.21048445  0.21048445
#> effect_145  0.00000000  0.00000000 -1.21061233 -1.21061233 -1.21061233
#> effect_146  0.00000000  0.00000000 -1.53851400 -1.53851400 -1.53851400
#> effect_147  0.00000000  0.00000000 -0.70373316 -0.70373316 -0.70373316
#> effect_148  0.00000000  0.00000000 -0.73418493 -0.73418493 -0.73418493
#> effect_149  0.00000000  0.00000000 -1.12808582 -1.12808582 -1.12808582
#> effect_150  0.00000000  0.00000000 -0.58195958 -0.58195958 -0.58195958
#> effect_151  0.00000000  0.00000000  2.26802476  2.26802476  2.26802476
#> effect_152  0.00000000  0.00000000 -0.36674271 -0.36674271 -0.36674271
#> effect_153  0.00000000  0.00000000  1.64707706  1.64707706  1.64707706
#> effect_154  0.00000000  0.00000000 -0.39541487 -0.39541487 -0.39541487
#> effect_155  0.00000000  0.00000000 -0.09244672 -0.09244672 -0.09244672
#> effect_156  0.00000000  0.00000000  0.47225617  0.47225617  0.47225617
#> effect_157  0.00000000  0.00000000  0.39093748  0.39093748  0.39093748
#> effect_158  0.00000000  0.00000000 -0.20857905 -0.20857905 -0.20857905
#> effect_159  0.00000000  0.00000000  0.72897762  0.72897762  0.72897762
#> effect_160  0.00000000  0.00000000 -1.92353314 -1.92353314 -1.92353314
#> effect_161  0.00000000  0.00000000  0.34093423  0.34093423  0.34093423
#> effect_162  0.00000000  0.00000000  0.77623513  0.77623513  0.77623513
#> effect_163  0.00000000  0.00000000 -0.14498074 -0.14498074 -0.14498074
#> effect_164  0.00000000  0.00000000 -0.27775349 -0.27775349 -0.27775349
#> effect_165  0.00000000  0.00000000 -0.07862984 -0.07862984 -0.07862984
#> effect_166  0.00000000  0.00000000 -1.07143620 -1.07143620 -1.07143620
#> effect_167  0.00000000  0.00000000 -0.39867896 -0.39867896 -0.39867896
#> effect_168  0.00000000  0.00000000  1.17391248  1.17391248  1.17391248
#> effect_169  0.00000000  0.00000000 -0.78329705 -0.78329705 -0.78329705
#> effect_170  0.00000000  0.00000000  1.09201318  1.09201318  1.09201318
#> effect_171  0.00000000  0.00000000 -0.16253710 -0.16253710 -0.16253710
#> effect_172  0.00000000  0.00000000  0.16128046  0.16128046  0.16128046
#> effect_173  0.00000000  0.00000000 -0.65695206 -0.65695206 -0.65695206
#> effect_174  0.00000000  0.00000000 -0.78377543 -0.78377543 -0.78377543
#> effect_175  0.00000000  0.00000000  1.21465793  1.21465793  1.21465793
#> effect_176  0.00000000  0.00000000 -0.23764758 -0.23764758 -0.23764758
#> effect_177  0.00000000  0.00000000 -0.30882324 -0.30882324 -0.30882324
#> effect_178  0.00000000  0.00000000  0.62196387  0.62196387  0.62196387
#> effect_179  0.00000000  0.00000000  0.76555504  0.76555504  0.76555504
#> effect_180  0.00000000  0.00000000  0.76539903  0.76539903  0.76539903
#> effect_181  0.00000000  0.00000000  0.82962526  0.82962526  0.82962526
#> effect_182  0.00000000  0.00000000 -0.19194039 -0.19194039 -0.19194039
#> effect_183  0.00000000  0.00000000 -0.55918919 -0.55918919 -0.55918919
#> effect_184  0.00000000  0.00000000 -1.11336101 -1.11336101 -1.11336101
#> effect_185  0.00000000  0.00000000  0.65074310  0.65074310  0.65074310
#> effect_186  0.00000000  0.00000000  0.50551542  0.50551542  0.50551542
#> effect_187  0.00000000  0.00000000  1.17103160  1.17103160  1.17103160
#> effect_188  0.00000000  0.00000000  0.18402876  0.18402876  0.18402876
#> effect_189  0.00000000  0.00000000 -0.96662517 -0.96662517 -0.96662517
#> effect_190  0.00000000  0.00000000 -0.30890299 -0.30890299 -0.30890299
#> effect_191  0.00000000  0.00000000 -1.29250380 -1.29250380 -1.29250380
#> effect_192  0.00000000  0.00000000  0.42698275  0.42698275  0.42698275
#> effect_193  0.00000000  0.00000000 -1.27842010 -1.27842010 -1.27842010
#> effect_194  0.00000000  0.00000000 -0.09458059 -0.09458059 -0.09458059
#> effect_195  0.00000000  0.00000000  1.14194421  1.14194421  1.14194421
#> effect_196  0.00000000  0.00000000  0.32364047  0.32364047  0.32364047
#> effect_197  0.00000000  0.00000000 -0.22914288 -0.22914288 -0.22914288
#> effect_198  0.00000000  0.00000000  0.94738768  0.94738768  0.94738768
#> effect_199  0.00000000  0.00000000  1.44934078  1.44934078  1.44934078
#> effect_200  0.00000000  0.00000000  0.43060410  0.43060410  0.43060410
#> 
#> $Bhat
#>             condition_1 condition_2  condition_3 condition_4  condition_5
#> effect_1     2.19911622  -0.6074422  -4.60321188  -0.4217871 -11.90054496
#> effect_2     0.01492194   2.6532015   2.77026616  -0.2577228   5.12038848
#> effect_3   -11.30577947  -1.6288799   2.94227746   1.0363703   7.00617240
#> effect_4     1.69412057   0.2929484  -3.72290622   4.4136389   0.98568697
#> effect_5     2.40783634   9.5719865  -6.94079219   8.7900778  -0.28943596
#> effect_6    -2.41356757  -0.8416951   4.30350651  -1.0643348  -1.06089091
#> effect_7     3.32887462   5.0271075  -5.26957447   1.7007926  -6.56856404
#> effect_8    -2.39443004  -0.6559669   1.89052684   7.0839888   4.86464872
#> effect_9    -3.96160902  -2.5696010  -0.39315978   3.4792849   2.29217447
#> effect_10   -2.54654531  -3.1663280  -5.04813346  -3.4773312   3.28843627
#> effect_11  -14.64716492  -3.7692011  -0.19254257  -0.9554658   0.97267439
#> effect_12    8.04582488  -1.7686104  -1.10941533   9.3729644   1.40362239
#> effect_13   -0.52621807  -3.4915231  -8.25706741  -4.8299741   5.66466976
#> effect_14    8.36628362   0.4709118  -1.02710179   4.9167075  -5.87312588
#> effect_15   -1.65367800  -1.1223171  -5.80356373  10.2703819   5.58608266
#> effect_16    3.09060810  -1.3151810  -2.77247043  -2.5529861   6.32329377
#> effect_17    4.59840036   6.4134062  -0.09316949  -1.6265390  -5.38621631
#> effect_18    0.71595742   2.6061484   3.27248357   3.3477955  -6.22189387
#> effect_19   10.34153340   1.7961558  -9.62653124   4.8494006   1.55041899
#> effect_20   -3.64744891   2.7425847   1.86828353  -9.9291641  10.16126973
#> effect_21   -3.38617932   7.7321983  -9.06196631  -5.0169757  -6.49856903
#> effect_22    0.62590095  -5.2366998  -0.17436298  -5.1704795   3.39055232
#> effect_23    1.41054218  -0.4785865  16.20878136  -1.1233740   8.50879384
#> effect_24    5.15944436   4.6129244   1.15375048   4.0404344   3.74429431
#> effect_25    5.42121883  -2.0950769   3.21208336  -1.6195884 -10.38825466
#> effect_26   -0.41428868  -5.5415806  -0.29577852   6.3572397  -7.99017932
#> effect_27    0.75681968   4.3921408  -2.26108396  -3.7997212   1.93239979
#> effect_28    0.96120983  11.0833934  10.61677413  -4.9105608   0.40477694
#> effect_29   -4.01640517  -4.5983294  -2.86218799   1.3975503  -4.67406915
#> effect_30    3.43530821  -7.2179644  -0.30999142   4.0883726   3.15768718
#> effect_31   -0.14187927 -13.9224317   9.89414730 -14.9363292 -11.58748299
#> effect_32   -8.75359036  -7.2229326   7.83649420  -4.0542108   0.19408067
#> effect_33   -5.54269175   0.9208087   0.26892541   8.2037895  -7.01602240
#> effect_34   -1.42190057   0.9235930  -3.43898469   5.1016340  -2.03118488
#> effect_35   -2.13038061   2.3838432  -5.60095826   3.3319859 -12.21995257
#> effect_36   -3.82657104   2.2688488   9.55701024   7.5342653   3.03662703
#> effect_37   -8.88667376   9.5156886   1.57513249   7.0044127  -4.21977626
#> effect_38   -0.45169287  -0.2689205  -0.52620965   1.0704771   1.58149629
#> effect_39   -4.78652994  11.0562358  16.23578816   4.3497633  -2.88861040
#> effect_40   -5.93697907   2.3544457  -4.41794772   6.0958562   1.43418426
#> effect_41   -6.85339077  -4.3829384   0.40674476  -3.0755580  -3.79072966
#> effect_42   -4.80584168  -2.5334043   1.38634852  -7.3043821  -5.18365776
#> effect_43   -6.71727978   6.1035591   2.97849855  -1.4043800  -4.35473619
#> effect_44   -1.41365776  -3.2283589   4.05296123   8.6288620  11.80448632
#> effect_45    3.26367325  -7.1738806  -7.92986724   5.2224367   5.74172290
#> effect_46    8.11486650   1.2807730  -4.87369871   1.3409220  -0.68657355
#> effect_47   -6.44899757   8.8534684  -3.91084486  -7.6758803  -3.93641857
#> effect_48   -1.34671839  -0.3604926   6.01184782   4.7255390  -0.09742855
#> effect_49    7.56458095  -1.0750653   9.19634033  -3.0722221   2.75305253
#> effect_50   -3.75043992  -2.8509183   2.80214397   0.4001195   1.45801199
#> effect_51    2.58348145   0.4751204   2.02626875   3.5093886   6.83358881
#> effect_52    2.84314373  -7.3891495   5.82268105   0.7978187   3.23114581
#> effect_53    3.82640182   1.6113511   8.87646902   6.4990221 -12.96216409
#> effect_54    1.89311997  -7.4559861  -1.53961115   7.7732431  -2.47049750
#> effect_55   -7.95137834   2.0346396   2.39680118  -6.5448155   2.86514107
#> effect_56    0.89614819  -4.0225037   5.35159991  -4.7895666  -4.47391624
#> effect_57   -3.07191992  -1.1515988   1.50738752  -3.7543901   2.27159081
#> effect_58   -6.02201295  -3.1656778   1.66337894  -3.2909070 -11.27376535
#> effect_59   -4.35130754  -5.0234495   3.95031452  10.3832050  -3.46198615
#> effect_60    3.24313621   1.4407112   2.02610321  -5.3709841   5.39812880
#> effect_61    5.06441577   3.1734727  -1.78090772   8.7222916  -7.51616921
#> effect_62   -4.38471391   0.8061632  -5.14705317 -10.3724055  -7.75013075
#> effect_63   -6.31659831  -2.5825834  -6.61667289  -5.4836231  -5.11873803
#> effect_64    1.00744514  -0.4661701   1.96055448   2.1747150   4.59760313
#> effect_65   -1.63829818  11.4562443  -1.94716854   2.7860036   8.37402394
#> effect_66    0.23542193   6.2619468   2.78987678   3.2474886  -0.71555428
#> effect_67    6.14008139   4.4785768   5.98299240   4.3607704  -1.87089761
#> effect_68    1.97441648   1.7375217  -6.65247346  -0.1322437 -10.45072292
#> effect_69    4.26765695   3.8194088  -0.45223474   6.2240130  -3.89884282
#> effect_70   -4.23047018   0.7198935   3.31426292  10.9275368   8.69118156
#> effect_71    2.28421054   4.8875110   0.12274729   7.6563524   2.99424891
#> effect_72    2.12013661  -5.4773227   7.69397038   4.8602645   4.21661016
#> effect_73   -2.28627917  -9.3088247   4.41868273   7.1943228  -4.43304258
#> effect_74   -7.03180903  -1.7779965  -0.28665689  13.7827544   1.41369665
#> effect_75    3.44891814 -11.1999176  -3.71073301  -0.2774951  -4.19354627
#> effect_76    1.57909983   7.3542147   7.21390578  -1.7096163   1.31502943
#> effect_77    2.39880170   1.8757777  -0.39302011  -0.6771877  -1.01500178
#> effect_78    3.29666365   4.9010357  -5.16587666  -9.7826155  -0.25065355
#> effect_79   -3.93446056  -5.9239911  -4.03721819   1.7930449  -3.80937888
#> effect_80    0.93352345   4.6146026  -3.54417394   2.8897648  -7.06745351
#> effect_81    7.03867854   0.5769305   4.62583697   2.1298994  -0.79845821
#> effect_82    5.12576903   3.3975837  -6.88774706   0.1866597  -6.88298722
#> effect_83   10.34672460  -4.1638790  -9.57855634  -4.1038802   4.02695575
#> effect_84    4.33377082  -9.6394120  -4.91341053   0.4651676  -1.54901834
#> effect_85    0.68830667   5.0945984   6.23274778  -2.2658476   1.23015978
#> effect_86    1.20168610  -2.0744315   6.70514699  -0.7077840   2.43507956
#> effect_87   -5.17005580   6.1534030   3.33095944  -7.6509664  -9.05499335
#> effect_88    6.62850298 -11.6650184   5.96133063   2.4092473   1.97442369
#> effect_89    4.67209383  -2.3682814  -4.88376335   6.5117383   3.15502379
#> effect_90    4.33754437   1.5051833  -1.46224304   2.6970346   3.65779700
#> effect_91    4.86101168   4.9650209  -4.22312120  -0.2570431   5.63486225
#> effect_92    2.82520607  -4.3894081  -7.34946608   3.6268781  -3.38236813
#> effect_93    2.79484116   3.3959363  -5.93214702  -0.7930916  -4.18116229
#> effect_94   -4.77884334   3.0232354   0.30682486   2.2851419  -7.94776681
#> effect_95   -5.69573837   0.7180635   0.63354440  -2.4838207   3.21606235
#> effect_96   -2.72343251  -4.2755732  -4.25382596  -0.8675804   4.97533415
#> effect_97    6.85120404  -0.8979633   3.27993762   2.1257320  -6.89199685
#> effect_98    1.19824757   0.3110836   0.74910285  -5.3293603  -3.32817093
#> effect_99    3.69711599  -4.8193998  -0.35698451   1.3401949  -7.89185340
#> effect_100  -5.19207023   7.1192474  -3.01866866   3.0329878  -2.19654859
#> effect_101   1.36455352   1.3303352   4.30255337  10.0288237  -1.66342777
#> effect_102  -5.13429679  -2.5691631  -0.54874151   2.0389922   0.23182927
#> effect_103  -5.71208116   0.4806900   1.84818792   6.8591175   1.17089057
#> effect_104  -1.88638702   2.6414117  -0.59688307   4.0282436   7.14016103
#> effect_105  -4.26770050  -8.6352946  -7.79112698  -6.3090511   0.81984220
#> effect_106  -4.02457187  -0.6342054  -2.25744790  -2.0713075  -2.03143255
#> effect_107   7.74773904  -5.1706197   3.49208250  -7.1695744  -4.24715803
#> effect_108  -3.28276382   6.8736655  -4.96188628  13.1186815   1.00715781
#> effect_109  -5.24708778   0.8375907   2.89421314  -2.1057490  -0.23387262
#> effect_110  -5.96962856  -3.1240411  -0.12387380   0.1067075   7.96411506
#> effect_111   2.12639399  -0.2838810  -6.33428166  -5.2368941  -5.69239853
#> effect_112  -4.18270617   6.6489178  -5.08267934  -1.1788693   2.54716999
#> effect_113   2.46642424  -1.2732573   6.39976899   1.7170889   2.58833344
#> effect_114  -5.39702909   4.0240029  -1.32894225   3.9483140  -5.57311579
#> effect_115   0.05732980  -2.1421596  -1.06510074  -4.6348678   8.02336838
#> effect_116   0.02052226   9.0124609   8.86309068   7.6820312   2.57849550
#> effect_117   2.52809346  -0.3224842  -2.36150255   1.6818943  -0.09885763
#> effect_118   4.66415534  -8.7625559  -3.44482269   4.3134436   1.54281584
#> effect_119   3.26011145   8.3560217   3.12417494   2.8546639  -3.14735157
#> effect_120  -0.34158904  -3.0189374  -3.35112783   7.0503724   8.99007768
#> effect_121   6.77727507  -2.2843697   0.11700024  -4.1290260  -0.07704292
#> effect_122   4.42992638  -6.5773210  12.74142912   4.0249043   0.64406317
#> effect_123  -3.51399732  -7.2211854  -0.52338994   6.0567125   1.87953905
#> effect_124   1.89883940   4.5453662   5.79572219   0.5209062   3.05722032
#> effect_125   0.87966864  -0.3856871  -3.14748340   1.2133340  -2.95521685
#> effect_126  -2.26289172   7.3618928  -0.42905298 -10.9561974  -7.90297040
#> effect_127   2.49759218  -5.5235338  -5.60234703  -5.1604273   3.05528005
#> effect_128   0.47226832  -4.2605420  -5.84428148  -1.9414575   5.05160349
#> effect_129   5.64885160   7.3576386   2.90154428  -2.4685599  -2.85497863
#> effect_130  -4.40843529  -4.7521743   0.31627120  -4.2860857  -1.58704758
#> effect_131   0.55782842  -4.9629836   1.53347052  -1.3420802   4.01198874
#> effect_132   0.29044057   5.8384990   5.67516929  -2.4874334   1.38865717
#> effect_133  -4.86092749  -0.5616688  -5.89427915  -3.5903144   9.71987540
#> effect_134  -1.10894290   5.7340560   0.47112541  -1.4172504  -4.80312931
#> effect_135  -6.08268127  -4.0069171  -2.61502901   4.5157895  -3.10540882
#> effect_136   7.97804710   0.4263386  -3.48284672  -5.3826536  -3.63230356
#> effect_137  -6.92542889  -4.9939016  -2.71413883  -5.7379390  -0.16046560
#> effect_138   6.91519041  -5.0473894   2.90909738   7.9909952  -2.40564869
#> effect_139  -6.57863687 -10.4511873   0.22386473  -2.7796408  -6.70759625
#> effect_140  -3.76297127  12.5578563   1.61847020  -3.1955185  -4.53762625
#> effect_141  -0.13486954  -6.3724496  -1.41696209   1.8537805   7.06849350
#> effect_142   4.32735619   0.9652820   4.40150940  -1.3957764   0.69292972
#> effect_143  -5.06735313   0.8696176  -0.44877014  -1.5982155 -12.62824116
#> effect_144  -7.43762552   7.3079673  -6.75465131   2.9880703   4.51995593
#> effect_145  -0.16341543  -1.9700253   2.62624813  -0.2468940   1.49166276
#> effect_146  -2.33095798   5.5900828  -7.14911457  -5.2234029  -1.41177945
#> effect_147  14.40998290   3.1580946  -4.51030784   3.3119290  13.96630191
#> effect_148   6.56828523   2.8679009 -10.43987187   2.2340885  -2.56430137
#> effect_149   4.91187874  -4.9203745   6.20135422  -5.3976786   1.95713585
#> effect_150  -0.44289897  -3.3368063  -5.54114435  -8.4334095   2.34912827
#> effect_151  -2.29007243  -8.3127803  -8.63651120   0.2817953   1.04410067
#> effect_152   9.99325576   2.2827666  -4.16791638  -0.4449641  -7.14636063
#> effect_153  -5.74253421  -8.1409843   2.62951148   8.3922044  11.35935985
#> effect_154   2.79335012  -3.2776736  -2.67953618  -1.6188472  -0.93501953
#> effect_155   4.04594674  -7.5085299  -2.02199450 -11.7711776  -9.11379996
#> effect_156   3.85373230   3.1355480  -8.00479148   7.4676621   0.94992882
#> effect_157  -7.20074884  -5.0606143   2.39746967   9.4700087   0.27608782
#> effect_158  -0.77175997   5.9456974  -2.33240124   4.5574271  -6.41988137
#> effect_159 -13.25434427   1.7571224  -0.67055149   6.8220598   0.23805424
#> effect_160  -5.32467418 -10.3350884   4.41796706   0.3634535  -9.57597862
#> effect_161  -9.86048924  -1.0565638   8.40327024  -2.1953169  -0.78939497
#> effect_162   0.67134528  -7.8042770   8.01126873  -5.8545416   5.63621488
#> effect_163   1.81320967  -4.0852262   0.26423686   3.0627207  -2.23092746
#> effect_164   1.62667928  -1.8602050   0.20345281  -2.0957480  -4.56644079
#> effect_165  -3.99148687   2.2162707   2.48797339   7.7003605   0.75913688
#> effect_166 -10.43597662   7.7095438  -7.63348331   0.6033868  -6.68264366
#> effect_167   2.99692426   0.7034287   3.98229070  -3.7166305  -1.47922895
#> effect_168  -6.62584931  -4.8420132  13.93154920   7.2290812  -2.06064274
#> effect_169   5.16446252   0.8416446   3.86108548  -0.9280881   6.19953695
#> effect_170  -3.09224768  -0.5641838   2.22482613  -5.4059954  -2.62172257
#> effect_171  -9.88301469  -3.7663308   6.48895345   1.3192049  -9.71421388
#> effect_172   6.48139359  -5.0013006   6.18405804  -4.2802794  -9.09405357
#> effect_173  -6.53552935  -1.3857568   2.20785687  -4.6925142  -7.32668173
#> effect_174   5.30349302  -9.5091772  -0.29931920  -0.6063651   7.13736114
#> effect_175   8.23476933   2.9028756   3.13356497  -8.5173625   4.08068847
#> effect_176  -7.49955850   1.9717753   5.93693142  -9.1876859   6.73652727
#> effect_177   0.87814014   1.6021147   3.60168786  -9.4149963   0.28864613
#> effect_178   5.70980455  -0.3160348   2.82256782   0.2749121   6.93752864
#> effect_179  -7.37507463   0.8981226   0.95600329  -1.6425152  -0.18456508
#> effect_180  12.17699791  -0.6933632   3.24769656   1.4725823   5.12479968
#> effect_181  -0.02043668  -2.3975649   1.04322408 -13.0713972  -2.02106300
#> effect_182  -4.05978366   4.7450706   2.12535323  -6.5524020  -2.57378195
#> effect_183  -6.47270235   1.8362230  -4.04003402   0.2857247   0.88740735
#> effect_184   3.95604047   5.4606707  -0.81749529  -6.0951960  -5.14440661
#> effect_185  -0.17767038  -1.7724641  -0.42764411   8.2585298   2.61827113
#> effect_186   1.50402054   4.1846013  -6.11794563   3.7892418  -3.95094054
#> effect_187  10.14071122   1.3903701   2.06549526   4.3614222  -7.40265450
#> effect_188   1.71257285   4.4321224  -1.71839159  -0.8725116  -3.36586416
#> effect_189   1.41072357   0.8332296  -8.10340912   6.0888650  -0.10032659
#> effect_190   2.46398714  -4.0955520   3.88222749  -6.2617716  -0.09552610
#> effect_191   4.13555009   7.9640695  -0.50908720   5.5649591   3.97365149
#> effect_192   7.52035420 -12.1854303  -3.44447371   1.8562328   3.58663006
#> effect_193   4.41786492   3.9931723   0.20243496   1.1945271  -0.61156441
#> effect_194   6.41621409  -6.9274033   4.82365215  -9.1791043  -0.94860161
#> effect_195   0.19079545   3.6357152   0.53051464   4.7834580  -3.82766443
#> effect_196  -2.90899593  -1.5304937  10.40229517  -6.2704697  -3.52017501
#> effect_197   2.00848209   3.8405280  -6.06062594  -1.9434889   3.29046222
#> effect_198   0.42198675  -1.4093751   5.69524234  -0.6609479  -3.79294747
#> effect_199  -6.51485152   0.7497252   6.18252383   0.7756140   6.93588206
#> effect_200   2.02050815   1.0473737   4.39745114   5.1898900  -1.43163666
#> 
#> $Shat
#>            condition_1 condition_2 condition_3 condition_4 condition_5
#> effect_1             5           5           5           5           5
#> effect_2             5           5           5           5           5
#> effect_3             5           5           5           5           5
#> effect_4             5           5           5           5           5
#> effect_5             5           5           5           5           5
#> effect_6             5           5           5           5           5
#> effect_7             5           5           5           5           5
#> effect_8             5           5           5           5           5
#> effect_9             5           5           5           5           5
#> effect_10            5           5           5           5           5
#> effect_11            5           5           5           5           5
#> effect_12            5           5           5           5           5
#> effect_13            5           5           5           5           5
#> effect_14            5           5           5           5           5
#> effect_15            5           5           5           5           5
#> effect_16            5           5           5           5           5
#> effect_17            5           5           5           5           5
#> effect_18            5           5           5           5           5
#> effect_19            5           5           5           5           5
#> effect_20            5           5           5           5           5
#> effect_21            5           5           5           5           5
#> effect_22            5           5           5           5           5
#> effect_23            5           5           5           5           5
#> effect_24            5           5           5           5           5
#> effect_25            5           5           5           5           5
#> effect_26            5           5           5           5           5
#> effect_27            5           5           5           5           5
#> effect_28            5           5           5           5           5
#> effect_29            5           5           5           5           5
#> effect_30            5           5           5           5           5
#> effect_31            5           5           5           5           5
#> effect_32            5           5           5           5           5
#> effect_33            5           5           5           5           5
#> effect_34            5           5           5           5           5
#> effect_35            5           5           5           5           5
#> effect_36            5           5           5           5           5
#> effect_37            5           5           5           5           5
#> effect_38            5           5           5           5           5
#> effect_39            5           5           5           5           5
#> effect_40            5           5           5           5           5
#> effect_41            5           5           5           5           5
#> effect_42            5           5           5           5           5
#> effect_43            5           5           5           5           5
#> effect_44            5           5           5           5           5
#> effect_45            5           5           5           5           5
#> effect_46            5           5           5           5           5
#> effect_47            5           5           5           5           5
#> effect_48            5           5           5           5           5
#> effect_49            5           5           5           5           5
#> effect_50            5           5           5           5           5
#> effect_51            5           5           5           5           5
#> effect_52            5           5           5           5           5
#> effect_53            5           5           5           5           5
#> effect_54            5           5           5           5           5
#> effect_55            5           5           5           5           5
#> effect_56            5           5           5           5           5
#> effect_57            5           5           5           5           5
#> effect_58            5           5           5           5           5
#> effect_59            5           5           5           5           5
#> effect_60            5           5           5           5           5
#> effect_61            5           5           5           5           5
#> effect_62            5           5           5           5           5
#> effect_63            5           5           5           5           5
#> effect_64            5           5           5           5           5
#> effect_65            5           5           5           5           5
#> effect_66            5           5           5           5           5
#> effect_67            5           5           5           5           5
#> effect_68            5           5           5           5           5
#> effect_69            5           5           5           5           5
#> effect_70            5           5           5           5           5
#> effect_71            5           5           5           5           5
#> effect_72            5           5           5           5           5
#> effect_73            5           5           5           5           5
#> effect_74            5           5           5           5           5
#> effect_75            5           5           5           5           5
#> effect_76            5           5           5           5           5
#> effect_77            5           5           5           5           5
#> effect_78            5           5           5           5           5
#> effect_79            5           5           5           5           5
#> effect_80            5           5           5           5           5
#> effect_81            5           5           5           5           5
#> effect_82            5           5           5           5           5
#> effect_83            5           5           5           5           5
#> effect_84            5           5           5           5           5
#> effect_85            5           5           5           5           5
#> effect_86            5           5           5           5           5
#> effect_87            5           5           5           5           5
#> effect_88            5           5           5           5           5
#> effect_89            5           5           5           5           5
#> effect_90            5           5           5           5           5
#> effect_91            5           5           5           5           5
#> effect_92            5           5           5           5           5
#> effect_93            5           5           5           5           5
#> effect_94            5           5           5           5           5
#> effect_95            5           5           5           5           5
#> effect_96            5           5           5           5           5
#> effect_97            5           5           5           5           5
#> effect_98            5           5           5           5           5
#> effect_99            5           5           5           5           5
#> effect_100           5           5           5           5           5
#> effect_101           5           5           5           5           5
#> effect_102           5           5           5           5           5
#> effect_103           5           5           5           5           5
#> effect_104           5           5           5           5           5
#> effect_105           5           5           5           5           5
#> effect_106           5           5           5           5           5
#> effect_107           5           5           5           5           5
#> effect_108           5           5           5           5           5
#> effect_109           5           5           5           5           5
#> effect_110           5           5           5           5           5
#> effect_111           5           5           5           5           5
#> effect_112           5           5           5           5           5
#> effect_113           5           5           5           5           5
#> effect_114           5           5           5           5           5
#> effect_115           5           5           5           5           5
#> effect_116           5           5           5           5           5
#> effect_117           5           5           5           5           5
#> effect_118           5           5           5           5           5
#> effect_119           5           5           5           5           5
#> effect_120           5           5           5           5           5
#> effect_121           5           5           5           5           5
#> effect_122           5           5           5           5           5
#> effect_123           5           5           5           5           5
#> effect_124           5           5           5           5           5
#> effect_125           5           5           5           5           5
#> effect_126           5           5           5           5           5
#> effect_127           5           5           5           5           5
#> effect_128           5           5           5           5           5
#> effect_129           5           5           5           5           5
#> effect_130           5           5           5           5           5
#> effect_131           5           5           5           5           5
#> effect_132           5           5           5           5           5
#> effect_133           5           5           5           5           5
#> effect_134           5           5           5           5           5
#> effect_135           5           5           5           5           5
#> effect_136           5           5           5           5           5
#> effect_137           5           5           5           5           5
#> effect_138           5           5           5           5           5
#> effect_139           5           5           5           5           5
#> effect_140           5           5           5           5           5
#> effect_141           5           5           5           5           5
#> effect_142           5           5           5           5           5
#> effect_143           5           5           5           5           5
#> effect_144           5           5           5           5           5
#> effect_145           5           5           5           5           5
#> effect_146           5           5           5           5           5
#> effect_147           5           5           5           5           5
#> effect_148           5           5           5           5           5
#> effect_149           5           5           5           5           5
#> effect_150           5           5           5           5           5
#> effect_151           5           5           5           5           5
#> effect_152           5           5           5           5           5
#> effect_153           5           5           5           5           5
#> effect_154           5           5           5           5           5
#> effect_155           5           5           5           5           5
#> effect_156           5           5           5           5           5
#> effect_157           5           5           5           5           5
#> effect_158           5           5           5           5           5
#> effect_159           5           5           5           5           5
#> effect_160           5           5           5           5           5
#> effect_161           5           5           5           5           5
#> effect_162           5           5           5           5           5
#> effect_163           5           5           5           5           5
#> effect_164           5           5           5           5           5
#> effect_165           5           5           5           5           5
#> effect_166           5           5           5           5           5
#> effect_167           5           5           5           5           5
#> effect_168           5           5           5           5           5
#> effect_169           5           5           5           5           5
#> effect_170           5           5           5           5           5
#> effect_171           5           5           5           5           5
#> effect_172           5           5           5           5           5
#> effect_173           5           5           5           5           5
#> effect_174           5           5           5           5           5
#> effect_175           5           5           5           5           5
#> effect_176           5           5           5           5           5
#> effect_177           5           5           5           5           5
#> effect_178           5           5           5           5           5
#> effect_179           5           5           5           5           5
#> effect_180           5           5           5           5           5
#> effect_181           5           5           5           5           5
#> effect_182           5           5           5           5           5
#> effect_183           5           5           5           5           5
#> effect_184           5           5           5           5           5
#> effect_185           5           5           5           5           5
#> effect_186           5           5           5           5           5
#> effect_187           5           5           5           5           5
#> effect_188           5           5           5           5           5
#> effect_189           5           5           5           5           5
#> effect_190           5           5           5           5           5
#> effect_191           5           5           5           5           5
#> effect_192           5           5           5           5           5
#> effect_193           5           5           5           5           5
#> effect_194           5           5           5           5           5
#> effect_195           5           5           5           5           5
#> effect_196           5           5           5           5           5
#> effect_197           5           5           5           5           5
#> effect_198           5           5           5           5           5
#> effect_199           5           5           5           5           5
#> effect_200           5           5           5           5           5
#>