Estimate null correlations (simple)

Estimates a null correlation matrix from data using simple z score threshold

estimate_null_correlation_simple(data, z_thresh = 2, est_cor = TRUE)

Arguments

data
a mash data object, eg as created by mash_set_data

z_thresh
the z score threshold below which to call an effect null

est_cor
whether to estimate correlation matrix (TRUE) or the covariance matrix (FALSE).

Details

Returns a simple estimate of the correlation matrix (or covariance matrix) among conditions under the null. Specifically, the simple estimate is the empirical correlation (or covariance) matrix of the z scores for those effects that have (absolute) z score < z_thresh in all conditions.

Examples

simdata = simple_sims(50,5,1)
data = mash_set_data(simdata$Bhat, simdata$Shat)
estimate_null_correlation_simple(data)

#>             condition_1 condition_2 condition_3 condition_4 condition_5
#> condition_1   1.0000000 -0.06902590  0.03823190  0.12797704  0.16615204
#> condition_2  -0.0690259  1.00000000  0.06848560  0.29195746  0.05490694
#> condition_3   0.0382319  0.06848560  1.00000000 -0.04648041 -0.02602396
#> condition_4   0.1279770  0.29195746 -0.04648041  1.00000000 -0.07451916
#> condition_5   0.1661520  0.05490694 -0.02602396 -0.07451916  1.00000000