Return the estimated mixture proportions

Return the estimated mixture proportions

get_estimated_pi(m, dimension = c("cov", "grid", "all"))

Arguments

m
the mash result

dimension
indicates whether you want the mixture proportions for the covariances, grid, or all

Value

a named vector containing the estimated mixture proportions.

Details

If the fit was done with `usepointmass=TRUE` then the first element of the returned vector will correspond to the null, and the remaining elements to the non-null covariance matrices. Suppose the fit was done with $K$ covariances and a grid of length $L$. If `dimension=cov` then the returned vector will be of length $K$ (or $K+1$ if `usepointmass=TRUE`). If `dimension=grid` then the returned vector will be of length $L$ (or $L+1$). If `dimension=all` then the returned vector will be of length $LK$ (or $LK+1$). The names of the vector will be informative for which combination each element corresponds to.

Examples

simdata = simple_sims(50,5,1)
data = mash_set_data(simdata$Bhat, simdata$Shat)
m = mash(data, cov_canonical(data))

#>  - Computing 200 x 141 likelihood matrix.
#>  - Likelihood calculations took 0.01 seconds.
#>  - Fitting model with 141 mixture components.
#>  - Model fitting took 0.22 seconds.
#>  - Computing posterior matrices.
#>  - Computation allocated took 0.00 seconds.

get_estimated_pi(m)

#>          null      identity   condition_1   condition_2   condition_3 
#>    0.55632080    0.10786812    0.05117083    0.00000000    0.01931244 
#>   condition_4   condition_5 equal_effects  simple_het_1  simple_het_2 
#>    0.00000000    0.00000000    0.00000000    0.00000000    0.00000000 
#>  simple_het_3 
#>    0.26532780