Perform condition-by-condition analyses

Performs simple "condition-by-condition" analysis by running ash from package ashr on data from each condition, one at a time. May be a useful first step to identify top hits in each condition before a mash analysis.

mash_1by1(data, alpha = 0, ...)

Arguments

data
A list with the following two elements: Bhat an n by R matrix of observations (n units in R conditions); and Shat, an n by R matrix of standard errors (n units in R conditions),

alpha
Numeric value of alpha parameter in the model. alpha = 0 for Exchangeable Effects (EE), alpha = 1 for Exchangeable Z-scores (EZ).

...
optionally, other parameters to be passed to ash

Value

A list similar to the output of mash, particularly including posterior matrices.

Examples

simdata = simple_sims(50,5,1)
mash_1by1(simdata)

#> $result
#> $result$PosteriorMean
#>              condition_1  condition_2   condition_3   condition_4  condition_5
#> effect_1    1.074451e-01  0.064344973  0.2912904287 -0.3670974651  0.036246799
#> effect_2    1.052052e-02  0.001408967  0.0128595525 -0.3721018468 -0.654203482
#> effect_3   -3.348870e-02  0.008395729 -0.0712892864 -0.0970400129  0.676185031
#> effect_4   -2.460727e-01  0.181654196  0.0953283252  0.1622659022  0.221073655
#> effect_5   -1.001677e-01 -0.393280910 -0.0075183499  0.0323717564 -0.385369670
#> effect_6    1.827590e-02  0.041323057 -0.0342623453 -0.0225557906 -0.120114371
#> effect_7    6.513069e-02 -0.022289499 -0.0003366789 -0.0200523967 -0.545731138
#> effect_8   -1.229361e-01 -0.239743822 -0.1082082410 -0.0335999833  0.278729433
#> effect_9   -3.882621e-03 -0.326883760  0.0183072670 -0.0032575256  0.030263068
#> effect_10  -1.327041e-01 -0.369134080 -0.0288665055  0.0046555141  0.081140157
#> effect_11   9.562457e-02 -0.044477076 -0.0545264505 -0.0016889800 -0.006260572
#> effect_12   9.612292e-03  0.009626487 -0.0002622963 -0.0606926182 -0.004401513
#> effect_13  -5.998638e-02 -0.042382317  0.0135596268 -0.1148523316 -0.210400695
#> effect_14   5.429008e-02 -0.099711169  0.0709496939 -0.1296986015 -1.144704599
#> effect_15  -2.954176e-02 -0.092680006  0.0147066188  0.0192906846  0.121771484
#> effect_16  -1.069917e-02 -0.139843835 -0.0123364957 -0.0331263210  0.048374465
#> effect_17  -5.096817e-02  0.267459976  0.0286868544 -0.0009306705 -0.433224827
#> effect_18   1.790109e-01  0.010104670  0.0433475446  0.1945222886  0.015620349
#> effect_19   1.940649e-01  0.963759818  0.0264648732  0.0155438376  0.014331244
#> effect_20   8.866940e-02 -0.008032529  0.0051808459 -0.1027045379  0.309910887
#> effect_21  -4.585732e-02  0.143466611  0.0138472216  0.2518268956  0.348830424
#> effect_22  -9.118607e-02 -0.108388935 -0.0201067278  0.0610845728  0.049069306
#> effect_23  -7.502539e-02 -0.104761183 -0.0550666858 -0.0755866766 -0.448773586
#> effect_24   2.511196e-01 -0.091760441  0.0061022664 -0.0399707268 -0.109362427
#> effect_25   1.815013e-02  0.276592761 -0.0377457712 -0.0067719297 -0.108313201
#> effect_26   6.412708e-02 -0.080249573  0.0332465478  0.0995387663  0.043922852
#> effect_27   3.753515e-02  0.062075201  0.0181951133 -0.1845591798 -0.138374411
#> effect_28   4.928781e-02 -0.040024362  0.0287914219  0.3332429970 -0.038111279
#> effect_29   4.034161e-03  0.097981758  0.0280840322 -0.0680484616 -0.092177195
#> effect_30  -2.284360e-02 -0.042585800 -0.0285710916 -0.0243982958 -0.202973334
#> effect_31  -4.972072e-02  0.089568292  0.0134213357  0.0771015913 -0.012976870
#> effect_32   1.109089e-01  0.098581506 -0.0158960641 -0.0703983716 -0.411260915
#> effect_33   2.605023e-01  0.035982230 -0.0163662549 -0.0022211512 -0.110081296
#> effect_34  -4.164708e-02  0.138547011  0.1057514631 -0.0624427874  0.212617991
#> effect_35  -4.171682e-02 -0.018374701  0.0318870469  0.0233179411  0.012056263
#> effect_36   1.740436e-05 -0.009804006 -0.0114721835  0.1211451700 -0.004044850
#> effect_37  -1.145433e-02 -0.077626192 -0.0074446384 -0.0040436863  0.008340432
#> effect_38  -4.589154e-02  0.172004182  0.0162232596 -0.0026083087 -0.128996931
#> effect_39   1.179394e-03  0.379987368  0.0241841226  0.0206123418  0.040345505
#> effect_40   1.999476e-01 -0.129685485 -0.0173878082  0.0535079544 -0.040332076
#> effect_41   2.275963e-02  0.130255328 -0.0367455239  0.0089714237  0.028046602
#> effect_42   4.986299e-02  0.370022969 -0.0023162566  0.0455353019 -0.150534982
#> effect_43  -2.379896e-01 -0.029067370  0.0140996334 -0.0538577959 -0.018692388
#> effect_44   8.982302e-02 -0.009259332 -0.0665559599 -0.0037222981  0.170241056
#> effect_45  -4.277246e-02  0.226113664  0.0432554427  0.0576188018  0.072235402
#> effect_46   5.514213e-04 -0.226021053  0.0019502522  0.0662278007  0.110054118
#> effect_47   4.010337e-03  0.019498041 -0.0233063859  0.0099180614 -0.227000722
#> effect_48  -1.327246e-01 -0.007250305 -0.0688668511  0.0255942177  0.053963636
#> effect_49  -1.205319e-01  0.080174690  0.0222057204  0.0476904792  0.153132325
#> effect_50  -4.673818e-02 -0.026185957  1.7489340474  0.0428690994 -0.114561005
#> effect_51  -3.443669e-01 -0.179650001  0.2088282473 -0.0402171270 -0.143012113
#> effect_52   3.315897e-01 -0.324665877 -0.0990594667  0.4148290170  0.041333869
#> effect_53   4.368069e-02  0.117689751  0.0830107063 -0.0304825707 -0.036430303
#> effect_54  -8.924080e-03 -0.047115790  0.1463520464 -0.1049542867 -0.026531972
#> effect_55   2.283289e-01  0.061419487 -0.5440472206  0.0562799433  1.229748643
#> effect_56  -9.845056e-01  0.002300825 -0.2043885061 -0.0125344208  0.208232067
#> effect_57  -1.817490e-04 -1.247040126  1.6550946014  0.0296826590  1.674841891
#> effect_58  -2.133284e-02 -0.098649773  0.0953153850 -0.0130844315  0.106241143
#> effect_59  -1.014561e-01 -0.078269379  0.0098559527  0.3834236792  0.210135396
#> effect_60  -3.772809e-02 -0.085670612  0.0089908503  0.0721512809 -0.309474987
#> effect_61  -1.183043e+00 -0.008314774  0.0694980132  0.0135520828  0.321046529
#> effect_62  -1.098961e-01 -0.304522790  0.0234502550 -2.5140993369 -0.059160986
#> effect_63   1.354080e-01  0.198802211 -0.0134158650  0.7206351040  0.198862979
#> effect_64   3.579307e-02 -0.057948200  0.1221661028 -0.1320934276  0.030359020
#> effect_65   1.274266e-02 -0.109071598  0.0983740042 -2.3731503040 -0.359130256
#> effect_66  -1.110095e-01 -0.049426237  0.0671411311 -0.2690115335  0.173495029
#> effect_67   2.402334e+00 -0.050017086  0.0603738108  0.0971249600  0.023505067
#> effect_68   4.833391e-02 -0.136421281  0.0262800533 -0.0079143266  0.988076443
#> effect_69  -6.949594e-02  0.458610444  0.2382126169  0.1285148481  0.068832778
#> effect_70   6.570188e-02  0.071052342  0.0546103529 -0.3327887160 -2.232978220
#> effect_71  -1.044357e-01 -0.105941984 -0.0241948830 -0.0958546330 -0.615983324
#> effect_72   7.611833e-02 -0.016024236 -0.0455178066 -0.0708059257 -0.155210637
#> effect_73  -1.252598e+00  0.014786283  0.0787354043 -0.0303107223  0.212686160
#> effect_74   4.684089e-01 -0.040764569  1.2103804883 -0.0086195394 -0.002959214
#> effect_75  -6.022851e-03 -0.394964453  0.2756582573  0.3532910329  0.020724140
#> effect_76   2.731951e-02  0.657346871 -0.1152913289 -0.2148345324  0.305170404
#> effect_77  -9.800925e-01  0.026243350  0.0160396602  0.0754111718  0.907734884
#> effect_78   1.254688e-01  0.003366351  0.0396715867  0.0537801474 -0.824445767
#> effect_79   1.330722e-01 -0.028168668 -0.0192465146 -0.0932753979 -0.042086968
#> effect_80  -1.764574e-01 -0.006107879 -0.8741272547 -1.4979154606 -0.503297454
#> effect_81   8.918357e-02  0.167094731 -0.2529270765  0.0371796960  0.027403951
#> effect_82   1.459960e-02  0.271330721 -0.0531102154  0.0370358985  0.013178671
#> effect_83  -8.007905e-03  0.094807758 -0.0238451387 -0.0259281196  0.831613303
#> effect_84   1.051062e+00 -0.400674651 -0.0938205921  0.0730571642 -0.348619223
#> effect_85  -3.840994e-03 -0.132764117 -0.0073272842 -1.1861438207 -0.167780896
#> effect_86   1.145375e-03  0.044244177  0.1400359114  0.0304900954  0.604564241
#> effect_87  -3.103363e-01 -0.056337733 -0.7575027883 -0.0401919243  0.041232124
#> effect_88  -1.330588e-01  0.120123309 -0.0039109463  0.0164436743  0.051734007
#> effect_89  -1.022143e-01 -0.042420535  0.0379987393  0.0070639957  0.070465612
#> effect_90   8.309514e-03  0.130833251 -0.0140072681 -0.0369119005  0.209948286
#> effect_91  -3.924304e-01 -0.290869432  0.0350212817 -2.1849798141 -0.026444273
#> effect_92  -5.870777e-01  0.068639776 -0.0694355925 -0.2744477769  0.000691603
#> effect_93   7.573239e-02  0.027281763 -0.3382218435  0.0490856709 -0.202216997
#> effect_94  -1.267251e-01  0.010470373 -0.0521305297  0.0005383084  0.374122297
#> effect_95   1.026018e-03 -0.009232498 -0.7275991008  0.0454884062  2.093285900
#> effect_96  -5.454189e-03  0.126491468 -0.0517653691 -1.5596607092  0.806221545
#> effect_97  -7.166069e-02  0.004049706 -0.5136100458  0.0824753633 -0.124992006
#> effect_98  -1.588298e-01 -0.294363676  3.1615941386  0.0369174280  0.602464927
#> effect_99  -7.785822e-02  0.036055162 -0.6857726710  0.0771108224 -0.583741712
#> effect_100 -6.512289e-02  0.010504542 -0.0687175000 -0.0876862114  0.431823075
#> effect_101  5.903740e-02 -0.058202339 -0.1021473312 -0.1185329544  0.015047859
#> effect_102  6.234449e-02  0.034169009 -0.0013718822 -0.2471809487 -0.123088529
#> effect_103 -1.048973e-01  0.326828720 -0.0486275538  0.0216884588  0.065826235
#> effect_104 -1.089362e+00  0.033479221  0.0469320952 -0.0597640870 -0.008051915
#> effect_105 -2.923615e-02  0.025638296  0.0195840561  0.1044431548 -0.055331351
#> effect_106 -4.860615e-01  0.312394211  0.0268070066  0.0051457639 -0.130756314
#> effect_107  2.758390e-01 -0.001308131 -0.0087944716 -0.0048318506  0.116692555
#> effect_108 -1.547677e-01  0.023846407 -0.0392359050 -0.0188234651  0.107518917
#> effect_109  1.502884e-01  0.033246459 -0.1177227732 -0.0122716026 -0.050879655
#> effect_110  4.050130e-02 -0.379482064 -0.0194105534 -0.1344655521  0.226746863
#> effect_111 -1.415041e-01  0.134711707 -0.0198550243  0.4101781191  0.028309611
#> effect_112  6.264758e-02  0.061594429 -0.0243292921  0.0290164559  0.193524758
#> effect_113 -7.103267e-02  0.051701735  0.0102749552  0.0028977522  0.026662445
#> effect_114  4.989807e-02 -0.296198733  0.0193001483  0.1088700649 -0.074404710
#> effect_115 -2.193638e-01  0.019118747 -0.1216693638 -0.0177155672 -0.108696017
#> effect_116  6.125474e-02 -0.146764275  0.0741965469  0.0285775531 -0.073157530
#> effect_117 -5.444405e-02  0.034726935 -0.0306096775 -0.0772896727  0.097453041
#> effect_118  5.018765e-02 -0.082170193  0.0481579680  0.1599491135 -0.400303333
#> effect_119  1.677166e-01 -0.030537808 -0.3166545602  0.0778547538  0.386415673
#> effect_120  1.618974e-01 -0.170084387  0.0140239190 -0.0865319107 -0.092826057
#> effect_121  1.149804e-01 -0.035830456 -0.0226138028  0.0715156064 -0.049649560
#> effect_122 -9.609302e-01 -0.204123288 -0.0749049812  0.0704094759  0.007651884
#> effect_123 -5.295272e-02 -0.076002490 -0.6930871286  0.0998795156  0.115811926
#> effect_124  9.684306e-02 -0.027997153 -0.0037575609  0.1183306179  0.150268699
#> effect_125  1.331551e-01 -0.026701428 -0.1326777051 -0.0282836190 -0.079472701
#> effect_126 -3.962996e-01  0.043705697  0.0099790936  0.0507647178 -0.089825093
#> effect_127 -8.206791e-02 -0.068208376 -0.0788014996  0.3007290251 -0.263252140
#> effect_128  1.578961e-02  0.050166583 -0.0864241978  0.1664388553 -1.453472812
#> effect_129  3.592476e-02  0.020260226  0.0097464152 -0.0113235776  0.002385190
#> effect_130  8.570222e-02 -0.016978668 -0.0152772195 -0.0938676040 -0.194203792
#> effect_131  2.144611e-03  0.141903193 -0.0318779631 -0.1141837353 -0.119270322
#> effect_132 -7.844157e-02  0.147202121 -0.0085786313  0.1489166902 -0.396627763
#> effect_133 -2.744102e-01 -0.029119006 -0.0100544755  0.4826610288  0.022918507
#> effect_134 -7.243927e-01  0.065556861  0.0453532638  0.0658505924  0.178760170
#> effect_135  3.114789e-01 -0.035956597  0.1077308554  0.0338171621  0.154587607
#> effect_136 -3.295029e-02  0.044902269 -0.0378631728  0.1698268978  0.044864277
#> effect_137  5.650328e-02 -0.011250030 -0.0035210833 -0.0021534005 -0.092195952
#> effect_138  2.419363e-02  0.006810341 -0.0041430654  0.0400971390 -0.021061274
#> effect_139 -4.027309e+00 -0.048831198 -0.0103850438  0.0697710464 -0.338916865
#> effect_140 -1.631531e+00 -0.134522721 -0.0553790923  0.0406606545 -1.285892017
#> effect_141  6.143512e-02  0.034338465  0.1006291849 -0.0468162104  0.109379660
#> effect_142 -3.308259e-01 -0.042582458 -0.0220857491 -1.9423141927  0.002138330
#> effect_143  2.457466e-02 -0.097758138 -0.1409048531  0.0364459436  0.185991286
#> effect_144 -6.027477e-02 -0.073864405  0.0122815476 -0.0583587344  0.146545985
#> effect_145  1.891338e-01  0.008936113 -0.0159771303 -0.0230358663 -0.134576619
#> effect_146  3.790986e-01 -0.011203658 -0.0046856416  0.1392198685 -0.183435954
#> effect_147  7.645080e-03 -0.131181242  0.0042286176  1.0516051436  0.128621726
#> effect_148  7.070719e-02 -0.141508660 -0.0371163428 -0.0117442295 -0.823412119
#> effect_149  1.176719e+00  0.075306372  0.1045801017 -0.2461264177  0.091476185
#> effect_150  9.779520e-02  0.074881922 -0.0235384230  0.0513121335 -0.586298534
#> effect_151 -3.447104e-01 -0.966408268 -2.9889356003 -2.2078278972 -2.775429244
#> effect_152  1.537388e-02 -0.161782017  0.0520799347  0.1159636821  0.165696255
#> effect_153 -4.773017e-01 -0.932220697 -0.0983225581 -1.4164847337 -1.024403289
#> effect_154 -1.065126e-03  0.024138621  0.0106742313 -0.0212447185  0.032375192
#> effect_155 -4.695171e-02  0.231521538  0.0333295383  0.0102448452  0.146487302
#> effect_156  5.621339e-02  0.186243583 -0.0093941370  0.1636056067 -0.059570596
#> effect_157  2.956974e-01  0.190122789  0.0085097663  0.0585281006  0.050897135
#> effect_158 -1.560054e-01 -0.205476568 -0.0241263857 -0.0357207806  0.012082300
#> effect_159  6.635582e-02  0.071929727 -0.0359028958  0.0316484249  0.055188632
#> effect_160  1.243724e-02  0.097377333 -0.0206884303  0.0470582298 -0.013498732
#> effect_161  3.134988e-02 -0.077373684 -0.0060474679 -0.0776852394 -0.286819839
#> effect_162 -3.393327e-01 -0.954052864 -0.1123354805  0.0569491373 -0.461393813
#> effect_163 -1.029483e-02 -0.311723750  0.0415909350 -0.0091049683 -0.094203471
#> effect_164  2.945371e+00  0.176681966  0.1859784265  0.3371016515  0.193085894
#> effect_165 -6.808293e-02 -0.108982050  0.0734480704 -0.0132436331 -0.041055695
#> effect_166 -9.560504e-02  0.352681273 -0.0431836484  0.0518856439 -0.057265744
#> effect_167 -9.023597e-02 -0.024872920  0.0524073336 -0.0324090541 -0.931295347
#> effect_168 -2.390835e+00 -1.849013110 -3.4140815023 -1.5258668561 -0.762584555
#> effect_169 -6.001917e-02 -0.078892246 -0.0249242486 -0.0432925383 -2.098367689
#> effect_170  3.792512e-02 -0.216038456 -0.0550791231 -0.0359022299 -0.181198796
#> effect_171 -2.283582e-03  0.181476310 -0.0640540398  1.2147528909  0.079382049
#> effect_172 -1.020787e-01  0.002781093 -0.0210371439 -0.0203327694  0.016729540
#> effect_173 -2.267498e+00 -0.075831772 -0.1834807682  0.0520642812 -0.003054781
#> effect_174 -5.798206e-02  0.257965020  0.1355101406  0.0798543740  0.104408199
#> effect_175  8.922646e-03 -0.027922127  0.0146083319  0.3690791342  1.593235964
#> effect_176 -9.054802e-03 -0.349053227 -0.0570811312 -0.1895275596 -0.352885844
#> effect_177  5.690489e-03 -0.021362504  0.1182210150 -0.0064057952  0.041649986
#> effect_178 -2.427672e-02 -0.396627059 -0.2265433236 -0.0029741421 -0.103560692
#> effect_179 -1.388841e-01  0.154755289 -1.7334739832 -0.0418340728 -0.073169646
#> effect_180  1.311983e-01  0.031017252  0.0084964303 -0.0192313364  0.111065386
#> effect_181 -3.114425e-02  0.079656077 -0.0119098766 -0.0713415049 -0.047928287
#> effect_182 -5.250991e-03 -0.858509239  0.0031412524 -0.4261706689 -0.538943622
#> effect_183  5.602471e-02 -0.220101930  0.0048131301 -0.6649885832 -0.023273964
#> effect_184 -8.225831e-01 -0.464486441 -0.3049631613 -0.1116583526 -0.277160153
#> effect_185 -7.671390e-02  0.011469874 -0.1053599604 -0.1101718029 -0.052202086
#> effect_186  2.762780e-01  0.387730896  0.0078674467  0.5052950136  1.010153965
#> effect_187 -8.330334e-03  0.079256988  0.0455747679  0.0357974968  0.260340770
#> effect_188  4.603613e-02  0.120464371  0.2797751708  0.0330989213  0.420080483
#> effect_189  1.253793e-01 -0.125920953  0.0381248164  0.6145849644  0.209240318
#> effect_190  4.030771e-01  0.258121776  1.1482577166  1.2022095828  1.588047324
#> effect_191  8.798318e-02  0.023676292 -0.1250593473  0.0089903523 -0.180067485
#> effect_192 -2.197114e-01 -0.228957523 -0.1737895311 -0.0239765811  0.019630599
#> effect_193 -8.509873e-02  0.013510797 -0.0529426467 -0.0167172849  0.214608317
#> effect_194 -1.388455e-02 -0.082087349 -0.0258166369 -0.0477518038  0.202184649
#> effect_195  1.190701e-01  0.157885666  0.2523457914  0.3090611005  0.387265879
#> effect_196 -1.044364e+00 -0.161581657 -0.6132250594 -0.0303600179 -0.080917249
#> effect_197  9.731983e-02 -0.015427994  0.0255825207 -0.0007097020 -0.186329483
#> effect_198 -4.916684e-02 -0.069965736  0.0846952733 -0.0709886605  0.358239892
#> effect_199  5.078583e-02  0.681607268  0.0087313529 -0.3476272194  0.226906792
#> effect_200  2.544206e-02 -0.012263271 -1.0758591347 -0.0034631033 -0.136631488
#> 
#> $result$PosteriorSD
#>            condition_1 condition_2 condition_3 condition_4 condition_5
#> effect_1     0.4209143   0.3344638   0.7188996   0.7604594   0.3597664
#> effect_2     0.2634632   0.2769986   0.2147647   0.7653880   0.8910615
#> effect_3     0.2891130   0.2782448   0.3511835   0.3994958   0.9021701
#> effect_4     0.6278692   0.4986728   0.4048833   0.5087328   0.5602062
#> effect_5     0.4080312   0.7150153   0.2080319   0.2821899   0.7150831
#> effect_6     0.2697056   0.3043339   0.2604632   0.2685219   0.4445585
#> effect_7     0.3439739   0.2856796   0.2043058   0.2655873   0.8295876
#> effect_8     0.4476495   0.5678495   0.4317296   0.2840979   0.6195446
#> effect_9     0.2606660   0.6568801   0.2242810   0.2540013   0.3561520
#> effect_10    0.4640060   0.6947327   0.2474055   0.2543485   0.4001275
#> effect_11    0.3998915   0.3081025   0.3109107   0.2537571   0.3478057
#> effect_12    0.2629343   0.2786452   0.2043028   0.3319413   0.3476104
#> effect_13    0.3345224   0.3055827   0.2158555   0.4311153   0.5486068
#> effect_14    0.3241734   0.3858988   0.3503899   0.4564427   1.0528336
#> effect_15    0.2833591   0.3755844   0.2177335   0.2647471   0.4465090
#> effect_16    0.2635725   0.4432498   0.2139785   0.2833577   0.3686440
#> effect_17    0.3182280   0.5979294   0.2469801   0.2536945   0.7525838
#> effect_18    0.5363657   0.2788150   0.2830867   0.5564301   0.3498058
#> effect_19    0.5581685   1.0014368   0.2417867   0.2610024   0.3494312
#> effect_20    0.3873001   0.2781369   0.2060932   0.4096917   0.6493163
#> effect_21    0.3092692   0.4482419   0.2163160   0.6326229   0.6843037
#> effect_22    0.3918732   0.3985622   0.2278700   0.3326743   0.3692089
#> effect_23    0.3622506   0.3932803   0.3122406   0.3598911   0.7641233
#> effect_24    0.6342116   0.3742338   0.2067776   0.2945173   0.4319668
#> effect_25    0.2695814   0.6074632   0.2690875   0.2551030   0.4307462
#> effect_26    0.3421249   0.3573524   0.2579707   0.4040089   0.3651624
#> effect_27    0.2954013   0.3312856   0.2240637   0.5421028   0.4661012
#> effect_28    0.3152541   0.3028275   0.2472276   0.7258788   0.3610020
#> effect_29    0.2607016   0.3833649   0.2455585   0.3457399   0.4122745
#> effect_30    0.2746729   0.3058246   0.2467064   0.2708410   0.5404250
#> effect_31    0.3160179   0.3710138   0.2156366   0.3627271   0.3490713
#> effect_32    0.4269754   0.3842440   0.2197928   0.3501553   0.7357529
#> effect_33    0.6458214   0.2983261   0.2206364   0.2538227   0.4328042
#> effect_34    0.3021192   0.4414543   0.4267062   0.3352172   0.5510320
#> effect_35    0.3022356   0.2829654   0.2546564   0.2694655   0.3488465
#> effect_36    0.2602176   0.2787073   0.2127357   0.4419668   0.3475807
#> effect_37    0.2640533   0.3535258   0.2079607   0.2541816   0.3481043
#> effect_38    0.3093282   0.4863108   0.2203781   0.2538816   0.4550301
#> effect_39    0.2602591   0.7039652   0.2366096   0.2662210   0.3625473
#> effect_40    0.5664775   0.4290694   0.2225231   0.3186014   0.3625378
#> effect_41    0.2745740   0.4298718   0.2666012   0.2561731   0.3549557
#> effect_42    0.3162692   0.6954961   0.2046603   0.3041760   0.4804187
#> effect_43    0.6175648   0.2913813   0.2167259   0.3192453   0.3508229
#> effect_44    0.3893986   0.2785202   0.3400357   0.2541033   0.5033736
#> effect_45    0.3040062   0.5523918   0.2828561   0.3262084   0.3907650
#> effect_46    0.2602267   0.5522852   0.2045550   0.3423192   0.4327725
#> effect_47    0.2606960   0.2836977   0.2346669   0.2567210   0.5665659
#> effect_48    0.4640400   0.2779203   0.3455016   0.2724133   0.3733417
#> effect_49    0.4435636   0.3572430   0.2322753   0.3080197   0.4834650
#> effect_50    0.3107938   0.2888115   1.3594829   0.2994950   0.4380385
#> effect_51    0.7402714   0.4961266   0.6066249   0.2949354   0.4715702
#> effect_52    0.7268670   0.6548048   0.4127920   0.8056932   0.3632528
#> effect_53    0.3055425   0.4120021   0.3779736   0.2793302   0.3598858
#> effect_54    0.2625637   0.3113648   0.5044694   0.4137056   0.3541846
#> effect_55    0.6050057   0.3303733   0.9690853   0.3237220   1.0654161
#> effect_56    1.1455264   0.2770591   0.5999295   0.2584997   0.5462271
#> effect_57    0.2602186   1.0501700   1.3584294   0.2781486   1.0734095
#> effect_58    0.2729353   0.3843440   0.4048556   0.2589215   0.4283411
#> effect_59    0.4103262   0.3544627   0.2106103   0.7763727   0.5483161
#> effect_60    0.2957097   0.3652914   0.2095852   0.3534476   0.6489110
#> effect_61    1.2027774   0.2782204   0.3469867   0.2592929   0.6595678
#> effect_62    0.4252081   0.6355289   0.2349832   1.0939526   0.3780029
#> effect_63    0.4684647   0.5200109   0.2156279   1.0229044   0.5358593
#> effect_64    0.2926507   0.3255931   0.4594484   0.4604392   0.3562055
#> effect_65    0.2649432   0.3995539   0.4113473   1.1326296   0.6931767
#> effect_66    0.4271507   0.3142947   0.3414239   0.6536338   0.5071233
#> effect_67    1.1410817   0.3150542   0.3252018   0.3996496   0.3527595
#> effect_68    0.3135776   0.4385017   0.2413609   0.2556230   1.0188184
#> effect_69    0.3520325   0.7654729   0.6491160   0.4544581   0.3873180
#> effect_70    0.3450270   0.3440097   0.3111173   0.7253995   0.9827069
#> effect_71    0.4156107   0.3950016   0.2366336   0.3973464   0.8707001
#> effect_72    0.3642685   0.2815593   0.2885185   0.3509210   0.4858986
#> effect_73    1.2173341   0.2808893   0.3683375   0.2790748   0.5511064
#> effect_74    0.8554713   0.3036827   1.2959011   0.2559827   0.3475057
#> effect_75    0.2612932   0.7163949   0.6991860   0.7466225   0.3515902
#> effect_76    0.2803132   0.8864019   0.4459665   0.5845977   0.6448922
#> effect_77    1.1439709   0.2888607   0.2200487   0.3595623   0.9954480
#> effect_78    0.4519280   0.2771695   0.2738901   0.3191023   0.9665737
#> effect_79    0.4646148   0.2905578   0.2261309   0.3926517   0.3637992
#> effect_80    0.5325879   0.2776431   1.1743161   1.2388093   0.8022590
#> effect_81    0.3882359   0.4799225   0.6692983   0.2898541   0.3546239
#> effect_82    0.2663757   0.6019923   0.3074160   0.2896177   0.3491227
#> effect_83    0.2621111   0.3787087   0.2358558   0.2728612   0.9692539
#> effect_84    1.1674404   0.7210415   0.4016560   0.3551480   0.6841201
#> effect_85    0.2606565   0.4333947   0.2078487   1.1961130   0.5005301
#> effect_86    0.2602567   0.3078192   0.4930535   0.2793414   0.8643503
#> effect_87    0.7038293   0.3234065   1.1133422   0.2948926   0.3631796
#> effect_88    0.4645927   0.4154906   0.2053261   0.2618405   0.3714264
#> effect_89    0.4116741   0.3056281   0.2697173   0.2552285   0.3889622
#> effect_90    0.2622549   0.4306847   0.2165753   0.2894142   0.5481110
#> effect_91    0.7879399   0.6220099   0.2623330   1.1778541   0.3541411
#> effect_92    0.9451593   0.3405530   0.3468400   0.6601218   0.3474240
#> effect_93    0.3635561   0.2897645   0.7744208   0.3105326   0.5395869
#> effect_94    0.4540406   0.2789502   0.3049919   0.2536763   0.7058139
#> effect_95    0.2602490   0.2785112   1.0958566   0.3040930   1.0121113
#> effect_96    0.2611005   0.4245575   0.3040870   1.2414795   0.9595876
#> effect_97    0.3560335   0.2772621   0.9443453   0.3727500   0.4503044
#> effect_98    0.5058524   0.6255057   1.0315231   0.2894233   0.8631693
#> effect_99    0.3674782   0.2984047   1.0700210   0.3627444   0.8524453
#> effect_100   0.3439595   0.2789630   0.3451496   0.3823981   0.7515284
#> effect_101   0.3327879   0.3259400   0.4192550   0.4374829   0.3496356
#> effect_102   0.3388461   0.2964070   0.2044254   0.6268070   0.4480605
#> effect_103   0.4164265   0.6568287   0.2962856   0.2674760   0.3843415
#> effect_104   1.1787765   0.2956944   0.2920543   0.3302064   0.3480579
#> effect_105   0.2829314   0.2883468   0.2268084   0.4127955   0.3745424
#> effect_106   0.8699469   0.6431544   0.2425776   0.2544988   0.4571072
#> effect_107   0.6643092   0.2769935   0.2093638   0.2544008   0.4405376
#> effect_108   0.4995239   0.2868816   0.2728021   0.2642444   0.4298233
#> effect_109   0.4924708   0.2954561   0.4507717   0.2583039   0.3707068
#> effect_110   0.3002182   0.7035396   0.2264594   0.4643737   0.5662947
#> effect_111   0.4784073   0.4361186   0.2273569   0.8014540   0.3550935
#> effect_112   0.3394031   0.3306164   0.2369338   0.2771786   0.5298906
#> effect_113   0.3548726   0.3172410   0.2111355   0.2539317   0.3542495
#> effect_114   0.3163312   0.6273316   0.2262382   0.4206424   0.3930027
#> effect_115   0.5931051   0.2834462   0.4584850   0.2630922   0.4311913
#> effect_116   0.3368458   0.4527555   0.3579372   0.2765467   0.3917126
#> effect_117   0.3244509   0.2969905   0.2515689   0.3630789   0.4182399
#> effect_118   0.3168435   0.3601611   0.2951146   0.5051559   0.7271169
#> effect_119   0.5194764   0.2927700   0.7495529   0.3641356   0.7159361
#> effect_120   0.5105891   0.4838207   0.2166024   0.3802678   0.4130034
#> effect_121   0.4340394   0.2981630   0.2331559   0.3522539   0.3696849
#> effect_122   1.1370649   0.5264716   0.3595721   0.3501762   0.3479961
#> effect_123   0.3217703   0.3511644   1.0746587   0.4046225   0.4395046
#> effect_124   0.4020815   0.2904029   0.2052475   0.4371344   0.4801061
#> effect_125   0.4647517   0.2892560   0.4794575   0.2761269   0.3983405
#> effect_126   0.7915997   0.3071671   0.2107627   0.3135791   0.4096446
#> effect_127   0.3752256   0.3399374   0.3684876   0.6904766   0.6041686
#> effect_128   0.2673837   0.3152470   0.3855565   0.5151224   1.0807446
#> effect_129   0.2928564   0.2842162   0.2104760   0.2576291   0.3474754
#> effect_130   0.3818860   0.2821094   0.2187077   0.3937318   0.5306522
#> effect_131   0.2603547   0.4460918   0.2546344   0.4299524   0.4435658
#> effect_132   0.3685535   0.4533526   0.2091256   0.4878288   0.7241836
#> effect_133   0.6626118   0.2914292   0.2108569   0.8637825   0.3525016
#> effect_134   1.0296168   0.3361727   0.2881069   0.3416107   0.5131619
#> effect_135   0.7050923   0.2982986   0.4307570   0.2844391   0.4851695
#> effect_136   0.2883044   0.3086217   0.2693798   0.5202606   0.3658783
#> effect_137   0.3281749   0.2792538   0.2051324   0.2538133   0.4122955
#> effect_138   0.2762994   0.2778080   0.2054508   0.2947317   0.3517247
#> effect_139   0.8787887   0.3135340   0.2112766   0.3489767   0.6756134
#> effect_140   1.2527712   0.4358547   0.3130088   0.2956906   1.0716074
#> effect_141   0.3371767   0.2965835   0.4160867   0.3064546   0.4319869
#> effect_142   0.7260555   0.3058206   0.2320178   1.2204871   0.3474644
#> effect_143   0.2767705   0.3830371   0.4946378   0.2886522   0.5213941
#> effect_144   0.3350502   0.3480651   0.2138974   0.3275856   0.4757310
#> effect_145   0.5511135   0.2784141   0.2199371   0.2691137   0.4616182
#> effect_146   0.7751338   0.2792352   0.2057695   0.4721872   0.5184934
#> effect_147   0.2619448   0.4311738   0.2054985   1.1610689   0.4545871
#> effect_148   0.3542710   0.4455482   0.2675222   0.2579224   0.9661841
#> effect_149   1.2013187   0.3501540   0.4242953   0.6254787   0.4114886
#> effect_150   0.4037893   0.3495386   0.2351775   0.3145771   0.8539301
#> effect_151   0.7406274   1.0021255   1.0852360   1.1728602   0.8738094
#> effect_152   0.2670238   0.4729339   0.3048666   0.4330441   0.4981152
#> effect_153   0.8628171   0.9928686   0.4112387   1.2324814   1.0280243
#> effect_154   0.2602515   0.2871147   0.2116531   0.2669527   0.3573649
#> effect_155   0.3111647   0.5585790   0.2581739   0.2569222   0.4756620
#> effect_156   0.3276492   0.5044619   0.2100530   0.5107915   0.3783812
#> effect_157   0.6873938   0.5093104   0.2090506   0.3279011   0.3707214
#> effect_158   0.5014590   0.5281028   0.2364809   0.2874751   0.3488526
#> effect_159   0.3462333   0.3452722   0.2645122   0.2810838   0.3744162
#> effect_160   0.2647244   0.3824788   0.2290688   0.3068871   0.3492059
#> effect_161   0.2859437   0.3531582   0.2067340   0.3638187   0.6274226
#> effect_162   0.7350288   0.9988721   0.4400690   0.3249638   0.7732672
#> effect_163   0.2633276   0.6425096   0.2786894   0.2562472   0.4145553
#> effect_164   0.9904186   0.4923354   0.5713177   0.7299332   0.5293979
#> effect_165   0.3494219   0.3994238   0.3562053   0.2590466   0.3630529
#> effect_166   0.3998563   0.6803627   0.2826763   0.3156251   0.3762726
#> effect_167   0.3901489   0.2877105   0.3056773   0.2822473   1.0027427
#> effect_168   1.1440797   1.0164471   0.9687267   1.2402413   0.9418343
#> effect_169   0.3345824   0.3553709   0.2382703   0.3002324   1.0110908
#> effect_170   0.2960254   0.5406697   0.3122712   0.2877686   0.5159463
#> effect_171   0.2603730   0.4984473   0.3340688   1.2021957   0.3982438
#> effect_172   0.4114332   0.2771037   0.2297959   0.2659029   0.3501530
#> effect_173   1.1744106   0.3509165   0.5673254   0.3159519   0.3475114
#> effect_174   0.3308632   0.5878226   0.4847298   0.3678695   0.4262201
#> effect_175   0.2625629   0.2903353   0.2175683   0.7624165   1.0784967
#> effect_176   0.2626320   0.6771317   0.3171831   0.5492911   0.6878161
#> effect_177   0.2611783   0.2849963   0.4517514   0.2549529   0.3634812
#> effect_178   0.2764017   0.7177530   0.6326062   0.2539458   0.4252417
#> effect_179   0.4741526   0.4635700   1.3595791   0.2977031   0.3917250
#> effect_180   0.4615101   0.2932336   0.2090362   0.2646827   0.4339517
#> effect_181   0.2856452   0.3564856   0.2133562   0.3519269   0.3682842
#> effect_182   0.2610362   0.9701157   0.2049628   0.8158865   0.8253476
#> effect_183   0.3273073   0.5454282   0.2058498   0.9910056   0.3526572
#> effect_184   1.0794920   0.7697176   0.7356205   0.4255426   0.6180039
#> effect_185   0.3653676   0.2793432   0.4259016   0.4229342   0.3718241
#> effect_186   0.6648297   0.7104360   0.2083777   0.8817220   1.0245115
#> effect_187   0.2622650   0.3559030   0.2886610   0.2875992   0.6012311
#> effect_188   0.3095779   0.4159785   0.7044417   0.2833151   0.7425873
#> effect_189   0.4517773   0.4237490   0.2700314   0.9595223   0.5473344
#> effect_190   0.7979500   0.5879911   1.2788077   1.1995859   1.0787294
#> effect_191   0.3860501   0.2867470   0.4650270   0.2561835   0.5146556
#> effect_192   0.5935710   0.5556545   0.5515694   0.2702989   0.3511680
#> effect_193   0.3807821   0.2802513   0.3070017   0.2621032   0.5532021
#> effect_194   0.2658035   0.3600399   0.2402978   0.3081298   0.5395510
#> effect_195   0.4410678   0.4677585   0.6685139   0.6997665   0.7166282
#> effect_196   1.1653639   0.4726688   1.0211061   0.2791480   0.3998878
#> effect_197   0.4029371   0.2812305   0.2397634   0.2536831   0.5217773
#> effect_198   0.3150410   0.3424500   0.3817280   0.3512642   0.6924159
#> effect_199   0.3179041   0.8983438   0.2092936   0.7408419   0.5664656
#> effect_200   0.2778623   0.2796795   1.2559823   0.2540448   0.4640442
#> 
#> $result$lfsr
#>             condition_1 condition_2 condition_3 condition_4 condition_5
#> effect_1   0.8921234174   0.8945272  0.81894988  0.75311814  0.89069969
#> effect_2   0.9507816259   0.9348499  0.96774138  0.75068670  0.54798859
#> effect_3   0.9377131853   0.9310493  0.93247177  0.89888529  0.53791540
#> effect_4   0.8144676849   0.8153118  0.91856982  0.86031898  0.77483824
#> effect_5   0.8965176597   0.6908750  0.97063108  0.93932912  0.68145959
#> effect_6   0.9465951290   0.9103746  0.95492290  0.94521296  0.83787357
#> effect_7   0.9181088574   0.9227872  0.97411322  0.94666467  0.59942980
#> effect_8   0.8828861909   0.7790534  0.91133594  0.93857620  0.74081606
#> effect_9   0.9541065367   0.7278210  0.96460889  0.95561622  0.89417535
#> effect_10  0.8771428201   0.7041241  0.95824150  0.95493439  0.86290877
#> effect_11  0.8992783749   0.9082387  0.94249889  0.95636546  0.90719877
#> effect_12  0.9512517595   0.9303533  0.97414642  0.92157236  0.90813595
#> effect_13  0.9213203646   0.9096591  0.96734780  0.88806908  0.78129811
#> effect_14  0.9248772863   0.8699128  0.93267215  0.87922128  0.34727378
#> effect_15  0.9400807793   0.8747803  0.96669670  0.94710153  0.83681009
#> effect_16  0.9506886138   0.8426480  0.96803344  0.93886691  0.88344210
#> effect_17  0.9269488949   0.7623810  0.95835161  0.95672157  0.65615074
#> effect_18  0.8507188401   0.9300808  0.94932609  0.84221736  0.90231475
#> effect_19  0.8423905019   0.4227319  0.95971008  0.94921385  0.90300325
#> effect_20  0.9035298771   0.9312532  0.97182129  0.89542180  0.72300512
#> effect_21  0.9301266507   0.8402356  0.96718526  0.81134495  0.70131374
#> effect_22  0.9019880786   0.8639382  0.96354484  0.92132466  0.88301884
#> effect_23  0.9119503381   0.8664311  0.94217153  0.91219273  0.64808735
#> effect_24  0.8118218556   0.8754184  0.97135819  0.93463019  0.84478378
#> effect_25  0.9466652553   0.7569680  0.95277525  0.95387774  0.84545874
#> effect_26  0.9187350844   0.8834300  0.95554884  0.89735475  0.88613585
#> effect_27  0.9352541583   0.8961075  0.96467480  0.84774749  0.82619030
#> effect_28  0.9279952508   0.9112491  0.95828752  0.76976095  0.88960140
#> effect_29  0.9540336236   0.8711080  0.95872079  0.91692985  0.85583643
#> effect_30  0.9440055169   0.9095214  0.95842253  0.94413005  0.78582389
#> effect_31  0.9277258014   0.8769410  0.96742578  0.91124402  0.90372130
#> effect_32  0.8900441272   0.8706933  0.96601390  0.91545046  0.66767251
#> effect_33  0.8069304788   0.9139500  0.96574200  0.95611315  0.84432137
#> effect_34  0.9327301745   0.8435136  0.91270505  0.92046653  0.77995185
#> effect_35  0.9326871864   0.9251993  0.95638594  0.94476643  0.90420623
#> effect_36  0.9559158531   0.9302522  0.96851209  0.88430040  0.90831446
#> effect_37  0.9502935365   0.8852603  0.97066938  0.95523440  0.90613704
#> effect_38  0.9301054273   0.8215242  0.96582481  0.95592838  0.83217978
#> effect_39  0.9553822694   0.6981436  0.96109638  0.94634205  0.88827640
#> effect_40  0.8391678734   0.8494569  0.96514770  0.92611452  0.88828440
#> effect_41  0.9440538264   0.8490732  0.95339196  0.95274968  0.89544241
#> effect_42  0.9276372340   0.7036327  0.97320683  0.93114361  0.81846441
#> effect_43  0.8187275165   0.9184816  0.96704220  0.92589340  0.90065463
#> effect_44  0.9028226590   0.9305617  0.93527516  0.95539100  0.80606174
#> effect_45  0.9320358920   0.7873946  0.94938269  0.92351552  0.86857743
#> effect_46  0.9556717547   0.7874517  0.97337766  0.91807747  0.84433886
#> effect_47  0.9540450962   0.9245133  0.96162711  0.95225519  0.77127301
#> effect_48  0.8771308235   0.9316901  0.93390371  0.94342118  0.88001807
#> effect_49  0.8843093900   0.8834823  0.96229003  0.92978686  0.81682102
#> effect_50  0.9295800808   0.9203305  0.29645585  0.93281764  0.84144088
#> effect_51  0.7646389358   0.8165974  0.85865100  0.93447645  0.82323796
#> effect_52  0.7709300273   0.7290841  0.91645975  0.73020803  0.88768729
#> effect_53  0.9314746525   0.8575805  0.92562469  0.94048110  0.89059190
#> effect_54  0.9516049232   0.9064401  0.89065462  0.89405236  0.89630153
#> effect_55  0.8238556575   0.8965637  0.70843095  0.92436207  0.31716299
#> effect_56  0.4937148113   0.9343795  0.86086246  0.95086180  0.78261697
#> effect_57  0.9558409353   0.3142893  0.32325410  0.94096644  0.18106636
#> effect_58  0.9448708207   0.8706461  0.91857716  0.95056411  0.84679184
#> effect_59  0.8957371935   0.8848115  0.96939301  0.74521196  0.78145934
#> effect_60  0.9351362405   0.8796522  0.96985641  0.91434841  0.72325140
#> effect_61  0.4223857307   0.9310948  0.93353006  0.95030973  0.71673919
#> effect_62  0.8906512660   0.7406474  0.96154024  0.07424623  0.87679788
#> effect_63  0.8755638756   0.8044104  0.96742887  0.59537520  0.78833917
#> effect_64  0.9363162577   0.8989750  0.90364749  0.87780804  0.89412024
#> effect_65  0.9496125363   0.8634699  0.91684649  0.09806513  0.69566863
#> effect_66  0.8899838782   0.9048578  0.93492737  0.80236469  0.80402895
#> effect_67  0.0962816385   0.9044521  0.93897009  0.89883319  0.89800135
#> effect_68  0.9285886330   0.8449348  0.95982276  0.95329555  0.40621294
#> effect_69  0.9153878851   0.6560040  0.84422018  0.87992125  0.87073057
#> effect_70  0.9177525395   0.8898490  0.94244803  0.76998666  0.06471635
#> effect_71  0.8939363671   0.8656189  0.96108987  0.89961285  0.56577778
#> effect_72  0.9112723577   0.9266178  0.94799364  0.91519411  0.81550786
#> effect_73  0.3985746904   0.9273553  0.92810660  0.94058549  0.77991050
#> effect_74  0.7059233860   0.9107510  0.46199318  0.95293214  0.90885521
#> effect_75  0.9530634059   0.6899588  0.82629548  0.75986382  0.89954213
#> effect_76  0.9413973488   0.5573425  0.90741565  0.83109949  0.72568745
#> effect_77  0.4953584473   0.9202939  0.96593097  0.91230271  0.43824642
#> effect_78  0.8813910459   0.9338117  0.95158831  0.92594248  0.47280045
#> effect_79  0.8769275769   0.9190610  0.96405499  0.90119920  0.88723721
#> effect_80  0.8521435661   0.9323225  0.57992138  0.31624710  0.62039259
#> effect_81  0.9032145680   0.8247067  0.83711835  0.93636639  0.89580761
#> effect_82  0.9486160590   0.7600821  0.94335830  0.93645555  0.90361466
#> effect_83  0.9520708575   0.8733050  0.96130156  0.94322247  0.46977152
#> effect_84  0.4692377472   0.6868585  0.91942596  0.91377942  0.70142990
#> effect_85  0.9541265399   0.8473865  0.97073024  0.41878972  0.80760157
#> effect_86  0.9553980191   0.9083969  0.89401029  0.94047653  0.57116204
#> effect_87  0.7815096290   0.9000913  0.62371034  0.93449218  0.88774801
#> effect_88  0.8769354167   0.8559254  0.97244591  0.94871238  0.88138918
#> effect_89  0.8952783767   0.9096332  0.95261929  0.95372967  0.86969837
#> effect_90  0.9519180068   0.8486843  0.96709459  0.93653242  0.78157307
#> effect_91  0.7414055848   0.7485826  0.95445507  0.13526681  0.89635110
#> effect_92  0.6530900517   0.8915326  0.93356699  0.79954798  0.90997203
#> effect_93  0.9115117055   0.9196304  0.79732623  0.92890723  0.78628615
#> effect_94  0.8806509558   0.9298717  0.94395382  0.95690424  0.68752040
#> effect_95  0.9554532155   0.9305769  0.63520472  0.93117310  0.08776909
#> effect_96  0.9533433484   0.8516113  0.94417601  0.29738224  0.48054972
#> effect_97  0.9140407148   0.9334443  0.72106776  0.90788892  0.83474483
#> effect_98  0.8620782944   0.7465441  0.02691528  0.93652899  0.57215543
#> effect_99  0.9101941883   0.9139015  0.65147539  0.91123824  0.58106448
#> effect_100 0.9181137201   0.9298521  0.93399218  0.90465194  0.65688142
#> effect_101 0.9219130486   0.8987987  0.91472255  0.88586150  0.90262111
#> effect_102 0.9198478119   0.9151500  0.97364463  0.81379336  0.83596523
#> effect_103 0.8936578785   0.7278524  0.94608972  0.94571860  0.87262570
#> effect_104 0.4554133261   0.9156044  0.94712692  0.92215924  0.90628515
#> effect_105 0.9402625973   0.9206789  0.96385512  0.89436318  0.87917380
#> effect_106 0.6978761819   0.7361087  0.95950136  0.95469222  0.83105419
#> effect_107 0.7990097544   0.9349028  0.96996077  0.95484747  0.84007109
#> effect_108 0.8643932021   0.9218119  0.95185676  0.94736830  0.84596974
#> effect_109 0.8669572382   0.9157575  0.90607889  0.95100348  0.88191277
#> effect_110 0.9334356157   0.6984211  0.96395791  0.87641196  0.77142539
#> effect_111 0.8720212636   0.8460798  0.96369438  0.73241360  0.89529267
#> effect_112 0.9196585839   0.8964420  0.96100845  0.94136947  0.79161714
#> effect_113 0.9144313762   0.9032935  0.96916648  0.95578956  0.89622774
#> effect_114 0.9276153924   0.7454756  0.96402326  0.89167717  0.86720061
#> effect_115 0.8286517503   0.9247455  0.90391858  0.94799716  0.84521247
#> effect_116 0.9205282640   0.8380470  0.93076093  0.94163438  0.86799254
#> effect_117 0.9247811986   0.9147815  0.95717153  0.91112632  0.85244555
#> effect_118 0.9274350810   0.8820908  0.94637678  0.86164188  0.67347981
#> effect_119 0.8570476752   0.9175286  0.80718703  0.91077281  0.68089817
#> effect_120 0.8603366547   0.8227669  0.96708515  0.90536754  0.85541964
#> effect_121 0.8876102383   0.9140507  0.96204460  0.91474790  0.88266484
#> effect_122 0.5025255706   0.8010625  0.93034530  0.91544347  0.90649007
#> effect_123 0.9257116832   0.8863935  0.64861339  0.89714636  0.84063693
#> effect_124 0.8985366442   0.9191713  0.97252026  0.88598262  0.81863303
#> effect_125 0.8768791589   0.9200016  0.89795213  0.94181152  0.86397340
#> effect_126 0.7395632798   0.9087625  0.96932658  0.92784757  0.85734656
#> effect_127 0.9075917000   0.8918336  0.92806809  0.78608533  0.74980697
#> effect_128 0.9479686440   0.9043494  0.92365559  0.85794420  0.24421310
#> effect_129 0.9362361389   0.9240449  0.96945201  0.95151141  0.90913955
#> effect_130 0.9053524885   0.9260446  0.96637007  0.90083457  0.79119947
#> effect_131 0.9549322252   0.8412757  0.95639153  0.88847110  0.83841544
#> effect_132 0.9098330250   0.8377570  0.97007511  0.86798607  0.67543680
#> effect_133 0.7997438300   0.9184483  0.96928585  0.69864319  0.89832804
#> effect_134 0.5952511185   0.8936826  0.94809457  0.91831538  0.80074934
#> effect_135 0.7809370694   0.9139670  0.91160159  0.93844277  0.81590134
#> effect_136 0.9380381259   0.9079495  0.95270287  0.85602357  0.88556885
#> effect_137 0.9234956524   0.9294236  0.97263440  0.95614538  0.85582439
#> effect_138 0.9432255410   0.9319344  0.97233295  0.93455133  0.89935644
#> effect_139 0.0002694865   0.9052659  0.96910675  0.91584518  0.70678416
#> effect_140 0.2793463518   0.8462065  0.94198235  0.93419953  0.29800410
#> effect_141 0.9204156264   0.9150381  0.91557563  0.93033757  0.84477269
#> effect_142 0.7713076732   0.9095237  0.96236210  0.19159537  0.90926150
#> effect_143 0.9430043048   0.8712626  0.89354713  0.93682104  0.79626474
#> effect_144 0.9211402505   0.8878860  0.96806402  0.92304769  0.82099311
#> effect_145 0.8451054234   0.9307446  0.96596710  0.94493193  0.82861297
#> effect_146 0.7477844357   0.9294503  0.97206678  0.87362503  0.79784690
#> effect_147 0.9522540265   0.8484503  0.97229118  0.46681474  0.83241993
#> effect_148 0.9146338961   0.8415384  0.95316332  0.95128665  0.47323813
#> effect_149 0.4245804962   0.8868794  0.91335956  0.81435035  0.85628663
#> effect_150 0.8979577119   0.8871757  0.96148698  0.92750189  0.57984261
#> effect_151 0.7644704736   0.4216495  0.04328847  0.13043505  0.01301969
#> effect_152 0.9481955821   0.8281676  0.94398460  0.88740152  0.80890833
#> effect_153 0.7018607180   0.4357240  0.91687553  0.34185970  0.39213497
#> effect_154 0.9554351404   0.9216278  0.96894940  0.94597621  0.89295742
#> effect_155 0.9294475047   0.7840736  0.95549771  0.95208327  0.82103035
#> effect_156 0.9236766651   0.8123772  0.96964112  0.85955545  0.87654277
#> effect_157 0.7888843245   0.8099064  0.97011150  0.92294061  0.88190207
#> effect_158 0.8636868130   0.8002136  0.96113135  0.93726956  0.90419255
#> effect_159 0.9173446924   0.8892365  0.95391152  0.93977110  0.87926201
#> effect_160 0.9497747608   0.8715260  0.96319845  0.93018516  0.90344527
#> effect_161 0.9390004567   0.8854365  0.97138596  0.91087884  0.73615645
#> effect_162 0.7671115875   0.4267105  0.90904683  0.92393895  0.64159788
#> effect_163 0.9508988802   0.7364943  0.95040641  0.95268024  0.85453462
#> effect_164 0.0264137518   0.8185057  0.87012762  0.76784641  0.79188719
#> effect_165 0.9162680671   0.8635313  0.93120059  0.95047764  0.88785327
#> effect_166 0.8992902673   0.7132716  0.94942682  0.92713964  0.87797582
#> effect_167 0.9025696941   0.9211641  0.94378548  0.93930630  0.42872268
#> effect_168 0.0983381899   0.1336218  0.01167888  0.30764747  0.49938876
#> effect_169 0.9212998816   0.8843769  0.96064758  0.93255216  0.08685746
#> effect_170 0.9350157606   0.7936238  0.94216400  0.93715741  0.79923440
#> effect_171 0.9548669239   0.8154258  0.93676596  0.40887408  0.86403124
#> effect_172 0.8953604094   0.9341244  0.96299027  0.94650328  0.90171846
#> effect_173 0.1218055164   0.8865127  0.87139694  0.92702679  0.90880777
#> effect_174 0.9225721657   0.7680504  0.89643053  0.90952328  0.84797120
#> effect_175 0.9516056561   0.9192195  0.96675278  0.75215446  0.20327984
#> effect_176 0.9515380484   0.7153023  0.94095323  0.84498319  0.69908642
#> effect_177 0.9532272632   0.9233637  0.90580552  0.95406260  0.88749849
#> effect_178 0.9431773364   0.6890550  0.84990976  0.95575283  0.84851651
#> effect_179 0.8735409451   0.8327723  0.30080981  0.93346589  0.86798485
#> effect_180 0.8780241361   0.9172165  0.97011855  0.94713547  0.84368843
#> effect_181 0.9391237024   0.8838440  0.96827033  0.91485734  0.88371349
#> effect_182 0.9534428957   0.4668328  0.97281652  0.72485267  0.60274964
#> effect_183 0.9237944725   0.7911049  0.97200381  0.61860493  0.89813017
#> effect_184 0.5557700611   0.6529330  0.81258589  0.88999238  0.74172304
#> effect_185 0.9109031399   0.9292967  0.91292368  0.89088994  0.88110188
#> effect_186 0.7987843598   0.6939021  0.97044909  0.68834623  0.39762734
#> effect_187 0.9519074359   0.8841224  0.94795871  0.93722215  0.75150966
#> effect_188 0.9300157448   0.8556937  0.82435350  0.93888372  0.66302716
#> effect_189 0.8814438180   0.8519967  0.95254157  0.64010203  0.78200354
#> effect_190 0.7363458443   0.7679564  0.48290753  0.41320880  0.20473466
#> effect_191 0.9039509349   0.9219188  0.90207208  0.95273985  0.79993688
#> effect_192 0.8284651153   0.7856463  0.87635160  0.94437892  0.90014232
#> effect_193 0.9057238084   0.9281080  0.94346010  0.94855915  0.77874529
#> effect_194 0.9490018036   0.8821486  0.96010505  0.92974822  0.78630593
#> effect_195 0.8851765996   0.8307171  0.83739742  0.78186839  0.68044212
#> effect_196 0.4716747515   0.8282984  0.68026112  0.94055556  0.86305116
#> effect_197 0.8982466833   0.9269738  0.96024752  0.95682458  0.79605556
#> effect_198 0.9280705276   0.8906074  0.92465156  0.91507919  0.69615509
#> effect_199 0.9270624935   0.5459653  0.96999424  0.76264743  0.77132939
#> effect_200 0.9424989514   0.9288370  0.50776821  0.95551678  0.82730160
#> 
#> 
#> $loglik
#> [1] -1639.133
#> 
#> attr(,"class")
#> [1] "mash_1by1"