Perform condition-by-condition analyses

Performs simple "condition-by-condition" analysis by running ash from package ashr on data from each condition, one at a time. May be a useful first step to identify top hits in each condition before a mash analysis.

mash_1by1(data, alpha = 0, ...)

Arguments

data
A list with the following two elements: Bhat an n by R matrix of observations (n units in R conditions); and Shat, an n by R matrix of standard errors (n units in R conditions),

alpha
Numeric value of alpha parameter in the model. alpha = 0 for Exchangeable Effects (EE), alpha = 1 for Exchangeable Z-scores (EZ).

...
optionally, other parameters to be passed to ash

Value

A list similar to the output of mash, particularly including posterior matrices.

Examples

simdata = simple_sims(50,5,1)
mash_1by1(simdata)

#> $result
#> $result$PosteriorMean
#>             condition_1   condition_2   condition_3   condition_4  condition_5
#> effect_1    0.400699614 -0.0193895175  0.0835234760 -0.0015421215 -0.016411677
#> effect_2   -0.268059544 -0.0095085715 -0.0978120565  0.0005080524 -0.055389178
#> effect_3    0.019209030 -0.0326804417 -0.0486515199 -0.0056280192 -0.054703789
#> effect_4    0.024041099  0.0078079804  0.0399944324 -0.0120225734 -0.660422999
#> effect_5   -0.020377048 -0.0465503658  0.1426153559  0.0282070680 -0.018793465
#> effect_6   -0.319787417  0.0423797492 -0.0514224275  0.0016362165 -0.001071294
#> effect_7    0.125990217 -0.0804841537  0.0128042901 -0.0406551358  0.189402517
#> effect_8   -0.711387308 -0.1433959597 -0.2978110695  0.0083761375  0.045623546
#> effect_9    0.141392325 -0.0053515431 -0.6311611849 -0.0230974393  0.087337755
#> effect_10  -0.487663754 -0.0303378410 -0.4634886672 -0.2257967050 -0.135820321
#> effect_11  -0.063840415  0.0297789799 -0.0975687724 -0.0022936625  0.043407281
#> effect_12  -0.027200061 -0.2034392554  0.1382034942 -0.0019039529 -0.079334041
#> effect_13  -0.137696451  0.0165321004 -0.1208809147  0.0093866075  0.022180693
#> effect_14  -0.112810054  0.0836906417 -0.1274870794  0.0529306950  0.308459625
#> effect_15   0.091238838 -0.0724193850 -0.0301145744  0.0237894227 -0.177883303
#> effect_16   0.346164844  0.0589324200 -0.1538489311 -0.0182896521 -0.001359080
#> effect_17   0.006753702  0.0645107015  0.0658332261  0.0121695373 -0.005989304
#> effect_18  -0.118854550  0.0410772515  0.0205323556  0.0014557167  0.054933810
#> effect_19  -0.135644593 -0.1019228786  0.0799606051  0.0104690705  0.530859065
#> effect_20   0.148367769 -0.0200279746 -0.1200361453 -0.0078736011 -0.674005920
#> effect_21  -0.089717359  0.0207823382 -0.0092712186 -0.0152327420  0.141401099
#> effect_22  -0.469452739 -0.1762782405  0.0009458062  0.0291818574 -0.065545391
#> effect_23   0.008083148 -0.0667318929  0.0435291451  0.0021213487  0.006806519
#> effect_24   0.202285207  0.0324130540  0.0120654795  0.0153038445  0.124160427
#> effect_25   0.089224130  0.0978825268 -0.0859378311 -0.0142986467  0.247989673
#> effect_26   0.567708419  0.0209698680  0.0860607172 -0.7194788548  0.225796159
#> effect_27   1.208858368 -0.0334174448  0.0381855477 -0.0096958438  0.356181296
#> effect_28   0.125917784 -0.0450263292 -0.0120116584  0.0027550845 -0.567959360
#> effect_29  -0.033158507  0.3827132856 -0.0453082824  0.1695237825 -0.047927221
#> effect_30   0.332438341 -0.0218704319  0.0941068153 -0.0005906256 -0.025555512
#> effect_31   0.156236504  0.0503645270  0.3872227071 -0.2845717017  0.050248172
#> effect_32   0.880931787 -0.1071706178  0.2642705799 -0.0100110021 -0.095348322
#> effect_33   0.065591654 -0.0875062880 -0.0497032093 -0.0272854139 -0.038338731
#> effect_34  -0.030443947 -0.0261506381 -0.0110156118  0.0105560581 -0.002151680
#> effect_35   0.182268477  0.0231221130 -0.0508852000 -0.0049845429 -0.388229982
#> effect_36   0.041163226  0.0688793771  0.0178657424  0.0143105558  0.031669158
#> effect_37   0.077432301  0.0005227554 -0.0347124236  0.0233833281 -0.004108641
#> effect_38   0.366140822  0.0314233692  0.3474082393 -0.0244601514  0.106865754
#> effect_39  -0.332603633 -0.0061173221 -0.0361861487  0.0307316615 -0.117091444
#> effect_40  -0.119390129  0.0221159197 -0.0471563134 -0.0167584267 -0.006106074
#> effect_41  -0.148813361 -0.1764662010 -0.0117275827 -0.0035296870  0.133125645
#> effect_42  -0.203576845  0.0034618468 -0.0191155655 -0.0071883617 -0.397618389
#> effect_43   0.010557348 -0.0129521258  0.0789448454 -0.2026315096  0.071037986
#> effect_44  -0.092209118 -0.0114981741  0.0584935173 -0.0071477383 -0.006551006
#> effect_45   0.140246374 -0.0438430101 -0.3079445533 -0.1911811860 -0.176207896
#> effect_46   0.340316387  0.2796807912 -0.0752913918 -0.0661987402 -0.069009022
#> effect_47  -0.139815755 -0.0659657460 -0.0459240470  0.0038750033 -0.091359367
#> effect_48   0.079082432 -0.0178159893  0.1220515870  0.0066574013  0.017926125
#> effect_49   0.059420406 -0.0656792870 -0.0020738983  0.0155240819  0.268287807
#> effect_50  -0.209327846 -0.0174315203  0.0541043952 -0.0149189822  0.019139322
#> effect_51  -0.041263714 -0.1732145092  0.0494199913  0.0338171904 -0.050830367
#> effect_52  -0.559836550 -1.8775669809  0.0605319849  0.1432032157  0.024991311
#> effect_53  -1.297068756  0.0725883183  0.0958043379  0.0076716989  0.076528188
#> effect_54  -2.297156913  0.0049535662  0.6125814387 -0.0015779850 -0.177715898
#> effect_55   1.168804364 -0.1672292274 -0.4094120053 -0.1787861539  0.111029882
#> effect_56   0.136882610  0.0040986484 -2.3049782418 -0.1310249946 -0.035364394
#> effect_57  -0.512729963  0.2020417278 -0.4350537047 -0.0005893659  0.003814162
#> effect_58   0.073359440  0.0089043620 -0.2958192893 -0.0084865343  0.007097779
#> effect_59   0.957226851 -0.0514683490 -1.1621906145 -0.3071043342 -0.081925415
#> effect_60   0.351682328  0.4092893714 -0.0103353205 -0.0574347412 -0.038526110
#> effect_61   0.107477537 -0.0362760104 -0.1222254900 -0.0001068110 -0.088880790
#> effect_62  -0.083575677 -2.4516182636  0.4459903756  0.0344635955  0.010725839
#> effect_63  -0.942446478 -0.0515845385  1.7700804367  0.0897859238  0.061681484
#> effect_64   0.314761129  2.0224756827 -0.2175490942 -0.2113091949  0.543277900
#> effect_65   0.386310692 -0.0715331799  0.0481669964  0.1227731385  0.034952347
#> effect_66  -0.926578377  0.0409371598  1.2694100188 -0.0135111102 -0.043363580
#> effect_67   0.099671021 -0.0577050634 -0.0106223992 -0.0092058461 -0.140874884
#> effect_68   0.218237679  0.0625474272 -0.0786599965  1.7790640814  0.117469019
#> effect_69  -0.007156847  0.4442014824 -0.0840821788 -0.0148063378  0.008103180
#> effect_70  -0.198578033 -0.3875682843 -0.0694242255 -0.0097936238 -1.636260307
#> effect_71   0.518418280 -0.0240036708  0.9739481198 -0.0339422320 -0.405355280
#> effect_72   0.338217717 -0.0442463029 -0.0120226842 -0.0591859206 -0.155495979
#> effect_73   0.006789461  0.0445413914  0.5625323119 -0.0127317349  0.230685319
#> effect_74  -0.012768504  0.0326172817  0.0428961837 -0.0028529871 -0.020784417
#> effect_75  -0.468132372  0.1351405581 -0.0494339388  0.0180961136 -0.050246983
#> effect_76  -0.469781367  0.1464291550  0.1528130816  0.0106725736 -0.287998600
#> effect_77   0.501052403 -0.1682654007  0.8904509406  0.0719379572 -0.110482067
#> effect_78  -0.071132013 -0.1088915144 -0.1866385928  0.0012222979  0.179703782
#> effect_79   0.097113013 -1.7316260629 -0.5608833900  0.0257194857  0.008742128
#> effect_80  -0.005248752 -2.4939917913  0.0504781981  0.0306605037  0.350979841
#> effect_81  -1.213784307  0.4105586885 -0.0285127454  0.0093786822 -1.207214272
#> effect_82  -0.688220209  0.0125808202 -0.0833812274 -0.0520285565 -0.163177306
#> effect_83   0.057969744 -0.0717076160 -0.3243937710 -0.0315198219  1.151039273
#> effect_84  -0.590569547  0.0173914316 -0.0226579202 -0.0034078267  0.037605511
#> effect_85  -0.052677170 -0.0006664301 -0.0665119348  0.0246444774  0.060361944
#> effect_86  -1.237977736  0.3741079209  0.9948653373 -0.0047898973 -0.273175324
#> effect_87  -0.146845343  0.2357716178  0.0217646388 -0.0173149346  0.022702404
#> effect_88   0.560721151 -0.0602792581 -0.0075110391  0.0162258868  0.003732451
#> effect_89  -0.164573021 -0.0340390462 -2.7631693716  0.6685130945  0.112982775
#> effect_90  -0.326724546  0.1984808290 -0.0617686130  0.0014698963 -0.020365393
#> effect_91   0.027705521  0.0213247890 -1.8738307364  0.0029169707  0.059975394
#> effect_92   0.020082190  0.6740592915  0.0591317195 -0.0139191096 -0.184590707
#> effect_93   0.280558745  0.1607706032 -0.1389649837 -0.0098341419  0.609059392
#> effect_94  -0.192283460 -0.0249558637 -0.0194489313  0.0483566043 -0.066385038
#> effect_95  -0.174364979 -0.1833361885  0.4176439634  0.0071864353 -0.198124473
#> effect_96   0.043154298 -0.1252520648  0.9943780810 -0.0346445781  0.018927286
#> effect_97  -2.331606978 -0.0171772290 -0.0097490767 -0.0276508108 -0.207486442
#> effect_98   0.262770271  0.0217639780  0.0101296374 -0.0415908265 -0.069665332
#> effect_99  -0.256046997 -0.0413018681 -0.0178102931  0.0049709058  0.021271248
#> effect_100 -0.008696654 -0.1614020094 -0.0790518030 -0.6987603229  0.114991375
#> effect_101  0.105676469  0.0148649141 -0.0119446066 -0.0130563383  0.077218415
#> effect_102 -1.049299117 -0.0254711540  0.0161203191 -0.0257453150  0.089970334
#> effect_103  0.019236889 -0.0387690793 -0.1281811748 -0.0027823123 -0.020220707
#> effect_104  1.322528329 -0.1016775999  0.1214423230  0.0129253039 -0.048822540
#> effect_105 -0.428612869 -0.0578670018 -0.0656433148 -0.0067513191 -2.059669721
#> effect_106  0.530932993  0.0661326390 -0.1383833749 -0.0259124661  0.062158535
#> effect_107  1.381320492  0.3566256836 -0.0384868254  0.0129351701 -0.096342202
#> effect_108 -0.184496355  0.0457887019 -0.0898147212  0.0021873157  0.330460179
#> effect_109 -0.077420530 -0.1154148672  0.1502963365  0.0592168553 -0.018702300
#> effect_110 -1.601014082 -0.0518276866 -0.0731837676 -0.0016725092  0.009674428
#> effect_111 -0.322528252  0.5101476827  0.8590177553  0.0282142877  0.100295097
#> effect_112 -0.203316977 -0.0307497033 -0.0169634793 -0.0076831414 -0.060343607
#> effect_113 -0.504524558 -0.0600822215  0.1188741835  0.0248165775  0.037100591
#> effect_114 -0.031379892 -0.0304839877 -0.0628108915  0.0019934308  0.144966208
#> effect_115 -0.143350965  0.7794953116  0.0230395262 -0.0113586595  0.017660643
#> effect_116 -0.024260164  0.0229685700  0.1625790338  0.0013506626  0.388407966
#> effect_117 -0.761739423 -0.0184011309  0.2329637233  0.0046761109 -0.034201056
#> effect_118  0.430711970  0.1003606278 -0.0238530425  0.0935137325  0.083601517
#> effect_119  0.321958274 -0.1643393170 -0.3171691606  0.0290283392  0.067446619
#> effect_120 -0.524080013 -0.0073153097 -0.0308756227 -0.0085490842  0.017846543
#> effect_121 -0.121872269  0.0512665613 -0.0284161439 -0.0007048046  0.174280055
#> effect_122  0.303492146 -0.1057154574 -0.0030602666  0.0181986172 -0.011210116
#> effect_123  1.881741148 -0.0325211028  0.1525717748 -0.0666313126 -0.012913934
#> effect_124  2.649314605  0.0277652532 -0.1343401424  0.0041478284  0.010429231
#> effect_125 -0.017907253 -0.1630958593 -0.3314566190  0.0047690226  0.062521821
#> effect_126 -0.016690338 -0.4463905056  0.0923468097  0.0378501250 -0.089982370
#> effect_127  0.591110825 -0.0184992424 -0.0028465061  0.0140507448 -0.161279033
#> effect_128  0.087058414  0.4553865152  0.0372656331  0.0185084900 -0.108777362
#> effect_129 -0.101369649 -0.0610630644 -0.2158679259  0.0082041334 -0.043553247
#> effect_130  0.401736956 -0.1610250911 -0.0817624168  0.0226511935  0.086473041
#> effect_131 -0.048021931 -0.0076166142 -0.0225269520  0.4487617889  0.031990241
#> effect_132  0.009738147  0.1266929757  0.6125443868  0.0024750450  0.005781389
#> effect_133  0.263468577 -0.0785334906  0.3790872228  0.0252195385  0.078372149
#> effect_134  0.102179343  0.0434378483  0.0036144536  0.0363649209  0.123014460
#> effect_135 -0.073183081 -0.0179287049  0.1470741812 -0.0434366259 -0.263982280
#> effect_136  0.107869235 -0.4089547553 -0.0627625676 -0.0014352407 -0.009589754
#> effect_137 -0.283798623 -0.0037280636 -0.0396556571 -0.0076001304 -0.016805176
#> effect_138  0.609881860  0.0220097460 -0.0090190421 -0.0197707422  0.173123573
#> effect_139  0.115611145 -0.0138280205 -0.2454337585 -0.0190878183 -0.032584555
#> effect_140 -0.038496122  0.0078747472 -0.0007381769 -0.0139168215  0.011129916
#> effect_141 -0.765740358 -0.0692329850  0.0238623021 -0.0141830397  0.135344723
#> effect_142  0.037940406 -0.0507078296 -0.1594467082  0.0241521100  0.215186314
#> effect_143 -0.178988990 -0.0084895428 -0.2644199460  0.0154431282  0.074338914
#> effect_144  0.355099292  0.1462356848  1.6254405901 -0.0014499037 -0.004033778
#> effect_145  1.762268475 -0.1528413591 -0.0512814040  0.0002592641  0.028395194
#> effect_146  1.156698803 -0.0599761417  0.0482464553  0.1297067503 -0.576440616
#> effect_147 -0.307308632 -0.0026929261 -0.0563212724 -0.0497495366 -0.197482006
#> effect_148 -0.233739273 -0.1272157110  0.2954032492  0.0035201650 -0.301894655
#> effect_149 -0.070287717 -0.1501401238 -0.0394891091  0.0145753185  0.118859309
#> effect_150  0.054813878 -0.0183495446  0.0498162032  0.0050740958  0.076570500
#> effect_151 -0.670541300  0.0004554463  0.1944235686  0.0011515790  0.281013935
#> effect_152 -0.035305890  0.2167959645  0.0513948562  0.0633327585  0.348969755
#> effect_153  0.072128913 -0.0461635250 -0.1124181923 -0.0092363012 -0.037846247
#> effect_154  0.359147088 -0.3092645400  0.0884885100 -0.0244691978  0.014882245
#> effect_155 -0.065507417  2.3576249158  0.8525265088  0.0294234113  0.316789917
#> effect_156 -0.214603988 -0.5727784441 -0.2675371237 -0.0683720845 -0.036406220
#> effect_157 -0.010268699  1.2727314060 -0.1105249594  0.0081686236 -0.032572382
#> effect_158  0.198150986 -0.0648117159  0.1195533049  0.0458452836  0.038443797
#> effect_159  2.056134342  0.5211208524  0.1026608102  0.4437396528  0.022381676
#> effect_160 -0.279236270 -0.0642035572 -0.0357138128  0.0149164118 -0.519738587
#> effect_161 -0.685359094  0.1342867472  0.3021648851 -0.0095029003 -0.177890452
#> effect_162 -0.484276900 -0.3682278628 -1.4849235323 -0.0076280359 -1.973594168
#> effect_163  0.065392226 -0.0233632494  0.0422420110 -0.0057528203 -0.153925907
#> effect_164 -0.028160200 -0.0386375752 -0.0981495970 -0.0050113406 -0.846570378
#> effect_165  0.259247194  0.0165123178  0.1077244830  0.0255050210  0.899331789
#> effect_166 -0.009358086 -0.0448384549  0.0359352334 -0.0072357775  0.084731963
#> effect_167  1.475833994 -0.0369053093 -0.0722452880 -0.0026670782 -0.050975188
#> effect_168  0.311529589 -0.2270694572  0.0437638128 -0.0056847892 -0.010894538
#> effect_169  0.223683220  0.1123978969  0.1383730204  0.0299503596  1.452396452
#> effect_170 -0.094581873 -0.1171128173 -0.0438685496  0.0297050692 -0.033449035
#> effect_171  0.359030399  0.0373904377 -0.0093860608  0.1470664435 -0.019994857
#> effect_172  0.095069309 -0.2220616274  0.0361707156 -0.0077844097 -0.007402834
#> effect_173 -0.260090979 -0.0290285304  0.0438129786 -0.0725490972 -0.111596815
#> effect_174 -0.399690183 -0.2549465837 -1.0476199436 -0.0080407388 -0.329703056
#> effect_175  0.586108529  0.0114877934  0.0940418846  0.0426990600  2.337347558
#> effect_176  0.215649684  1.3511370932  0.1292486625  0.1327632056  1.022466016
#> effect_177  0.221798458  0.0999676746  0.8837879476  0.0370030904  0.383621407
#> effect_178 -0.034502908 -1.1788759651 -0.2121607542  0.0038245191 -0.236327193
#> effect_179 -0.166571480  0.1174797817  0.0699836172 -0.0013770957  0.145365761
#> effect_180  1.314593266  0.1557592096  0.4668151828  0.0956309416  1.349620423
#> effect_181 -0.068431041  0.0792398071  0.0112436666  0.0820119429 -0.094481412
#> effect_182  0.481846012  0.6629911872  0.2730611425  0.0801017790 -0.001687994
#> effect_183  1.846459537  0.8865174750  0.1763746065  0.3171877915  2.300452720
#> effect_184 -0.489908661 -0.0967785897 -0.3189098681  0.0078522396 -1.684220890
#> effect_185 -0.832437944 -0.0135326451 -0.1239837521 -0.0228000079 -0.095012562
#> effect_186 -0.042018492 -0.0938613712  0.0452566678 -0.0168582754 -0.032673555
#> effect_187 -0.031788236  0.0319833369  0.3614814308 -0.0215313337 -0.255882991
#> effect_188 -1.386659045 -0.0707639589 -0.9257974205 -0.0034347082  0.013602470
#> effect_189  0.141883738 -0.0363159829 -0.0917703473  0.0011056136 -0.037919353
#> effect_190 -0.985991974 -0.0549989186 -0.1904776969 -0.0090819643 -0.053470892
#> effect_191 -0.558578980 -0.0258804838 -0.7068900248 -2.7762977121 -1.038403366
#> effect_192  2.209428614  0.0739657472  0.5694716082  0.1614606339  0.417195956
#> effect_193  0.182561017  0.0588692323  0.0213687403 -0.0018338827 -0.045137994
#> effect_194 -1.040312530 -0.2154758876 -1.3098906138  0.0071646155 -1.468057114
#> effect_195 -1.903384609 -0.1402900172 -0.1016940901 -0.0608518943 -0.046766021
#> effect_196 -1.219291281 -0.1553432701 -0.4975989534 -0.1031030392 -0.214577159
#> effect_197 -0.350480908 -0.0414067800 -0.7749050848  0.0065089857 -0.093706933
#> effect_198  0.080188849  0.3322262504  0.3561438743  0.0070636239 -0.021022811
#> effect_199 -0.081912315 -0.0077507306 -0.0475050244 -0.0027658277 -0.040124006
#> effect_200  0.080884048 -0.1981888252  0.0999157873  0.0210160281  0.024432053
#> 
#> $result$PosteriorSD
#>            condition_1 condition_2 condition_3 condition_4 condition_5
#> effect_1     0.7047980   0.2604573   0.3871147   0.1385210   0.2861580
#> effect_2     0.6049643   0.2515857   0.4058271   0.1380622   0.3337242
#> effect_3     0.4305180   0.2791525   0.3456969   0.1444937   0.3326644
#> effect_4     0.4314493   0.2506222   0.3372902   0.1630354   0.9637278
#> effect_5     0.4307246   0.3033676   0.4651657   0.2222805   0.2879184
#> effect_6     0.6458845   0.2957755   0.3485998   0.1385854   0.2803226
#> effect_7     0.4874239   0.3680146   0.3198005   0.2653752   0.5410053
#> effect_8     0.8737131   0.4793805   0.6436567   0.1514841   0.3190826
#> effect_9     0.4995710   0.2495531   0.8935086   0.2035441   0.3855503
#> effect_10    0.7608840   0.2754411   0.7859097   0.6592523   0.4632482
#> effect_11    0.4462114   0.2745777   0.4055051   0.1391386   0.3159257
#> effect_12    0.4321666   0.5700635   0.4593950   0.1387890   0.3723546
#> effect_13    0.4966147   0.2573612   0.4365083   0.1544785   0.2907701
#> effect_14    0.4774645   0.3740869   0.4452723   0.3041428   0.6847084
#> effect_15    0.4623660   0.3526214   0.3291710   0.2061079   0.5250552
#> effect_16    0.6658014   0.3267231   0.4797205   0.1856434   0.2803383
#> effect_17    0.4290650   0.3374291   0.3649877   0.1635372   0.2810967
#> effect_18    0.4819740   0.2934490   0.3231351   0.1384652   0.3330196
#> effect_19    0.4949839   0.4078978   0.3825384   0.1578787   0.8797771
#> effect_20    0.5052101   0.2612017   0.4353853   0.1500698   0.9715362
#> effect_21    0.4613701   0.2621047   0.3187718   0.1743839   0.4717749
#> effect_22    0.7497479   0.5307554   0.3176394   0.2257943   0.3498188
#> effect_23    0.4291540   0.3416992   0.3405924   0.1389763   0.2813287
#> effect_24    0.5502340   0.2787217   0.3195589   0.1746423   0.4451198
#> effect_25    0.4610495   0.4005213   0.3902408   0.1710114   0.6161862
#> effect_26    0.8061815   0.2623330   0.3904003   1.1395263   0.5888186
#> effect_27    0.9982281   0.2803489   0.3356762   0.1554312   0.7333475
#> effect_28    0.4873680   0.3005704   0.3195419   0.1396325   0.9056725
#> effect_29    0.4337475   0.7778307   0.3423246   0.5673092   0.3224369
#> effect_30    0.6555210   0.2634506   0.4009334   0.1380819   0.2939886
#> effect_31    0.5116508   0.3104618   0.7256296   0.7421148   0.3258839
#> effect_32    0.9334380   0.4173720   0.6092524   0.1564179   0.3987285
#> effect_33    0.4471187   0.3812687   0.3467876   0.2189386   0.3090048
#> effect_34    0.4329907   0.2692023   0.3192397   0.1581595   0.2804005
#> effect_35    0.5333557   0.2650596   0.3480298   0.1431582   0.7636744
#> effect_36    0.4363255   0.3458266   0.3218211   0.1710541   0.3006687
#> effect_37    0.4537282   0.2486049   0.3327320   0.2046041   0.2806741
#> effect_38    0.6804304   0.2771430   0.6907679   0.2085863   0.4175078
#> effect_39    0.6556459   0.2498448   0.3339557   0.2313345   0.4339316
#> effect_40    0.4823785   0.2637611   0.3441703   0.1799718   0.2811280
#> effect_41    0.5055727   0.5310367   0.3194531   0.1406505   0.4590959
#> effect_42    0.5513240   0.2489975   0.3224161   0.1482318   0.7722311
#> effect_43    0.4293620   0.2540707   0.3812426   0.6231450   0.3587379
#> effect_44    0.4630064   0.2529366   0.3563981   0.1481264   0.2812530
#> effect_45    0.4986519   0.2984161   0.6536352   0.6044404   0.5227011
#> effect_46    0.6614438   0.6679441   0.3766198   0.3424709   0.3554310
#> effect_47    0.4983071   0.3402263   0.3429344   0.1411776   0.3921741
#> effect_48    0.4547109   0.2587034   0.4380638   0.1468860   0.2872532
#> effect_49    0.4440074   0.3396755   0.3176850   0.1754443   0.6401278
#> effect_50    0.5561746   0.2582929   0.3514953   0.1732462   0.2881912
#> effect_51    0.4363611   0.5261514   0.3464925   0.2421879   0.3267583
#> effect_52    0.8019829   1.2330060   0.3587382   0.5186105   0.2934257
#> effect_53    1.0058902   0.3529452   0.4031727   0.1495170   0.3677367
#> effect_54    0.9078210   0.2494166   0.8830583   0.1385451   0.5248204
#> effect_55    0.9934802   0.5170524   0.7440093   0.5834857   0.4242279
#> effect_56    0.4959670   0.2491585   0.9871193   0.4945192   0.3051674
#> effect_57    0.7757000   0.5681044   0.7643715   0.1380816   0.2806220
#> effect_58    0.4513649   0.2512228   0.6416733   0.1518018   0.2814184
#> effect_59    0.9540775   0.3125356   1.0767747   0.7710913   0.3766255
#> effect_60    0.6698816   0.8028837   0.3190479   0.3175350   0.3092525
#> effect_61    0.4735778   0.2851058   0.4382947   0.1380086   0.3880931
#> effect_62    0.4574577   1.1158288   0.7727820   0.2444310   0.2828404
#> effect_63    0.9503616   0.3127544   1.0883343   0.4034540   0.3436211
#> effect_64    0.6420136   1.2134002   0.5575758   0.6369315   0.8886228
#> effect_65    0.6947975   0.3509219   0.3451990   0.4775513   0.3046504
#> effect_66    0.9462230   0.2932002   1.0918653   0.1682072   0.3158642
#> effect_67    0.4680618   0.3243754   0.3191273   0.1539291   0.4709751
#> effect_68    0.5636771   0.3336565   0.3808801   1.4293645   0.4345328
#> effect_69    0.4290903   0.8340509   0.3878365   0.1728389   0.2817562
#> effect_70    0.5471054   0.7824976   0.3693393   0.1557356   1.2002743
#> effect_71    0.7789809   0.2662302   1.0342043   0.2426226   0.7791758
#> effect_72    0.6598719   0.2991485   0.3195454   0.3226322   0.4928657
#> effect_73    0.4290671   0.2996856   0.8531816   0.1654756   0.5949589
#> effect_74    0.4295943   0.2790506   0.3399872   0.1397483   0.2895467
#> effect_75    0.7489282   0.4657682   0.3465070   0.1849243   0.3258821
#> effect_76    0.7499517   0.4843074   0.4783874   0.1585372   0.6624471
#> effect_77    0.7688708   0.5186377   1.0080981   0.3580727   0.4233463
#> effect_78    0.4501114   0.4204522   0.5208531   0.1383302   0.5276031
#> effect_79    0.4663034   1.2443553   0.8521528   0.2132195   0.2819935
#> effect_80    0.4289834   1.1039061   0.3476001   0.2310814   0.7282554
#> effect_81    0.9987561   0.8040507   0.3280268   0.1544543   1.1654687
#> effect_82    0.8639408   0.2537695   0.3869311   0.3014097   0.5040903
#> effect_83    0.4433117   0.3512565   0.6694401   0.2341296   1.1538250
#> effect_84    0.8180688   0.2582505   0.3243008   0.1404751   0.3080422
#> effect_85    0.4408939   0.2486106   0.3658034   0.2092662   0.3415232
#> effect_86    1.0011750   0.7694565   1.0400167   0.1427804   0.6457451
#> effect_87    0.5039732   0.6136301   0.3237983   0.1820277   0.2912439
#> effect_88    0.8024574   0.3293035   0.3183798   0.1780123   0.2806083
#> effect_89    0.5185475   0.2813679   0.8860408   1.1059241   0.4273648
#> effect_90    0.6511877   0.5630823   0.3601756   0.1384741   0.2891925
#> effect_91    0.4322892   0.2627694   1.0747012   0.1398260   0.3409107
#> effect_92    0.4306713   0.9994204   0.3571267   0.1696550   0.5343922
#> effect_93    0.6150622   0.5070748   0.4603930   0.1558622   0.9325699
#> effect_94    0.5417939   0.2675277   0.3225810   0.2901026   0.3511750
#> effect_95    0.5267201   0.5412265   0.7506466   0.1482268   0.5528141
#> effect_96    0.4370448   0.4490638   1.0398845   0.2450572   0.2880235
#> effect_97    0.9019005   0.2580254   0.3188921   0.2202658   0.5652393
#> effect_98    0.6006549   0.2633167   0.3189922   0.2684560   0.3564993
#> effect_99    0.5951478   0.2938485   0.3217956   0.1431314   0.2899658
#> effect_100   0.4292005   0.5080577   0.3813789   1.1261750   0.4305808
#> effect_101   0.4722860   0.2557434   0.3195208   0.1666075   0.3688718
#> effect_102   0.9743052   0.2682437   0.3210531   0.2133143   0.3898877
#> effect_103   0.4305228   0.2893887   0.4461909   0.1396644   0.2890716
#> effect_104   1.0074142   0.4074520   0.4372543   0.1661495   0.3237590
#> effect_105   0.7236089   0.3246850   0.3647600   0.1471188   1.1554729
#> effect_106   0.7860944   0.3405471   0.4596309   0.2139270   0.3443820
#> effect_107   1.0098078   0.7520287   0.3359413   0.1661840   0.4003583
#> effect_108   0.5352306   0.3019667   0.3952957   0.1390370   0.7076784
#> effect_109   0.4537212   0.4320064   0.4751404   0.3227217   0.2878472
#> effect_110   1.0059787   0.3132124   0.3739824   0.1386112   0.2823709
#> effect_111   0.6479853   0.8880620   0.9970160   0.2223066   0.4068248
#> effect_112   0.5511047   0.2760831   0.3214149   0.1495481   0.3414942
#> effect_113   0.7709146   0.3289257   0.4338400   0.2099005   0.3073855
#> effect_114   0.4332448   0.2756684   0.3613971   0.1388637   0.4771699
#> effect_115   0.5011469   1.0568821   0.3245209   0.1607967   0.2870551
#> effect_116   0.4314963   0.2648589   0.4908786   0.1384015   0.7638380
#> effect_117   0.8935659   0.2593420   0.5751270   0.1425653   0.3037173
#> effect_118   0.7249922   0.4050539   0.3250010   0.4124055   0.3793894
#> effect_119   0.6475490   0.5126086   0.6625574   0.2252424   0.3528937
#> effect_120   0.7822168   0.2503768   0.3297326   0.1519828   0.2871935
#> effect_121   0.4842631   0.3121559   0.3279594   0.1381141   0.5199815
#> effect_122   0.6332493   0.4147571   0.3177527   0.1853051   0.2830722
#> effect_123   0.9770223   0.2788956   0.4780765   0.3436660   0.2839642
#> effect_124   0.8549502   0.2715429   0.4543197   0.1416248   0.2827033
#> effect_125   0.4303018   0.5106863   0.6760809   0.1427406   0.3449624
#> effect_126   0.4301132   0.8359427   0.3986170   0.2560078   0.3899075
#> effect_127   0.8183448   0.2594506   0.3177358   0.1701244   0.5013340
#> effect_128   0.4596547   0.8436428   0.3348762   0.1864572   0.4205979
#> effect_129   0.4692432   0.3308068   0.5556311   0.1509940   0.3161313
#> effect_130   0.7055110   0.5074712   0.3848469   0.2018876   0.3841247
#> effect_131   0.4389294   0.2505251   0.3242260   0.9268748   0.3010468
#> effect_132   0.4292871   0.4515253   0.8830371   0.1393225   0.2810423
#> effect_133   0.6012250   0.3643058   0.7187072   0.2113835   0.3707705
#> effect_134   0.4698102   0.2976822   0.3178021   0.2509669   0.4433159
#> effect_135   0.4512647   0.2588252   0.4709677   0.2744701   0.6351327
#> effect_136   0.4738601   0.8025756   0.3613403   0.1384524   0.2823351
#> effect_137   0.6176590   0.2490615   0.3369839   0.1493233   0.2864344
#> effect_138   0.8277623   0.2636265   0.3187107   0.1911589   0.5183445
#> effect_139   0.4795412   0.2548117   0.5889604   0.1886137   0.3017533
#> effect_140   0.4354101   0.2506566   0.3176348   0.1696468   0.2830331
#> effect_141   0.8950640   0.3465059   0.3250066   0.1705973   0.4625170
#> effect_142   0.4352264   0.3111059   0.4868915   0.2074489   0.5752695
#> effect_143   0.5305990   0.2509868   0.6094104   0.1751492   0.3641411
#> effect_144   0.6723933   0.4839943   1.1006459   0.1384615   0.2806605
#> effect_145   0.9920993   0.4945933   0.3484499   0.1380207   0.2969637
#> effect_146   0.9918837   0.3287224   0.3452805   0.4918450   0.9113728
#> effect_147   0.6362306   0.2488390   0.3539474   0.2944270   0.5519520
#> effect_148   0.5766676   0.4524160   0.6412580   0.1406366   0.6776630
#> effect_149   0.4496437   0.4902814   0.3368340   0.1720056   0.4367430
#> effect_150   0.4418469   0.2592850   0.3469056   0.1433362   0.3678063
#> effect_151   0.8562048   0.2486026   0.5302943   0.1382938   0.6546391
#> effect_152   0.4343889   0.5884615   0.3485705   0.3344709   0.7262744
#> effect_153   0.4506690   0.3026554   0.4252448   0.1540213   0.3083571
#> effect_154   0.6753538   0.7017993   0.3935621   0.2086197   0.2851400
#> effect_155   0.4470747   1.1412080   0.9946386   0.2266616   0.6935201
#> effect_156   0.5606197   0.9341484   0.6126988   0.3484430   0.3064909
#> effect_157   0.4293349   1.2172300   0.4227230   0.1508936   0.3017388
#> effect_158   0.5467450   0.3380077   0.4347433   0.2821943   0.3091436
#> effect_159   0.9498079   0.8964835   0.4122577   0.9220013   0.2909515
#> effect_160   0.6139997   0.3368387   0.3335593   0.1732369   0.8716999
#> effect_161   0.8627053   0.4643431   0.6479669   0.1548349   0.5250652
#> effect_162   0.7588368   0.7636577   1.1042610   0.1493987   1.1699601
#> effect_163   0.4470145   0.2653768   0.3393680   0.1447675   0.4905480
#> effect_164   0.4324013   0.2891600   0.4062739   0.1432112   1.0566776
#> effect_165   0.5977731   0.2573411   0.4189934   0.2124325   1.0779936
#> effect_166   0.4292543   0.3002273   0.3337446   0.1483555   0.3812537
#> effect_167   1.0105188   0.2861759   0.3728157   0.1395317   0.3269764
#> effect_168   0.6395123   0.6022243   0.3408183   0.1446176   0.2829200
#> effect_169   0.5682513   0.4266867   0.4596173   0.2285487   1.1964146
#> effect_170   0.4645892   0.4349821   0.3409194   0.2276711   0.3027961
#> effect_171   0.6752687   0.2870061   0.3188002   0.5260335   0.2888844
#> effect_172   0.4649173   0.5955567   0.3339427   0.1498245   0.2815163
#> effect_173   0.5984641   0.2734324   0.3408657   0.3597027   0.4251396
#> effect_174   0.7041032   0.6379886   1.0534385   0.1505342   0.7069039
#> effect_175   0.8157846   0.2529289   0.4008478   0.2720769   1.0943528
#> effect_176   0.5614999   1.2292278   0.4476028   0.4980249   1.1201581
#> effect_177   0.5666692   0.4043369   1.0058079   0.2531398   0.7594212
#> effect_178   0.4341447   1.1984628   0.5513214   0.1410978   0.6019654
#> effect_179   0.5202100   0.4356235   0.3700248   0.1384171   0.4777720
#> effect_180   1.0069714   0.4992165   0.7883605   0.4174166   1.1872243
#> effect_181   0.4486297   0.3656499   0.3193066   0.3842278   0.3973057
#> effect_182   0.7573608   0.9927964   0.6184792   0.3793809   0.2803607
#> effect_183   0.9818373   1.1057365   0.5082104   0.7836227   1.1035026
#> effect_184   0.7622350   0.3984937   0.6642241   0.1500109   1.1987817
#> effect_185   0.9183797   0.2545570   0.4406289   0.2024403   0.3981776
#> effect_186   0.4366307   0.3931113   0.3422737   0.1803401   0.3018599
#> effect_187   0.4333579   0.2780332   0.7033768   0.1977218   0.6256159
#> effect_188   1.0099492   0.3494458   1.0197281   0.1405133   0.2843580
#> effect_189   0.4999658   0.2851735   0.3978597   0.1382714   0.3084530
#> effect_190   0.9609313   0.3192166   0.5255250   0.1535559   0.3307687
#> effect_191   0.8013071   0.2688192   0.9327596   1.2205477   1.1248840
#> effect_192   0.9231800   0.3555839   0.8574793   0.5528394   0.7896219
#> effect_193   0.5336018   0.3266021   0.3235815   0.1387328   0.3183848
#> effect_194   0.9725468   0.5866690   1.0959637   0.1481701   1.1973588
#> effect_195   0.9739345   0.4742942   0.4109737   0.3274268   0.3207373
#> effect_196   0.9993321   0.4985596   0.8103839   0.4347028   0.5744821
#> effect_197   0.6689957   0.2940354   0.9637574   0.1465229   0.3960338
#> effect_198   0.4553778   0.7267264   0.6986323   0.1479093   0.2897509
#> effect_199   0.4564290   0.2505929   0.3445237   0.1396451   0.3113912
#> effect_200   0.4558000   0.5626686   0.4086141   0.1958018   0.2928774
#> 
#> $result$lfsr
#>            condition_1 condition_2 condition_3 condition_4 condition_5
#> effect_1    0.62490090  0.94864843  0.87490628  0.98708867  0.93826052
#> effect_2    0.70066516  0.95410777  0.86560269  0.98758180  0.91484441
#> effect_3    0.85078791  0.94069222  0.89732319  0.98491447  0.91527617
#> effect_4    0.84822239  0.95499254  0.90269391  0.98105662  0.60215057
#> effect_5    0.85017140  0.93201787  0.83674577  0.97086296  0.93694515
#> effect_6    0.67036317  0.93464682  0.89557994  0.98704252  0.94610191
#> effect_7    0.78791763  0.91062440  0.91843394  0.96330642  0.83322061
#> effect_8    0.47008706  0.87278388  0.74405448  0.98330152  0.92096304
#> effect_9    0.77832444  0.95623777  0.57496753  0.97405428  0.89464475
#> effect_10   0.57857235  0.94212951  0.65592684  0.86995922  0.86471968
#> effect_11   0.82578528  0.94247064  0.86576099  0.98671437  0.92233848
#> effect_12   0.84652414  0.83903817  0.83955230  0.98691008  0.89969435
#> effect_13   0.78062838  0.95027742  0.85065430  0.98268849  0.93503681
#> effect_14   0.79609582  0.90862902  0.84640566  0.95614392  0.76829075
#> effect_15   0.80935355  0.91567076  0.90864486  0.97361935  0.83985641
#> effect_16   0.65528408  0.92417995  0.82964140  0.97709378  0.94596608
#> effect_17   0.85721101  0.92065234  0.88637850  0.98096462  0.94371802
#> effect_18   0.79235036  0.93546520  0.91417951  0.98713086  0.91513130
#> effect_19   0.78190709  0.89741623  0.87722454  0.98202320  0.65986122
#> effect_20   0.77397488  0.94827964  0.85119881  0.98360296  0.59629151
#> effect_21   0.81028013  0.94784177  0.92030826  0.97903294  0.86136162
#> effect_22   0.58806593  0.85405084  0.92453253  0.97025983  0.90842454
#> effect_23   0.85653899  0.91925048  0.90051622  0.98680133  0.94330932
#> effect_24   0.74051165  0.94085684  0.91882972  0.97898789  0.87179542
#> effect_25   0.81058021  0.89988086  0.87333438  0.97962413  0.80049912
#> effect_26   0.53808114  0.94773257  0.87325436  0.66976544  0.81270273
#> effect_27   0.27417977  0.94023777  0.90379923  0.98249925  0.74381686
#> effect_28   0.78796268  0.93297983  0.91885848  0.98647829  0.64298313
#> effect_29   0.84327755  0.74763962  0.89941188  0.89613309  0.91952732
#> effect_30   0.66309956  0.94720630  0.86801444  0.98754338  0.93309557
#> effect_31   0.76906895  0.92960598  0.69533826  0.84373812  0.91807547
#> effect_32   0.39682012  0.89423257  0.76313562  0.98230569  0.88961513
#> effect_33   0.82475389  0.90626336  0.89666277  0.97143496  0.92545808
#> effect_34   0.84476352  0.94466636  0.91938867  0.98196941  0.94558959
#> effect_35   0.75287935  0.94646990  0.89591872  0.98527739  0.72779158
#> effect_36   0.83883191  0.91789707  0.91567038  0.97961660  0.92949197
#> effect_37   0.81770681  0.95856460  0.90590257  0.97387450  0.94464516
#> effect_38   0.64403179  0.94146491  0.71666895  0.97319854  0.88243719
#> effect_39   0.66300504  0.95585386  0.90501323  0.96930490  0.87612432
#> effect_40   0.79201803  0.94706232  0.89825940  0.97806543  0.94365984
#> effect_41   0.77369702  0.85394558  0.91901006  0.98607188  0.86634775
#> effect_42   0.73971679  0.95716801  0.91497451  0.98400973  0.72315513
#> effect_43   0.85527951  0.95226324  0.87788504  0.88058873  0.90494508
#> effect_44   0.80876196  0.95305036  0.89108937  0.98403368  0.94343748
#> effect_45   0.77903889  0.93372578  0.73838138  0.88591463  0.84082753
#> effect_46   0.65860568  0.79873732  0.88025848  0.94867218  0.90623044
#> effect_47   0.77930734  0.91973380  0.89902847  0.98588683  0.89211625
#> effect_48   0.81671547  0.94954999  0.84990017  0.98432119  0.93742679
#> effect_49   0.82837463  0.91991457  0.92397662  0.97884832  0.78952520
#> effect_50   0.73618299  0.94976862  0.89388329  0.97923166  0.93675226
#> effect_51   0.83877552  0.85576927  0.89684077  0.96741838  0.91771053
#> effect_52   0.54197646  0.21053117  0.88978538  0.90884279  0.93342269
#> effect_53   0.24523288  0.91556468  0.86690933  0.98372334  0.90146888
#> effect_54   0.03589839  0.95643573  0.58359460  0.98707110  0.83995337
#> effect_55   0.28788010  0.85914181  0.68368331  0.89173852  0.87985933
#> effect_56   0.78113560  0.95685730  0.06188251  0.91484852  0.92726839
#> effect_57   0.56567879  0.83980130  0.67041089  0.98754397  0.94478861
#> effect_58   0.82014445  0.95442419  0.74517445  0.98323498  0.94316282
#> effect_59   0.36611674  0.92890719  0.35770672  0.83392980  0.89805722
#> effect_60   0.65216177  0.73496610  0.91974867  0.95357881  0.92534349
#> effect_61   0.79939095  0.93846624  0.84978818  0.98776601  0.89367383
#> effect_62   0.81400588  0.08596198  0.66481152  0.96702564  0.94130199
#> effect_63   0.37195902  0.92883362  0.17001213  0.93593026  0.91087055
#> effect_64   0.67326517  0.17427607  0.79054980  0.87658497  0.65417822
#> effect_65   0.63281501  0.91622751  0.89762693  0.91896899  0.92751781
#> effect_66   0.37828747  0.93555313  0.31998223  0.98012136  0.92236554
#> effect_67   0.80419561  0.92495706  0.91959696  0.98279878  0.86167746
#> effect_68   0.73072594  0.92189288  0.87807021  0.32463368  0.87589236
#> effect_69   0.85700758  0.71859143  0.87454258  0.97930297  0.94265376
#> effect_70   0.74279529  0.74531032  0.88406009  0.98243927  0.25326993
#> effect_71   0.56278039  0.94594797  0.42872754  0.96734234  0.71935330
#> effect_72   0.65980063  0.93347165  0.91885259  0.95258975  0.85296235
#> effect_73   0.85719298  0.93328564  0.60724159  0.98061193  0.80999515
#> effect_74   0.85414437  0.94073112  0.90090782  0.98642761  0.93582858
#> effect_75   0.58875844  0.87759301  0.89683200  0.97721657  0.91807621
#> effect_76   0.58789366  0.87102806  0.83029394  0.98189729  0.77902919
#> effect_77   0.57166073  0.85855648  0.46224602  0.94551391  0.88019792
#> effect_78   0.82147174  0.89319286  0.80926367  0.98724396  0.83880294
#> effect_79   0.80576416  0.24998488  0.60803109  0.97241059  0.94232758
#> effect_80   0.85796772  0.07894520  0.89617513  0.96934864  0.74644774
#> effect_81   0.27251893  0.73436542  0.90958791  0.98269333  0.38987740
#> effect_82   0.48066693  0.95246537  0.87499888  0.95666152  0.84843363
#> effect_83   0.82921997  0.91611789  0.72925905  0.96882115  0.40964632
#> effect_84   0.52687230  0.94979138  0.91297493  0.98613662  0.92590589
#> effect_85   0.83228414  0.95849779  0.88594090  0.97308303  0.91170510
#> effect_86   0.26443810  0.75178415  0.42052852  0.98538586  0.78690825
#> effect_87   0.77492424  0.82164237  0.91348287  0.97771227  0.93473921
#> effect_88   0.54153780  0.92332750  0.92122437  0.97840330  0.94482833
#> effect_89   0.76387543  0.93985371  0.01618584  0.68884432  0.87865371
#> effect_90   0.66637297  0.84175018  0.88899257  0.98712395  0.93606482
#> effect_91   0.84625091  0.94752553  0.14457512  0.98639438  0.91194948
#> effect_92   0.85032726  0.61726339  0.89068152  0.97986392  0.83598387
#> effect_93   0.69325532  0.86280450  0.83906725  0.98241440  0.62462355
#> effect_94   0.74668030  0.94538134  0.91478803  0.95878208  0.90789272
#> effect_95   0.75778482  0.85011195  0.67939991  0.98401087  0.82824285
#> effect_96   0.83771202  0.88341352  0.42071862  0.96691584  0.93687057
#> effect_97   0.03255280  0.94991285  0.92005750  0.97120798  0.82294314
#> effect_98   0.70381727  0.94726867  0.91985717  0.96275089  0.90581465
#> effect_99   0.70783793  0.93532418  0.91570113  0.98528501  0.93555332
#> effect_100  0.85622775  0.86244534  0.87781550  0.67748777  0.87741596
#> effect_101  0.80050163  0.95121031  0.91889428  0.98040782  0.90103219
#> effect_102  0.33084391  0.94507351  0.91663344  0.97239446  0.89298882
#> effect_103  0.85077323  0.93691136  0.84596028  0.98646422  0.93614624
#> effect_104  0.23719035  0.89756551  0.85029258  0.98049025  0.91896785
#> effect_105  0.60975318  0.92485452  0.88650089  0.98426635  0.14455025
#> effect_106  0.55643908  0.91962850  0.83943769  0.97229011  0.91056872
#> effect_107  0.21915161  0.76026823  0.90361559  0.98048405  0.88899310
#> effect_108  0.75149860  0.93249879  0.87080953  0.98676812  0.75691230
#> effect_109  0.81771388  0.88927089  0.83188155  0.95257232  0.93699591
#> effect_110  0.15835146  0.92867965  0.88162574  0.98702466  0.94184798
#> effect_111  0.66878457  0.68843716  0.47519846  0.97085849  0.88652388
#> effect_112  0.73987668  0.94187766  0.91616951  0.98371653  0.91171670
#> effect_113  0.56987759  0.92345219  0.85194820  0.97297524  0.92621370
#> effect_114  0.84425246  0.94204018  0.88832341  0.98686545  0.85922610
#> effect_115  0.77710322  0.57392218  0.91275720  0.98147114  0.93757362
#> effect_116  0.84810515  0.94656053  0.82416299  0.98718192  0.72770345
#> effect_117  0.44757605  0.94921599  0.78139252  0.98544897  0.92797166
#> effect_118  0.60862485  0.89836782  0.91229158  0.93396882  0.89699938
#> effect_119  0.66911262  0.86077763  0.73325264  0.97035469  0.90722025
#> effect_120  0.55990554  0.95524546  0.90819450  0.98319723  0.93747083
#> effect_121  0.79047701  0.92903495  0.90964457  0.98748999  0.84194687
#> effect_122  0.67980484  0.89511340  0.92348623  0.97715153  0.94104871
#> effect_123  0.09605549  0.94079034  0.83044602  0.94843269  0.94014883
#> effect_124  0.01156815  0.94369355  0.84201719  0.98573902  0.94145656
#> effect_125  0.85147223  0.86148300  0.72537421  0.98539746  0.91033883
#> effect_126  0.85210926  0.71757461  0.86916048  0.96498193  0.89298126
#> effect_127  0.52660877  0.94915984  0.92359285  0.97978079  0.84954965
#> effect_128  0.81189609  0.71340775  0.90435883  0.97695496  0.88125259
#> effect_129  0.80315234  0.92283159  0.79155545  0.98340497  0.92224809
#> effect_130  0.62433314  0.86265971  0.87605242  0.97433517  0.89518926
#> effect_131  0.83495191  0.95509097  0.91304956  0.77467505  0.92930000
#> effect_132  0.85569778  0.88256122  0.58361188  0.98662212  0.94382145
#> effect_133  0.70340056  0.91184146  0.69965230  0.97272306  0.90030248
#> effect_134  0.80265460  0.93398097  0.92320896  0.96587538  0.87249520
#> effect_135  0.82024970  0.94948577  0.83391864  0.96165988  0.79183857
#> effect_136  0.79914924  0.73512452  0.88835445  0.98714083  0.94189171
#> effect_137  0.69134367  0.95703846  0.90290140  0.98376591  0.93804480
#> effect_138  0.51752190  0.94712462  0.92044024  0.97615485  0.84261933
#> effect_139  0.79436136  0.95178348  0.77406676  0.97658760  0.92894401
#> effect_140  0.84032343  0.95495815  0.92463427  0.97986537  0.94109073
#> effect_141  0.44581524  0.91767441  0.91228627  0.97969720  0.86500671
#> effect_142  0.84063294  0.92938867  0.82612426  0.97339170  0.81861700
#> effect_143  0.75491353  0.95464011  0.76304957  0.97889963  0.90285456
#> effect_144  0.65023364  0.87113988  0.20880599  0.98713369  0.94468167
#> effect_145  0.12043092  0.86733547  0.89566890  0.98769649  0.93143460
#> effect_146  0.29208948  0.92351932  0.89757714  0.91550389  0.63916604
#> effect_147  0.67758488  0.95753938  0.89247463  0.95797500  0.82860817
#> effect_148  0.72128705  0.88225238  0.74540859  0.98607695  0.77171935
#> effect_149  0.82197372  0.86888788  0.90300333  0.97944916  0.87503898
#> effect_150  0.83105094  0.94924549  0.89659172  0.98522728  0.90144211
#> effect_151  0.48883693  0.95859585  0.80450549  0.98727796  0.78273107
#> effect_152  0.84209413  0.83179277  0.89559733  0.95026509  0.74746675
#> effect_153  0.82087823  0.93226216  0.85612042  0.98278021  0.92575897
#> effect_154  0.64795498  0.78371515  0.87167328  0.97319287  0.93909278
#> effect_155  0.82480357  0.10267515  0.47789651  0.97011067  0.76396249
#> effect_156  0.73294880  0.66064801  0.76125612  0.94747136  0.92663621
#> effect_157  0.85542703  0.39184855  0.85734646  0.98342629  0.92895131
#> effect_158  0.74305857  0.92046225  0.85151015  0.96024561  0.92539384
#> effect_159  0.06620093  0.68351124  0.86244961  0.77671471  0.93492228
#> effect_160  0.69403670  0.92084633  0.90529879  0.97923329  0.66497921
#> effect_161  0.48198344  0.87809297  0.74161199  0.98261741  0.83985227
#> effect_162  0.58032984  0.75462799  0.25007220  0.98374932  0.16439089
#> effect_163  0.82487149  0.94632739  0.90131181  0.98484334  0.85389222
#> effect_164  0.84600477  0.93699357  0.86538306  0.98526241  0.52467174
#> effect_165  0.70592221  0.95028857  0.85916192  0.97254458  0.50374883
#> effect_166  0.85589141  0.93309832  0.90516499  0.98398176  0.89628644
#> effect_167  0.19171410  0.93807454  0.88223401  0.98652368  0.91761971
#> effect_168  0.67513580  0.82627713  0.90037086  0.98488214  0.94121388
#> effect_169  0.72740190  0.89108098  0.83944429  0.96978570  0.30869044
#> effect_170  0.80731306  0.88825502  0.90030595  0.96993690  0.92842474
#> effect_171  0.64802059  0.93777221  0.92024808  0.90695514  0.93627317
#> effect_172  0.80701503  0.82895968  0.90502257  0.98365619  0.94300891
#> effect_173  0.70541768  0.94292755  0.90034039  0.94517998  0.87950912
#> effect_174  0.62545373  0.81154453  0.40019501  0.98350298  0.75730110
#> effect_175  0.52904752  0.95305594  0.86805672  0.96209507  0.08919782
#> effect_176  0.73230872  0.36575838  0.84527569  0.91398598  0.45656444
#> effect_177  0.72855140  0.89860745  0.46497609  0.96549094  0.73007694
#> effect_178  0.84253746  0.42407692  0.79377792  0.98591402  0.80688436
#> effect_179  0.76263141  0.88803573  0.88369830  0.98716910  0.85898723
#> effect_180  0.23968201  0.86566367  0.65424794  0.93286041  0.34171605
#> effect_181  0.82307535  0.91140050  0.91926760  0.94006344  0.89015808
#> effect_182  0.58159353  0.62191732  0.75808830  0.94108838  0.94581027
#> effect_183  0.10292226  0.53164632  0.81558393  0.82957939  0.09575658
#> effect_184  0.57740947  0.90055629  0.73228858  0.98361571  0.23960106
#> effect_185  0.41705118  0.95194573  0.84865663  0.97424146  0.88982538
#> effect_186  0.83835152  0.90234533  0.89944399  0.97800207  0.92889063
#> effect_187  0.84402906  0.94112111  0.70906579  0.97504137  0.79621111
#> effect_188  0.21755049  0.91671109  0.44790109  0.98612237  0.93978139
#> effect_189  0.77801806  0.93844138  0.86953589  0.98730000  0.92571433
#> effect_190  0.35489041  0.92667106  0.80691338  0.98287421  0.91605213
#> effect_191  0.54260047  0.94482840  0.54060792  0.09215136  0.45062053
#> effect_192  0.04550150  0.91470028  0.60392651  0.89999033  0.71356747
#> effect_193  0.75269799  0.92421995  0.91370720  0.98694490  0.92126490
#> effect_194  0.33420250  0.83250504  0.30624744  0.98402374  0.30378604
#> effect_195  0.09198089  0.87458798  0.86307792  0.95165293  0.92025173
#> effect_196  0.27066840  0.86590167  0.63886048  0.92897936  0.81895820
#> effect_197  0.65284071  0.93525829  0.51079105  0.98440762  0.89064343
#> effect_198  0.81604961  0.77225982  0.71194111  0.98408321  0.93569390
#> effect_199  0.81501063  0.95502201  0.89804135  0.98647274  0.92436385
#> effect_200  0.81563077  0.84191027  0.86423416  0.97536690  0.93374594
#> 
#> 
#> $loglik
#> [1] -1641.129
#> 
#> attr(,"class")
#> [1] "mash_1by1"