Update the data object for mash analysis.

This function can update two parts of the mash data. The first one is setting the reference group, so the mash data can be used for commonbaseline analysis. The other one is updating the null correlation matrix.

mash_update_data(mashdata, ref = NULL, V = NULL)

Arguments

mashdata
mash data object ontaining the Bhat matrix, standard errors, V; created using mash_set_data

ref
the reference group. It could be a number between 1,..., R, R is number of conditions, or the name of reference group. If there is no reference group, it can be the string 'mean'.

V
an R by R matrix / [R x R x N] array of correlation matrix of error correlations

Value

a updated mash data object

Examples

simdata = simple_sims(50,5,1)
data = mash_set_data(simdata$Bhat, simdata$Shat)
mash_update_data(data, 'mean')

#> $Bhat
#>            condition_1-mean condition_2-mean condition_3-mean condition_4-mean
#> effect_1        0.149046793    -0.1300291792      -0.56490986      0.332875711
#> effect_2        1.184917431    -0.3997862079       0.52851723      0.059111384
#> effect_3        1.314164141    -2.3931334676       2.17930042     -1.512954360
#> effect_4       -0.465449933    -1.2468710981      -0.17603600      0.052299869
#> effect_5       -1.210698042     1.5724085253       0.61972202     -0.758376993
#> effect_6        0.209271049     1.0181493801       1.09519270     -1.330593856
#> effect_7       -0.710851704     0.3470898892       0.64582675      0.927486654
#> effect_8        0.512135897    -0.4236642605      -0.46077547     -0.357151858
#> effect_9       -0.309848478    -0.9736272272       0.63469327      0.478581279
#> effect_10      -1.110562161     0.2616370982       0.38703790      0.927406749
#> effect_11       1.079559718     0.4428828262      -1.72983572     -0.781154227
#> effect_12      -0.482962496    -0.1001585423       1.30727523     -2.085142170
#> effect_13       0.129551936     0.9859734098      -0.15198096     -0.372455918
#> effect_14      -0.186504866    -0.6418055093       1.15159076     -0.035497302
#> effect_15       0.232939892    -0.7128987839      -0.82481930     -0.629208502
#> effect_16      -0.268607678     0.4010904590       0.49528763     -1.361380601
#> effect_17       0.140333184    -0.3195701794       0.72419073      0.448550123
#> effect_18       1.458148236    -1.1744841656      -1.08621181     -0.407439757
#> effect_19       0.840211867    -0.3442518770      -0.86400635      1.754975549
#> effect_20      -0.322182353    -1.0376377638      -0.16120195      0.431172388
#> effect_21      -0.904104286     0.3026266226       1.96401755     -0.439926020
#> effect_22       0.069028430    -0.5839396436      -1.10899717      0.598192543
#> effect_23       0.010684824     0.2589896193       0.85548443     -1.550608307
#> effect_24      -0.123318985    -0.2654762489      -0.54463865      1.760179386
#> effect_25       0.035459649     1.3961481794      -1.39193008      0.350573869
#> effect_26       0.476546604    -0.7524459166       0.57519211      0.030670944
#> effect_27       1.892390803    -0.8358385114      -0.63871624      0.018707533
#> effect_28      -1.017137266    -0.7408548333       0.03075207      0.321890559
#> effect_29       0.762247910     0.1412897512       0.88738701     -1.468095873
#> effect_30      -0.193493805    -0.5340057715       0.03783011     -0.120391466
#> effect_31       0.437120939    -0.7349529862       0.34463651      0.100451314
#> effect_32      -0.713065420     0.9113555649      -1.38064829     -0.031298872
#> effect_33       0.091010687    -0.2454355084      -0.17079437      0.761338696
#> effect_34       1.546192074     0.0574069411       0.52687807     -0.933802968
#> effect_35      -2.209087844     0.7950896206       0.42973463      1.930353698
#> effect_36       0.253412300     0.5043288020       0.76406583     -0.668176206
#> effect_37       0.991163311     0.2774442606      -0.90540355     -0.169957083
#> effect_38      -1.222232573     1.3884077306       0.13694487     -0.581244658
#> effect_39      -0.207535628     0.7442682146       0.50312945     -1.552365855
#> effect_40       0.528055791     0.8708267143      -1.33990442     -0.218828821
#> effect_41      -0.237820976     0.1132571728       0.11106182      1.620448639
#> effect_42      -0.255690265    -0.0494983848       1.20939940     -0.430361249
#> effect_43      -0.057571857     0.7026683729       0.87728663     -1.455342844
#> effect_44       0.673623605    -0.1356101765      -0.75671972      0.230565784
#> effect_45       0.553625587     0.1254569922      -0.44233850      0.862444707
#> effect_46      -0.483550460    -0.3732829448       0.39967933      1.234778111
#> effect_47       0.933201709    -0.3393709932      -0.12283253     -1.562588883
#> effect_48      -0.033325173     1.0437249691      -0.88524930     -1.054604561
#> effect_49       1.061495259    -1.2020884333       1.21662153     -0.488124526
#> effect_50       0.930780170    -2.8001041708       1.18442522      1.028888597
#> effect_51       2.970387860    -0.6379932866      -1.29361872     -2.014247930
#> effect_52       0.329065728     1.9673448969      -1.14395410     -2.483120922
#> effect_53      -0.670491670     0.4463261025       0.21753926      0.844583776
#> effect_54       0.990001972     1.2384227302      -2.05008973     -1.526393368
#> effect_55      -1.223807989     1.6777763187       0.63577504     -0.481583948
#> effect_56      -0.235627440    -1.3674206019       0.63245662     -0.303974279
#> effect_57       0.211940833     1.2611128472      -1.63962135      0.262150072
#> effect_58       1.383255022     2.5181184915      -0.06843406     -1.305668219
#> effect_59      -0.796906413     0.4881883049       3.75825600     -1.828822184
#> effect_60      -0.770641825     2.5301585090      -0.37400382     -2.709342188
#> effect_61      -1.962336590     1.9283777145      -0.00266522     -1.043793038
#> effect_62       0.627971299     0.8026961815       0.23323883     -0.683230578
#> effect_63       0.824509373    -2.3686680355       0.65901358      2.152050368
#> effect_64       2.538139393    -2.9426179863      -1.58451631      1.202590528
#> effect_65      -0.032928143     0.4562750363       1.00895336     -0.992974575
#> effect_66       0.620767481     0.4324719234       0.30804420     -1.753470369
#> effect_67      -0.772305170     1.2036287993       0.61421097      0.083361724
#> effect_68       0.795323005    -1.8842687955       2.23628678     -0.544682288
#> effect_69       1.461979843    -0.3149202685       0.07982646     -1.718826107
#> effect_70      -1.110837625     0.4729346303       0.22551564      0.357086808
#> effect_71      -1.532943798    -0.1401250985       1.93157942     -1.079819309
#> effect_72      -2.429758989    -0.8528011395      -0.21542637      1.411303821
#> effect_73       1.285440870    -0.9944158846      -0.22543236     -0.205640872
#> effect_74      -1.722333497     1.7085905327       2.30046988     -1.767198769
#> effect_75       0.834359333     0.4530099957       0.35104766     -1.154522598
#> effect_76       2.857415269     0.3651642106      -0.29261786     -1.831785140
#> effect_77       1.455144672     0.6048161106       0.43121729     -0.506120390
#> effect_78      -0.133385927    -0.1753933649       0.60826054      1.171839294
#> effect_79       1.045810040     3.2752714196       0.20107210     -1.041305585
#> effect_80       1.458435767    -0.9084204591      -2.33149954      0.668996211
#> effect_81      -0.616632837    -1.5899567173       2.95451220     -0.231333987
#> effect_82      -0.912864797     0.1863796244       0.86487586     -0.614536314
#> effect_83       1.284715322    -0.4447646842       0.78336140     -0.928538624
#> effect_84      -0.740458541     1.4800470547      -0.70864759      0.316125087
#> effect_85       2.221073234     0.1116970387      -1.15417898     -0.788114661
#> effect_86       1.093633252    -0.1108696710      -1.88137269      2.194824434
#> effect_87       2.293272377     0.7017327232      -0.53383791     -0.513097565
#> effect_88       0.695609345     0.5720101513      -0.40216243      0.989348844
#> effect_89       1.728459864     1.3214189004      -0.76277989     -0.299001604
#> effect_90       0.026352446     0.4817947108      -0.77066098     -2.511844390
#> effect_91      -0.111347105    -2.2772908224       4.90080822      0.359811317
#> effect_92      -1.523140508     0.0660183871       1.25983747      0.549143323
#> effect_93      -0.450459066    -0.4731479757       1.05877342      0.169522739
#> effect_94      -0.297446811    -1.3836630582       1.33489428      0.125003782
#> effect_95       0.980687064     0.2977961493       0.56675901     -1.481913694
#> effect_96      -0.626836493     1.5623084155      -1.60483856     -0.262204556
#> effect_97      -0.893881124    -0.3047513426       0.72138522      0.234411828
#> effect_98      -0.969099289     2.3116619014      -1.70331261     -0.506605535
#> effect_99       3.145301824     0.3589639307      -0.83494734     -1.009964660
#> effect_100     -0.193101152    -0.0797179484       2.03794269     -0.343041100
#> effect_101      1.366562414    -0.9848041325      -0.47473324     -0.060088786
#> effect_102      0.946743278    -0.3606043358      -1.42384436     -0.174250548
#> effect_103     -0.469639485     1.6903268622       0.42879323     -0.731368948
#> effect_104     -0.998951019     0.1687296926       0.19673724      0.453454368
#> effect_105      1.171711081     1.3222479478      -1.21338676     -0.813160299
#> effect_106     -2.573369868    -0.0004512952      -0.13887582      1.275611164
#> effect_107      1.721064329    -1.0951586165      -0.76682561     -0.057893927
#> effect_108      0.225829517     0.7550582006      -0.04469374     -0.133994146
#> effect_109     -0.807170468     0.3181995282      -0.83421815      1.219880218
#> effect_110     -0.592665702    -0.4054900559      -0.14975772      0.820867257
#> effect_111     -1.027353301    -0.2952947674       0.12272082      0.801473622
#> effect_112      1.997448168    -1.8636182311      -0.14862221      1.153698098
#> effect_113     -3.192729813     1.2621311978       2.07030027     -1.567786081
#> effect_114      0.965607189    -0.3562607911       0.28779399     -0.904696930
#> effect_115      0.813660332    -0.8211255979      -0.50787150      1.059504276
#> effect_116     -0.392560558     1.1038962764      -0.26190426      0.257963114
#> effect_117     -0.838855531    -1.1771638281       1.57101939     -1.294925607
#> effect_118     -0.973770759    -0.6899322361       1.71243076     -0.262057273
#> effect_119     -0.005820702     0.8902570022       1.71242184     -1.126672754
#> effect_120     -3.693902812    -0.2558640828       2.80356647     -0.033939334
#> effect_121     -1.285316152     0.4629373991      -0.23947122      1.319516672
#> effect_122      2.693109050    -1.6341376713      -0.87057648     -0.214782746
#> effect_123      1.598157939     1.0101016333      -2.09324185     -0.267897080
#> effect_124     -0.036480521     0.0395778149       0.05365574      0.714683400
#> effect_125      2.280162216    -1.6318339850      -0.55688839      0.242128896
#> effect_126      0.314014726    -0.8561775297       0.01925154      0.646307955
#> effect_127     -0.115244242     1.0617076250       0.18983401     -0.649217116
#> effect_128     -0.416371788    -1.4940731537      -0.71895822      0.145154545
#> effect_129     -1.409677116     0.7887397371       1.15159656      0.894860287
#> effect_130      2.319474069    -0.3718720964      -0.71276960     -0.511436993
#> effect_131      0.964395636     0.9273091198      -1.28763521     -1.363316027
#> effect_132      0.340444685    -0.3719428326      -0.36486679     -0.344882869
#> effect_133      0.594739519     1.1639448745       0.57024820     -1.217995007
#> effect_134     -0.573304728     1.1062250551       0.64155549     -0.276636533
#> effect_135      0.715876878     0.6912101524      -0.79949855      1.563455600
#> effect_136     -0.043754867    -0.9215559936      -0.90761162      0.719021257
#> effect_137     -0.512708192    -1.0158382158      -0.20290192      0.953989069
#> effect_138      0.896177297    -0.8430658088      -0.52236678      0.855013859
#> effect_139      1.222982642    -0.3478376340      -0.46967349     -0.592409655
#> effect_140     -1.086637879     1.3450142565      -1.47694233      0.642709804
#> effect_141      1.739667002     0.3906327318      -0.73856991     -0.493779408
#> effect_142     -0.110829874     0.4254016295       0.04038214     -1.420137066
#> effect_143      0.259201203     0.2150371113       0.83450372     -0.191533703
#> effect_144      0.966938433    -0.1759618331       0.05459655     -1.203947435
#> effect_145      0.240763417    -0.7017830491      -0.17831617     -0.510273117
#> effect_146     -0.108840473    -0.0232958665       0.05880187      0.507609927
#> effect_147      1.532776111     0.0857600601      -2.31355086      0.041220914
#> effect_148     -0.252889522    -0.7067435896       0.34581645     -0.410468192
#> effect_149      0.268063668    -0.2957992011       0.41328717     -0.113939458
#> effect_150      1.213550625     0.1747380524      -1.53522379      0.078375810
#> effect_151      0.209618795    -0.2242811382       0.36998238     -0.785909502
#> effect_152     -1.143828833    -1.6721630177       1.97289493      0.012289719
#> effect_153      0.749438956    -1.3483040389      -0.08343031      1.243753447
#> effect_154     -0.578692313     1.1854670037      -0.58316976      0.264749467
#> effect_155     -1.034953073     0.7070808920       0.55537113     -0.119396844
#> effect_156     -0.707646157    -0.2077282673      -1.27358395      1.578730015
#> effect_157      0.527164216    -0.5921347236       0.93414064     -1.434043266
#> effect_158     -0.745432740     0.3221010854      -0.36510716     -0.335194564
#> effect_159     -1.305635335    -0.0741916418       0.17120322      0.324711553
#> effect_160      0.993422924    -0.3153651945      -0.09479663      0.644915485
#> effect_161      0.118643343    -1.6312114757       1.17040182     -0.416263148
#> effect_162     -0.005658679    -0.7045402262       1.28114392     -0.267698280
#> effect_163      0.367034712    -0.5246487878      -0.52235595      1.045366542
#> effect_164      0.975015037    -0.1178265171      -1.09653556      1.275816380
#> effect_165      0.551482095     0.4934253798      -1.11935418      1.211630386
#> effect_166     -0.342683829    -0.7512714577       0.95007092     -0.382312599
#> effect_167     -0.695897860     2.0253315149       0.01337345     -1.170855568
#> effect_168     -0.092114181     0.7540874987       0.30328852     -0.851922818
#> effect_169      0.969737437     0.0169505483      -0.61771301      0.632457824
#> effect_170      0.692236774    -1.0771496060       0.67281186     -1.289099740
#> effect_171     -0.690474869     0.6863137574      -0.29945941     -0.304752248
#> effect_172      1.212892269     1.3219462794      -0.64802671     -0.362874242
#> effect_173      0.797118762     0.6532795221      -0.89552160      0.302690827
#> effect_174     -0.388538636    -0.9148550728       0.55349839      0.590464475
#> effect_175     -0.551685862     0.2104284198      -0.05326767     -1.107238862
#> effect_176     -1.074558478    -0.0314383084       0.49675727     -0.312583943
#> effect_177      0.513135657     0.4650320029      -1.30705361     -0.702688874
#> effect_178     -0.603545594    -2.3578097802       1.61852088      0.429203163
#> effect_179     -0.147444442    -1.3266134735      -0.87886975      1.756321363
#> effect_180      0.766435393     0.2234853740      -0.54333002      0.252173424
#> effect_181      1.856909056    -1.4741187455       0.55316061      0.304108318
#> effect_182      0.016065836     0.7221910759      -0.45784711     -0.228447734
#> effect_183     -1.013421679     0.0335853290       1.61356177     -0.612187802
#> effect_184     -0.696834016    -1.5068732939      -0.38526853      2.016180683
#> effect_185      0.534394315    -0.4783414658      -1.11363325      1.056057426
#> effect_186     -0.467401702     0.9563456200      -1.23553230      0.125286979
#> effect_187     -0.537921118    -0.5626728241       0.03117742      1.418130955
#> effect_188     -1.310120850    -0.6744740399       1.22593033      1.212616850
#> effect_189      0.328441716    -0.7429047830      -0.03868814      0.216474434
#> effect_190      0.200862688     0.2044164739      -0.55475917      0.725664272
#> effect_191      1.139549584    -0.5128588052       0.57117234     -0.314898802
#> effect_192      0.226026171    -0.5611844755       0.79947731     -1.170623058
#> effect_193      0.003104987    -0.4775366236      -0.43125614     -0.638785108
#> effect_194      0.427623367     0.0259502987      -0.75208639      0.673570729
#> effect_195      0.144932928     0.9918596066       0.07873947     -1.520103620
#> effect_196      0.022638657     1.2934084585      -1.36201792      0.004925233
#> effect_197      0.913991781    -1.0733324660       0.32487269     -0.513160667
#> effect_198      0.798500215    -1.1951223916      -0.29588003      1.339720439
#> effect_199     -0.553544236     1.7849645358      -0.22886022      0.008234427
#> effect_200     -0.159295811     0.1146373390       0.21566972      0.228044267
#> 
#> $Shat
#>             [,1]      [,2]      [,3]      [,4]
#>   [1,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [2,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [3,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [4,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [5,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [6,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [7,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [8,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [9,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [10,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [11,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [12,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [13,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [14,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [15,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [16,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [17,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [18,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [19,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [20,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [21,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [22,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [23,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [24,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [25,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [26,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [27,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [28,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [29,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [30,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [31,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [32,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [33,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [34,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [35,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [36,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [37,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [38,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [39,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [40,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [41,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [42,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [43,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [44,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [45,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [46,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [47,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [48,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [49,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [50,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [51,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [52,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [53,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [54,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [55,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [56,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [57,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [58,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [59,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [60,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [61,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [62,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [63,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [64,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [65,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [66,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [67,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [68,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [69,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [70,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [71,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [72,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [73,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [74,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [75,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [76,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [77,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [78,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [79,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [80,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [81,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [82,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [83,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [84,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [85,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [86,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [87,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [88,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [89,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [90,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [91,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [92,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [93,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [94,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [95,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [96,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [97,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [98,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [99,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [100,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [101,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [102,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [103,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [104,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [105,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [106,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [107,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [108,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [109,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [110,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [111,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [112,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [113,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [114,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [115,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [116,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [117,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [118,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [119,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [120,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [121,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [122,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [123,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [124,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [125,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [126,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [127,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [128,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [129,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [130,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [131,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [132,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [133,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [134,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [135,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [136,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [137,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [138,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [139,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [140,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [141,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [142,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [143,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [144,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [145,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [146,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [147,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [148,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [149,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [150,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [151,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [152,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [153,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [154,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [155,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [156,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [157,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [158,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [159,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [160,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [161,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [162,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [163,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [164,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [165,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [166,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [167,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [168,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [169,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [170,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [171,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [172,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [173,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [174,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [175,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [176,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [177,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [178,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [179,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [180,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [181,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [182,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [183,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [184,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [185,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [186,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [187,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [188,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [189,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [190,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [191,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [192,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [193,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [194,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [195,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [196,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [197,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [198,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [199,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [200,] 0.8944272 0.8944272 0.8944272 0.8944272
#> 
#> $Shat_orig
#>            condition_1 condition_2 condition_3 condition_4 condition_5
#> effect_1             1           1           1           1           1
#> effect_2             1           1           1           1           1
#> effect_3             1           1           1           1           1
#> effect_4             1           1           1           1           1
#> effect_5             1           1           1           1           1
#> effect_6             1           1           1           1           1
#> effect_7             1           1           1           1           1
#> effect_8             1           1           1           1           1
#> effect_9             1           1           1           1           1
#> effect_10            1           1           1           1           1
#> effect_11            1           1           1           1           1
#> effect_12            1           1           1           1           1
#> effect_13            1           1           1           1           1
#> effect_14            1           1           1           1           1
#> effect_15            1           1           1           1           1
#> effect_16            1           1           1           1           1
#> effect_17            1           1           1           1           1
#> effect_18            1           1           1           1           1
#> effect_19            1           1           1           1           1
#> effect_20            1           1           1           1           1
#> effect_21            1           1           1           1           1
#> effect_22            1           1           1           1           1
#> effect_23            1           1           1           1           1
#> effect_24            1           1           1           1           1
#> effect_25            1           1           1           1           1
#> effect_26            1           1           1           1           1
#> effect_27            1           1           1           1           1
#> effect_28            1           1           1           1           1
#> effect_29            1           1           1           1           1
#> effect_30            1           1           1           1           1
#> effect_31            1           1           1           1           1
#> effect_32            1           1           1           1           1
#> effect_33            1           1           1           1           1
#> effect_34            1           1           1           1           1
#> effect_35            1           1           1           1           1
#> effect_36            1           1           1           1           1
#> effect_37            1           1           1           1           1
#> effect_38            1           1           1           1           1
#> effect_39            1           1           1           1           1
#> effect_40            1           1           1           1           1
#> effect_41            1           1           1           1           1
#> effect_42            1           1           1           1           1
#> effect_43            1           1           1           1           1
#> effect_44            1           1           1           1           1
#> effect_45            1           1           1           1           1
#> effect_46            1           1           1           1           1
#> effect_47            1           1           1           1           1
#> effect_48            1           1           1           1           1
#> effect_49            1           1           1           1           1
#> effect_50            1           1           1           1           1
#> effect_51            1           1           1           1           1
#> effect_52            1           1           1           1           1
#> effect_53            1           1           1           1           1
#> effect_54            1           1           1           1           1
#> effect_55            1           1           1           1           1
#> effect_56            1           1           1           1           1
#> effect_57            1           1           1           1           1
#> effect_58            1           1           1           1           1
#> effect_59            1           1           1           1           1
#> effect_60            1           1           1           1           1
#> effect_61            1           1           1           1           1
#> effect_62            1           1           1           1           1
#> effect_63            1           1           1           1           1
#> effect_64            1           1           1           1           1
#> effect_65            1           1           1           1           1
#> effect_66            1           1           1           1           1
#> effect_67            1           1           1           1           1
#> effect_68            1           1           1           1           1
#> effect_69            1           1           1           1           1
#> effect_70            1           1           1           1           1
#> effect_71            1           1           1           1           1
#> effect_72            1           1           1           1           1
#> effect_73            1           1           1           1           1
#> effect_74            1           1           1           1           1
#> effect_75            1           1           1           1           1
#> effect_76            1           1           1           1           1
#> effect_77            1           1           1           1           1
#> effect_78            1           1           1           1           1
#> effect_79            1           1           1           1           1
#> effect_80            1           1           1           1           1
#> effect_81            1           1           1           1           1
#> effect_82            1           1           1           1           1
#> effect_83            1           1           1           1           1
#> effect_84            1           1           1           1           1
#> effect_85            1           1           1           1           1
#> effect_86            1           1           1           1           1
#> effect_87            1           1           1           1           1
#> effect_88            1           1           1           1           1
#> effect_89            1           1           1           1           1
#> effect_90            1           1           1           1           1
#> effect_91            1           1           1           1           1
#> effect_92            1           1           1           1           1
#> effect_93            1           1           1           1           1
#> effect_94            1           1           1           1           1
#> effect_95            1           1           1           1           1
#> effect_96            1           1           1           1           1
#> effect_97            1           1           1           1           1
#> effect_98            1           1           1           1           1
#> effect_99            1           1           1           1           1
#> effect_100           1           1           1           1           1
#> effect_101           1           1           1           1           1
#> effect_102           1           1           1           1           1
#> effect_103           1           1           1           1           1
#> effect_104           1           1           1           1           1
#> effect_105           1           1           1           1           1
#> effect_106           1           1           1           1           1
#> effect_107           1           1           1           1           1
#> effect_108           1           1           1           1           1
#> effect_109           1           1           1           1           1
#> effect_110           1           1           1           1           1
#> effect_111           1           1           1           1           1
#> effect_112           1           1           1           1           1
#> effect_113           1           1           1           1           1
#> effect_114           1           1           1           1           1
#> effect_115           1           1           1           1           1
#> effect_116           1           1           1           1           1
#> effect_117           1           1           1           1           1
#> effect_118           1           1           1           1           1
#> effect_119           1           1           1           1           1
#> effect_120           1           1           1           1           1
#> effect_121           1           1           1           1           1
#> effect_122           1           1           1           1           1
#> effect_123           1           1           1           1           1
#> effect_124           1           1           1           1           1
#> effect_125           1           1           1           1           1
#> effect_126           1           1           1           1           1
#> effect_127           1           1           1           1           1
#> effect_128           1           1           1           1           1
#> effect_129           1           1           1           1           1
#> effect_130           1           1           1           1           1
#> effect_131           1           1           1           1           1
#> effect_132           1           1           1           1           1
#> effect_133           1           1           1           1           1
#> effect_134           1           1           1           1           1
#> effect_135           1           1           1           1           1
#> effect_136           1           1           1           1           1
#> effect_137           1           1           1           1           1
#> effect_138           1           1           1           1           1
#> effect_139           1           1           1           1           1
#> effect_140           1           1           1           1           1
#> effect_141           1           1           1           1           1
#> effect_142           1           1           1           1           1
#> effect_143           1           1           1           1           1
#> effect_144           1           1           1           1           1
#> effect_145           1           1           1           1           1
#> effect_146           1           1           1           1           1
#> effect_147           1           1           1           1           1
#> effect_148           1           1           1           1           1
#> effect_149           1           1           1           1           1
#> effect_150           1           1           1           1           1
#> effect_151           1           1           1           1           1
#> effect_152           1           1           1           1           1
#> effect_153           1           1           1           1           1
#> effect_154           1           1           1           1           1
#> effect_155           1           1           1           1           1
#> effect_156           1           1           1           1           1
#> effect_157           1           1           1           1           1
#> effect_158           1           1           1           1           1
#> effect_159           1           1           1           1           1
#> effect_160           1           1           1           1           1
#> effect_161           1           1           1           1           1
#> effect_162           1           1           1           1           1
#> effect_163           1           1           1           1           1
#> effect_164           1           1           1           1           1
#> effect_165           1           1           1           1           1
#> effect_166           1           1           1           1           1
#> effect_167           1           1           1           1           1
#> effect_168           1           1           1           1           1
#> effect_169           1           1           1           1           1
#> effect_170           1           1           1           1           1
#> effect_171           1           1           1           1           1
#> effect_172           1           1           1           1           1
#> effect_173           1           1           1           1           1
#> effect_174           1           1           1           1           1
#> effect_175           1           1           1           1           1
#> effect_176           1           1           1           1           1
#> effect_177           1           1           1           1           1
#> effect_178           1           1           1           1           1
#> effect_179           1           1           1           1           1
#> effect_180           1           1           1           1           1
#> effect_181           1           1           1           1           1
#> effect_182           1           1           1           1           1
#> effect_183           1           1           1           1           1
#> effect_184           1           1           1           1           1
#> effect_185           1           1           1           1           1
#> effect_186           1           1           1           1           1
#> effect_187           1           1           1           1           1
#> effect_188           1           1           1           1           1
#> effect_189           1           1           1           1           1
#> effect_190           1           1           1           1           1
#> effect_191           1           1           1           1           1
#> effect_192           1           1           1           1           1
#> effect_193           1           1           1           1           1
#> effect_194           1           1           1           1           1
#> effect_195           1           1           1           1           1
#> effect_196           1           1           1           1           1
#> effect_197           1           1           1           1           1
#> effect_198           1           1           1           1           1
#> effect_199           1           1           1           1           1
#> effect_200           1           1           1           1           1
#> 
#> $Shat_alpha
#>        [,1] [,2] [,3] [,4]
#>   [1,]    1    1    1    1
#>   [2,]    1    1    1    1
#>   [3,]    1    1    1    1
#>   [4,]    1    1    1    1
#>   [5,]    1    1    1    1
#>   [6,]    1    1    1    1
#>   [7,]    1    1    1    1
#>   [8,]    1    1    1    1
#>   [9,]    1    1    1    1
#>  [10,]    1    1    1    1
#>  [11,]    1    1    1    1
#>  [12,]    1    1    1    1
#>  [13,]    1    1    1    1
#>  [14,]    1    1    1    1
#>  [15,]    1    1    1    1
#>  [16,]    1    1    1    1
#>  [17,]    1    1    1    1
#>  [18,]    1    1    1    1
#>  [19,]    1    1    1    1
#>  [20,]    1    1    1    1
#>  [21,]    1    1    1    1
#>  [22,]    1    1    1    1
#>  [23,]    1    1    1    1
#>  [24,]    1    1    1    1
#>  [25,]    1    1    1    1
#>  [26,]    1    1    1    1
#>  [27,]    1    1    1    1
#>  [28,]    1    1    1    1
#>  [29,]    1    1    1    1
#>  [30,]    1    1    1    1
#>  [31,]    1    1    1    1
#>  [32,]    1    1    1    1
#>  [33,]    1    1    1    1
#>  [34,]    1    1    1    1
#>  [35,]    1    1    1    1
#>  [36,]    1    1    1    1
#>  [37,]    1    1    1    1
#>  [38,]    1    1    1    1
#>  [39,]    1    1    1    1
#>  [40,]    1    1    1    1
#>  [41,]    1    1    1    1
#>  [42,]    1    1    1    1
#>  [43,]    1    1    1    1
#>  [44,]    1    1    1    1
#>  [45,]    1    1    1    1
#>  [46,]    1    1    1    1
#>  [47,]    1    1    1    1
#>  [48,]    1    1    1    1
#>  [49,]    1    1    1    1
#>  [50,]    1    1    1    1
#>  [51,]    1    1    1    1
#>  [52,]    1    1    1    1
#>  [53,]    1    1    1    1
#>  [54,]    1    1    1    1
#>  [55,]    1    1    1    1
#>  [56,]    1    1    1    1
#>  [57,]    1    1    1    1
#>  [58,]    1    1    1    1
#>  [59,]    1    1    1    1
#>  [60,]    1    1    1    1
#>  [61,]    1    1    1    1
#>  [62,]    1    1    1    1
#>  [63,]    1    1    1    1
#>  [64,]    1    1    1    1
#>  [65,]    1    1    1    1
#>  [66,]    1    1    1    1
#>  [67,]    1    1    1    1
#>  [68,]    1    1    1    1
#>  [69,]    1    1    1    1
#>  [70,]    1    1    1    1
#>  [71,]    1    1    1    1
#>  [72,]    1    1    1    1
#>  [73,]    1    1    1    1
#>  [74,]    1    1    1    1
#>  [75,]    1    1    1    1
#>  [76,]    1    1    1    1
#>  [77,]    1    1    1    1
#>  [78,]    1    1    1    1
#>  [79,]    1    1    1    1
#>  [80,]    1    1    1    1
#>  [81,]    1    1    1    1
#>  [82,]    1    1    1    1
#>  [83,]    1    1    1    1
#>  [84,]    1    1    1    1
#>  [85,]    1    1    1    1
#>  [86,]    1    1    1    1
#>  [87,]    1    1    1    1
#>  [88,]    1    1    1    1
#>  [89,]    1    1    1    1
#>  [90,]    1    1    1    1
#>  [91,]    1    1    1    1
#>  [92,]    1    1    1    1
#>  [93,]    1    1    1    1
#>  [94,]    1    1    1    1
#>  [95,]    1    1    1    1
#>  [96,]    1    1    1    1
#>  [97,]    1    1    1    1
#>  [98,]    1    1    1    1
#>  [99,]    1    1    1    1
#> [100,]    1    1    1    1
#> [101,]    1    1    1    1
#> [102,]    1    1    1    1
#> [103,]    1    1    1    1
#> [104,]    1    1    1    1
#> [105,]    1    1    1    1
#> [106,]    1    1    1    1
#> [107,]    1    1    1    1
#> [108,]    1    1    1    1
#> [109,]    1    1    1    1
#> [110,]    1    1    1    1
#> [111,]    1    1    1    1
#> [112,]    1    1    1    1
#> [113,]    1    1    1    1
#> [114,]    1    1    1    1
#> [115,]    1    1    1    1
#> [116,]    1    1    1    1
#> [117,]    1    1    1    1
#> [118,]    1    1    1    1
#> [119,]    1    1    1    1
#> [120,]    1    1    1    1
#> [121,]    1    1    1    1
#> [122,]    1    1    1    1
#> [123,]    1    1    1    1
#> [124,]    1    1    1    1
#> [125,]    1    1    1    1
#> [126,]    1    1    1    1
#> [127,]    1    1    1    1
#> [128,]    1    1    1    1
#> [129,]    1    1    1    1
#> [130,]    1    1    1    1
#> [131,]    1    1    1    1
#> [132,]    1    1    1    1
#> [133,]    1    1    1    1
#> [134,]    1    1    1    1
#> [135,]    1    1    1    1
#> [136,]    1    1    1    1
#> [137,]    1    1    1    1
#> [138,]    1    1    1    1
#> [139,]    1    1    1    1
#> [140,]    1    1    1    1
#> [141,]    1    1    1    1
#> [142,]    1    1    1    1
#> [143,]    1    1    1    1
#> [144,]    1    1    1    1
#> [145,]    1    1    1    1
#> [146,]    1    1    1    1
#> [147,]    1    1    1    1
#> [148,]    1    1    1    1
#> [149,]    1    1    1    1
#> [150,]    1    1    1    1
#> [151,]    1    1    1    1
#> [152,]    1    1    1    1
#> [153,]    1    1    1    1
#> [154,]    1    1    1    1
#> [155,]    1    1    1    1
#> [156,]    1    1    1    1
#> [157,]    1    1    1    1
#> [158,]    1    1    1    1
#> [159,]    1    1    1    1
#> [160,]    1    1    1    1
#> [161,]    1    1    1    1
#> [162,]    1    1    1    1
#> [163,]    1    1    1    1
#> [164,]    1    1    1    1
#> [165,]    1    1    1    1
#> [166,]    1    1    1    1
#> [167,]    1    1    1    1
#> [168,]    1    1    1    1
#> [169,]    1    1    1    1
#> [170,]    1    1    1    1
#> [171,]    1    1    1    1
#> [172,]    1    1    1    1
#> [173,]    1    1    1    1
#> [174,]    1    1    1    1
#> [175,]    1    1    1    1
#> [176,]    1    1    1    1
#> [177,]    1    1    1    1
#> [178,]    1    1    1    1
#> [179,]    1    1    1    1
#> [180,]    1    1    1    1
#> [181,]    1    1    1    1
#> [182,]    1    1    1    1
#> [183,]    1    1    1    1
#> [184,]    1    1    1    1
#> [185,]    1    1    1    1
#> [186,]    1    1    1    1
#> [187,]    1    1    1    1
#> [188,]    1    1    1    1
#> [189,]    1    1    1    1
#> [190,]    1    1    1    1
#> [191,]    1    1    1    1
#> [192,]    1    1    1    1
#> [193,]    1    1    1    1
#> [194,]    1    1    1    1
#> [195,]    1    1    1    1
#> [196,]    1    1    1    1
#> [197,]    1    1    1    1
#> [198,]    1    1    1    1
#> [199,]    1    1    1    1
#> [200,]    1    1    1    1
#> 
#> $V
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    1    0    0    0    0
#> [2,]    0    1    0    0    0
#> [3,]    0    0    1    0    0
#> [4,]    0    0    0    1    0
#> [5,]    0    0    0    0    1
#> 
#> $commonV
#> [1] TRUE
#> 
#> $alpha
#> [1] 0
#> 
#> $L
#>                  condition_1 condition_2 condition_3 condition_4 condition_5
#> condition_1-mean         0.8        -0.2        -0.2        -0.2        -0.2
#> condition_2-mean        -0.2         0.8        -0.2        -0.2        -0.2
#> condition_3-mean        -0.2        -0.2         0.8        -0.2        -0.2
#> condition_4-mean        -0.2        -0.2        -0.2         0.8        -0.2
#> attr(,"reference")
#> [1] "mean"
#> 
#> $LSVSLt
#>                  condition_1-mean condition_2-mean condition_3-mean
#> condition_1-mean             1.00            -0.25            -0.25
#> condition_2-mean            -0.25             1.00            -0.25
#> condition_3-mean            -0.25            -0.25             1.00
#> condition_4-mean            -0.25            -0.25            -0.25
#>                  condition_4-mean
#> condition_1-mean            -0.25
#> condition_2-mean            -0.25
#> condition_3-mean            -0.25
#> condition_4-mean             1.00
#> 
#> attr(,"class")
#> [1] "mash"