Update the data object for mash analysis.

This function can update two parts of the mash data. The first one is setting the reference group, so the mash data can be used for commonbaseline analysis. The other one is updating the null correlation matrix.

mash_update_data(mashdata, ref = NULL, V = NULL)

Arguments

mashdata
mash data object ontaining the Bhat matrix, standard errors, V; created using mash_set_data

ref
the reference group. It could be a number between 1,..., R, R is number of conditions, or the name of reference group. If there is no reference group, it can be the string 'mean'.

V
an R by R matrix / [R x R x J] array of correlation matrix of error correlations

Value

a updated mash data object

Examples

simdata = simple_sims(50,5,1)
data = mash_set_data(simdata$Bhat, simdata$Shat)
mash_update_data(data, 'mean')

#> $Bhat
#>            condition_1-mean condition_2-mean condition_3-mean condition_4-mean
#> effect_1       -0.921496107      2.040083015     -0.066666750     -0.655800996
#> effect_2        1.265825782      0.271446574     -0.430925400      0.189782660
#> effect_3       -0.618278665     -1.771888111      1.332757443      0.762902276
#> effect_4        0.316299353      0.368728135     -0.080324205     -0.640378273
#> effect_5       -0.453594516      0.765725201     -0.263211037      0.060397434
#> effect_6       -0.420856019      1.952176560     -0.982818451      0.517682410
#> effect_7       -0.872685804      1.788072523     -0.037459574     -0.631517691
#> effect_8        0.129597235      0.466102266     -0.866788526     -0.275801317
#> effect_9       -0.228917761      0.418969845      0.990171717     -0.640753631
#> effect_10      -0.055354620      1.532502179      0.327802441     -1.576913546
#> effect_11       1.025390493     -1.117042791      0.508847521      0.341690365
#> effect_12      -0.403683696      0.545692759      0.469242756      0.086914078
#> effect_13      -0.067689264     -0.493034706      0.191128645     -0.080524530
#> effect_14       0.456413108      0.766222258     -0.269610737     -0.031901276
#> effect_15       0.397981144     -1.828001040     -0.389912078      2.225625461
#> effect_16       0.858196935      1.138621586      0.267227739     -1.325691823
#> effect_17      -0.875907822      1.194637654      2.268433028      0.204051199
#> effect_18      -1.301245766      0.738726938      1.302288087     -0.831197772
#> effect_19       0.585896907      1.393322058      0.602922269     -0.293301523
#> effect_20      -0.468635333      0.860213482      0.243423239     -0.002556807
#> effect_21      -0.668562248      0.854078585      0.214515734     -0.696984260
#> effect_22       0.395847803      0.255202165     -0.670263000      0.012386114
#> effect_23       0.137454530      0.352705857     -0.525104369     -0.714619142
#> effect_24      -1.558471420      0.844790394      1.034801496     -0.220160834
#> effect_25      -1.461400015      0.350454250     -0.318846108      1.627715759
#> effect_26      -1.260271829      0.577699246      0.681864624     -0.191173341
#> effect_27       0.604671733      0.534999769      0.223660957     -1.073223747
#> effect_28       0.325750229     -1.486763134      0.933340078      0.314809798
#> effect_29      -0.297651911     -0.865617396      0.040914411      1.150613719
#> effect_30       0.079557612      0.675907050     -0.698859636      0.168298392
#> effect_31      -1.660410843      1.851022032      0.127734719     -1.679145852
#> effect_32       0.983882119     -0.598299829     -0.103788605     -0.243849638
#> effect_33       0.069996091      0.959527940     -0.604748599      1.557684510
#> effect_34      -0.826769148      0.216264544      1.060992031     -1.890047163
#> effect_35       0.977598046      1.639513636     -0.051119865     -1.493244655
#> effect_36      -0.499533014      1.283499223     -0.460703730     -0.887481421
#> effect_37      -0.285114029     -0.497659833      0.533410834      0.123126605
#> effect_38      -0.376126483      0.476466728     -1.230985120     -0.242481000
#> effect_39      -0.249550984     -0.457407701      0.535313791     -0.610108314
#> effect_40      -1.548407171      0.314807227      0.916809201      0.578976845
#> effect_41       0.533938793     -0.321178598     -0.589708940      1.173310463
#> effect_42       1.092177161      0.841899455      0.086488416     -0.557113733
#> effect_43      -1.132785601     -0.078635870     -0.183638833      0.276744145
#> effect_44      -0.137420215      0.187801064      0.606873899      0.997874924
#> effect_45      -0.185178116     -1.104766650      0.442995551     -0.499646451
#> effect_46      -0.160878683      0.432422172     -1.160642034      0.455674649
#> effect_47       0.152015535      2.001310311     -2.060088905      0.792540184
#> effect_48       1.180988277     -1.613179463      0.832200739     -1.173893750
#> effect_49      -0.945216471      1.994143312      0.479927312      0.237675123
#> effect_50       0.633552633     -0.157813123     -1.407167863      1.296309308
#> effect_51      -0.539147619     -1.087453814      0.060022336      1.074689441
#> effect_52      -0.181565230     -0.950207017     -0.026594215      0.414771166
#> effect_53       0.034093161     -0.916012359      2.517623141     -0.765967657
#> effect_54      -1.401708354     -1.077120116      3.887893758     -1.681360738
#> effect_55       2.100061896      0.060046928     -2.109319983     -0.766725841
#> effect_56       1.102989813     -2.598308476     -0.342104481     -0.282016589
#> effect_57      -1.097308259      0.604969480      0.329791117      1.384736278
#> effect_58       1.461093685     -1.470554896      0.647659554     -0.238606234
#> effect_59      -0.280884478     -1.234334883      0.388002146     -0.838431306
#> effect_60      -0.533710603     -0.340632854      2.302313094     -2.035859855
#> effect_61       1.228359570      0.164316945      1.119774627     -1.929499697
#> effect_62       0.782187527     -1.795879447     -0.080637318     -0.388696316
#> effect_63       0.120769594      1.334305663     -0.949307669      0.482475371
#> effect_64      -1.052862134      0.708067859      0.528210683     -1.050843758
#> effect_65      -0.223635228      0.352031045     -0.041567237      0.877161312
#> effect_66       0.735839059     -1.542732170     -0.932041957     -1.171260265
#> effect_67      -0.057316243     -1.123159031     -0.073330544     -0.246157417
#> effect_68       0.073229325      0.817912709     -0.191085087     -0.608433370
#> effect_69      -1.386914760      0.981304760      0.533617279      1.933912648
#> effect_70      -0.661355790     -0.352331077      0.159840721     -0.287935748
#> effect_71      -0.764139263      0.191490402      0.572853868      1.436498394
#> effect_72      -0.069282658      0.616690585      0.313233092     -1.081189200
#> effect_73       1.434191774     -0.147273861     -1.068215718      2.797756542
#> effect_74       0.570022454     -0.139569424     -0.241538295      0.866341746
#> effect_75       0.381186099     -0.884259260      1.688838993      1.007016135
#> effect_76      -1.240473502     -0.613490496     -0.158397860     -0.478568248
#> effect_77      -0.183218850      1.475453877      1.097136626     -0.016256827
#> effect_78      -0.904202605     -0.156725021     -0.360621482     -0.375303807
#> effect_79      -1.202087124     -0.208841356     -0.623926010      1.166430756
#> effect_80       0.199674697      0.524093428     -0.332798204     -2.005807844
#> effect_81      -0.738547198      0.435039859      0.881536469     -1.275916304
#> effect_82      -0.316394536      1.394074162     -1.373975698      1.554127698
#> effect_83      -0.127390315      1.153398804     -1.207009127     -0.632859726
#> effect_84      -2.619686070      0.211643077     -0.006666018     -0.339071622
#> effect_85      -2.835063841      1.549391091     -0.620579442      1.774603816
#> effect_86       0.047846927     -0.941693465     -0.560083886      1.329845662
#> effect_87       2.363139676     -1.716225853     -0.881592969      0.461705414
#> effect_88      -1.539160576      0.515471876      0.637926901     -0.342414468
#> effect_89       0.174189769     -1.124471609      1.106184318     -1.041973927
#> effect_90      -1.748912618      1.712859562      0.722555522     -1.642546295
#> effect_91      -0.850850425     -1.624617590      0.789386550      2.901653695
#> effect_92      -1.011766502      1.119441766     -1.624564822      2.513453043
#> effect_93      -0.172375980     -0.675916242      1.022852526      0.960043936
#> effect_94       0.248211652     -0.312935246     -0.753578413     -0.339087388
#> effect_95      -0.349470742      1.214068507     -0.513841913      1.630548430
#> effect_96      -1.286714351     -0.981123184      2.018767083     -0.459632759
#> effect_97      -1.541303282     -1.137888285      2.205009916      0.315664207
#> effect_98       1.273800771      1.731924033     -1.667190487     -0.872591203
#> effect_99      -0.037881098      0.760779122     -0.495312474     -0.212397951
#> effect_100     -1.602097313      0.090216148     -0.380793625      0.507191967
#> effect_101      2.872782760     -0.951853191     -0.648224905     -1.381014525
#> effect_102      0.285201563      1.328305533      0.833918008     -1.384850914
#> effect_103     -1.004050268     -0.332901458     -0.804575667      1.152243752
#> effect_104     -1.993169623      0.540640684     -0.236038392      1.145279854
#> effect_105      0.603352281     -0.297762135     -1.283688648      0.183329280
#> effect_106     -0.458964154     -0.088902806     -0.954576372      0.493403988
#> effect_107     -0.364031660      0.879959566     -0.384411287     -1.304338358
#> effect_108      0.793446382     -0.700118997      0.452497240      0.005537958
#> effect_109     -1.058834798      0.105719772      0.926854467      0.216460581
#> effect_110      1.443992358      0.178317782     -0.745446900     -0.966486919
#> effect_111     -0.730402503      0.152436785     -0.471362425     -0.173075727
#> effect_112     -0.825215640     -0.369852049      1.477585363     -0.159457337
#> effect_113     -2.263431317     -0.007877658     -0.004416039      1.995248556
#> effect_114     -2.292150749      1.140860829      1.073660830     -0.696948477
#> effect_115      0.911878761      0.763358665     -0.697599425     -1.480434538
#> effect_116      0.099379113      0.025665641      0.089032415     -1.094950780
#> effect_117     -0.254961921     -1.340851643      0.897598747      0.195958216
#> effect_118      0.970526859      1.117813792      1.049365766     -1.023207845
#> effect_119     -0.056755875     -2.082607958     -0.427562163      0.608975333
#> effect_120     -1.370406512      0.564681378      0.963625156     -1.166186230
#> effect_121     -1.698068772      1.889060835      0.378882584     -0.631844427
#> effect_122     -0.158167102      0.506388359      0.334378290     -2.006672334
#> effect_123     -2.379176643      0.939042673     -0.698864909      1.254208833
#> effect_124      1.450293779     -1.801859571     -0.894494976     -0.158834344
#> effect_125     -0.565724036      1.096121605     -0.065739707     -0.025699024
#> effect_126     -0.052463070      0.009141990     -0.179558858      0.494170813
#> effect_127      0.296742433     -1.546419639      1.285953127      0.179426403
#> effect_128     -0.961652211      0.930422351      0.964779469      0.793464187
#> effect_129     -1.130778875     -0.585496904      1.371416323     -0.210723641
#> effect_130      3.291113649     -1.218696874     -0.746976765      0.190700406
#> effect_131     -2.058618983      1.341347810     -1.277391488      0.093319349
#> effect_132      1.323029227      0.316489661     -1.110769119     -1.013247930
#> effect_133     -1.786215438      0.284345693      0.243289198      0.908219464
#> effect_134      0.582382022     -0.045651822      0.760668710     -1.172979406
#> effect_135     -0.594142862     -0.558671156      0.519210354     -1.102641758
#> effect_136      2.478123514     -1.615692936     -1.784636314      1.046228300
#> effect_137     -2.033567674     -0.430903364     -0.735767710      1.305167465
#> effect_138      0.339305594     -0.409424321      0.157780190     -0.623562414
#> effect_139      1.109037158     -1.781636620      1.266457927     -0.339547165
#> effect_140     -1.755198846      0.007190519      0.465186680      1.096814781
#> effect_141     -0.636946245      0.178588078     -0.164307626     -0.342058536
#> effect_142     -2.644442586      0.037457327      1.794169637     -0.427602975
#> effect_143      0.905026424      0.058308568     -0.971943785     -0.148338427
#> effect_144     -0.292601764      0.922966068      0.143478791      0.502565156
#> effect_145      1.174562308     -2.406837023     -0.786652781      1.187361606
#> effect_146     -1.488838402      0.945227825     -0.287869171      0.516380014
#> effect_147      2.241884605     -1.075802454     -1.260802408      1.434803267
#> effect_148     -0.218237589      0.276693961      2.088392464     -1.487767168
#> effect_149      0.275485811     -0.264769463      0.090386548     -0.296835043
#> effect_150      1.029841224     -0.304945143      1.169982032     -0.653345913
#> effect_151      1.757053823      0.165014419      0.208427008     -2.673735385
#> effect_152     -0.981549837      0.056218361      1.643145878     -0.078411225
#> effect_153      3.256044120     -1.092907887     -1.007071654     -0.168072073
#> effect_154      0.791359555     -0.891283719      0.292173098      0.371158432
#> effect_155      0.627910533     -0.181355231     -0.056035755      0.185337814
#> effect_156     -0.565109699      0.677728880     -0.715268770      0.306377671
#> effect_157     -0.416676791      0.393267825     -0.721002735     -0.427377922
#> effect_158     -1.058634835     -0.211556972      0.992557136     -0.252239886
#> effect_159      1.928327989      0.236320585      1.157223983     -1.213255295
#> effect_160     -0.430167725     -0.980958013      1.931926466     -0.613184734
#> effect_161     -2.108827754      0.701492072      1.094046053     -0.804204320
#> effect_162     -0.532987625      1.547563540     -0.014434534     -0.125955158
#> effect_163     -0.819105799      0.307329308      0.321098832      0.736545157
#> effect_164      2.304922772     -0.351102040      0.276525076     -1.933494641
#> effect_165      1.212905425     -0.905891690     -0.202677809     -0.445053788
#> effect_166      0.333644653     -0.414782107      0.929693973     -0.196868831
#> effect_167      1.027679875      1.528741193      0.065539594     -1.766520159
#> effect_168     -2.253431117     -0.166941946      0.366806798      0.985794155
#> effect_169     -0.735623285     -0.679803805      2.098630195      0.268404076
#> effect_170      0.843824059      1.106135604     -0.957323995      0.859626201
#> effect_171     -0.521015589     -0.281413408      0.453136249      0.368532002
#> effect_172      0.677014076      0.662915002     -0.114145842      0.135670232
#> effect_173     -0.282315750     -0.793492828      1.198723985     -0.384236846
#> effect_174     -0.360241919     -0.900957998     -0.990161663      1.368813229
#> effect_175     -1.076120280      2.430712470     -0.495931351     -0.386792572
#> effect_176     -0.329128007     -0.718662844     -0.174339708      0.082980230
#> effect_177     -0.278629823     -0.135412552     -0.332275450      0.182695199
#> effect_178     -0.104552908     -1.128404895      0.720140535      0.041102311
#> effect_179      0.133232011      0.026114183     -0.920002945     -0.109307197
#> effect_180     -0.657511357      1.396442963     -0.775086870      0.114871217
#> effect_181      0.985112188      0.460436410      0.095199996     -1.087376468
#> effect_182      1.217395657     -0.475111145     -0.394557435      0.348369923
#> effect_183      0.423196936     -0.202296927     -0.698175699      0.812061133
#> effect_184      0.190101374     -0.408537962      0.415065630     -0.288825997
#> effect_185     -0.526233934      0.248493925     -0.163494835     -0.297143720
#> effect_186      0.359820167      0.019035399     -1.319297817      0.918678456
#> effect_187     -0.487429927     -1.045445726      1.365882275     -0.596409118
#> effect_188     -0.266145308      0.416278274     -0.040896954      0.112377442
#> effect_189      0.659678336      0.424362048      0.116772256     -0.894867966
#> effect_190     -1.146656199      0.208250754      1.717211949     -2.081462382
#> effect_191     -0.932489608      0.950840990      0.024171104      0.040924225
#> effect_192      1.959670257     -0.692837112     -1.465657559      0.392268278
#> effect_193     -1.414734148      0.473893588      0.532383669     -0.434509784
#> effect_194      0.762153867      0.050904469     -0.185023257      0.120701165
#> effect_195     -0.522030405      0.780295702      0.874834764      0.202477450
#> effect_196     -0.009081831      0.790168910     -2.221151056      0.006479554
#> effect_197     -0.834741831      0.239608236     -0.077596680      0.323042924
#> effect_198      1.739977922      0.619557697     -0.536796825     -0.250281246
#> effect_199     -0.486208720      0.021019147     -0.058402645      0.690195195
#> effect_200      1.609722156      0.910636953      0.130030183     -1.616677588
#> 
#> $Shat
#>             [,1]      [,2]      [,3]      [,4]
#>   [1,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [2,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [3,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [4,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [5,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [6,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [7,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [8,] 0.8944272 0.8944272 0.8944272 0.8944272
#>   [9,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [10,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [11,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [12,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [13,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [14,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [15,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [16,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [17,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [18,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [19,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [20,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [21,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [22,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [23,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [24,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [25,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [26,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [27,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [28,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [29,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [30,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [31,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [32,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [33,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [34,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [35,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [36,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [37,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [38,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [39,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [40,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [41,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [42,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [43,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [44,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [45,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [46,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [47,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [48,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [49,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [50,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [51,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [52,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [53,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [54,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [55,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [56,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [57,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [58,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [59,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [60,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [61,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [62,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [63,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [64,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [65,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [66,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [67,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [68,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [69,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [70,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [71,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [72,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [73,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [74,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [75,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [76,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [77,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [78,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [79,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [80,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [81,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [82,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [83,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [84,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [85,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [86,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [87,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [88,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [89,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [90,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [91,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [92,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [93,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [94,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [95,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [96,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [97,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [98,] 0.8944272 0.8944272 0.8944272 0.8944272
#>  [99,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [100,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [101,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [102,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [103,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [104,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [105,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [106,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [107,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [108,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [109,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [110,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [111,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [112,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [113,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [114,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [115,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [116,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [117,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [118,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [119,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [120,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [121,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [122,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [123,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [124,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [125,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [126,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [127,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [128,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [129,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [130,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [131,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [132,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [133,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [134,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [135,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [136,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [137,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [138,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [139,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [140,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [141,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [142,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [143,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [144,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [145,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [146,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [147,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [148,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [149,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [150,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [151,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [152,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [153,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [154,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [155,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [156,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [157,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [158,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [159,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [160,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [161,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [162,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [163,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [164,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [165,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [166,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [167,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [168,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [169,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [170,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [171,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [172,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [173,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [174,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [175,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [176,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [177,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [178,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [179,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [180,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [181,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [182,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [183,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [184,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [185,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [186,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [187,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [188,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [189,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [190,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [191,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [192,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [193,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [194,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [195,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [196,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [197,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [198,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [199,] 0.8944272 0.8944272 0.8944272 0.8944272
#> [200,] 0.8944272 0.8944272 0.8944272 0.8944272
#> 
#> $Shat_orig
#>            condition_1 condition_2 condition_3 condition_4 condition_5
#> effect_1             1           1           1           1           1
#> effect_2             1           1           1           1           1
#> effect_3             1           1           1           1           1
#> effect_4             1           1           1           1           1
#> effect_5             1           1           1           1           1
#> effect_6             1           1           1           1           1
#> effect_7             1           1           1           1           1
#> effect_8             1           1           1           1           1
#> effect_9             1           1           1           1           1
#> effect_10            1           1           1           1           1
#> effect_11            1           1           1           1           1
#> effect_12            1           1           1           1           1
#> effect_13            1           1           1           1           1
#> effect_14            1           1           1           1           1
#> effect_15            1           1           1           1           1
#> effect_16            1           1           1           1           1
#> effect_17            1           1           1           1           1
#> effect_18            1           1           1           1           1
#> effect_19            1           1           1           1           1
#> effect_20            1           1           1           1           1
#> effect_21            1           1           1           1           1
#> effect_22            1           1           1           1           1
#> effect_23            1           1           1           1           1
#> effect_24            1           1           1           1           1
#> effect_25            1           1           1           1           1
#> effect_26            1           1           1           1           1
#> effect_27            1           1           1           1           1
#> effect_28            1           1           1           1           1
#> effect_29            1           1           1           1           1
#> effect_30            1           1           1           1           1
#> effect_31            1           1           1           1           1
#> effect_32            1           1           1           1           1
#> effect_33            1           1           1           1           1
#> effect_34            1           1           1           1           1
#> effect_35            1           1           1           1           1
#> effect_36            1           1           1           1           1
#> effect_37            1           1           1           1           1
#> effect_38            1           1           1           1           1
#> effect_39            1           1           1           1           1
#> effect_40            1           1           1           1           1
#> effect_41            1           1           1           1           1
#> effect_42            1           1           1           1           1
#> effect_43            1           1           1           1           1
#> effect_44            1           1           1           1           1
#> effect_45            1           1           1           1           1
#> effect_46            1           1           1           1           1
#> effect_47            1           1           1           1           1
#> effect_48            1           1           1           1           1
#> effect_49            1           1           1           1           1
#> effect_50            1           1           1           1           1
#> effect_51            1           1           1           1           1
#> effect_52            1           1           1           1           1
#> effect_53            1           1           1           1           1
#> effect_54            1           1           1           1           1
#> effect_55            1           1           1           1           1
#> effect_56            1           1           1           1           1
#> effect_57            1           1           1           1           1
#> effect_58            1           1           1           1           1
#> effect_59            1           1           1           1           1
#> effect_60            1           1           1           1           1
#> effect_61            1           1           1           1           1
#> effect_62            1           1           1           1           1
#> effect_63            1           1           1           1           1
#> effect_64            1           1           1           1           1
#> effect_65            1           1           1           1           1
#> effect_66            1           1           1           1           1
#> effect_67            1           1           1           1           1
#> effect_68            1           1           1           1           1
#> effect_69            1           1           1           1           1
#> effect_70            1           1           1           1           1
#> effect_71            1           1           1           1           1
#> effect_72            1           1           1           1           1
#> effect_73            1           1           1           1           1
#> effect_74            1           1           1           1           1
#> effect_75            1           1           1           1           1
#> effect_76            1           1           1           1           1
#> effect_77            1           1           1           1           1
#> effect_78            1           1           1           1           1
#> effect_79            1           1           1           1           1
#> effect_80            1           1           1           1           1
#> effect_81            1           1           1           1           1
#> effect_82            1           1           1           1           1
#> effect_83            1           1           1           1           1
#> effect_84            1           1           1           1           1
#> effect_85            1           1           1           1           1
#> effect_86            1           1           1           1           1
#> effect_87            1           1           1           1           1
#> effect_88            1           1           1           1           1
#> effect_89            1           1           1           1           1
#> effect_90            1           1           1           1           1
#> effect_91            1           1           1           1           1
#> effect_92            1           1           1           1           1
#> effect_93            1           1           1           1           1
#> effect_94            1           1           1           1           1
#> effect_95            1           1           1           1           1
#> effect_96            1           1           1           1           1
#> effect_97            1           1           1           1           1
#> effect_98            1           1           1           1           1
#> effect_99            1           1           1           1           1
#> effect_100           1           1           1           1           1
#> effect_101           1           1           1           1           1
#> effect_102           1           1           1           1           1
#> effect_103           1           1           1           1           1
#> effect_104           1           1           1           1           1
#> effect_105           1           1           1           1           1
#> effect_106           1           1           1           1           1
#> effect_107           1           1           1           1           1
#> effect_108           1           1           1           1           1
#> effect_109           1           1           1           1           1
#> effect_110           1           1           1           1           1
#> effect_111           1           1           1           1           1
#> effect_112           1           1           1           1           1
#> effect_113           1           1           1           1           1
#> effect_114           1           1           1           1           1
#> effect_115           1           1           1           1           1
#> effect_116           1           1           1           1           1
#> effect_117           1           1           1           1           1
#> effect_118           1           1           1           1           1
#> effect_119           1           1           1           1           1
#> effect_120           1           1           1           1           1
#> effect_121           1           1           1           1           1
#> effect_122           1           1           1           1           1
#> effect_123           1           1           1           1           1
#> effect_124           1           1           1           1           1
#> effect_125           1           1           1           1           1
#> effect_126           1           1           1           1           1
#> effect_127           1           1           1           1           1
#> effect_128           1           1           1           1           1
#> effect_129           1           1           1           1           1
#> effect_130           1           1           1           1           1
#> effect_131           1           1           1           1           1
#> effect_132           1           1           1           1           1
#> effect_133           1           1           1           1           1
#> effect_134           1           1           1           1           1
#> effect_135           1           1           1           1           1
#> effect_136           1           1           1           1           1
#> effect_137           1           1           1           1           1
#> effect_138           1           1           1           1           1
#> effect_139           1           1           1           1           1
#> effect_140           1           1           1           1           1
#> effect_141           1           1           1           1           1
#> effect_142           1           1           1           1           1
#> effect_143           1           1           1           1           1
#> effect_144           1           1           1           1           1
#> effect_145           1           1           1           1           1
#> effect_146           1           1           1           1           1
#> effect_147           1           1           1           1           1
#> effect_148           1           1           1           1           1
#> effect_149           1           1           1           1           1
#> effect_150           1           1           1           1           1
#> effect_151           1           1           1           1           1
#> effect_152           1           1           1           1           1
#> effect_153           1           1           1           1           1
#> effect_154           1           1           1           1           1
#> effect_155           1           1           1           1           1
#> effect_156           1           1           1           1           1
#> effect_157           1           1           1           1           1
#> effect_158           1           1           1           1           1
#> effect_159           1           1           1           1           1
#> effect_160           1           1           1           1           1
#> effect_161           1           1           1           1           1
#> effect_162           1           1           1           1           1
#> effect_163           1           1           1           1           1
#> effect_164           1           1           1           1           1
#> effect_165           1           1           1           1           1
#> effect_166           1           1           1           1           1
#> effect_167           1           1           1           1           1
#> effect_168           1           1           1           1           1
#> effect_169           1           1           1           1           1
#> effect_170           1           1           1           1           1
#> effect_171           1           1           1           1           1
#> effect_172           1           1           1           1           1
#> effect_173           1           1           1           1           1
#> effect_174           1           1           1           1           1
#> effect_175           1           1           1           1           1
#> effect_176           1           1           1           1           1
#> effect_177           1           1           1           1           1
#> effect_178           1           1           1           1           1
#> effect_179           1           1           1           1           1
#> effect_180           1           1           1           1           1
#> effect_181           1           1           1           1           1
#> effect_182           1           1           1           1           1
#> effect_183           1           1           1           1           1
#> effect_184           1           1           1           1           1
#> effect_185           1           1           1           1           1
#> effect_186           1           1           1           1           1
#> effect_187           1           1           1           1           1
#> effect_188           1           1           1           1           1
#> effect_189           1           1           1           1           1
#> effect_190           1           1           1           1           1
#> effect_191           1           1           1           1           1
#> effect_192           1           1           1           1           1
#> effect_193           1           1           1           1           1
#> effect_194           1           1           1           1           1
#> effect_195           1           1           1           1           1
#> effect_196           1           1           1           1           1
#> effect_197           1           1           1           1           1
#> effect_198           1           1           1           1           1
#> effect_199           1           1           1           1           1
#> effect_200           1           1           1           1           1
#> 
#> $Shat_alpha
#>        [,1] [,2] [,3] [,4]
#>   [1,]    1    1    1    1
#>   [2,]    1    1    1    1
#>   [3,]    1    1    1    1
#>   [4,]    1    1    1    1
#>   [5,]    1    1    1    1
#>   [6,]    1    1    1    1
#>   [7,]    1    1    1    1
#>   [8,]    1    1    1    1
#>   [9,]    1    1    1    1
#>  [10,]    1    1    1    1
#>  [11,]    1    1    1    1
#>  [12,]    1    1    1    1
#>  [13,]    1    1    1    1
#>  [14,]    1    1    1    1
#>  [15,]    1    1    1    1
#>  [16,]    1    1    1    1
#>  [17,]    1    1    1    1
#>  [18,]    1    1    1    1
#>  [19,]    1    1    1    1
#>  [20,]    1    1    1    1
#>  [21,]    1    1    1    1
#>  [22,]    1    1    1    1
#>  [23,]    1    1    1    1
#>  [24,]    1    1    1    1
#>  [25,]    1    1    1    1
#>  [26,]    1    1    1    1
#>  [27,]    1    1    1    1
#>  [28,]    1    1    1    1
#>  [29,]    1    1    1    1
#>  [30,]    1    1    1    1
#>  [31,]    1    1    1    1
#>  [32,]    1    1    1    1
#>  [33,]    1    1    1    1
#>  [34,]    1    1    1    1
#>  [35,]    1    1    1    1
#>  [36,]    1    1    1    1
#>  [37,]    1    1    1    1
#>  [38,]    1    1    1    1
#>  [39,]    1    1    1    1
#>  [40,]    1    1    1    1
#>  [41,]    1    1    1    1
#>  [42,]    1    1    1    1
#>  [43,]    1    1    1    1
#>  [44,]    1    1    1    1
#>  [45,]    1    1    1    1
#>  [46,]    1    1    1    1
#>  [47,]    1    1    1    1
#>  [48,]    1    1    1    1
#>  [49,]    1    1    1    1
#>  [50,]    1    1    1    1
#>  [51,]    1    1    1    1
#>  [52,]    1    1    1    1
#>  [53,]    1    1    1    1
#>  [54,]    1    1    1    1
#>  [55,]    1    1    1    1
#>  [56,]    1    1    1    1
#>  [57,]    1    1    1    1
#>  [58,]    1    1    1    1
#>  [59,]    1    1    1    1
#>  [60,]    1    1    1    1
#>  [61,]    1    1    1    1
#>  [62,]    1    1    1    1
#>  [63,]    1    1    1    1
#>  [64,]    1    1    1    1
#>  [65,]    1    1    1    1
#>  [66,]    1    1    1    1
#>  [67,]    1    1    1    1
#>  [68,]    1    1    1    1
#>  [69,]    1    1    1    1
#>  [70,]    1    1    1    1
#>  [71,]    1    1    1    1
#>  [72,]    1    1    1    1
#>  [73,]    1    1    1    1
#>  [74,]    1    1    1    1
#>  [75,]    1    1    1    1
#>  [76,]    1    1    1    1
#>  [77,]    1    1    1    1
#>  [78,]    1    1    1    1
#>  [79,]    1    1    1    1
#>  [80,]    1    1    1    1
#>  [81,]    1    1    1    1
#>  [82,]    1    1    1    1
#>  [83,]    1    1    1    1
#>  [84,]    1    1    1    1
#>  [85,]    1    1    1    1
#>  [86,]    1    1    1    1
#>  [87,]    1    1    1    1
#>  [88,]    1    1    1    1
#>  [89,]    1    1    1    1
#>  [90,]    1    1    1    1
#>  [91,]    1    1    1    1
#>  [92,]    1    1    1    1
#>  [93,]    1    1    1    1
#>  [94,]    1    1    1    1
#>  [95,]    1    1    1    1
#>  [96,]    1    1    1    1
#>  [97,]    1    1    1    1
#>  [98,]    1    1    1    1
#>  [99,]    1    1    1    1
#> [100,]    1    1    1    1
#> [101,]    1    1    1    1
#> [102,]    1    1    1    1
#> [103,]    1    1    1    1
#> [104,]    1    1    1    1
#> [105,]    1    1    1    1
#> [106,]    1    1    1    1
#> [107,]    1    1    1    1
#> [108,]    1    1    1    1
#> [109,]    1    1    1    1
#> [110,]    1    1    1    1
#> [111,]    1    1    1    1
#> [112,]    1    1    1    1
#> [113,]    1    1    1    1
#> [114,]    1    1    1    1
#> [115,]    1    1    1    1
#> [116,]    1    1    1    1
#> [117,]    1    1    1    1
#> [118,]    1    1    1    1
#> [119,]    1    1    1    1
#> [120,]    1    1    1    1
#> [121,]    1    1    1    1
#> [122,]    1    1    1    1
#> [123,]    1    1    1    1
#> [124,]    1    1    1    1
#> [125,]    1    1    1    1
#> [126,]    1    1    1    1
#> [127,]    1    1    1    1
#> [128,]    1    1    1    1
#> [129,]    1    1    1    1
#> [130,]    1    1    1    1
#> [131,]    1    1    1    1
#> [132,]    1    1    1    1
#> [133,]    1    1    1    1
#> [134,]    1    1    1    1
#> [135,]    1    1    1    1
#> [136,]    1    1    1    1
#> [137,]    1    1    1    1
#> [138,]    1    1    1    1
#> [139,]    1    1    1    1
#> [140,]    1    1    1    1
#> [141,]    1    1    1    1
#> [142,]    1    1    1    1
#> [143,]    1    1    1    1
#> [144,]    1    1    1    1
#> [145,]    1    1    1    1
#> [146,]    1    1    1    1
#> [147,]    1    1    1    1
#> [148,]    1    1    1    1
#> [149,]    1    1    1    1
#> [150,]    1    1    1    1
#> [151,]    1    1    1    1
#> [152,]    1    1    1    1
#> [153,]    1    1    1    1
#> [154,]    1    1    1    1
#> [155,]    1    1    1    1
#> [156,]    1    1    1    1
#> [157,]    1    1    1    1
#> [158,]    1    1    1    1
#> [159,]    1    1    1    1
#> [160,]    1    1    1    1
#> [161,]    1    1    1    1
#> [162,]    1    1    1    1
#> [163,]    1    1    1    1
#> [164,]    1    1    1    1
#> [165,]    1    1    1    1
#> [166,]    1    1    1    1
#> [167,]    1    1    1    1
#> [168,]    1    1    1    1
#> [169,]    1    1    1    1
#> [170,]    1    1    1    1
#> [171,]    1    1    1    1
#> [172,]    1    1    1    1
#> [173,]    1    1    1    1
#> [174,]    1    1    1    1
#> [175,]    1    1    1    1
#> [176,]    1    1    1    1
#> [177,]    1    1    1    1
#> [178,]    1    1    1    1
#> [179,]    1    1    1    1
#> [180,]    1    1    1    1
#> [181,]    1    1    1    1
#> [182,]    1    1    1    1
#> [183,]    1    1    1    1
#> [184,]    1    1    1    1
#> [185,]    1    1    1    1
#> [186,]    1    1    1    1
#> [187,]    1    1    1    1
#> [188,]    1    1    1    1
#> [189,]    1    1    1    1
#> [190,]    1    1    1    1
#> [191,]    1    1    1    1
#> [192,]    1    1    1    1
#> [193,]    1    1    1    1
#> [194,]    1    1    1    1
#> [195,]    1    1    1    1
#> [196,]    1    1    1    1
#> [197,]    1    1    1    1
#> [198,]    1    1    1    1
#> [199,]    1    1    1    1
#> [200,]    1    1    1    1
#> 
#> $V
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    1    0    0    0    0
#> [2,]    0    1    0    0    0
#> [3,]    0    0    1    0    0
#> [4,]    0    0    0    1    0
#> [5,]    0    0    0    0    1
#> 
#> $commonV
#> [1] TRUE
#> 
#> $alpha
#> [1] 0
#> 
#> $L
#>                  condition_1 condition_2 condition_3 condition_4 condition_5
#> condition_1-mean         0.8        -0.2        -0.2        -0.2        -0.2
#> condition_2-mean        -0.2         0.8        -0.2        -0.2        -0.2
#> condition_3-mean        -0.2        -0.2         0.8        -0.2        -0.2
#> condition_4-mean        -0.2        -0.2        -0.2         0.8        -0.2
#> attr(,"reference")
#> [1] "mean"
#> 
#> attr(,"class")
#> [1] "mash"