Create simulation with signal data used for contrast analysis.

Create simulation with signal data used for contrast analysis.

sim_contrast2(nsamp = 1000, ncond = 5, err_sd = sqrt(0.5))

Arguments

nsamp
Number of samples of each type.

ncond
Number of conditions.

err_sd
The standard deviation of the errors.

Details

The first condition is the reference group. The deviations are the difference between the subsequent conditions with the reference group. The simulation consists of 90 10 different types of deviations: equal among conditions, present only in the first subsequent condition, independent across conditions.

Examples

sim_contrast2(100,5)

#> $C
#>            condition_1 condition_2 condition_3 condition_4 condition_5
#> sample_1      8.250563    8.250563    8.250563    8.250563    8.250563
#> sample_2      9.050843    9.050843    9.050843    9.050843    9.050843
#> sample_3      7.877350    7.877350    7.877350    7.877350    7.877350
#> sample_4     10.344145   10.344145   10.344145   10.344145   10.344145
#> sample_5     11.147874   11.147874   11.147874   11.147874   11.147874
#> sample_6      9.587952    9.587952    9.587952    9.587952    9.587952
#> sample_7      9.070101    9.070101    9.070101    9.070101    9.070101
#> sample_8      8.538504    8.538504    8.538504    8.538504    8.538504
#> sample_9     11.220393   11.220393   11.220393   11.220393   11.220393
#> sample_10     9.320573    9.320573    9.320573    9.320573    9.320573
#> sample_11     9.674030    9.674030    9.674030    9.674030    9.674030
#> sample_12     8.518756    8.518756    8.518756    8.518756    8.518756
#> sample_13     9.827655    9.827655    9.827655    9.827655    9.827655
#> sample_14     7.838286    7.838286    7.838286    7.838286    7.838286
#> sample_15    10.498800   10.498800   10.498800   10.498800   10.498800
#> sample_16     8.458642    8.458642    8.458642    8.458642    8.458642
#> sample_17     9.553298    9.553298    9.553298    9.553298    9.553298
#> sample_18    11.053848   11.053848   11.053848   11.053848   11.053848
#> sample_19    11.112678   11.112678   11.112678   11.112678   11.112678
#> sample_20    10.550760   10.550760   10.550760   10.550760   10.550760
#> sample_21    10.358819   10.358819   10.358819   10.358819   10.358819
#> sample_22    11.419718   11.419718   11.419718   11.419718   11.419718
#> sample_23    11.138562   11.138562   11.138562   11.138562   11.138562
#> sample_24     9.077515    9.077515    9.077515    9.077515    9.077515
#> sample_25     9.912349    9.912349    9.912349    9.912349    9.912349
#> sample_26    10.100811   10.100811   10.100811   10.100811   10.100811
#> sample_27     9.325250    9.325250    9.325250    9.325250    9.325250
#> sample_28    11.260660   11.260660   11.260660   11.260660   11.260660
#> sample_29     9.173412    9.173412    9.173412    9.173412    9.173412
#> sample_30     9.014301    9.014301    9.014301    9.014301    9.014301
#> sample_31    11.812208   11.812208   11.812208   11.812208   11.812208
#> sample_32    10.950458   10.950458   10.950458   10.950458   10.950458
#> sample_33     8.914908    8.914908    8.914908    8.914908    8.914908
#> sample_34     8.157080    8.157080    8.157080    8.157080    8.157080
#> sample_35     9.471836    9.471836    9.471836    9.471836    9.471836
#> sample_36     9.112267    9.112267    9.112267    9.112267    9.112267
#> sample_37    10.154976   10.154976   10.154976   10.154976   10.154976
#> sample_38    11.363630   11.363630   11.363630   11.363630   11.363630
#> sample_39    10.252750   10.252750   10.252750   10.252750   10.252750
#> sample_40     9.699885    9.699885    9.699885    9.699885    9.699885
#> sample_41     9.974266    9.974266    9.974266    9.974266    9.974266
#> sample_42     9.737293    9.737293    9.737293    9.737293    9.737293
#> sample_43     8.689101    8.689101    8.689101    8.689101    8.689101
#> sample_44     9.153022    9.153022    9.153022    9.153022    9.153022
#> sample_45    10.051662   10.051662   10.051662   10.051662   10.051662
#> sample_46    11.017037   11.017037   11.017037   11.017037   11.017037
#> sample_47     9.311240    9.311240    9.311240    9.311240    9.311240
#> sample_48    10.672175   10.672175   10.672175   10.672175   10.672175
#> sample_49    11.060153   11.060153   11.060153   11.060153   11.060153
#> sample_50    10.940465   10.940465   10.940465   10.940465   10.940465
#> sample_51    11.244268   11.244268   11.244268   11.244268   11.244268
#> sample_52    10.729483   10.729483   10.729483   10.729483   10.729483
#> sample_53     9.978971    9.978971    9.978971    9.978971    9.978971
#> sample_54     9.491573    9.491573    9.491573    9.491573    9.491573
#> sample_55     8.938709    8.938709    8.938709    8.938709    8.938709
#> sample_56    10.171365   10.171365   10.171365   10.171365   10.171365
#> sample_57    11.985450   11.985450   11.985450   11.985450   11.985450
#> sample_58    10.145902   10.145902   10.145902   10.145902   10.145902
#> sample_59    10.593955   10.593955   10.593955   10.593955   10.593955
#> sample_60     8.949053    8.949053    8.949053    8.949053    8.949053
#> sample_61    10.737366   10.737366   10.737366   10.737366   10.737366
#> sample_62    10.852686   10.852686   10.852686   10.852686   10.852686
#> sample_63    11.129226   11.129226   11.129226   11.129226   11.129226
#> sample_64    10.998682   10.998682   10.998682   10.998682   10.998682
#> sample_65    10.593254   10.593254   10.593254   10.593254   10.593254
#> sample_66    10.251759   10.251759   10.251759   10.251759   10.251759
#> sample_67    10.547863   10.547863   10.547863   10.547863   10.547863
#> sample_68    11.441785   11.441785   11.441785   11.441785   11.441785
#> sample_69     9.709485    9.709485    9.709485    9.709485    9.709485
#> sample_70     9.599464    9.599464    9.599464    9.599464    9.599464
#> sample_71     9.392665    9.392665    9.392665    9.392665    9.392665
#> sample_72     9.919102    9.919102    9.919102    9.919102    9.919102
#> sample_73    10.181439   10.181439   10.181439   10.181439   10.181439
#> sample_74     9.485022    9.485022    9.485022    9.485022    9.485022
#> sample_75     9.775012    9.775012    9.775012    9.775012    9.775012
#> sample_76    11.240330   11.240330   11.240330   11.240330   11.240330
#> sample_77     9.359297    9.359297    9.359297    9.359297    9.359297
#> sample_78     9.347115    9.347115    9.347115    9.347115    9.347115
#> sample_79    10.801058   10.801058   10.801058   10.801058   10.801058
#> sample_80    11.011877   11.011877   11.011877   11.011877   11.011877
#> sample_81     9.357450    9.357450    9.357450    9.357450    9.357450
#> sample_82     8.758381    8.758381    8.758381    8.758381    8.758381
#> sample_83    10.089116   10.089116   10.089116   10.089116   10.089116
#> sample_84     9.179200    9.179200    9.179200    9.179200    9.179200
#> sample_85     8.970087    8.970087    8.970087    8.970087    8.970087
#> sample_86    10.990749   10.990749   10.990749   10.990749   10.990749
#> sample_87    11.295997   11.295997   11.295997   11.295997   11.295997
#> sample_88    11.722164   11.722164   11.722164   11.722164   11.722164
#> sample_89     7.403776    7.403776    7.403776    7.403776    7.403776
#> sample_90     9.852243    9.852243    9.852243    9.852243    9.852243
#> sample_91    11.817259   10.854097    9.605502   11.222950   11.363200
#> sample_92    11.330413   10.326209    8.406221    9.685983    9.649797
#> sample_93     6.883008    7.846961    9.356228    8.627822    8.394306
#> sample_94    12.766506   12.766506   12.766506   12.766506   11.091120
#> sample_95     9.735950    9.735950    9.735950    9.735950   10.227468
#> sample_96    10.903316   10.903316   10.903316   10.903316   10.928960
#> sample_97    11.846257   10.686147   10.686147   10.686147   10.686147
#> sample_98    10.029655    9.548900    9.548900    9.548900    9.548900
#> sample_99    10.380611   11.238650   11.238650   11.238650   11.238650
#> sample_100    8.501050   10.090357   10.090357   10.090357   10.090357
#> 
#> $Chat
#>            condition_1 condition_2 condition_3 condition_4 condition_5
#> sample_1      8.750145    7.601537    7.270550    9.016675    7.325318
#> sample_2      8.956067    9.684558    9.298858    8.350834   10.908594
#> sample_3      8.341249    6.916340    9.119516    6.956300    6.487140
#> sample_4      9.595762   10.065774   11.618364   10.864668   10.191656
#> sample_5     12.172193   11.135928   10.361825   10.752279   11.435870
#> sample_6      9.759622    9.624272    9.356493    9.075333    8.543432
#> sample_7      9.729376    9.564692    8.889309    9.260072    8.742297
#> sample_8      8.118998    9.881806    9.033465    8.259508    8.738236
#> sample_9     10.052032   11.589855   11.859101   10.523539   12.279696
#> sample_10     9.503047    9.028558   10.004456    8.370683   10.512154
#> sample_11     8.471305    9.800171   10.366588    9.244082    9.953022
#> sample_12     8.185030    7.916693    9.377549    9.003281    9.122156
#> sample_13    11.198656    9.544046    8.671010   10.422457    9.881765
#> sample_14     7.746374    7.780033    8.491482    8.301232    7.540014
#> sample_15    10.521436   10.387957   10.435900   10.559605   11.743776
#> sample_16     9.447916    8.779428    8.165658    8.683246    8.591396
#> sample_17     9.486002    9.420560   10.747850    8.264075    9.591289
#> sample_18    11.575723   11.645930   10.305908    9.961501   11.379618
#> sample_19    11.009616   12.733959   10.109610   11.043846   10.465849
#> sample_20    11.005323   10.697277   11.285521   10.524514   11.700674
#> sample_21    10.384334    9.779492   10.939781    9.741301   10.486976
#> sample_22    12.155402   10.528164   11.299787   12.444704    9.553421
#> sample_23    10.137474   12.006052   11.439848   11.069595   11.239138
#> sample_24     9.515884    8.675314    8.922656    9.119117    9.728324
#> sample_25     9.565800    9.689441    9.679743   10.064022    9.667259
#> sample_26     9.449437   10.497043   10.068543    9.465309   11.040557
#> sample_27     9.472099    9.209830    9.088534    9.391572    9.026402
#> sample_28    10.596349   11.346529   11.757752   11.722454   11.163462
#> sample_29     8.691977    9.439442   10.018357    9.618978    8.876401
#> sample_30     9.498067    9.795689    9.282908    9.609451    8.840002
#> sample_31    10.902008   10.251046   11.398386   13.301458   12.693407
#> sample_32    11.653851   11.943043   10.208881   10.195254   10.641150
#> sample_33     9.056630    9.140960    8.945192    8.217446    7.789009
#> sample_34     8.641991    8.052672    7.923789    7.511668    8.722676
#> sample_35     9.460597    8.608157    9.067040    9.406992   10.340649
#> sample_36     8.492207    9.060320    9.412860    8.202729    9.446595
#> sample_37     9.518053   10.556173    9.968051   10.316751    9.596575
#> sample_38    11.269297   11.901634   11.040843   11.401700   11.407701
#> sample_39     9.368206   10.692291   10.979714    9.177392   11.331745
#> sample_40     9.548024    9.552227    9.764823   10.984393    8.879256
#> sample_41     9.815754    9.702344   10.144395    9.443089   11.091563
#> sample_42    11.406933   10.326369   10.148420    9.461638   10.513347
#> sample_43     8.748714    9.631712    8.625044   10.014146    8.788334
#> sample_44     7.982946   10.042066    8.022958    8.823236    8.819867
#> sample_45    10.096988   10.294543    9.768678   10.973936   10.576731
#> sample_46    11.242297   10.928317   12.232380   11.539520   10.140672
#> sample_47     8.808915    9.825378   10.733617    8.378644    8.690187
#> sample_48    10.438372   10.496444    9.899601   11.212596   11.707145
#> sample_49    10.224749   11.092505   11.701471   10.148836   10.878567
#> sample_50    11.529293   11.735790   10.869071   10.837653   10.758710
#> sample_51    11.892057   11.225775   11.588924   11.037764   10.669982
#> sample_52    10.402730   10.288587   10.401020   10.768347   11.513450
#> sample_53    10.370903   10.682131    9.961816   10.090513   10.104529
#> sample_54     9.821219    9.221833   10.002848    9.945160   10.303618
#> sample_55     8.967524    8.107712    8.732756    9.608147   10.252133
#> sample_56    11.158716   10.144238    9.487777   10.218436    9.359144
#> sample_57    11.750186   11.829518   12.140715   10.457789   12.123934
#> sample_58     9.864680   10.599834   10.002657   10.590288    9.419853
#> sample_59     9.225782   11.572498   10.183923   10.402138   10.867556
#> sample_60     8.711116    8.453780    8.918701    9.355961    8.852932
#> sample_61    10.921411   11.756199   11.021228   10.857958   10.314185
#> sample_62    10.418080   10.402758   12.420080   10.458790   11.609704
#> sample_63    10.062799   11.441321    9.646768   10.307671   10.970554
#> sample_64    10.665472   10.722471   11.392795   12.096659    9.525976
#> sample_65    10.775187    9.702575    9.770359   11.357482    9.733496
#> sample_66     9.775622   10.769207   10.187661   10.474032   10.060805
#> sample_67     9.791387   12.026497   11.680603   10.785652   10.213560
#> sample_68    12.430598   11.834627   10.752069   10.909779   11.490035
#> sample_69    10.159950    8.412198    8.667236    9.116866    9.568723
#> sample_70    10.321787   10.136840    9.016249    8.950378   10.761986
#> sample_71     8.997193    9.294440    9.542657   10.100616    8.176733
#> sample_72    10.042604    8.527077   11.653315   11.081460    9.597514
#> sample_73    10.998270    9.077673   10.511058   10.483190    9.880516
#> sample_74     9.102760    8.926719    7.952741    9.757854    9.955984
#> sample_75     8.518148    9.263549    9.069008   10.270025    9.447324
#> sample_76    10.694910   10.955120   11.658669   11.021701   12.724900
#> sample_77     9.744893    8.503935    8.908955    8.637740    9.077893
#> sample_78     9.633104    9.713725    8.824537    9.223342    9.498308
#> sample_79    10.822093   10.860560    9.522851   11.318650   11.867381
#> sample_80    11.114873   10.587396   10.330182   11.007547   11.490360
#> sample_81    10.361588   10.321521    9.438803    8.285627    9.325467
#> sample_82     7.915073    9.550236    8.863468    9.070764    9.090901
#> sample_83    11.281960   11.336796   10.182578   10.172648    8.993420
#> sample_84     9.532833   10.260347    8.636996    9.127333    7.673040
#> sample_85     8.184814    7.959130    8.945172    9.773596    7.428046
#> sample_86    10.103280   12.663609   11.952254   10.749395   12.476938
#> sample_87    11.393829   11.009603   10.958184   10.350159   11.200634
#> sample_88    11.378944   11.347555   11.454079   11.754342   11.229710
#> sample_89     7.770752    7.289155    6.649606    7.775780    7.930099
#> sample_90    10.744754    9.073016    9.132455    9.574488   10.549831
#> sample_91    12.236162   10.649579    9.770179   12.029421   11.175281
#> sample_92    10.003944   10.854573    7.627739    9.392663    8.537950
#> sample_93     5.229784    8.723702    9.049663    7.663481    7.419730
#> sample_94    13.847960   12.161758   13.622464   13.054836   12.015236
#> sample_95     8.938280   10.704552   10.568739    8.091262    9.614941
#> sample_96    10.763391   11.093970   11.106731   10.728789   12.133183
#> sample_97    11.841234   10.746854   10.788356    9.268596   11.220402
#> sample_98     9.815558    9.499992    8.745511    9.172000    9.453615
#> sample_99    10.806782   11.357256   12.033159   10.799755   12.074213
#> sample_100    7.706364   10.722403    9.088520   10.814836    9.134673
#> 
#> $Shat
#>            condition_1 condition_2 condition_3 condition_4 condition_5
#> sample_1     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_2     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_3     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_4     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_5     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_6     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_7     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_8     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_9     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_10    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_11    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_12    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_13    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_14    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_15    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_16    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_17    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_18    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_19    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_20    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_21    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_22    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_23    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_24    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_25    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_26    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_27    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_28    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_29    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_30    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_31    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_32    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_33    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_34    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_35    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_36    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_37    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_38    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_39    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_40    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_41    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_42    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_43    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_44    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_45    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_46    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_47    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_48    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_49    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_50    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_51    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_52    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_53    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_54    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_55    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_56    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_57    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_58    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_59    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_60    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_61    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_62    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_63    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_64    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_65    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_66    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_67    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_68    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_69    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_70    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_71    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_72    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_73    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_74    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_75    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_76    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_77    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_78    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_79    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_80    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_81    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_82    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_83    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_84    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_85    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_86    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_87    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_88    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_89    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_90    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_91    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_92    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_93    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_94    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_95    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_96    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_97    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_98    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_99    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_100   0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#>