Create simulation with signal data used for contrast analysis.

Create simulation with signal data used for contrast analysis.

sim_contrast2(nsamp = 1000, ncond = 5, err_sd = sqrt(0.5))

Arguments

nsamp
Number of samples of each type.

ncond
Number of conditions.

err_sd
The standard deviation of the errors.

Details

The first condition is the reference group. The deviations are the difference between the subsequent conditions with the reference group. The simulation consists of 90 10 different types of deviations: equal among conditions, present only in the first subsequent condition, independent across conditions.

Examples

sim_contrast2(100,5)

#> $C
#>            condition_1 condition_2 condition_3 condition_4 condition_5
#> sample_1      9.335040    9.335040    9.335040    9.335040    9.335040
#> sample_2      9.501823    9.501823    9.501823    9.501823    9.501823
#> sample_3      9.923486    9.923486    9.923486    9.923486    9.923486
#> sample_4     10.579343   10.579343   10.579343   10.579343   10.579343
#> sample_5     10.691646   10.691646   10.691646   10.691646   10.691646
#> sample_6     11.038588   11.038588   11.038588   11.038588   11.038588
#> sample_7     10.241835   10.241835   10.241835   10.241835   10.241835
#> sample_8     10.732969   10.732969   10.732969   10.732969   10.732969
#> sample_9     10.866747   10.866747   10.866747   10.866747   10.866747
#> sample_10     8.731347    8.731347    8.731347    8.731347    8.731347
#> sample_11     9.333163    9.333163    9.333163    9.333163    9.333163
#> sample_12     8.838986    8.838986    8.838986    8.838986    8.838986
#> sample_13     9.846291    9.846291    9.846291    9.846291    9.846291
#> sample_14     9.888220    9.888220    9.888220    9.888220    9.888220
#> sample_15     9.498877    9.498877    9.498877    9.498877    9.498877
#> sample_16     8.907580    8.907580    8.907580    8.907580    8.907580
#> sample_17     8.436729    8.436729    8.436729    8.436729    8.436729
#> sample_18     8.830608    8.830608    8.830608    8.830608    8.830608
#> sample_19    10.003372   10.003372   10.003372   10.003372   10.003372
#> sample_20    11.461109   11.461109   11.461109   11.461109   11.461109
#> sample_21    10.810956   10.810956   10.810956   10.810956   10.810956
#> sample_22    11.153107   11.153107   11.153107   11.153107   11.153107
#> sample_23     8.348012    8.348012    8.348012    8.348012    8.348012
#> sample_24     9.804540    9.804540    9.804540    9.804540    9.804540
#> sample_25    10.432830   10.432830   10.432830   10.432830   10.432830
#> sample_26    10.313061   10.313061   10.313061   10.313061   10.313061
#> sample_27     9.273144    9.273144    9.273144    9.273144    9.273144
#> sample_28     8.768103    8.768103    8.768103    8.768103    8.768103
#> sample_29    10.846161   10.846161   10.846161   10.846161   10.846161
#> sample_30    11.002441   11.002441   11.002441   11.002441   11.002441
#> sample_31     7.902669    7.902669    7.902669    7.902669    7.902669
#> sample_32    10.085002   10.085002   10.085002   10.085002   10.085002
#> sample_33     8.866740    8.866740    8.866740    8.866740    8.866740
#> sample_34    11.256399   11.256399   11.256399   11.256399   11.256399
#> sample_35    10.656066   10.656066   10.656066   10.656066   10.656066
#> sample_36     9.425182    9.425182    9.425182    9.425182    9.425182
#> sample_37    11.106108   11.106108   11.106108   11.106108   11.106108
#> sample_38    11.404706   11.404706   11.404706   11.404706   11.404706
#> sample_39     9.974318    9.974318    9.974318    9.974318    9.974318
#> sample_40    11.169898   11.169898   11.169898   11.169898   11.169898
#> sample_41     9.926486    9.926486    9.926486    9.926486    9.926486
#> sample_42    10.413521   10.413521   10.413521   10.413521   10.413521
#> sample_43     9.021433    9.021433    9.021433    9.021433    9.021433
#> sample_44     9.598045    9.598045    9.598045    9.598045    9.598045
#> sample_45     9.787401    9.787401    9.787401    9.787401    9.787401
#> sample_46     9.676835    9.676835    9.676835    9.676835    9.676835
#> sample_47    11.349053   11.349053   11.349053   11.349053   11.349053
#> sample_48     9.499017    9.499017    9.499017    9.499017    9.499017
#> sample_49    10.787761   10.787761   10.787761   10.787761   10.787761
#> sample_50    10.739053   10.739053   10.739053   10.739053   10.739053
#> sample_51     9.049877    9.049877    9.049877    9.049877    9.049877
#> sample_52     9.565823    9.565823    9.565823    9.565823    9.565823
#> sample_53     9.461733    9.461733    9.461733    9.461733    9.461733
#> sample_54    10.813057   10.813057   10.813057   10.813057   10.813057
#> sample_55     9.583962    9.583962    9.583962    9.583962    9.583962
#> sample_56     9.272542    9.272542    9.272542    9.272542    9.272542
#> sample_57     9.652521    9.652521    9.652521    9.652521    9.652521
#> sample_58    11.597470   11.597470   11.597470   11.597470   11.597470
#> sample_59     9.567444    9.567444    9.567444    9.567444    9.567444
#> sample_60    10.903758   10.903758   10.903758   10.903758   10.903758
#> sample_61     9.486022    9.486022    9.486022    9.486022    9.486022
#> sample_62     9.818143    9.818143    9.818143    9.818143    9.818143
#> sample_63     8.701175    8.701175    8.701175    8.701175    8.701175
#> sample_64    11.586647   11.586647   11.586647   11.586647   11.586647
#> sample_65     8.758266    8.758266    8.758266    8.758266    8.758266
#> sample_66    11.255683   11.255683   11.255683   11.255683   11.255683
#> sample_67     9.941164    9.941164    9.941164    9.941164    9.941164
#> sample_68     9.739645    9.739645    9.739645    9.739645    9.739645
#> sample_69     9.891012    9.891012    9.891012    9.891012    9.891012
#> sample_70    10.139878   10.139878   10.139878   10.139878   10.139878
#> sample_71     9.524576    9.524576    9.524576    9.524576    9.524576
#> sample_72    10.005521   10.005521   10.005521   10.005521   10.005521
#> sample_73    10.847152   10.847152   10.847152   10.847152   10.847152
#> sample_74     9.335701    9.335701    9.335701    9.335701    9.335701
#> sample_75    10.874232   10.874232   10.874232   10.874232   10.874232
#> sample_76    12.118111   12.118111   12.118111   12.118111   12.118111
#> sample_77     9.030472    9.030472    9.030472    9.030472    9.030472
#> sample_78    10.563102   10.563102   10.563102   10.563102   10.563102
#> sample_79     8.751211    8.751211    8.751211    8.751211    8.751211
#> sample_80     9.927527    9.927527    9.927527    9.927527    9.927527
#> sample_81    10.478341   10.478341   10.478341   10.478341   10.478341
#> sample_82     9.107429    9.107429    9.107429    9.107429    9.107429
#> sample_83     8.162294    8.162294    8.162294    8.162294    8.162294
#> sample_84     9.535283    9.535283    9.535283    9.535283    9.535283
#> sample_85     8.214502    8.214502    8.214502    8.214502    8.214502
#> sample_86     8.316661    8.316661    8.316661    8.316661    8.316661
#> sample_87     9.797672    9.797672    9.797672    9.797672    9.797672
#> sample_88     8.951306    8.951306    8.951306    8.951306    8.951306
#> sample_89     9.426364    9.426364    9.426364    9.426364    9.426364
#> sample_90    10.355710   10.355710   10.355710   10.355710   10.355710
#> sample_91    10.834172   10.321366    9.780439   11.616578   10.727807
#> sample_92    10.526937    8.671181   12.260314   10.437342   10.719204
#> sample_93     9.377693    9.349745   10.767425    8.542574    9.238982
#> sample_94    11.524059   11.524059   11.524059   11.524059   11.576349
#> sample_95    10.196001   10.196001   10.196001   10.196001   10.762842
#> sample_96     9.254927    9.254927    9.254927    9.254927    9.652937
#> sample_97    10.622898   10.199035   10.199035   10.199035   10.199035
#> sample_98    12.092314   10.415417   10.415417   10.415417   10.415417
#> sample_99    10.381096   10.752322   10.752322   10.752322   10.752322
#> sample_100    9.282049   10.882167   10.882167   10.882167   10.882167
#> 
#> $Chat
#>            condition_1 condition_2 condition_3 condition_4 condition_5
#> sample_1      9.142662    9.205115    9.171393    9.913120    7.778133
#> sample_2      9.365224    8.999281   10.069109    9.566438    9.515791
#> sample_3      9.414114    9.538456   10.183859    8.746580   10.126230
#> sample_4     11.449269   10.015215   10.026913    9.888337   10.370691
#> sample_5     11.066938   10.346021   11.125840   11.370924   11.213037
#> sample_6      9.922509   10.810195   10.879001   10.469898   11.480406
#> sample_7     11.090277   10.765892   10.002890   11.026698    9.923983
#> sample_8      9.989503   10.299558    9.823983    9.790030   10.954112
#> sample_9     10.447507    9.438738   10.369275   10.627473   11.281282
#> sample_10     9.789464    8.968950   10.089441    9.282456    8.662119
#> sample_11     9.273219    8.980800    8.709362    9.693970    9.515382
#> sample_12     8.282194    9.307768    8.794731    9.013626    9.504888
#> sample_13     8.491908    9.014594    8.959337    9.510964   10.562887
#> sample_14     9.861423    9.620828    9.362818    9.491536   10.341755
#> sample_15     8.782647   10.149904   10.617966    9.701878   10.005464
#> sample_16     9.169968    8.869446    9.531578    8.370236    9.900471
#> sample_17     8.147193    7.343502    8.517877    8.732664    9.198625
#> sample_18     8.645777    9.778184    8.774964    7.979724    8.182586
#> sample_19    10.005266    9.944136    9.001669   10.587429    9.186390
#> sample_20    10.490933   11.428320   11.776464   11.876150   11.576684
#> sample_21    10.983475    9.957364   11.930415   11.070867   10.849894
#> sample_22    11.209932   10.409380   11.240117   10.752882   11.237332
#> sample_23     7.789576    7.560435    8.077867    9.262912    8.371593
#> sample_24     9.763053   10.410358    9.190899   10.671842    8.913962
#> sample_25    10.646680   10.880239   10.645674   10.900370   11.794968
#> sample_26    10.069833   10.301226   10.418253   10.312621   11.365758
#> sample_27     8.122244    7.919614    8.550355   10.806118   11.340667
#> sample_28     9.849249    8.351261   10.314882    8.361045    9.607425
#> sample_29    10.295705   11.456817   10.297286   10.564081   11.039276
#> sample_30    11.742669   11.262915   10.555227    9.622429   10.675891
#> sample_31     9.182904    8.576206    7.070128    8.495226    7.334911
#> sample_32     9.498424   10.619989    9.534622   10.341304    9.937549
#> sample_33     9.137337    8.101077    9.041352    8.616976    8.295495
#> sample_34    12.102766    9.547178   11.267241   11.519080   11.425768
#> sample_35    11.620926    9.299556    9.418393   11.111472   10.938737
#> sample_36     9.107080    9.544295   10.369586    9.704686    7.853826
#> sample_37    11.724668   11.532720   11.160037   10.672554   10.887158
#> sample_38    10.666632   11.938005   11.010213   10.984057   12.232950
#> sample_39     8.598171    9.790533    9.859231   10.163711   10.822290
#> sample_40    10.844198   10.359663   11.759985   10.108372   10.924589
#> sample_41     9.081683    9.525596    9.581664    9.064279   10.279490
#> sample_42     9.861169   12.781664   10.642671   10.749688   10.583412
#> sample_43     8.907980    7.636204    9.303984    8.305832    9.678604
#> sample_44     9.597092    9.849169   10.030959    9.222317   10.750961
#> sample_45    10.908408    9.851350   10.015093   10.921905    9.818347
#> sample_46     8.378839    9.663133    9.875465    9.826161   10.804979
#> sample_47    11.098688   12.027275   12.160644   11.703182   12.077156
#> sample_48     9.330103    9.831336    9.824062   10.796903    9.421824
#> sample_49    11.977560   10.594240   10.864276   10.903780   11.457433
#> sample_50    10.730988   10.725553   11.337991   10.212842    9.616705
#> sample_51     9.880791    9.616992    9.158564    9.586567   10.100306
#> sample_52    10.417387    8.484380    9.054076    8.916258    9.568821
#> sample_53    10.280802    9.690363    9.727380    9.110426    9.054345
#> sample_54    11.938331    9.588060   11.345156   10.770332   11.246833
#> sample_55     8.953681   10.094274    9.203404    9.552419    9.699374
#> sample_56     8.683383    8.867028    9.546834    9.765878    9.408146
#> sample_57     9.550515   10.109080    8.273401    9.030711    9.081348
#> sample_58    10.613528   11.514423   12.184148   11.805039   11.411181
#> sample_59     9.640674   10.349849    9.952259   11.575042   10.040755
#> sample_60    10.970254   12.528394   11.055503   10.214225   11.811888
#> sample_61    10.593814    8.460262    9.151376    9.969542    8.501909
#> sample_62     8.709381    9.699226    8.705174    9.357189   10.232436
#> sample_63     9.821118   10.716751    9.268688    9.239340    7.574012
#> sample_64    11.828301   11.250761   11.039156   12.696539   11.968856
#> sample_65     9.520772    8.424833    9.876551    8.292227    9.379160
#> sample_66    12.349727   11.043200   11.174493   10.449028   11.262341
#> sample_67    10.489506   10.942147    9.645599    9.569416    9.953339
#> sample_68     9.833710    8.964703    9.790072    9.745576    9.804751
#> sample_69    10.050206    9.342425    8.505687   10.391759    9.666285
#> sample_70     9.776440    9.609706    9.428902    9.607515   10.921903
#> sample_71     9.503560    9.838238   10.077142    9.461292    8.705688
#> sample_72    10.696992    9.156933   10.674894   10.625479   10.370666
#> sample_73    11.031516   11.603296   11.013284   10.299538   10.829506
#> sample_74     9.353488    8.909361    9.868878    8.856732    8.504760
#> sample_75    10.215215   10.111079   11.037360   10.897658   12.197656
#> sample_76    12.243505   11.319223   12.011213   11.199569   11.538590
#> sample_77     7.784626    8.244733    8.903889    9.490874    8.113765
#> sample_78     9.729343    8.987713   10.163919   10.718150   10.065163
#> sample_79     8.440132    9.034795    8.938776    8.557617    8.903481
#> sample_80     9.567100    9.867863   10.100328   10.549190    9.121195
#> sample_81     9.649655   11.031906    9.306685   11.544097   10.780691
#> sample_82    10.314870    8.856343    8.150841    8.011961    8.751236
#> sample_83     9.037231    7.622676    8.414278    8.049147    7.534440
#> sample_84     8.858992    9.978653    9.539076    9.342331    9.832533
#> sample_85     8.486119    8.573788    9.435623    8.807198    7.628482
#> sample_86     9.665256    7.955299    8.161205    8.911677    7.868102
#> sample_87     9.925666    8.562432    9.083373    9.803736    9.178029
#> sample_88     9.356384    7.874962    9.630468    8.510875    9.086539
#> sample_89    10.090644    9.234885   10.562144    8.781030   10.737715
#> sample_90    11.456074   11.693962   11.681498   11.710989   10.646376
#> sample_91    10.768801   10.452664   10.374795   11.620082   12.211807
#> sample_92    11.167012    7.334455   11.519585   10.456079   10.565979
#> sample_93     9.002742    8.858940   10.093234    7.151958   10.220378
#> sample_94    12.101630   10.558234   10.678090   11.357507   10.935737
#> sample_95     9.546755   10.533025   10.623723   11.183139    9.620307
#> sample_96     8.861115    8.421484    9.118687    9.120883    9.680449
#> sample_97    11.853171   11.003518   11.381679    9.757181   10.673621
#> sample_98    12.669448   10.647358    9.663868   10.257140   10.003625
#> sample_99     9.289353   13.809014   11.113123   11.194250   10.175980
#> sample_100    8.862304   10.476672   11.483424   10.022075   10.859443
#> 
#> $Shat
#>            condition_1 condition_2 condition_3 condition_4 condition_5
#> sample_1     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_2     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_3     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_4     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_5     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_6     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_7     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_8     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_9     0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_10    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_11    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_12    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_13    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_14    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_15    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_16    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_17    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_18    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_19    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_20    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_21    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_22    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_23    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_24    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_25    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_26    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_27    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_28    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_29    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_30    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_31    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_32    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_33    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_34    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_35    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_36    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_37    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_38    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_39    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_40    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_41    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_42    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_43    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_44    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_45    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_46    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_47    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_48    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_49    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_50    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_51    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_52    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_53    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_54    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_55    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_56    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_57    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_58    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_59    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_60    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_61    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_62    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_63    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_64    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_65    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_66    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_67    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_68    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_69    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_70    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_71    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_72    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_73    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_74    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_75    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_76    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_77    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_78    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_79    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_80    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_81    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_82    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_83    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_84    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_85    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_86    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_87    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_88    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_89    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_90    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_91    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_92    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_93    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_94    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_95    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_96    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_97    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_98    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_99    0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#> sample_100   0.7071068   0.7071068   0.7071068   0.7071068   0.7071068
#>