Create some more simple simulated data for testing purposes

Create some more simple simulated data for testing purposes

simple_sims2(nsamp = 100, err_sd = 0.01)

Arguments

nsamp
number of samples of each type

err_sd
the standard deviation of the errors

Details

The simulation consists of five conditions with two types of effecc those present (and identical) in first two conditions and those present (and identical) in last three conditions

Examples

simple_sims2(100, 5)

#> $B
#>              condition_1   condition_2 condition_3 condition_4 condition_5
#> effect_1    1.5329409093  1.5329409093  0.00000000  0.00000000  0.00000000
#> effect_2    1.9735197822  1.9735197822  0.00000000  0.00000000  0.00000000
#> effect_3   -1.6021096223 -1.6021096223  0.00000000  0.00000000  0.00000000
#> effect_4   -0.1902448515 -0.1902448515  0.00000000  0.00000000  0.00000000
#> effect_5    0.0587437208  0.0587437208  0.00000000  0.00000000  0.00000000
#> effect_6   -0.3070393887 -0.3070393887  0.00000000  0.00000000  0.00000000
#> effect_7   -0.3202120435 -0.3202120435  0.00000000  0.00000000  0.00000000
#> effect_8   -2.0675736772 -2.0675736772  0.00000000  0.00000000  0.00000000
#> effect_9   -0.2381833028 -0.2381833028  0.00000000  0.00000000  0.00000000
#> effect_10   1.8479000186  1.8479000186  0.00000000  0.00000000  0.00000000
#> effect_11   1.7541434981  1.7541434981  0.00000000  0.00000000  0.00000000
#> effect_12   0.2290141964  0.2290141964  0.00000000  0.00000000  0.00000000
#> effect_13   0.0050640664  0.0050640664  0.00000000  0.00000000  0.00000000
#> effect_14  -0.7884400878 -0.7884400878  0.00000000  0.00000000  0.00000000
#> effect_15  -0.0005658213 -0.0005658213  0.00000000  0.00000000  0.00000000
#> effect_16   0.8289295108  0.8289295108  0.00000000  0.00000000  0.00000000
#> effect_17  -0.1375809653 -0.1375809653  0.00000000  0.00000000  0.00000000
#> effect_18   1.6161599135  1.6161599135  0.00000000  0.00000000  0.00000000
#> effect_19  -1.1428148387 -1.1428148387  0.00000000  0.00000000  0.00000000
#> effect_20   1.2865632879  1.2865632879  0.00000000  0.00000000  0.00000000
#> effect_21   0.3318319879  0.3318319879  0.00000000  0.00000000  0.00000000
#> effect_22   0.1951731244  0.1951731244  0.00000000  0.00000000  0.00000000
#> effect_23   0.1761853543  0.1761853543  0.00000000  0.00000000  0.00000000
#> effect_24   1.6595817616  1.6595817616  0.00000000  0.00000000  0.00000000
#> effect_25   1.4099004615  1.4099004615  0.00000000  0.00000000  0.00000000
#> effect_26   0.9161915892  0.9161915892  0.00000000  0.00000000  0.00000000
#> effect_27   1.1503390755  1.1503390755  0.00000000  0.00000000  0.00000000
#> effect_28   1.2257965687  1.2257965687  0.00000000  0.00000000  0.00000000
#> effect_29  -0.2870601510 -0.2870601510  0.00000000  0.00000000  0.00000000
#> effect_30   0.5100599240  0.5100599240  0.00000000  0.00000000  0.00000000
#> effect_31  -0.3072482850 -0.3072482850  0.00000000  0.00000000  0.00000000
#> effect_32   0.6682042102  0.6682042102  0.00000000  0.00000000  0.00000000
#> effect_33  -0.7396265086 -0.7396265086  0.00000000  0.00000000  0.00000000
#> effect_34   0.7171519656  0.7171519656  0.00000000  0.00000000  0.00000000
#> effect_35   0.0239914957  0.0239914957  0.00000000  0.00000000  0.00000000
#> effect_36   1.2073887021  1.2073887021  0.00000000  0.00000000  0.00000000
#> effect_37  -0.9172046310 -0.9172046310  0.00000000  0.00000000  0.00000000
#> effect_38   0.3403124158  0.3403124158  0.00000000  0.00000000  0.00000000
#> effect_39  -1.4576800168 -1.4576800168  0.00000000  0.00000000  0.00000000
#> effect_40   0.8996282803  0.8996282803  0.00000000  0.00000000  0.00000000
#> effect_41  -0.4840212635 -0.4840212635  0.00000000  0.00000000  0.00000000
#> effect_42  -0.5966185739 -0.5966185739  0.00000000  0.00000000  0.00000000
#> effect_43  -1.8940052085 -1.8940052085  0.00000000  0.00000000  0.00000000
#> effect_44  -1.1560470146 -1.1560470146  0.00000000  0.00000000  0.00000000
#> effect_45   0.4111188581  0.4111188581  0.00000000  0.00000000  0.00000000
#> effect_46  -1.2924490117 -1.2924490117  0.00000000  0.00000000  0.00000000
#> effect_47   0.7435921631  0.7435921631  0.00000000  0.00000000  0.00000000
#> effect_48  -1.1379286211 -1.1379286211  0.00000000  0.00000000  0.00000000
#> effect_49  -0.8939760340 -0.8939760340  0.00000000  0.00000000  0.00000000
#> effect_50  -0.5420926472 -0.5420926472  0.00000000  0.00000000  0.00000000
#> effect_51  -0.7062708698 -0.7062708698  0.00000000  0.00000000  0.00000000
#> effect_52   0.9533891795  0.9533891795  0.00000000  0.00000000  0.00000000
#> effect_53  -0.9540376361 -0.9540376361  0.00000000  0.00000000  0.00000000
#> effect_54   0.6894286139  0.6894286139  0.00000000  0.00000000  0.00000000
#> effect_55   1.3830709236  1.3830709236  0.00000000  0.00000000  0.00000000
#> effect_56  -0.1764552060 -0.1764552060  0.00000000  0.00000000  0.00000000
#> effect_57  -0.1663701986 -0.1663701986  0.00000000  0.00000000  0.00000000
#> effect_58  -0.1135201099 -0.1135201099  0.00000000  0.00000000  0.00000000
#> effect_59  -0.1936064125 -0.1936064125  0.00000000  0.00000000  0.00000000
#> effect_60   0.2581472814  0.2581472814  0.00000000  0.00000000  0.00000000
#> effect_61  -0.5383125734 -0.5383125734  0.00000000  0.00000000  0.00000000
#> effect_62  -1.1790628113 -1.1790628113  0.00000000  0.00000000  0.00000000
#> effect_63   0.9006474226  0.9006474226  0.00000000  0.00000000  0.00000000
#> effect_64  -0.0162495475 -0.0162495475  0.00000000  0.00000000  0.00000000
#> effect_65   0.1786304898  0.1786304898  0.00000000  0.00000000  0.00000000
#> effect_66   0.6019090196  0.6019090196  0.00000000  0.00000000  0.00000000
#> effect_67   0.2611343796  0.2611343796  0.00000000  0.00000000  0.00000000
#> effect_68   0.3040656602  0.3040656602  0.00000000  0.00000000  0.00000000
#> effect_69  -0.6450103999 -0.6450103999  0.00000000  0.00000000  0.00000000
#> effect_70   0.4092446719  0.4092446719  0.00000000  0.00000000  0.00000000
#> effect_71  -1.0378114459 -1.0378114459  0.00000000  0.00000000  0.00000000
#> effect_72   0.2385615800  0.2385615800  0.00000000  0.00000000  0.00000000
#> effect_73   1.2160376895  1.2160376895  0.00000000  0.00000000  0.00000000
#> effect_74  -0.5933441907 -0.5933441907  0.00000000  0.00000000  0.00000000
#> effect_75  -1.3026824066 -1.3026824066  0.00000000  0.00000000  0.00000000
#> effect_76  -0.2056494610 -0.2056494610  0.00000000  0.00000000  0.00000000
#> effect_77  -1.0089977494 -1.0089977494  0.00000000  0.00000000  0.00000000
#> effect_78   0.3754811712  0.3754811712  0.00000000  0.00000000  0.00000000
#> effect_79   0.0232051627  0.0232051627  0.00000000  0.00000000  0.00000000
#> effect_80   1.1644853805  1.1644853805  0.00000000  0.00000000  0.00000000
#> effect_81  -0.0448651864 -0.0448651864  0.00000000  0.00000000  0.00000000
#> effect_82   0.4501868089  0.4501868089  0.00000000  0.00000000  0.00000000
#> effect_83  -0.0368351435 -0.0368351435  0.00000000  0.00000000  0.00000000
#> effect_84   1.1048666852  1.1048666852  0.00000000  0.00000000  0.00000000
#> effect_85  -0.3774209238 -0.3774209238  0.00000000  0.00000000  0.00000000
#> effect_86   0.2854741622  0.2854741622  0.00000000  0.00000000  0.00000000
#> effect_87   0.1170416669  0.1170416669  0.00000000  0.00000000  0.00000000
#> effect_88   2.3687635279  2.3687635279  0.00000000  0.00000000  0.00000000
#> effect_89   0.4639779553  0.4639779553  0.00000000  0.00000000  0.00000000
#> effect_90   1.3541388751  1.3541388751  0.00000000  0.00000000  0.00000000
#> effect_91  -1.2826011268 -1.2826011268  0.00000000  0.00000000  0.00000000
#> effect_92   0.2639631025  0.2639631025  0.00000000  0.00000000  0.00000000
#> effect_93  -0.8584770741 -0.8584770741  0.00000000  0.00000000  0.00000000
#> effect_94   0.3295136553  0.3295136553  0.00000000  0.00000000  0.00000000
#> effect_95   0.6210244859  0.6210244859  0.00000000  0.00000000  0.00000000
#> effect_96  -0.1109528220 -0.1109528220  0.00000000  0.00000000  0.00000000
#> effect_97  -1.2730599833 -1.2730599833  0.00000000  0.00000000  0.00000000
#> effect_98   0.0615776415  0.0615776415  0.00000000  0.00000000  0.00000000
#> effect_99  -0.0571990996 -0.0571990996  0.00000000  0.00000000  0.00000000
#> effect_100  1.6411026202  1.6411026202  0.00000000  0.00000000  0.00000000
#> effect_101  0.0000000000  0.0000000000  0.27917625  0.27917625  0.27917625
#> effect_102  0.0000000000  0.0000000000  0.74130971  0.74130971  0.74130971
#> effect_103  0.0000000000  0.0000000000 -1.84291687 -1.84291687 -1.84291687
#> effect_104  0.0000000000  0.0000000000  2.34960183  2.34960183  2.34960183
#> effect_105  0.0000000000  0.0000000000 -0.99590878 -0.99590878 -0.99590878
#> effect_106  0.0000000000  0.0000000000 -0.95011955 -0.95011955 -0.95011955
#> effect_107  0.0000000000  0.0000000000  1.51159974  1.51159974  1.51159974
#> effect_108  0.0000000000  0.0000000000  0.05142642  0.05142642  0.05142642
#> effect_109  0.0000000000  0.0000000000 -1.01016327 -1.01016327 -1.01016327
#> effect_110  0.0000000000  0.0000000000  2.02829156  2.02829156  2.02829156
#> effect_111  0.0000000000  0.0000000000 -0.47938788 -0.47938788 -0.47938788
#> effect_112  0.0000000000  0.0000000000 -2.60759390 -2.60759390 -2.60759390
#> effect_113  0.0000000000  0.0000000000  1.23311019  1.23311019  1.23311019
#> effect_114  0.0000000000  0.0000000000  0.17477714  0.17477714  0.17477714
#> effect_115  0.0000000000  0.0000000000  0.05563116  0.05563116  0.05563116
#> effect_116  0.0000000000  0.0000000000 -0.98324947 -0.98324947 -0.98324947
#> effect_117  0.0000000000  0.0000000000  0.20924235  0.20924235  0.20924235
#> effect_118  0.0000000000  0.0000000000  0.40491194  0.40491194  0.40491194
#> effect_119  0.0000000000  0.0000000000 -0.88145331 -0.88145331 -0.88145331
#> effect_120  0.0000000000  0.0000000000  0.85620360  0.85620360  0.85620360
#> effect_121  0.0000000000  0.0000000000  1.32074827  1.32074827  1.32074827
#> effect_122  0.0000000000  0.0000000000  0.65458507  0.65458507  0.65458507
#> effect_123  0.0000000000  0.0000000000  0.80501509  0.80501509  0.80501509
#> effect_124  0.0000000000  0.0000000000  1.37867950  1.37867950  1.37867950
#> effect_125  0.0000000000  0.0000000000 -1.02246389 -1.02246389 -1.02246389
#> effect_126  0.0000000000  0.0000000000  0.69744920  0.69744920  0.69744920
#> effect_127  0.0000000000  0.0000000000 -0.29411257 -0.29411257 -0.29411257
#> effect_128  0.0000000000  0.0000000000 -1.70915735 -1.70915735 -1.70915735
#> effect_129  0.0000000000  0.0000000000 -0.39824729 -0.39824729 -0.39824729
#> effect_130  0.0000000000  0.0000000000 -1.34952586 -1.34952586 -1.34952586
#> effect_131  0.0000000000  0.0000000000 -0.26989756 -0.26989756 -0.26989756
#> effect_132  0.0000000000  0.0000000000 -1.02630569 -1.02630569 -1.02630569
#> effect_133  0.0000000000  0.0000000000 -0.88977134 -0.88977134 -0.88977134
#> effect_134  0.0000000000  0.0000000000  0.15739619  0.15739619  0.15739619
#> effect_135  0.0000000000  0.0000000000  2.29191592  2.29191592  2.29191592
#> effect_136  0.0000000000  0.0000000000  0.07194875  0.07194875  0.07194875
#> effect_137  0.0000000000  0.0000000000  1.37079135  1.37079135  1.37079135
#> effect_138  0.0000000000  0.0000000000  0.77917821  0.77917821  0.77917821
#> effect_139  0.0000000000  0.0000000000 -1.40733258 -1.40733258 -1.40733258
#> effect_140  0.0000000000  0.0000000000 -0.95262326 -0.95262326 -0.95262326
#> effect_141  0.0000000000  0.0000000000  1.45243790  1.45243790  1.45243790
#> effect_142  0.0000000000  0.0000000000  0.36717937  0.36717937  0.36717937
#> effect_143  0.0000000000  0.0000000000  0.42150471  0.42150471  0.42150471
#> effect_144  0.0000000000  0.0000000000  2.16996295  2.16996295  2.16996295
#> effect_145  0.0000000000  0.0000000000 -1.45314523 -1.45314523 -1.45314523
#> effect_146  0.0000000000  0.0000000000  1.20742671  1.20742671  1.20742671
#> effect_147  0.0000000000  0.0000000000 -0.28177813 -0.28177813 -0.28177813
#> effect_148  0.0000000000  0.0000000000  0.69837491  0.69837491  0.69837491
#> effect_149  0.0000000000  0.0000000000  1.28276495  1.28276495  1.28276495
#> effect_150  0.0000000000  0.0000000000  1.50210009  1.50210009  1.50210009
#> effect_151  0.0000000000  0.0000000000  0.31424977  0.31424977  0.31424977
#> effect_152  0.0000000000  0.0000000000  0.46426977  0.46426977  0.46426977
#> effect_153  0.0000000000  0.0000000000 -1.23017954 -1.23017954 -1.23017954
#> effect_154  0.0000000000  0.0000000000  1.96092847  1.96092847  1.96092847
#> effect_155  0.0000000000  0.0000000000  0.19804968  0.19804968  0.19804968
#> effect_156  0.0000000000  0.0000000000  0.32778881  0.32778881  0.32778881
#> effect_157  0.0000000000  0.0000000000 -0.25423179 -0.25423179 -0.25423179
#> effect_158  0.0000000000  0.0000000000 -2.12721108 -2.12721108 -2.12721108
#> effect_159  0.0000000000  0.0000000000  0.56037077  0.56037077  0.56037077
#> effect_160  0.0000000000  0.0000000000  0.94489164  0.94489164  0.94489164
#> effect_161  0.0000000000  0.0000000000 -0.49037454 -0.49037454 -0.49037454
#> effect_162  0.0000000000  0.0000000000  0.68196215  0.68196215  0.68196215
#> effect_163  0.0000000000  0.0000000000 -0.32531591 -0.32531591 -0.32531591
#> effect_164  0.0000000000  0.0000000000  1.07424896  1.07424896  1.07424896
#> effect_165  0.0000000000  0.0000000000 -0.31242144 -0.31242144 -0.31242144
#> effect_166  0.0000000000  0.0000000000  1.27545459  1.27545459  1.27545459
#> effect_167  0.0000000000  0.0000000000  1.20278892  1.20278892  1.20278892
#> effect_168  0.0000000000  0.0000000000 -1.04831148 -1.04831148 -1.04831148
#> effect_169  0.0000000000  0.0000000000  1.81301588  1.81301588  1.81301588
#> effect_170  0.0000000000  0.0000000000 -0.48140438 -0.48140438 -0.48140438
#> effect_171  0.0000000000  0.0000000000  0.55395286  0.55395286  0.55395286
#> effect_172  0.0000000000  0.0000000000  0.30314990  0.30314990  0.30314990
#> effect_173  0.0000000000  0.0000000000  1.61954435  1.61954435  1.61954435
#> effect_174  0.0000000000  0.0000000000  1.68376695  1.68376695  1.68376695
#> effect_175  0.0000000000  0.0000000000 -0.37509320 -0.37509320 -0.37509320
#> effect_176  0.0000000000  0.0000000000 -1.56599511 -1.56599511 -1.56599511
#> effect_177  0.0000000000  0.0000000000  0.05805108  0.05805108  0.05805108
#> effect_178  0.0000000000  0.0000000000 -0.28348838 -0.28348838 -0.28348838
#> effect_179  0.0000000000  0.0000000000  0.23102389  0.23102389  0.23102389
#> effect_180  0.0000000000  0.0000000000 -0.81398149 -0.81398149 -0.81398149
#> effect_181  0.0000000000  0.0000000000  0.58799338  0.58799338  0.58799338
#> effect_182  0.0000000000  0.0000000000  0.47525212  0.47525212  0.47525212
#> effect_183  0.0000000000  0.0000000000  1.68212626  1.68212626  1.68212626
#> effect_184  0.0000000000  0.0000000000  0.94383928  0.94383928  0.94383928
#> effect_185  0.0000000000  0.0000000000  0.30671753  0.30671753  0.30671753
#> effect_186  0.0000000000  0.0000000000 -1.27285428 -1.27285428 -1.27285428
#> effect_187  0.0000000000  0.0000000000  0.34988199  0.34988199  0.34988199
#> effect_188  0.0000000000  0.0000000000  0.57311832  0.57311832  0.57311832
#> effect_189  0.0000000000  0.0000000000  0.26019444  0.26019444  0.26019444
#> effect_190  0.0000000000  0.0000000000  0.08372344  0.08372344  0.08372344
#> effect_191  0.0000000000  0.0000000000  0.10377362  0.10377362  0.10377362
#> effect_192  0.0000000000  0.0000000000  1.68989993  1.68989993  1.68989993
#> effect_193  0.0000000000  0.0000000000 -1.64694733 -1.64694733 -1.64694733
#> effect_194  0.0000000000  0.0000000000  0.06320843  0.06320843  0.06320843
#> effect_195  0.0000000000  0.0000000000  0.64722493  0.64722493  0.64722493
#> effect_196  0.0000000000  0.0000000000 -0.61962152 -0.61962152 -0.61962152
#> effect_197  0.0000000000  0.0000000000  1.13313298  1.13313298  1.13313298
#> effect_198  0.0000000000  0.0000000000  0.54358381  0.54358381  0.54358381
#> effect_199  0.0000000000  0.0000000000  1.93330001  1.93330001  1.93330001
#> effect_200  0.0000000000  0.0000000000  0.59619624  0.59619624  0.59619624
#> 
#> $Bhat
#>              condition_1  condition_2  condition_3  condition_4  condition_5
#> effect_1    -2.745468099  -1.61721909   2.05860374   5.76474265   4.28118222
#> effect_2     2.080449032   8.49427852  -2.23350593 -12.27805706   0.26318114
#> effect_3    10.569112790   1.06471211  -1.29476002   4.81390615   0.11240418
#> effect_4     8.731498616  -9.45334600  -7.20955253  -2.40818250  -4.36489955
#> effect_5     4.076381225   3.06089077   0.82332840  -2.04242763  -5.43189749
#> effect_6     9.261706602  -1.68958815   2.05244331   8.42430279   2.98750800
#> effect_7     1.805471242  -1.72657879   3.24349696   8.00537605  -1.56442012
#> effect_8    -6.238687959  -3.64196658  -1.49198695  -1.92316025  -6.72117780
#> effect_9    -0.698477140   3.08396201  -7.29795439  -1.85101731   1.14543822
#> effect_10    8.959137666 -11.72154579  -2.76487830  -6.73446659   1.89730881
#> effect_11   -2.962536728   8.02194927  -4.94341388  -2.47614971  -1.38739738
#> effect_12    4.863228001   9.16327651   1.47586227   0.46692043   4.36722756
#> effect_13   -0.044353396  -2.69617487  -3.93347114   1.76189913   3.42425132
#> effect_14   -0.998220176  17.01379142  -3.28630291   7.14399692  -0.05380048
#> effect_15   -0.743924582  -2.08926519   2.50230079   2.55476558  -3.49015271
#> effect_16    2.210425444   1.85256866   9.59215041  -2.73810080  -2.59421710
#> effect_17   -5.725005910  -4.35885097  -8.77090035   0.88703754   0.59755601
#> effect_18   -3.746399560   0.91612744   0.36381579  -4.56666683  -1.33187441
#> effect_19   -8.580259477  -0.26195407  -4.65296942  -3.64566212   2.91436949
#> effect_20   -5.847714543   3.36841857  -3.23488864   2.70075376  -0.11269103
#> effect_21    9.683876933  -5.00305002   4.27000079  -7.08812881   0.33770124
#> effect_22   -0.099387028   2.26665810   0.59815095   6.65716262  -4.71973154
#> effect_23    6.663002108  -6.73360122   4.13299412   8.20906984  -8.94304417
#> effect_24    0.144053863  -1.42014342   4.52903380  -0.77018729  -4.96962017
#> effect_25    0.529848628  -8.65136921   6.89128824  -1.30391458  -4.48425902
#> effect_26   -3.493504197   3.30506653   9.11444611  -2.11872843  -4.92001783
#> effect_27    1.723048265   3.18454053  -6.90466235  -3.56823417  -4.12406857
#> effect_28    4.372550400  -0.67465632  -6.17720062   1.20708377  -0.67689966
#> effect_29   -4.603632384   6.18581527  -2.36856808  -5.82413848   0.96450615
#> effect_30   -0.384924610   3.89788506  -5.42097346  -5.23616755   3.21499737
#> effect_31    4.404579883   0.44861716  -5.52120299  -2.75410376  -2.47624255
#> effect_32   -6.575276899  -2.08230410   3.18305896   0.02020546   6.08003450
#> effect_33   -0.123886233  -1.57747803  -0.53191230  -0.36382462  -8.15317866
#> effect_34  -11.314438206  -4.11989328   1.40286832  -8.84375743  -5.10447525
#> effect_35   -9.822083278  -2.20924234  -0.90339123   2.99193894  -1.47490137
#> effect_36   -0.248963649  -6.24885957  -2.69450269  -1.01794639   2.76372655
#> effect_37    1.713554933   0.12327195   4.07452953  -5.44912403  -5.28021375
#> effect_38    8.169571959  -0.80700790  -4.11727523  -2.00935922   4.12704806
#> effect_39    2.324746622  -1.85226046  10.62440301   4.83443657   1.27697427
#> effect_40    0.428027033  -0.91471245   6.82461741  -4.83343517   5.56951354
#> effect_41   -0.685413655   6.20013927  -4.18068064   0.95923565   1.17415146
#> effect_42    0.488430624   8.50424467  -5.21777110   8.03015206   4.54310668
#> effect_43   -4.469323686 -12.32080004   1.46555843  -1.42491856   1.09065945
#> effect_44    6.144802572  -1.63221556   5.30159768   1.76310088   7.09697336
#> effect_45    6.216982640   1.69221306  -5.09224238  -2.01116024  -4.97502686
#> effect_46  -11.595240652  -1.79252089  -0.51927521  -2.05279378   2.37826856
#> effect_47    2.268537620   7.92122913  -0.64043686   5.38130401   2.26858153
#> effect_48    1.755771400  -8.14182819  -0.08194576   1.22077613  -5.59980223
#> effect_49   -1.140477538  -4.04624318  -8.88163677   4.94813115   1.74113145
#> effect_50    2.895522854  -0.30908537  -0.52299544   1.18142529  -2.76126638
#> effect_51    3.776794423   6.07798243   1.93420135  -0.90310727   3.21629348
#> effect_52    5.343596939   1.28247115   9.86218914  -4.83579794   1.62995979
#> effect_53    2.469106398  -2.71166834   1.79617818  -3.23886696  -1.56392741
#> effect_54   -7.928179872   6.70701096  -0.19069514  -0.13000026  -4.73520823
#> effect_55    3.654654423   2.76169813   7.10872308  -7.34868857  -4.59104714
#> effect_56   -2.631731871  -1.43288351   5.40738950   8.68757807   2.10454717
#> effect_57    0.385090237  -0.05367261   2.31374331   4.30350579  -6.65143910
#> effect_58   -0.623425925   7.50475746   3.75677090   1.23152872  -0.77258666
#> effect_59    1.965481954  -6.57742197  -3.52169182  -3.59203715   4.76369012
#> effect_60   -0.457790268   7.19853436  -2.16742434  -4.72984476  -5.54033723
#> effect_61    3.293600018  -1.39214165   2.38046637   2.07568707   2.83780905
#> effect_62   -9.047514570  -3.36332121  10.90288349   8.37119360   1.78127657
#> effect_63    5.797288813   3.90719176  -2.87631613   0.98226289   2.16971269
#> effect_64   -2.097694275  -7.99873325  -4.93650559   0.71961255  -5.40333187
#> effect_65   -8.204432727   3.94041100   3.02240509   1.89952451   1.07355089
#> effect_66   -4.724380515   2.05944748   0.37845339  -2.83807601   4.12593404
#> effect_67   -4.673014597   3.97435439   1.58550670   1.33130148  -2.20674913
#> effect_68    4.787787347  13.48004383  -0.74653169   6.18189892   1.59769240
#> effect_69   -4.591747680   6.77230113   6.51281277 -12.64657900   5.16843432
#> effect_70   -0.358680997  -0.77574681   1.62372304  -5.36849903  -8.18125777
#> effect_71  -10.127549964  -6.12728906  -2.01069251   2.12193098  -6.98743944
#> effect_72    4.342397831  10.20344243  -5.71409340   0.32616577   4.85556171
#> effect_73    1.365181567   1.06637395   5.79228853  -6.46647748  -0.27298609
#> effect_74    9.277928411  -0.18647446  -0.43294144  -3.35236734 -12.31352699
#> effect_75   -4.081970420   2.15589559  -2.27712234   7.73502119   0.53139575
#> effect_76   -1.186395719  -1.59828660   3.19947797   1.76065863  -3.31954369
#> effect_77   -3.554769614  -9.65098790  -0.14107402  -3.90876247  -3.41627134
#> effect_78    2.736369284   5.86610466  -8.74884830  -0.23851011   4.62799562
#> effect_79   -8.924911487  -1.78259593   4.53023262   4.60788686  -1.06893273
#> effect_80   -5.115489031  -0.32859436   7.95517163  -0.93717001  -3.89938264
#> effect_81   -5.610643851   2.53822362   4.33471149  -2.81465517   7.80361901
#> effect_82    1.981084216   6.40796978  -0.02623233   4.49094407  -5.96029500
#> effect_83   -6.781468535   2.75744034  -4.42771534  -3.40989228   0.81190952
#> effect_84    6.792541686   1.09693835  -0.55161539   1.39705509   5.10743215
#> effect_85    0.941050464   2.07932164  -2.75046177  -1.10312233  -1.19914855
#> effect_86    0.969161692  -4.22172174   8.26440868  -2.81896259  -3.54916389
#> effect_87   -7.361992179   3.37865758   3.24517210  -0.32318783   3.91633647
#> effect_88    4.129254598  -6.31743146   3.33685884   1.19135199   7.97026812
#> effect_89    5.058476026  -0.17288870   0.98310037  -4.07457736  -5.51216959
#> effect_90   -6.304396312   0.66802363   0.04618440  -2.72175692  -7.67507962
#> effect_91    2.755974114  -4.78589900  11.91723831  -1.69997291  -3.06288639
#> effect_92   -4.084061313   0.07843657  -1.22205775  -2.08848656  -1.36986125
#> effect_93    3.078683732  -9.95152522   5.12358813   4.91944345  -2.24182601
#> effect_94    7.217606066   6.09906322  -2.96238799   1.07953755  -8.05248465
#> effect_95    0.580525248   5.35561725  -0.17747887   1.60252098   5.95705922
#> effect_96    1.514656306  -1.03215654   2.30120648  -4.23409953   5.15984051
#> effect_97    5.336938288  -2.96290264  -6.31766408   3.17334016   5.09292381
#> effect_98   -4.895068864  -2.15531200  -1.76436683   0.85411731  -3.81944727
#> effect_99   -2.603828947  -3.07750218  -0.05435164   7.42616619  -5.92731240
#> effect_100   8.281889431  -3.64987955  -2.48476429  -2.48632869   2.85479386
#> effect_101  -1.338046685  -0.69247702   2.65205898   5.04720010   6.61147346
#> effect_102  -7.090839484   1.27326389   2.71464555  -3.53187497   3.89813324
#> effect_103  -0.064936168  -1.56317192  -3.01615957  -7.11496628   1.52818421
#> effect_104  -3.962333454   6.24217321   6.23339105   3.95354642   0.99585829
#> effect_105   4.734848545   4.81175358  -0.28449359  -3.16857877  -2.27459891
#> effect_106   4.704471388  -6.42111295  -1.04414487  -3.78957560  -6.95090105
#> effect_107   0.502727891   5.38262285   1.35982483  -2.90868168  -1.04960779
#> effect_108  -2.159211649  -4.16889912   0.86443136  -6.77759854  -3.50361936
#> effect_109   0.559961317   8.75650770  -2.24805155  -8.50268510   9.07790170
#> effect_110  -1.459486650   3.22351958   9.72995250   9.36382355   6.11751609
#> effect_111  -2.842866579  -0.66127225  -4.89056598   2.79667204  -0.45528636
#> effect_112   5.389697630   0.34984160  -2.14414466   6.59608567   1.41440924
#> effect_113  -2.835519940  -2.51537948   1.75118119  -0.10319822  -2.62120115
#> effect_114  -1.021085922  -0.19952627   9.06559248  -1.89657953   2.82233331
#> effect_115   2.819028381   0.61271363  -0.63486641  -7.51570666  -2.27406724
#> effect_116  -3.194861531  -3.16136451   2.97692243   1.76089620  -2.28139247
#> effect_117  -3.747091125  -1.79098779 -12.04713847   2.12324290  -1.41635006
#> effect_118   3.345684796   0.48976064  -6.14517640   7.78185861  -0.65276070
#> effect_119  -5.070322683  -1.61173273  -2.61471822  -0.42098768  -3.12116981
#> effect_120   3.389086588  -4.63959609  -2.44737334   1.16914495   4.69760850
#> effect_121  11.462240557  -1.45397576 -10.43368595  -0.27461469   6.40662848
#> effect_122  -5.999123491  -2.06669979   5.32962047   1.89853490  -7.12108011
#> effect_123  -8.365808819   8.80007073  11.88594310  -1.58245666  -2.14038532
#> effect_124   3.133237534   7.62510795  -8.74490845   3.37847588   9.95302979
#> effect_125   3.564434671  -4.32292837  -3.14462465  -1.62968426   4.83619566
#> effect_126   1.394811761   4.81860928  -2.54694406   2.12334771  -0.05575632
#> effect_127   3.669462497   2.03512047   1.84794156   4.51760339   0.27304166
#> effect_128  -1.207762305  -2.52680660  -8.72043015  -2.14764010   1.75560787
#> effect_129   0.130748467  -7.32368270   3.48017544   0.77789617   1.35288610
#> effect_130  -0.978029876  -5.83996789   0.98843025  -1.76908226  -4.43794545
#> effect_131   0.731791781   4.49770706   2.11778740   7.52160257  -7.94072337
#> effect_132  10.988817807  -0.82744585  -2.34122687  -3.03206488  -2.19532544
#> effect_133  -0.018248810   4.06747165  -2.42733551   1.12321717   5.94808170
#> effect_134 -10.113343114   0.37290249   0.45939795  -6.09091086   8.93983429
#> effect_135   2.629709345  -1.35181705  -3.75647987   1.56387704   5.52688380
#> effect_136   0.266411697   4.22229877   0.25375834  -2.78012637   1.57207145
#> effect_137  -4.148699964  -3.16389881   1.21891917   3.58011182  12.24589365
#> effect_138  -3.751253529   4.61948697   3.02188835   2.85468743   8.29767596
#> effect_139   2.986266734  -5.98470782  -8.22392143   6.14014349   1.08436446
#> effect_140  -3.055041945   1.66263521   6.86616654   0.55815371   4.36938480
#> effect_141   5.745168772  -9.35273180   5.95485011  -4.44452621   2.58011853
#> effect_142  -1.872242820  -8.06080323   9.80121781   1.41636643   1.19183175
#> effect_143 -13.698859867  -3.72817043   6.42975136   0.54204359   3.26227416
#> effect_144  -5.692370528  -4.09453899  -3.25617046  -1.17661510  10.32827814
#> effect_145  -6.871824395  -1.82221584 -11.11478350   1.69797773  -5.33038494
#> effect_146  -7.120514209   5.02437608  -1.40964363  -1.83457724  10.29164132
#> effect_147   1.984545497  -0.51098851  -2.54976145  -1.93104257   8.61323740
#> effect_148   5.083350151  -2.27016573   5.96579105  13.56582694 -10.51263828
#> effect_149   3.860535697  10.84298270  -7.49507544  -6.67255478  -2.06263285
#> effect_150  -0.280787202  -3.23517287  -2.69843476  -3.20709193  -2.71287756
#> effect_151   3.611399242  -5.91246675   2.06496749  -3.15378905  -6.52570409
#> effect_152 -12.076086693   0.59498628  -3.73400062  -1.05772281  -0.12160966
#> effect_153  -3.341084432  10.48389363   9.04245053   4.37813365 -10.35663263
#> effect_154   1.615033304 -12.96184731  -3.26825487   8.68109504  -0.99760046
#> effect_155  -5.544986323   0.76212394   1.79316778  -2.20956433   1.09580982
#> effect_156   2.328028096   1.05805133   0.71142950 -10.76505277   4.34117750
#> effect_157   2.952721867   1.83845197  -9.01768514   5.52328366  -8.50230794
#> effect_158  -7.703011006   8.81154478  -2.57358177  -2.80290737  -5.28608028
#> effect_159  -1.787985244  -2.15503882   2.10973278   4.08499764   3.99749679
#> effect_160   5.410663237   3.29219658  -2.78392238  -2.14667268  -4.50913381
#> effect_161   4.171252767  -8.06441718   7.23531527   0.27307959   2.64174909
#> effect_162   7.125508505  -2.19554617   2.28775448   0.28953317   3.76537815
#> effect_163   9.686911746  10.63095315  -7.54320510   3.87946243  -4.06969969
#> effect_164  -0.693616651  -4.03632655   6.02503546  -0.18916198   0.19346225
#> effect_165  -3.142352103  -4.60125936  -8.79378177  -7.76856400  -6.95825257
#> effect_166   1.865072119   0.84932982  -7.99392521   1.13720835   1.83391103
#> effect_167  -7.902371839   7.14855301  -7.79886942  -4.26474050   6.77745743
#> effect_168  -1.211121039  -3.87268605   2.37715832  -7.16431748  -7.79569995
#> effect_169   0.289676179  -5.17996159   5.51664486   1.39500436   0.77566619
#> effect_170   0.931831728  -6.80085893  -1.89008736  -1.63117030  -1.76718730
#> effect_171  -6.362531617   4.84735306   0.12963384  -7.45429101  -8.81944243
#> effect_172  -5.483140721  -3.66135146   2.78639034  -1.73556417  -6.06972903
#> effect_173 -10.063713921  -5.37413070  -6.12605253   7.25168128  -0.47375301
#> effect_174   7.791514096  -3.25113469   5.43222302   1.60048633   4.03959187
#> effect_175  -4.725963234 -14.40708223  -3.80011083  -2.87613883  -2.08008894
#> effect_176  -9.985182508  -5.78643184   5.25344012   3.99345174  -2.82644561
#> effect_177   0.569387408  -7.14006682   4.37061334   5.88020269   0.46751337
#> effect_178  -1.204564448  -1.45225395  -9.75851061  -4.22468594  -1.48743851
#> effect_179 -10.223282492  -6.98650187  -3.09105895   2.43412782  -4.44689121
#> effect_180  10.439802375   1.43632695  -4.37454215  -5.19237085  -2.16625796
#> effect_181  -1.708543646  -2.16565675  -9.36896820  -7.91193258  -3.34410626
#> effect_182   1.976820968  -0.57911577   3.37470104   1.51708531  -0.88709091
#> effect_183  -0.056855570   4.70206112  -2.20832869   8.28933894  -3.67462629
#> effect_184   5.619931366  -1.77030078  -0.98566735  10.69185768   5.21494260
#> effect_185   2.540163218  -1.06630939  -1.98528009   1.62737402  -1.20588411
#> effect_186  10.690170366   0.16743255   3.60679282  -9.17957286   4.25371995
#> effect_187  -2.134406473   9.90454713  -2.01920726  -4.78368665   0.74040031
#> effect_188  -4.592969306  -3.65553318   4.74282677   6.98952296   1.93299599
#> effect_189  -4.257834019  -3.06122117   7.81776132  -1.75481857  -2.69613341
#> effect_190  -4.355682156   2.21025291  10.69328862  -2.48036919   0.62705692
#> effect_191   0.912219135  -2.14054846  -3.02659419   5.15852109   7.07676650
#> effect_192   4.158123479   0.76804747   2.43280084   0.84579920   2.58160148
#> effect_193  -1.801343999  -2.15413338   3.93716495   3.91969917   2.31367359
#> effect_194  -3.356539975  -0.91457837  -3.99002269  -2.48611468  -5.97887736
#> effect_195  -0.008789917   0.05456099   0.42990442  -5.14119343   0.18046270
#> effect_196   4.091645857  -5.77946504   7.74931039   3.47726257   7.64161869
#> effect_197   0.503120524   6.41041439   1.58398712  10.92842706   3.09812570
#> effect_198   2.941697404  -0.84424687  -3.90828985   7.40352206  -0.21569836
#> effect_199   1.315469800   4.99959138  -2.91203746   3.78517521  -4.94540728
#> effect_200  -3.198618784   6.66793475   3.68173852   3.90225206  -5.18286451
#> 
#> $Shat
#>            condition_1 condition_2 condition_3 condition_4 condition_5
#> effect_1             5           5           5           5           5
#> effect_2             5           5           5           5           5
#> effect_3             5           5           5           5           5
#> effect_4             5           5           5           5           5
#> effect_5             5           5           5           5           5
#> effect_6             5           5           5           5           5
#> effect_7             5           5           5           5           5
#> effect_8             5           5           5           5           5
#> effect_9             5           5           5           5           5
#> effect_10            5           5           5           5           5
#> effect_11            5           5           5           5           5
#> effect_12            5           5           5           5           5
#> effect_13            5           5           5           5           5
#> effect_14            5           5           5           5           5
#> effect_15            5           5           5           5           5
#> effect_16            5           5           5           5           5
#> effect_17            5           5           5           5           5
#> effect_18            5           5           5           5           5
#> effect_19            5           5           5           5           5
#> effect_20            5           5           5           5           5
#> effect_21            5           5           5           5           5
#> effect_22            5           5           5           5           5
#> effect_23            5           5           5           5           5
#> effect_24            5           5           5           5           5
#> effect_25            5           5           5           5           5
#> effect_26            5           5           5           5           5
#> effect_27            5           5           5           5           5
#> effect_28            5           5           5           5           5
#> effect_29            5           5           5           5           5
#> effect_30            5           5           5           5           5
#> effect_31            5           5           5           5           5
#> effect_32            5           5           5           5           5
#> effect_33            5           5           5           5           5
#> effect_34            5           5           5           5           5
#> effect_35            5           5           5           5           5
#> effect_36            5           5           5           5           5
#> effect_37            5           5           5           5           5
#> effect_38            5           5           5           5           5
#> effect_39            5           5           5           5           5
#> effect_40            5           5           5           5           5
#> effect_41            5           5           5           5           5
#> effect_42            5           5           5           5           5
#> effect_43            5           5           5           5           5
#> effect_44            5           5           5           5           5
#> effect_45            5           5           5           5           5
#> effect_46            5           5           5           5           5
#> effect_47            5           5           5           5           5
#> effect_48            5           5           5           5           5
#> effect_49            5           5           5           5           5
#> effect_50            5           5           5           5           5
#> effect_51            5           5           5           5           5
#> effect_52            5           5           5           5           5
#> effect_53            5           5           5           5           5
#> effect_54            5           5           5           5           5
#> effect_55            5           5           5           5           5
#> effect_56            5           5           5           5           5
#> effect_57            5           5           5           5           5
#> effect_58            5           5           5           5           5
#> effect_59            5           5           5           5           5
#> effect_60            5           5           5           5           5
#> effect_61            5           5           5           5           5
#> effect_62            5           5           5           5           5
#> effect_63            5           5           5           5           5
#> effect_64            5           5           5           5           5
#> effect_65            5           5           5           5           5
#> effect_66            5           5           5           5           5
#> effect_67            5           5           5           5           5
#> effect_68            5           5           5           5           5
#> effect_69            5           5           5           5           5
#> effect_70            5           5           5           5           5
#> effect_71            5           5           5           5           5
#> effect_72            5           5           5           5           5
#> effect_73            5           5           5           5           5
#> effect_74            5           5           5           5           5
#> effect_75            5           5           5           5           5
#> effect_76            5           5           5           5           5
#> effect_77            5           5           5           5           5
#> effect_78            5           5           5           5           5
#> effect_79            5           5           5           5           5
#> effect_80            5           5           5           5           5
#> effect_81            5           5           5           5           5
#> effect_82            5           5           5           5           5
#> effect_83            5           5           5           5           5
#> effect_84            5           5           5           5           5
#> effect_85            5           5           5           5           5
#> effect_86            5           5           5           5           5
#> effect_87            5           5           5           5           5
#> effect_88            5           5           5           5           5
#> effect_89            5           5           5           5           5
#> effect_90            5           5           5           5           5
#> effect_91            5           5           5           5           5
#> effect_92            5           5           5           5           5
#> effect_93            5           5           5           5           5
#> effect_94            5           5           5           5           5
#> effect_95            5           5           5           5           5
#> effect_96            5           5           5           5           5
#> effect_97            5           5           5           5           5
#> effect_98            5           5           5           5           5
#> effect_99            5           5           5           5           5
#> effect_100           5           5           5           5           5
#> effect_101           5           5           5           5           5
#> effect_102           5           5           5           5           5
#> effect_103           5           5           5           5           5
#> effect_104           5           5           5           5           5
#> effect_105           5           5           5           5           5
#> effect_106           5           5           5           5           5
#> effect_107           5           5           5           5           5
#> effect_108           5           5           5           5           5
#> effect_109           5           5           5           5           5
#> effect_110           5           5           5           5           5
#> effect_111           5           5           5           5           5
#> effect_112           5           5           5           5           5
#> effect_113           5           5           5           5           5
#> effect_114           5           5           5           5           5
#> effect_115           5           5           5           5           5
#> effect_116           5           5           5           5           5
#> effect_117           5           5           5           5           5
#> effect_118           5           5           5           5           5
#> effect_119           5           5           5           5           5
#> effect_120           5           5           5           5           5
#> effect_121           5           5           5           5           5
#> effect_122           5           5           5           5           5
#> effect_123           5           5           5           5           5
#> effect_124           5           5           5           5           5
#> effect_125           5           5           5           5           5
#> effect_126           5           5           5           5           5
#> effect_127           5           5           5           5           5
#> effect_128           5           5           5           5           5
#> effect_129           5           5           5           5           5
#> effect_130           5           5           5           5           5
#> effect_131           5           5           5           5           5
#> effect_132           5           5           5           5           5
#> effect_133           5           5           5           5           5
#> effect_134           5           5           5           5           5
#> effect_135           5           5           5           5           5
#> effect_136           5           5           5           5           5
#> effect_137           5           5           5           5           5
#> effect_138           5           5           5           5           5
#> effect_139           5           5           5           5           5
#> effect_140           5           5           5           5           5
#> effect_141           5           5           5           5           5
#> effect_142           5           5           5           5           5
#> effect_143           5           5           5           5           5
#> effect_144           5           5           5           5           5
#> effect_145           5           5           5           5           5
#> effect_146           5           5           5           5           5
#> effect_147           5           5           5           5           5
#> effect_148           5           5           5           5           5
#> effect_149           5           5           5           5           5
#> effect_150           5           5           5           5           5
#> effect_151           5           5           5           5           5
#> effect_152           5           5           5           5           5
#> effect_153           5           5           5           5           5
#> effect_154           5           5           5           5           5
#> effect_155           5           5           5           5           5
#> effect_156           5           5           5           5           5
#> effect_157           5           5           5           5           5
#> effect_158           5           5           5           5           5
#> effect_159           5           5           5           5           5
#> effect_160           5           5           5           5           5
#> effect_161           5           5           5           5           5
#> effect_162           5           5           5           5           5
#> effect_163           5           5           5           5           5
#> effect_164           5           5           5           5           5
#> effect_165           5           5           5           5           5
#> effect_166           5           5           5           5           5
#> effect_167           5           5           5           5           5
#> effect_168           5           5           5           5           5
#> effect_169           5           5           5           5           5
#> effect_170           5           5           5           5           5
#> effect_171           5           5           5           5           5
#> effect_172           5           5           5           5           5
#> effect_173           5           5           5           5           5
#> effect_174           5           5           5           5           5
#> effect_175           5           5           5           5           5
#> effect_176           5           5           5           5           5
#> effect_177           5           5           5           5           5
#> effect_178           5           5           5           5           5
#> effect_179           5           5           5           5           5
#> effect_180           5           5           5           5           5
#> effect_181           5           5           5           5           5
#> effect_182           5           5           5           5           5
#> effect_183           5           5           5           5           5
#> effect_184           5           5           5           5           5
#> effect_185           5           5           5           5           5
#> effect_186           5           5           5           5           5
#> effect_187           5           5           5           5           5
#> effect_188           5           5           5           5           5
#> effect_189           5           5           5           5           5
#> effect_190           5           5           5           5           5
#> effect_191           5           5           5           5           5
#> effect_192           5           5           5           5           5
#> effect_193           5           5           5           5           5
#> effect_194           5           5           5           5           5
#> effect_195           5           5           5           5           5
#> effect_196           5           5           5           5           5
#> effect_197           5           5           5           5           5
#> effect_198           5           5           5           5           5
#> effect_199           5           5           5           5           5
#> effect_200           5           5           5           5           5
#>