Create some more simple simulated data for testing purposes

Create some more simple simulated data for testing purposes

simple_sims2(nsamp = 100, err_sd = 0.01)

Arguments

nsamp
number of samples of each type

err_sd
the standard deviation of the errors

Details

The simulation consists of five conditions with two types of effecc those present (and identical) in first two conditions and those present (and identical) in last three conditions

Examples

simple_sims2(100, 5)

#> $B
#>             condition_1  condition_2  condition_3  condition_4  condition_5
#> effect_1    1.117451459  1.117451459  0.000000000  0.000000000  0.000000000
#> effect_2   -0.976468386 -0.976468386  0.000000000  0.000000000  0.000000000
#> effect_3    0.328119925  0.328119925  0.000000000  0.000000000  0.000000000
#> effect_4    0.266808456  0.266808456  0.000000000  0.000000000  0.000000000
#> effect_5    0.227352551  0.227352551  0.000000000  0.000000000  0.000000000
#> effect_6    1.133304529  1.133304529  0.000000000  0.000000000  0.000000000
#> effect_7    0.882689175  0.882689175  0.000000000  0.000000000  0.000000000
#> effect_8   -0.782776347 -0.782776347  0.000000000  0.000000000  0.000000000
#> effect_9   -1.352008379 -1.352008379  0.000000000  0.000000000  0.000000000
#> effect_10   0.407915481  0.407915481  0.000000000  0.000000000  0.000000000
#> effect_11  -0.109838787 -0.109838787  0.000000000  0.000000000  0.000000000
#> effect_12  -1.932123040 -1.932123040  0.000000000  0.000000000  0.000000000
#> effect_13   0.110085739  0.110085739  0.000000000  0.000000000  0.000000000
#> effect_14  -0.577936753 -0.577936753  0.000000000  0.000000000  0.000000000
#> effect_15   0.566764827  0.566764827  0.000000000  0.000000000  0.000000000
#> effect_16   0.825405849  0.825405849  0.000000000  0.000000000  0.000000000
#> effect_17  -1.062471470 -1.062471470  0.000000000  0.000000000  0.000000000
#> effect_18   0.912722922  0.912722922  0.000000000  0.000000000  0.000000000
#> effect_19  -0.021591197 -0.021591197  0.000000000  0.000000000  0.000000000
#> effect_20  -0.813040938 -0.813040938  0.000000000  0.000000000  0.000000000
#> effect_21  -0.100729168 -0.100729168  0.000000000  0.000000000  0.000000000
#> effect_22  -1.328003296 -1.328003296  0.000000000  0.000000000  0.000000000
#> effect_23  -1.556284521 -1.556284521  0.000000000  0.000000000  0.000000000
#> effect_24   0.331953398  0.331953398  0.000000000  0.000000000  0.000000000
#> effect_25  -1.466919004 -1.466919004  0.000000000  0.000000000  0.000000000
#> effect_26  -0.589500049 -0.589500049  0.000000000  0.000000000  0.000000000
#> effect_27  -0.204935920 -0.204935920  0.000000000  0.000000000  0.000000000
#> effect_28   0.871457972  0.871457972  0.000000000  0.000000000  0.000000000
#> effect_29   1.924887590  1.924887590  0.000000000  0.000000000  0.000000000
#> effect_30   0.002061406  0.002061406  0.000000000  0.000000000  0.000000000
#> effect_31  -1.603506645 -1.603506645  0.000000000  0.000000000  0.000000000
#> effect_32  -0.049076817 -0.049076817  0.000000000  0.000000000  0.000000000
#> effect_33   1.221564271  1.221564271  0.000000000  0.000000000  0.000000000
#> effect_34  -0.059825758 -0.059825758  0.000000000  0.000000000  0.000000000
#> effect_35   1.018886530  1.018886530  0.000000000  0.000000000  0.000000000
#> effect_36   1.286300100  1.286300100  0.000000000  0.000000000  0.000000000
#> effect_37   1.233854187  1.233854187  0.000000000  0.000000000  0.000000000
#> effect_38   0.020266523  0.020266523  0.000000000  0.000000000  0.000000000
#> effect_39  -0.375167949 -0.375167949  0.000000000  0.000000000  0.000000000
#> effect_40  -0.311472653 -0.311472653  0.000000000  0.000000000  0.000000000
#> effect_41  -0.770369155 -0.770369155  0.000000000  0.000000000  0.000000000
#> effect_42   0.132760830  0.132760830  0.000000000  0.000000000  0.000000000
#> effect_43  -2.200093010 -2.200093010  0.000000000  0.000000000  0.000000000
#> effect_44   0.266837188  0.266837188  0.000000000  0.000000000  0.000000000
#> effect_45   1.215466776  1.215466776  0.000000000  0.000000000  0.000000000
#> effect_46   0.116762523  0.116762523  0.000000000  0.000000000  0.000000000
#> effect_47   0.943152530  0.943152530  0.000000000  0.000000000  0.000000000
#> effect_48   2.006915896  2.006915896  0.000000000  0.000000000  0.000000000
#> effect_49  -1.272926699 -1.272926699  0.000000000  0.000000000  0.000000000
#> effect_50  -2.686772160 -2.686772160  0.000000000  0.000000000  0.000000000
#> effect_51   0.535409453  0.535409453  0.000000000  0.000000000  0.000000000
#> effect_52   0.294479970  0.294479970  0.000000000  0.000000000  0.000000000
#> effect_53  -2.300350019 -2.300350019  0.000000000  0.000000000  0.000000000
#> effect_54   0.723077245  0.723077245  0.000000000  0.000000000  0.000000000
#> effect_55  -0.098847780 -0.098847780  0.000000000  0.000000000  0.000000000
#> effect_56  -0.775283905 -0.775283905  0.000000000  0.000000000  0.000000000
#> effect_57  -1.257015634 -1.257015634  0.000000000  0.000000000  0.000000000
#> effect_58  -0.631607139 -0.631607139  0.000000000  0.000000000  0.000000000
#> effect_59  -0.539593627 -0.539593627  0.000000000  0.000000000  0.000000000
#> effect_60   0.602501858  0.602501858  0.000000000  0.000000000  0.000000000
#> effect_61  -0.494594419 -0.494594419  0.000000000  0.000000000  0.000000000
#> effect_62  -0.153121625 -0.153121625  0.000000000  0.000000000  0.000000000
#> effect_63   0.844099645  0.844099645  0.000000000  0.000000000  0.000000000
#> effect_64   1.786320852  1.786320852  0.000000000  0.000000000  0.000000000
#> effect_65  -0.512491075 -0.512491075  0.000000000  0.000000000  0.000000000
#> effect_66  -1.288856030 -1.288856030  0.000000000  0.000000000  0.000000000
#> effect_67   0.295719582  0.295719582  0.000000000  0.000000000  0.000000000
#> effect_68   1.329549466  1.329549466  0.000000000  0.000000000  0.000000000
#> effect_69  -0.337260490 -0.337260490  0.000000000  0.000000000  0.000000000
#> effect_70  -1.597325527 -1.597325527  0.000000000  0.000000000  0.000000000
#> effect_71   0.048718787  0.048718787  0.000000000  0.000000000  0.000000000
#> effect_72  -0.172682541 -0.172682541  0.000000000  0.000000000  0.000000000
#> effect_73   0.217724831  0.217724831  0.000000000  0.000000000  0.000000000
#> effect_74  -1.826188470 -1.826188470  0.000000000  0.000000000  0.000000000
#> effect_75   0.035175853  0.035175853  0.000000000  0.000000000  0.000000000
#> effect_76   0.560013990  0.560013990  0.000000000  0.000000000  0.000000000
#> effect_77  -0.618530981 -0.618530981  0.000000000  0.000000000  0.000000000
#> effect_78   0.015834754  0.015834754  0.000000000  0.000000000  0.000000000
#> effect_79   0.739191738  0.739191738  0.000000000  0.000000000  0.000000000
#> effect_80  -0.245336495 -0.245336495  0.000000000  0.000000000  0.000000000
#> effect_81  -0.629577668 -0.629577668  0.000000000  0.000000000  0.000000000
#> effect_82  -1.498951844 -1.498951844  0.000000000  0.000000000  0.000000000
#> effect_83  -0.036492817 -0.036492817  0.000000000  0.000000000  0.000000000
#> effect_84  -0.388590107 -0.388590107  0.000000000  0.000000000  0.000000000
#> effect_85  -1.011338371 -1.011338371  0.000000000  0.000000000  0.000000000
#> effect_86   1.982830186  1.982830186  0.000000000  0.000000000  0.000000000
#> effect_87  -2.141220083 -2.141220083  0.000000000  0.000000000  0.000000000
#> effect_88  -2.288445816 -2.288445816  0.000000000  0.000000000  0.000000000
#> effect_89  -0.042523492 -0.042523492  0.000000000  0.000000000  0.000000000
#> effect_90   0.929944122  0.929944122  0.000000000  0.000000000  0.000000000
#> effect_91   1.723335123  1.723335123  0.000000000  0.000000000  0.000000000
#> effect_92   0.139006467  0.139006467  0.000000000  0.000000000  0.000000000
#> effect_93  -1.451667731 -1.451667731  0.000000000  0.000000000  0.000000000
#> effect_94  -1.191199379 -1.191199379  0.000000000  0.000000000  0.000000000
#> effect_95  -0.100202157 -0.100202157  0.000000000  0.000000000  0.000000000
#> effect_96   0.424992470  0.424992470  0.000000000  0.000000000  0.000000000
#> effect_97   1.786011846  1.786011846  0.000000000  0.000000000  0.000000000
#> effect_98  -0.687733759 -0.687733759  0.000000000  0.000000000  0.000000000
#> effect_99   1.428659954  1.428659954  0.000000000  0.000000000  0.000000000
#> effect_100 -0.420352122 -0.420352122  0.000000000  0.000000000  0.000000000
#> effect_101  0.000000000  0.000000000 -0.152333768 -0.152333768 -0.152333768
#> effect_102  0.000000000  0.000000000 -1.234783997 -1.234783997 -1.234783997
#> effect_103  0.000000000  0.000000000  0.109738034  0.109738034  0.109738034
#> effect_104  0.000000000  0.000000000  0.879914041  0.879914041  0.879914041
#> effect_105  0.000000000  0.000000000  0.051076507  0.051076507  0.051076507
#> effect_106  0.000000000  0.000000000  1.482409691  1.482409691  1.482409691
#> effect_107  0.000000000  0.000000000 -1.775594492 -1.775594492 -1.775594492
#> effect_108  0.000000000  0.000000000 -1.552743228 -1.552743228 -1.552743228
#> effect_109  0.000000000  0.000000000 -0.567396547 -0.567396547 -0.567396547
#> effect_110  0.000000000  0.000000000 -0.892059485 -0.892059485 -0.892059485
#> effect_111  0.000000000  0.000000000 -1.794335705 -1.794335705 -1.794335705
#> effect_112  0.000000000  0.000000000 -1.011126800 -1.011126800 -1.011126800
#> effect_113  0.000000000  0.000000000 -2.009831577 -2.009831577 -2.009831577
#> effect_114  0.000000000  0.000000000  0.460694807  0.460694807  0.460694807
#> effect_115  0.000000000  0.000000000 -0.157209532 -0.157209532 -0.157209532
#> effect_116  0.000000000  0.000000000  0.528200903  0.528200903  0.528200903
#> effect_117  0.000000000  0.000000000 -0.423310984 -0.423310984 -0.423310984
#> effect_118  0.000000000  0.000000000  0.674391930  0.674391930  0.674391930
#> effect_119  0.000000000  0.000000000 -0.909983473 -0.909983473 -0.909983473
#> effect_120  0.000000000  0.000000000  0.247233244  0.247233244  0.247233244
#> effect_121  0.000000000  0.000000000 -0.293822767 -0.293822767 -0.293822767
#> effect_122  0.000000000  0.000000000 -0.986460811 -0.986460811 -0.986460811
#> effect_123  0.000000000  0.000000000 -1.041738882 -1.041738882 -1.041738882
#> effect_124  0.000000000  0.000000000 -0.725155522 -0.725155522 -0.725155522
#> effect_125  0.000000000  0.000000000  0.342781108  0.342781108  0.342781108
#> effect_126  0.000000000  0.000000000 -0.896099679 -0.896099679 -0.896099679
#> effect_127  0.000000000  0.000000000  0.960845732  0.960845732  0.960845732
#> effect_128  0.000000000  0.000000000  1.018852801  1.018852801  1.018852801
#> effect_129  0.000000000  0.000000000 -0.562235195 -0.562235195 -0.562235195
#> effect_130  0.000000000  0.000000000  0.020425027  0.020425027  0.020425027
#> effect_131  0.000000000  0.000000000  0.028343832  0.028343832  0.028343832
#> effect_132  0.000000000  0.000000000  0.493509049  0.493509049  0.493509049
#> effect_133  0.000000000  0.000000000  0.003629605  0.003629605  0.003629605
#> effect_134  0.000000000  0.000000000 -0.396652361 -0.396652361 -0.396652361
#> effect_135  0.000000000  0.000000000 -1.029287301 -1.029287301 -1.029287301
#> effect_136  0.000000000  0.000000000  0.540962186  0.540962186  0.540962186
#> effect_137  0.000000000  0.000000000  1.345128312  1.345128312  1.345128312
#> effect_138  0.000000000  0.000000000 -0.083836684 -0.083836684 -0.083836684
#> effect_139  0.000000000  0.000000000  1.200531527  1.200531527  1.200531527
#> effect_140  0.000000000  0.000000000  0.215733146  0.215733146  0.215733146
#> effect_141  0.000000000  0.000000000 -2.492890925 -2.492890925 -2.492890925
#> effect_142  0.000000000  0.000000000  1.017833662  1.017833662  1.017833662
#> effect_143  0.000000000  0.000000000 -1.080650804 -1.080650804 -1.080650804
#> effect_144  0.000000000  0.000000000 -1.203863813 -1.203863813 -1.203863813
#> effect_145  0.000000000  0.000000000 -0.212905932 -0.212905932 -0.212905932
#> effect_146  0.000000000  0.000000000 -0.064919543 -0.064919543 -0.064919543
#> effect_147  0.000000000  0.000000000  0.632746620  0.632746620  0.632746620
#> effect_148  0.000000000  0.000000000 -0.547743648 -0.547743648 -0.547743648
#> effect_149  0.000000000  0.000000000  0.558483411  0.558483411  0.558483411
#> effect_150  0.000000000  0.000000000  1.134124453  1.134124453  1.134124453
#> effect_151  0.000000000  0.000000000 -0.770071146 -0.770071146 -0.770071146
#> effect_152  0.000000000  0.000000000 -2.323800181 -2.323800181 -2.323800181
#> effect_153  0.000000000  0.000000000 -0.986952872 -0.986952872 -0.986952872
#> effect_154  0.000000000  0.000000000  0.248735004  0.248735004  0.248735004
#> effect_155  0.000000000  0.000000000 -0.569992200 -0.569992200 -0.569992200
#> effect_156  0.000000000  0.000000000 -0.154032131 -0.154032131 -0.154032131
#> effect_157  0.000000000  0.000000000  0.116756510  0.116756510  0.116756510
#> effect_158  0.000000000  0.000000000 -0.911446053 -0.911446053 -0.911446053
#> effect_159  0.000000000  0.000000000 -0.314834615 -0.314834615 -0.314834615
#> effect_160  0.000000000  0.000000000  0.276747647  0.276747647  0.276747647
#> effect_161  0.000000000  0.000000000  0.591081541  0.591081541  0.591081541
#> effect_162  0.000000000  0.000000000  0.676260401  0.676260401  0.676260401
#> effect_163  0.000000000  0.000000000  0.210890454  0.210890454  0.210890454
#> effect_164  0.000000000  0.000000000 -0.559642385 -0.559642385 -0.559642385
#> effect_165  0.000000000  0.000000000 -2.410674870 -2.410674870 -2.410674870
#> effect_166  0.000000000  0.000000000  1.033491818  1.033491818  1.033491818
#> effect_167  0.000000000  0.000000000  2.912885544  2.912885544  2.912885544
#> effect_168  0.000000000  0.000000000 -0.659474061 -0.659474061 -0.659474061
#> effect_169  0.000000000  0.000000000  1.073098497  1.073098497  1.073098497
#> effect_170  0.000000000  0.000000000 -0.813710433 -0.813710433 -0.813710433
#> effect_171  0.000000000  0.000000000 -2.187537438 -2.187537438 -2.187537438
#> effect_172  0.000000000  0.000000000 -0.971133087 -0.971133087 -0.971133087
#> effect_173  0.000000000  0.000000000 -0.052698658 -0.052698658 -0.052698658
#> effect_174  0.000000000  0.000000000 -0.693586230 -0.693586230 -0.693586230
#> effect_175  0.000000000  0.000000000 -1.814498105 -1.814498105 -1.814498105
#> effect_176  0.000000000  0.000000000 -1.136031839 -1.136031839 -1.136031839
#> effect_177  0.000000000  0.000000000  0.981704108  0.981704108  0.981704108
#> effect_178  0.000000000  0.000000000  0.744157139  0.744157139  0.744157139
#> effect_179  0.000000000  0.000000000  1.170712666  1.170712666  1.170712666
#> effect_180  0.000000000  0.000000000  0.950717818  0.950717818  0.950717818
#> effect_181  0.000000000  0.000000000 -1.496406931 -1.496406931 -1.496406931
#> effect_182  0.000000000  0.000000000  0.554974845  0.554974845  0.554974845
#> effect_183  0.000000000  0.000000000  1.588878557  1.588878557  1.588878557
#> effect_184  0.000000000  0.000000000 -0.811322799 -0.811322799 -0.811322799
#> effect_185  0.000000000  0.000000000  1.167449721  1.167449721  1.167449721
#> effect_186  0.000000000  0.000000000  0.328917037  0.328917037  0.328917037
#> effect_187  0.000000000  0.000000000 -0.755502050 -0.755502050 -0.755502050
#> effect_188  0.000000000  0.000000000  0.031106637  0.031106637  0.031106637
#> effect_189  0.000000000  0.000000000 -0.304815524 -0.304815524 -0.304815524
#> effect_190  0.000000000  0.000000000  0.024899643  0.024899643  0.024899643
#> effect_191  0.000000000  0.000000000 -0.577170222 -0.577170222 -0.577170222
#> effect_192  0.000000000  0.000000000  1.552500195  1.552500195  1.552500195
#> effect_193  0.000000000  0.000000000  1.223541800  1.223541800  1.223541800
#> effect_194  0.000000000  0.000000000  0.365369388  0.365369388  0.365369388
#> effect_195  0.000000000  0.000000000  1.825754285  1.825754285  1.825754285
#> effect_196  0.000000000  0.000000000  1.758058408  1.758058408  1.758058408
#> effect_197  0.000000000  0.000000000 -0.828645786 -0.828645786 -0.828645786
#> effect_198  0.000000000  0.000000000  1.178595905  1.178595905  1.178595905
#> effect_199  0.000000000  0.000000000  0.858398950  0.858398950  0.858398950
#> effect_200  0.000000000  0.000000000 -0.176401916 -0.176401916 -0.176401916
#> 
#> $Bhat
#>             condition_1  condition_2  condition_3  condition_4  condition_5
#> effect_1     1.15946370   3.09716612  -3.21274761  -3.18147131  -2.78526929
#> effect_2    -2.68218360   2.01390194  -6.20809679   2.20684850  -5.80927236
#> effect_3     8.04927987   8.30034926   0.44558151  -1.95310353   7.76387007
#> effect_4    -6.52591122  -3.06478796  -4.10399824  -6.29805480   5.40390473
#> effect_5    -0.87015659  -0.80445714  -5.14956526   3.65890635   1.57170768
#> effect_6     7.89877334  -5.53840150   4.95374569  10.45551994   1.68142383
#> effect_7     3.18828155   1.66769801   6.47998361   2.77781271  -3.76184622
#> effect_8     8.58076182  -2.06472298   8.61082145  -9.17320614  -4.19044905
#> effect_9    -2.30428994  -0.78435293 -12.98112044   3.79982381  -4.58972593
#> effect_10    2.94202666  13.37449558  -0.73878391  -0.69455383   5.00597365
#> effect_11    2.24042307   4.29289797   6.81600026  -9.84310826   8.21911236
#> effect_12   -5.81634325  -6.21449045  -1.75101360  -7.80480391   2.13369714
#> effect_13   -2.92999896  -1.38412556  -8.02847032  -3.94779899   1.92921436
#> effect_14   -5.97553177   0.93954052   5.45560024  -3.61659224   3.50026766
#> effect_15   -4.59544448   0.44487784   1.13733866  -2.01673710  -1.54594079
#> effect_16   11.76724581   5.30157939   4.64479907  -6.04832207  -5.10218006
#> effect_17    5.99630653  -4.79913614   3.43073731   2.59232037  -0.87521403
#> effect_18   -7.87168492  -4.43661596  -2.25550037   0.42074493   3.65992504
#> effect_19   -7.27519981   0.72681830   6.19325151  -3.00153271  -0.03062161
#> effect_20   -7.95586310   0.81749707   0.45178702   6.81700988  -7.57893964
#> effect_21   -3.14588828  -6.33485383   2.27029322   5.59926566   2.20888640
#> effect_22   -3.98783026  10.51627611   8.40307827   8.82242989   0.59066012
#> effect_23   -1.45716651  -4.63365086  -7.55649012   7.64486046  -0.36675459
#> effect_24    0.85957509   6.40589064  -2.54551678  -7.14854529   5.68167044
#> effect_25   -2.94859076  -0.06796441   3.38205888  11.82890699  -1.70663193
#> effect_26    0.86047733  -1.60860727  -2.73672590  -2.02510867  -6.68808261
#> effect_27   -8.72909653   4.90393151  -8.78848926  -2.64888583   0.22753343
#> effect_28    0.77278305  15.69410216  -6.21788230  -0.81049538   2.63046671
#> effect_29    4.19181855   2.65716770   4.80960921  -5.50997155  -1.96402504
#> effect_30    1.13307327   3.05109944  -0.70124857  -1.44615903   5.70261211
#> effect_31    2.08587224  -6.61545161   0.18092947   3.73609665  -2.07408345
#> effect_32    6.93013864   2.50218606   1.16758087   6.19949712  -6.81891801
#> effect_33    5.96575353  -0.74251262   8.37692861  -4.27621050   2.03880186
#> effect_34    3.16261774   9.05557664  -2.45758781   6.84904759 -11.62969852
#> effect_35  -10.22663982  -8.97233816  -0.12821728   1.34812239  -1.23409804
#> effect_36    0.66044616 -11.36270325   5.80054869   0.42925790 -10.02360006
#> effect_37   -0.17637666   4.34149168   2.40377732  -0.34583290  -2.66508183
#> effect_38    0.76915227  -1.95815106  -4.29019839   0.83867102  -3.10346085
#> effect_39   -5.76257606   0.95400638  -7.94653684   4.46923807   5.12283625
#> effect_40   -1.20023000   5.96041541   3.53258046  -6.92973709  -6.54247399
#> effect_41    0.44513090   1.76796733  -0.67017100   1.75373137  13.13627939
#> effect_42   -1.52284066  -3.16291453   3.28025778   8.78343676  -9.83026857
#> effect_43   -4.21013432  -9.78273280  -5.29186217   9.01009386  -1.07825729
#> effect_44    2.51266396   6.80938339   7.24302397  -5.55821139   2.03643673
#> effect_45    0.55267917  -9.39286770   1.21388949  -1.63666719  -7.38587779
#> effect_46    0.45767211   2.57724853   4.66177857  -1.27839180  -2.31792454
#> effect_47   -2.03600884   3.02688740  -2.96635657   3.49990243   1.41231786
#> effect_48    1.92635316  -5.04936778  -8.26156264   4.51634925   7.49040693
#> effect_49    3.68741678  -6.03894384   1.29028690   4.83578161   8.42575119
#> effect_50    4.17901553  -7.19931906  -8.50455132   4.89712822   1.97277443
#> effect_51   -7.18827628  -3.06599722  -2.35980166   6.07258366   4.26667821
#> effect_52    6.15420136   4.09258617   9.69444187  -8.17871964   0.38261403
#> effect_53   -7.18570160   0.92879130  -0.64991807   4.61878863  -2.10910553
#> effect_54   -7.64595323   1.48526485   0.16005905  -0.44476747   8.80330836
#> effect_55    0.11266310  -2.12803097   6.99522598  -2.07170847   0.93871592
#> effect_56    7.36380715   5.35132593  -0.47585901   8.44675361   0.26863513
#> effect_57   -4.60006675   6.50401192   3.69020894  -5.28873781   2.30353820
#> effect_58   -0.28281224  -1.43625051  -0.72875608  -7.09276518  -4.57377456
#> effect_59   -1.13908980   3.27188364   3.21424412   5.19554565   8.13111506
#> effect_60   -6.11934024   6.28420101   0.18041939   4.10802068   0.90620281
#> effect_61   -8.11542189  -6.47448561   5.20207070  -0.84804372 -13.19671410
#> effect_62   -3.02570073   5.25705943  -7.07876338   2.13041519   0.71117900
#> effect_63    7.53905212   4.95490024   3.09973102  -1.09502495   4.60191044
#> effect_64    2.80578335   3.95380970  -2.45047151  -1.64477526  -1.73304899
#> effect_65    6.16596082  -0.78567011  -4.60590621  -0.22816586   6.64501026
#> effect_66    2.50508234   2.75788539   1.03838110  -1.67383005  -2.11317195
#> effect_67    8.02561652   0.68334332  -4.69738363   3.51497118  -0.68728879
#> effect_68    5.31784848  -5.54465961  -3.40425861   5.97466072  -2.10018279
#> effect_69    5.87879275   5.27479558   3.42074258   1.89933660  -1.23248115
#> effect_70   -7.89368006  -2.96191954  -6.43609214  -2.92616627   6.23101436
#> effect_71   -6.21072785  -6.07676523   4.97374118  -5.24374385  -2.18713960
#> effect_72   -3.10270166   9.12071177   1.00212835   0.21414316  -7.96130691
#> effect_73   -4.33177081  -3.39546567   3.42884121  -1.64961243   3.99937240
#> effect_74   -3.72975906  -8.12939653  -0.07947003  -2.86233737   6.14344107
#> effect_75   -3.04414311   0.72307658  -4.38448311   2.12551323   2.36405769
#> effect_76   -0.49540960  -4.57213363  -4.50372687  -1.32175542  -3.94848781
#> effect_77    3.49918681  -8.05346735  -0.66703829  -2.28245085   0.31162754
#> effect_78    7.22617325   5.67831517  -6.25467116   5.14041056   7.62964394
#> effect_79   -1.12879532  -7.85613825  -1.07381895   0.45918333  -5.80272325
#> effect_80    6.08439113  -1.04873601  -1.12085206   1.20299369   7.90048302
#> effect_81    2.37354034   3.80550982  11.80613475   2.90710476   5.48753360
#> effect_82   -3.03609980 -11.94734130   0.42152658  -0.45295235   0.70167962
#> effect_83   -8.53202276   3.14370503  -8.27368491  -7.99075981  -2.35575891
#> effect_84    0.92643852  -1.77852357   0.32050208  -2.00099729   3.71279943
#> effect_85    0.42103993  -2.10731701   1.59283092   8.59377798  -6.19683690
#> effect_86  -12.04789752  -3.76918858  -3.55197556  10.05771969  -4.39150827
#> effect_87   -6.32853783  -5.32876189  -1.65323832  -5.46292152   7.31834008
#> effect_88   -3.34580266  -6.23137647  -5.90719828   4.53480591  -1.28400431
#> effect_89   -1.92644168  -0.94327908   4.16364727  -0.50482563  -1.28520184
#> effect_90   -3.94520634  -3.54256441   4.58055998   2.43708400  -4.06081103
#> effect_91   -5.64784788   2.38125352  -2.31049038  -2.32258263   5.54348493
#> effect_92   -1.16893101  -6.28914578   2.77137927  -0.12130362   0.88782722
#> effect_93   -3.79844239 -10.28800889   2.33094581   3.61526168   5.74202386
#> effect_94   -6.76595482   0.81958280   0.20374888  -1.45631198   9.28730800
#> effect_95   -3.72091109   0.25449656   6.98162427  -4.83370084  -5.74327134
#> effect_96    5.70829966  -4.14102578  -1.66356997   1.09788738   0.97923179
#> effect_97   -0.70346932   7.64084880  -1.98854084  -1.01289922  -5.13394563
#> effect_98    4.12528075   3.38038332  -9.67444413  -2.89936893   1.93465011
#> effect_99   -2.13516637   4.06057253  -1.68247272  -0.21462116  -0.67968234
#> effect_100   2.23192763   4.87664228   1.30139794   2.00720864  -2.99234297
#> effect_101   4.12850131   7.16311590  -3.22546298  10.93081274   5.20059360
#> effect_102  -7.03668947  -2.88018851  -8.77555886 -11.71734474  -2.35676220
#> effect_103  -0.49856898   0.74010539  -2.24641015   2.89654216 -10.30386212
#> effect_104  -4.20752034   0.22857541   2.16636952  -4.93883453  -5.19949609
#> effect_105   6.17326804  -2.02462481  -3.31572463  -0.40216792  -1.29917478
#> effect_106  -1.17738710  -0.80842644  -3.86668370   9.49208964  -0.88146923
#> effect_107   3.79427157  -4.41767757   5.21637034  -6.65262115  -1.43441069
#> effect_108  -2.43159171  -1.44318112   1.63252319  -8.92255338  -2.54807916
#> effect_109   1.46290824  -2.91292374   4.54020245  -4.69134532   7.65287738
#> effect_110   2.35527353   4.02625255  -3.68846460   0.16854588  -9.48999321
#> effect_111  -1.01159624 -14.85285815  -0.92104818  10.46839634  -4.06830806
#> effect_112  -0.72067681   3.14003330   4.76473962   1.31963157  -3.13897527
#> effect_113   1.14175024  -6.26625352  -4.71283234 -12.84469280   1.32037334
#> effect_114  -2.26863415  -2.25250898  -8.42667972  -4.53150926  -1.85641111
#> effect_115  -4.54702485  -1.91207968  -4.01391477   2.80089453  10.34028536
#> effect_116   5.45085348  11.98169943   3.25477140  -2.65620179  -1.46162514
#> effect_117   4.21464463  -4.08354739   1.59893103  -4.11850214   0.64578264
#> effect_118   2.00390792  -3.64897643   0.82313545  -8.36389784   8.21443113
#> effect_119   3.63709164   2.91182729  -0.18169111  -5.73030033   2.47340201
#> effect_120  -1.67925841   0.22326168   7.34756099   0.82248465   0.02107929
#> effect_121  -1.98415110  -8.74718447  -6.25690922   0.44925655   2.05744786
#> effect_122  -2.35548645  -4.74578291   7.44822020  -0.32558613  -8.73420120
#> effect_123  -2.90031229 -10.61324828   1.45882291  -4.87570048 -11.69189816
#> effect_124  -2.65610509   1.72072386  -6.27787364  -0.90132995 -11.62903482
#> effect_125  -0.31060905   5.73937203  -5.93257103   7.14164797  10.85172347
#> effect_126   3.42445623  -2.06023758  -0.20432026  -3.28479361  -1.57041257
#> effect_127  -7.19909717  -4.64949704  -1.46608927  -0.93480009  -2.52133127
#> effect_128   2.42973070  -7.30747920   3.61376815  -4.31393282   4.74051874
#> effect_129   4.80300375   6.10196466   5.74877194  -5.65190992   4.37045707
#> effect_130  -3.60844720  -3.39713469   2.98251909   1.18486399  -1.30836199
#> effect_131   1.97845434  -1.62985147  -9.35120906  -5.47635095  -7.83360452
#> effect_132 -10.77203938  -7.40621775 -11.19654893  -1.67423093  -6.39778009
#> effect_133   5.38362681  -0.86172373   7.65066569   6.05616510   6.53811516
#> effect_134  -2.33839238 -10.80856770  -6.03703160   5.49205422  -4.72787096
#> effect_135  -4.08432060   2.49399895  -2.01871171   0.40906955   7.48585622
#> effect_136   6.79734107  -7.70679049   0.50544482   1.26368627   4.31871189
#> effect_137   7.84350570  -2.23350882  -0.16876660  -4.33568834   0.67135949
#> effect_138   5.11850308   5.26924085   2.92964831   5.53418474   5.82448602
#> effect_139   1.83522323   5.56338988  -4.41874525  -5.88353000  -5.55718077
#> effect_140  -4.27477209   2.75380165  -4.37357412   5.63296449  -1.29236217
#> effect_141  -4.18849692   1.79409602   1.98814979  -7.44270597   3.45266831
#> effect_142  -3.47279939   7.09859185  -5.77753854  -5.49497748  -3.25011976
#> effect_143   4.35904333   5.69281063  -3.04903286   2.60000344  -6.03547707
#> effect_144  -4.59834373  -4.61242268  -1.28833620  -4.00114238  -4.54330653
#> effect_145   1.85425818  -0.43825338   0.04391322  -3.83766986   7.50086025
#> effect_146   9.15664623   0.50405323   3.43236681   1.27838415   8.38777899
#> effect_147   2.42382043  -3.37375210  10.13132244  -1.34005490  -4.68132273
#> effect_148  -0.39689482   6.30329848   2.06474906  -5.47524533   0.29640628
#> effect_149  -1.22810278  -4.13293997  -1.50637595  -6.15825323   5.36348942
#> effect_150   2.50382740  -4.92849559   2.02607604  -1.90606600   2.88864899
#> effect_151  -4.25230803   9.06104225  -5.02730135   2.65603646   0.39409727
#> effect_152  -4.07455741   4.75228977  -4.32921724   1.88208265  -2.23906790
#> effect_153  -5.68023234  -5.42546210  -1.39886379   2.28656352  -3.26285338
#> effect_154   5.15145245  -9.21460101  -0.53504471   0.67869184   7.08860367
#> effect_155  -5.30026656  -2.64082186   1.69831074   1.01820112   0.99084908
#> effect_156  -1.72033467  -4.43866503  -1.09263354  -9.27021473  -3.57168769
#> effect_157   1.58801897   0.77487889   4.30340544  -7.60730625  -7.92801052
#> effect_158   3.20633745   6.81814996  10.55274286  -1.39816221  -7.07033034
#> effect_159   1.72968770   1.26375235   0.72119788  -0.50042430   6.30906711
#> effect_160   0.52970785  -1.50057605  -3.81971544  -4.08976713  -0.35355036
#> effect_161  -7.96174028  -0.12866772  -5.71316211   7.83882432   3.92371110
#> effect_162  -1.32508183  -1.31353513   6.81034352   0.18859107  -4.32623598
#> effect_163  -1.17751129  -6.55449371  -2.63310646   0.50505784   3.66044345
#> effect_164   3.95935210  -4.23489096  -2.13584398   0.51284664   3.48171952
#> effect_165  -2.99017163   0.25830959   0.39110981  -6.90435123   6.63221027
#> effect_166   0.03294869   5.08518297   0.21735008   1.50246475  -7.32941241
#> effect_167   3.82591971  -3.44379766   3.52007478   6.17826747  -0.04444676
#> effect_168  -2.07080129   3.36087575   1.22164461   2.49115252  -1.29797728
#> effect_169  -5.24338880   5.30076305   6.59834884   5.28144744  -6.96648162
#> effect_170  -9.63330443   4.70232283 -11.85279761   9.71687612 -10.39243732
#> effect_171  -4.22264815   6.22133912   4.83109672  -7.52763698   6.91153070
#> effect_172   0.76317060   3.64741469   0.62729766  -5.90292841   0.98435556
#> effect_173   1.62451893  -0.10514413  -0.79097068  -4.61645177   3.63018809
#> effect_174  -0.47481569  -2.54213401  -6.80071915  -1.15209873 -10.55530950
#> effect_175   3.95516658  -5.30645333  -2.18181964  -8.24590183   6.34868604
#> effect_176   3.12406041   0.85682745   1.70085881   0.00788853  -0.58168722
#> effect_177  -8.60060211   9.92724976   4.78597029   1.25090133  11.85075423
#> effect_178  -1.97337942   0.72951246   3.85217766  -6.85977435   4.37355316
#> effect_179  -7.48505792   2.96977704   0.12661096  10.25355631   9.81517373
#> effect_180   4.31835469  -5.25473289  -0.97206561  -2.80527462   1.13761861
#> effect_181 -11.60998869   3.68683004   2.66898976  -3.44558266  -4.87209196
#> effect_182  -2.18852571   4.26343063   7.22023588   9.92445655  -6.15921344
#> effect_183   2.31061220   5.64612986   7.87537331  -0.74305843  -3.03751799
#> effect_184  -4.79809510   4.99340757   0.90610231   5.71013653  -4.63167529
#> effect_185   0.01004203   2.96627117   0.54010768   4.86196647  13.47914822
#> effect_186 -10.73418190   1.25879667   3.96441589  -6.26553666  -6.12401351
#> effect_187   1.60550608   2.73931447  -1.99810549   3.06584901  -3.14244808
#> effect_188   1.45984878   7.20892409   0.25986957  -6.41287541   3.78905940
#> effect_189  -5.86202555  -1.45257501   5.31898340  -1.03180421   0.97343430
#> effect_190   5.01029068  -2.00268137  -0.10586202  -1.43530403  -0.57229843
#> effect_191  -1.83005694  -3.03667608  -3.69477662  -0.30235817   6.31126900
#> effect_192   3.61440628  -0.40448768   6.52458801   2.34122106   6.70298838
#> effect_193   1.16153361   0.90719594  -0.68381078   4.43088446   0.34119653
#> effect_194   4.37770442  -2.57488988  -5.51067057   5.09900739   2.56501738
#> effect_195   0.02304389  -1.12493915   1.63393044   2.15859589   3.64826955
#> effect_196   1.16222386   6.20165031   0.65545610  -9.04413313  -2.87323047
#> effect_197  -0.99632763  -3.20351304   2.38113410   2.31363779  -3.02004520
#> effect_198  -5.82268257  -3.26442266   8.09793483  -0.17776008   1.16047061
#> effect_199  -1.26134656   4.00528975  -2.64371596   3.73567146  -0.18539508
#> effect_200  -6.97298733   5.05938638   7.02783212   0.67630933  -2.92441975
#> 
#> $Shat
#>            condition_1 condition_2 condition_3 condition_4 condition_5
#> effect_1             5           5           5           5           5
#> effect_2             5           5           5           5           5
#> effect_3             5           5           5           5           5
#> effect_4             5           5           5           5           5
#> effect_5             5           5           5           5           5
#> effect_6             5           5           5           5           5
#> effect_7             5           5           5           5           5
#> effect_8             5           5           5           5           5
#> effect_9             5           5           5           5           5
#> effect_10            5           5           5           5           5
#> effect_11            5           5           5           5           5
#> effect_12            5           5           5           5           5
#> effect_13            5           5           5           5           5
#> effect_14            5           5           5           5           5
#> effect_15            5           5           5           5           5
#> effect_16            5           5           5           5           5
#> effect_17            5           5           5           5           5
#> effect_18            5           5           5           5           5
#> effect_19            5           5           5           5           5
#> effect_20            5           5           5           5           5
#> effect_21            5           5           5           5           5
#> effect_22            5           5           5           5           5
#> effect_23            5           5           5           5           5
#> effect_24            5           5           5           5           5
#> effect_25            5           5           5           5           5
#> effect_26            5           5           5           5           5
#> effect_27            5           5           5           5           5
#> effect_28            5           5           5           5           5
#> effect_29            5           5           5           5           5
#> effect_30            5           5           5           5           5
#> effect_31            5           5           5           5           5
#> effect_32            5           5           5           5           5
#> effect_33            5           5           5           5           5
#> effect_34            5           5           5           5           5
#> effect_35            5           5           5           5           5
#> effect_36            5           5           5           5           5
#> effect_37            5           5           5           5           5
#> effect_38            5           5           5           5           5
#> effect_39            5           5           5           5           5
#> effect_40            5           5           5           5           5
#> effect_41            5           5           5           5           5
#> effect_42            5           5           5           5           5
#> effect_43            5           5           5           5           5
#> effect_44            5           5           5           5           5
#> effect_45            5           5           5           5           5
#> effect_46            5           5           5           5           5
#> effect_47            5           5           5           5           5
#> effect_48            5           5           5           5           5
#> effect_49            5           5           5           5           5
#> effect_50            5           5           5           5           5
#> effect_51            5           5           5           5           5
#> effect_52            5           5           5           5           5
#> effect_53            5           5           5           5           5
#> effect_54            5           5           5           5           5
#> effect_55            5           5           5           5           5
#> effect_56            5           5           5           5           5
#> effect_57            5           5           5           5           5
#> effect_58            5           5           5           5           5
#> effect_59            5           5           5           5           5
#> effect_60            5           5           5           5           5
#> effect_61            5           5           5           5           5
#> effect_62            5           5           5           5           5
#> effect_63            5           5           5           5           5
#> effect_64            5           5           5           5           5
#> effect_65            5           5           5           5           5
#> effect_66            5           5           5           5           5
#> effect_67            5           5           5           5           5
#> effect_68            5           5           5           5           5
#> effect_69            5           5           5           5           5
#> effect_70            5           5           5           5           5
#> effect_71            5           5           5           5           5
#> effect_72            5           5           5           5           5
#> effect_73            5           5           5           5           5
#> effect_74            5           5           5           5           5
#> effect_75            5           5           5           5           5
#> effect_76            5           5           5           5           5
#> effect_77            5           5           5           5           5
#> effect_78            5           5           5           5           5
#> effect_79            5           5           5           5           5
#> effect_80            5           5           5           5           5
#> effect_81            5           5           5           5           5
#> effect_82            5           5           5           5           5
#> effect_83            5           5           5           5           5
#> effect_84            5           5           5           5           5
#> effect_85            5           5           5           5           5
#> effect_86            5           5           5           5           5
#> effect_87            5           5           5           5           5
#> effect_88            5           5           5           5           5
#> effect_89            5           5           5           5           5
#> effect_90            5           5           5           5           5
#> effect_91            5           5           5           5           5
#> effect_92            5           5           5           5           5
#> effect_93            5           5           5           5           5
#> effect_94            5           5           5           5           5
#> effect_95            5           5           5           5           5
#> effect_96            5           5           5           5           5
#> effect_97            5           5           5           5           5
#> effect_98            5           5           5           5           5
#> effect_99            5           5           5           5           5
#> effect_100           5           5           5           5           5
#> effect_101           5           5           5           5           5
#> effect_102           5           5           5           5           5
#> effect_103           5           5           5           5           5
#> effect_104           5           5           5           5           5
#> effect_105           5           5           5           5           5
#> effect_106           5           5           5           5           5
#> effect_107           5           5           5           5           5
#> effect_108           5           5           5           5           5
#> effect_109           5           5           5           5           5
#> effect_110           5           5           5           5           5
#> effect_111           5           5           5           5           5
#> effect_112           5           5           5           5           5
#> effect_113           5           5           5           5           5
#> effect_114           5           5           5           5           5
#> effect_115           5           5           5           5           5
#> effect_116           5           5           5           5           5
#> effect_117           5           5           5           5           5
#> effect_118           5           5           5           5           5
#> effect_119           5           5           5           5           5
#> effect_120           5           5           5           5           5
#> effect_121           5           5           5           5           5
#> effect_122           5           5           5           5           5
#> effect_123           5           5           5           5           5
#> effect_124           5           5           5           5           5
#> effect_125           5           5           5           5           5
#> effect_126           5           5           5           5           5
#> effect_127           5           5           5           5           5
#> effect_128           5           5           5           5           5
#> effect_129           5           5           5           5           5
#> effect_130           5           5           5           5           5
#> effect_131           5           5           5           5           5
#> effect_132           5           5           5           5           5
#> effect_133           5           5           5           5           5
#> effect_134           5           5           5           5           5
#> effect_135           5           5           5           5           5
#> effect_136           5           5           5           5           5
#> effect_137           5           5           5           5           5
#> effect_138           5           5           5           5           5
#> effect_139           5           5           5           5           5
#> effect_140           5           5           5           5           5
#> effect_141           5           5           5           5           5
#> effect_142           5           5           5           5           5
#> effect_143           5           5           5           5           5
#> effect_144           5           5           5           5           5
#> effect_145           5           5           5           5           5
#> effect_146           5           5           5           5           5
#> effect_147           5           5           5           5           5
#> effect_148           5           5           5           5           5
#> effect_149           5           5           5           5           5
#> effect_150           5           5           5           5           5
#> effect_151           5           5           5           5           5
#> effect_152           5           5           5           5           5
#> effect_153           5           5           5           5           5
#> effect_154           5           5           5           5           5
#> effect_155           5           5           5           5           5
#> effect_156           5           5           5           5           5
#> effect_157           5           5           5           5           5
#> effect_158           5           5           5           5           5
#> effect_159           5           5           5           5           5
#> effect_160           5           5           5           5           5
#> effect_161           5           5           5           5           5
#> effect_162           5           5           5           5           5
#> effect_163           5           5           5           5           5
#> effect_164           5           5           5           5           5
#> effect_165           5           5           5           5           5
#> effect_166           5           5           5           5           5
#> effect_167           5           5           5           5           5
#> effect_168           5           5           5           5           5
#> effect_169           5           5           5           5           5
#> effect_170           5           5           5           5           5
#> effect_171           5           5           5           5           5
#> effect_172           5           5           5           5           5
#> effect_173           5           5           5           5           5
#> effect_174           5           5           5           5           5
#> effect_175           5           5           5           5           5
#> effect_176           5           5           5           5           5
#> effect_177           5           5           5           5           5
#> effect_178           5           5           5           5           5
#> effect_179           5           5           5           5           5
#> effect_180           5           5           5           5           5
#> effect_181           5           5           5           5           5
#> effect_182           5           5           5           5           5
#> effect_183           5           5           5           5           5
#> effect_184           5           5           5           5           5
#> effect_185           5           5           5           5           5
#> effect_186           5           5           5           5           5
#> effect_187           5           5           5           5           5
#> effect_188           5           5           5           5           5
#> effect_189           5           5           5           5           5
#> effect_190           5           5           5           5           5
#> effect_191           5           5           5           5           5
#> effect_192           5           5           5           5           5
#> effect_193           5           5           5           5           5
#> effect_194           5           5           5           5           5
#> effect_195           5           5           5           5           5
#> effect_196           5           5           5           5           5
#> effect_197           5           5           5           5           5
#> effect_198           5           5           5           5           5
#> effect_199           5           5           5           5           5
#> effect_200           5           5           5           5           5
#>